--- asm/non_matchings/ovl0/ovl0_2_5/func_80005378.s
@@ -1,13 +1,0 @@
-glabel func_80005378
-/* 005F78 80005378 8C8E0000 */  lw    $t6, ($a0)
-/* 005F7C 8000537C 3C018005 */  lui   $at, %hi(D_80049308) # $at, 0x8005
-/* 005F80 80005380 3C19DB06 */  lui   $t9, (0xDB06003C >> 16) # lui $t9, 0xdb06
-/* 005F84 80005384 25CF0004 */  addiu $t7, $t6, 4
-/* 005F88 80005388 AC2F9308 */  sw    $t7, %lo(D_80049308)($at)
-/* 005F8C 8000538C 8C820000 */  lw    $v0, ($a0)
-/* 005F90 80005390 3739003C */  ori   $t9, (0xDB06003C & 0xFFFF) # ori $t9, $t9, 0x3c
-/* 005F94 80005394 24580008 */  addiu $t8, $v0, 8
-/* 005F98 80005398 AC980000 */  sw    $t8, ($a0)
-/* 005F9C 8000539C AC400004 */  sw    $zero, 4($v0)
-/* 005FA0 800053A0 03E00008 */  jr    $ra
-/* 005FA4 800053A4 AC590000 */   sw    $t9, ($v0)
