--- asm/non_matchings/ovl0/ovl0_2_5/func_800075F8.s
@@ -1,12 +1,0 @@
-glabel func_800075F8
-/* 0081F8 800075F8 3C028005 */  lui   $v0, %hi(D_800492DC) # $v0, 0x8005
-/* 0081FC 800075FC 244292DC */  addiu $v0, %lo(D_800492DC) # addiu $v0, $v0, -0x6d24
-/* 008200 80007600 8C4E0000 */  lw    $t6, ($v0)
-/* 008204 80007604 24010002 */  li    $at, 2
-/* 008208 80007608 240F0001 */  li    $t7, 1
-/* 00820C 8000760C 11C10002 */  beq   $t6, $at, .L80007618_ovl0
-/* 008210 80007610 00000000 */   nop   
-/* 008214 80007614 AC4F0000 */  sw    $t7, ($v0)
-.L80007618_ovl0:
-/* 008218 80007618 03E00008 */  jr    $ra
-/* 00821C 8000761C 00000000 */   nop
