--- asm/non_matchings/ovl0/ovl0_2_5/func_80007620.s
@@ -1,13 +1,0 @@
-glabel func_80007620
-/* 008220 80007620 3C0E8005 */  lui   $t6, %hi(D_800492DC) # $t6, 0x8005
-/* 008224 80007624 8DCE92DC */  lw    $t6, %lo(D_800492DC)($t6)
-/* 008228 80007628 24010002 */  li    $at, 2
-/* 00822C 8000762C 00001025 */  move  $v0, $zero
-/* 008230 80007630 15C10003 */  bne   $t6, $at, .L80007640_ovl0
-/* 008234 80007634 00000000 */   nop   
-/* 008238 80007638 03E00008 */  jr    $ra
-/* 00823C 8000763C 24020001 */   li    $v0, 1
-
-.L80007640_ovl0:
-/* 008240 80007640 03E00008 */  jr    $ra
-/* 008244 80007644 00000000 */   nop
