--- asm/non_matchings/ovl0/ovl0_2_5/func_80007674.s
@@ -1,11 +1,0 @@
-glabel func_80007674
-/* 008274 80007674 24010001 */  li    $at, 1
-/* 008278 80007678 10810002 */  beq   $a0, $at, .L80007684_ovl0
-/* 00827C 8000767C 24010002 */   li    $at, 2
-/* 008280 80007680 14810002 */  bne   $a0, $at, .L8000768C_ovl0
-.L80007684_ovl0:
-/* 008284 80007684 3C018005 */   lui   $at, %hi(D_8004A460) # $at, 0x8005
-/* 008288 80007688 AC24A460 */  sw    $a0, %lo(D_8004A460)($at)
-.L8000768C_ovl0:
-/* 00828C 8000768C 03E00008 */  jr    $ra
-/* 008290 80007690 00000000 */   nop
