--- asm/non_matchings/ovl0/ovl0_2_5/func_80007694.s
@@ -1,20 +1,0 @@
-glabel func_80007694
-/* 008294 80007694 24010001 */  li    $at, 1
-/* 008298 80007698 10810003 */  beq   $a0, $at, .L800076A8_ovl0
-/* 00829C 8000769C 00047080 */   sll   $t6, $a0, 2
-/* 0082A0 800076A0 24010002 */  li    $at, 2
-/* 0082A4 800076A4 14810007 */  bne   $a0, $at, .L800076C4_ovl0
-.L800076A8_ovl0:
-/* 0082A8 800076A8 3C0F8005 */   lui   $t7, 0x8005
-/* 0082AC 800076AC 01EE7821 */  addu  $t7, $t7, $t6
-/* 0082B0 800076B0 8DEFA454 */  lw    $t7, -0x5bac($t7)
-/* 0082B4 800076B4 55E00004 */  bnezl $t7, .L800076C8_ovl0
-/* 0082B8 800076B8 00001025 */   move  $v0, $zero
-/* 0082BC 800076BC 03E00008 */  jr    $ra
-/* 0082C0 800076C0 24020001 */   li    $v0, 1
-
-.L800076C4_ovl0:
-/* 0082C4 800076C4 00001025 */  move  $v0, $zero
-.L800076C8_ovl0:
-/* 0082C8 800076C8 03E00008 */  jr    $ra
-/* 0082CC 800076CC 00000000 */   nop
