--- asm/non_matchings/ovl0/ovl0_2_5/func_80007824.s
@@ -1,7 +1,0 @@
-glabel func_80007824
-/* 008424 80007824 00000000 */  nop   
-/* 008428 80007828 00000000 */  nop   
-/* 00842C 8000782C 00000000 */  nop   
-/* 008430 80007830 8C8E0004 */  lw    $t6, 4($a0)
-/* 008434 80007834 03E00008 */  jr    $ra
-/* 008438 80007838 AC8E000C */   sw    $t6, 0xc($a0)
