--- asm/non_matchings/ovl0/ovl0_2_5/func_8000783C.s
@@ -1,31 +1,0 @@
-glabel func_8000783C
-/* 00843C 8000783C 27BDFFE8 */  addiu $sp, $sp, -0x18
-/* 008440 80007840 AFBF0014 */  sw    $ra, 0x14($sp)
-/* 008444 80007844 10C00007 */  beqz  $a2, .L80007864_ovl0
-/* 008448 80007848 00803825 */   move  $a3, $a0
-/* 00844C 8000784C 8C8E000C */  lw    $t6, 0xc($a0)
-/* 008450 80007850 24C2FFFF */  addiu $v0, $a2, -1
-/* 008454 80007854 0040C027 */  not   $t8, $v0
-/* 008458 80007858 01C27821 */  addu  $t7, $t6, $v0
-/* 00845C 8000785C 10000002 */  b     .L80007868_ovl0
-/* 008460 80007860 01F81824 */   and   $v1, $t7, $t8
-.L80007864_ovl0:
-/* 008464 80007864 8CE3000C */  lw    $v1, 0xc($a3)
-.L80007868_ovl0:
-/* 008468 80007868 8CF90008 */  lw    $t9, 8($a3)
-/* 00846C 8000786C 00651021 */  addu  $v0, $v1, $a1
-/* 008470 80007870 ACE2000C */  sw    $v0, 0xc($a3)
-/* 008474 80007874 0322082B */  sltu  $at, $t9, $v0
-/* 008478 80007878 10200007 */  beqz  $at, .L80007898_ovl0
-/* 00847C 8000787C 8FBF0014 */   lw    $ra, 0x14($sp)
-/* 008480 80007880 3C048004 */  lui   $a0, %hi(D_80040210) # $a0, 0x8004
-/* 008484 80007884 24840210 */  addiu $a0, %lo(D_80040210) # addiu $a0, $a0, 0x210
-/* 008488 80007888 0C008C27 */  jal   fatal_printf
-/* 00848C 8000788C 8CE50000 */   lw    $a1, ($a3)
-.L80007890_ovl0:
-/* 008490 80007890 1000FFFF */  b     .L80007890_ovl0
-/* 008494 80007894 00000000 */   nop   
-.L80007898_ovl0:
-/* 008498 80007898 00601025 */  move  $v0, $v1
-/* 00849C 8000789C 03E00008 */  jr    $ra
-/* 0084A0 800078A0 27BD0018 */   addiu $sp, $sp, 0x18
