--- asm/non_matchings/ovl0/ovl0_2_5/func_800079E4.s
@@ -1,7 +1,0 @@
-glabel func_800079E4
-/* 0085E4 800079E4 3C018005 */  lui   $at, %hi(D_8004A508) # $at, 0x8005
-/* 0085E8 800079E8 AC24A508 */  sw    $a0, %lo(D_8004A508)($at)
-/* 0085EC 800079EC 3C018005 */  lui   $at, 0x8005
-/* 0085F0 800079F0 240E0001 */  li    $t6, 1
-/* 0085F4 800079F4 03E00008 */  jr    $ra
-/* 0085F8 800079F8 AC2EA514 */   sw    $t6, -0x5aec($at)
