--- asm/non_matchings/ovl0/ovl0_2_5/func_800079FC.s
@@ -1,7 +1,0 @@
-glabel func_800079FC
-/* 0085FC 800079FC 3C018005 */  lui   $at, %hi(D_8004A50C) # $at, 0x8005
-/* 008600 80007A00 AC24A50C */  sw    $a0, %lo(D_8004A50C)($at)
-/* 008604 80007A04 3C018005 */  lui   $at, 0x8005
-/* 008608 80007A08 240E0001 */  li    $t6, 1
-/* 00860C 80007A0C 03E00008 */  jr    $ra
-/* 008610 80007A10 AC2EA514 */   sw    $t6, -0x5aec($at)
