+++ asm/non_matchings/ovl0/ovl0_4/func_80011180.s
@@ -81,9 +81,9 @@ glabel func_80011180
 /* 011EA0 800112A0 AFA20078 */   sw    $v0, 0x78($sp)
 /* 011EA4 800112A4 24010004 */  li    $at, 4
 /* 011EA8 800112A8 1461002B */  bne   $v1, $at, .L80011358_ovl0
-/* 011EAC 800112AC 3C098005 */   lui   $t1, 0x8005
+/* 011EAC 800112AC 3C098005 */   lui   $t1, %hi(D_8004A450)
 /* 011EB0 800112B0 8E2A0004 */  lw    $t2, 4($s1)
-/* 011EB4 800112B4 3C098004 */  lui   $t1, %hi(D_8003A450) # $t1, 0x8004
+/* 011EB4 800112B4 3C098004 */  lui   $t1, %hi(D_8003DCAB) # $t1, 0x8004
 /* 011EB8 800112B8 9129DCAB */  lbu   $t1, %lo(D_8003DCAB)($t1)
 /* 011EBC 800112BC 914B000E */  lbu   $t3, 0xe($t2)
 /* 011EC0 800112C0 512B000A */  beql  $t1, $t3, .L800112EC_ovl0
@@ -128,7 +128,7 @@ glabel func_80011180
 /* 011F50 80011350 1000077B */  b     .L80013140_ovl0
 /* 011F54 80011354 8D500000 */   lw    $s0, ($t2)
 .L80011358_ovl0:
-/* 011F58 80011358 8D29A450 */  lw    $t1, %lo(D_8003A450)($t1)
+/* 011F58 80011358 8D29A450 */  lw    $t1, %lo(D_8004A450)($t1)
 /* 011F5C 8001135C 59200008 */  blezl $t1, .L80011380_ovl0
 /* 011F60 80011360 8E2D0004 */   lw    $t5, 4($s1)
 /* 011F64 80011364 3C038005 */  lui   $v1, %hi(D_8004A3F8) # $v1, 0x8005
