+++ data/kirby.039DF0.0.5.c
@@ -1,0 +1,3509 @@
+#include <ultra64.h>
+
+#include "types.h"
+#include "segments.h"
+#include "banks.h"
+#include "src/ovl0/main.h"
+
+typedef struct {
+        s32             active;		/* Status flag */
+	OSThread	*thread;	/* Calling thread */
+        OSMesgQueue  	*cmdQueue;	/* Command queue */
+        OSMesgQueue  	*evtQueue;	/* Event queue */
+        OSMesgQueue  	*acsQueue;	/* Access queue */
+					/* Raw DMA routine */
+        s32             (*dma)(s32, u32, void *, u32);
+        s32             (*edma)(OSPiHandle *, s32, u32, void *, u32);
+} OSDevMgr; // 0x1C
+
+Overlay mainSegOverlay = {
+	_ovl1SegmentRomStart,
+	_ovl1DataSegmentRomEnd,
+	_ovl1SegmentStart,
+	_ovl1SegmentStart,
+	_ovl1SegmentEnd,
+	_ovl1SegmentEnd,
+	_ovl1DataSegmentEnd,
+	_ovl1DataSegmentEnd,
+	(void*)0x800F61A0,
+};
+
+u32 D_8003DC94 = 0x00000000;
+
+u32 D_8003DC98 = 0x00000000;
+
+u32 D_8003DC9C = 0x00000000;
+
+u32 D_8003DCA0 = 0x00000000;
+
+u32 D_8003DCA4 = 0x00000000;
+
+u32 D_8003DCA8 = 0x00000000;
+
+u32 D_8003DCAC = 0x80039E90;
+
+u32 D_8003DCB0 = 0x80041FF0;
+
+u32 D_8003DCB4 = 0x00000000;
+
+u32 D_8003DCB8 = 0x00000000;
+
+u32 D_8003DCBC = 0x00000000;
+
+u32 D_8003DCC0 = 0x00000000;
+
+u32 D_8003DCC4 = 0x00000000;
+
+u32 D_8003DCC8 = 0x00000000;
+
+u32 D_8003DCCC = 0x00000000;
+
+u32 D_8003DCD0 = 0x00000000;
+
+u32 D_8003DCD4 = 0x00000000;
+
+u32 D_8003DCD8 = 0x00000000;
+
+u32 D_8003DCDC = 0x00000000;
+
+u32 D_8003DCE0 = 0x00000000;
+
+u32 D_8003DCE4 = 0x00000000;
+
+u32 D_8003DCE8 = 0x00000000;
+
+u32 D_8003DCEC = 0x8003B220;
+
+u32 D_8003DCF0 = 0x80042410;
+
+u32 D_8003DCF4 = 0x00000000;
+
+u32 D_8003DCF8 = 0x00000000;
+
+u32 D_8003DCFC = 0x8003C3B0;
+
+u32 D_8003DD00 = 0x80042800;
+
+u32 D_8003DD04 = 0x00000000;
+
+u32 D_8003DD08 = 0x00000000;
+
+u32 D_8003DD0C = 0x00000000;
+
+u32 D_8003DD10 = 0x00000000;
+
+u32 D_8003DD14 = 0x00000000;
+
+u32 D_8003DD18 = 0x00000000;
+
+u32 D_8003DD1C = 0x00000000;
+
+u32 D_8003DD20 = 0x00000000;
+
+u32 D_8003DD24 = 0x00000000;
+
+u32 D_8003DD28 = 0x00000000;
+
+u32 D_8003DD2C = 0x00000000;
+
+u32 D_8003DD30 = 0x00000000;
+
+u32 D_8003DD34 = 0x00000000;
+
+Mtx D_8003DD38 = {
+	65536, 0, 1, 0,
+	0, 65536, 0, 1,
+	0, 0, 0, 0,
+	0, 0, 0, 0,
+};
+
+Gfx D_8003DD78[] = {
+	gsDPPipeSync(),
+	gsSPViewport(0x8004A530),
+	gsSPClearGeometryMode(G_ZBUFFER | G_SHADE | G_CULL_BOTH | G_FOG | G_LIGHTING | G_TEXTURE_GEN | G_TEXTURE_GEN_LINEAR | G_LOD | G_SHADING_SMOOTH),
+	gsSPClipRatio(FRUSTRATIO_1),
+	gsSPTexture(0, 0, 0, G_TX_RENDERTILE, G_OFF),
+	gsSPSetGeometryMode(G_ZBUFFER | G_SHADE | G_CULL_BACK | G_SHADING_SMOOTH),
+    gsSPMatrix(&D_8003DD38, G_MTX_NOPUSH | G_MTX_LOAD | G_MTX_PROJECTION),
+    gsSPMatrix(&D_8003DD38, G_MTX_NOPUSH | G_MTX_LOAD | G_MTX_MODELVIEW),
+    gsDPSetCycleType(G_CYC_1CYCLE),
+    gsDPPipelineMode(G_PM_NPRIMITIVE),
+    gsDPSetCombineMode(G_CC_SHADE, G_CC_SHADE),
+    gsDPSetTextureLOD(G_TL_TILE),
+    gsDPSetTextureLUT(G_TT_NONE),
+    gsDPSetTextureDetail(G_TD_CLAMP),
+    gsDPSetTexturePersp(G_TP_PERSP),
+    gsDPSetTextureFilter(G_TF_BILERP),
+    gsDPSetTextureConvert(G_TC_FILT),
+    gsDPSetCombineKey(G_CK_NONE),
+    gsDPSetAlphaCompare(G_AC_NONE),
+    gsDPSetRenderMode(G_RM_OPA_SURF, G_RM_OPA_SURF2),
+    gsDPSetColorDither(G_CD_MAGICSQ),
+    gsDPPipeSync(),
+    gsSPEndDisplayList(),
+};
+
+u32 D_8003DE48 = 0x00000000;
+
+u32 D_8003DE4C = 0x00000000;
+
+u32 D_8003DE50 = 0x00989680;
+
+u32 D_8003DE54 = 0x00000000;
+
+u32 D_8003DE58 = 0x00000000;
+
+u32 D_8003DE5C = 0x00000000;
+
+f32 D_8003DE60 = 30.0f;
+
+f32 D_8003DE64 = 4.0f / 3.0f;
+
+f32 D_8003DE68 = 100.0f;
+
+f32 D_8003DE6C = 12800.0f;
+
+f32 D_8003DE70 = 1.0f;
+
+u32 D_8003DE74 = 0x00000000;
+
+f32 D_8003DE78 = -160.0f;
+
+f32 D_8003DE7C = 160.0f;
+
+f32 D_8003DE80 = -120.0f;
+
+f32 D_8003DE84 = 120.0f;
+
+f32 D_8003DE88 = 100.0f;
+
+f32 D_8003DE8C = 12800.0f;
+
+f32 D_8003DE90 = 1.0f;
+
+u32 D_8003DE94 = 0x00000000;
+
+u32 D_8003DE98 = 0x00000000;
+
+u32 D_8003DE9C = 0x00000000;
+
+f32 D_8003DEA0 = 1500.0f;
+
+u32 D_8003DEA4 = 0x00000000;
+
+u32 D_8003DEA8 = 0x00000000;
+
+u32 D_8003DEAC = 0x00000000;
+
+u32 D_8003DEB0 = 0x00000000;
+
+f32 D_8003DEB4 = 1.0f;
+
+u32 D_8003DEB8 = 0x00000000;
+
+u32 D_8003DEBC = 0x00000000;
+
+u32 D_8003DEC0 = 0x00000000;
+
+u32 D_8003DEC4 = 0x00000000;
+
+u32 D_8003DEC8 = 0x00000000;
+
+u32 D_8003DECC = 0x00000000;
+
+u32 D_8003DED0 = 0x00000000;
+
+u32 D_8003DED4 = 0x00000000;
+
+u32 D_8003DED8 = 0x00000000;
+
+f32 D_8003DEDC = 1.0f;
+
+u32 D_8003DEE0 = 0x00000000;
+
+u32 D_8003DEE4 = 0x00000000;
+
+u32 D_8003DEE8 = 0x00000000;
+
+u32 D_8003DEEC = 0x00000000;
+
+u32 D_8003DEF0 = 0x00000000;
+
+u32 D_8003DEF4 = 0x00000000;
+
+f32 D_8003DEF8 = 1.0f;
+
+f32 D_8003DEFC = 1.0f;
+
+f32 D_8003DF00 = 1.0f;
+
+u32 D_8003DF04 = 0x00000000;
+
+u32 D_8003DF08 = 0x00000000;
+
+u32 D_8003DF0C = 0x00000000;
+
+u32 D_8003DF10 = 0x0000000A;
+
+u32 D_8003DF14 = 0x0000000A;
+
+u32 D_8003DF18 = 0x0000000A;
+
+u32 D_8003DF1C = 0x0000000A;
+
+u8 D_8003DF20[0x400] = {
+    0x2B, 0xE6, 0xDB, 0xB9, 0xB1, 0xF3, 0x53, 0x81, 0x37, 0x01, 0x7F, 0x44, 0x23, 0x3C, 0x11, 0xD0,
+    0x48, 0x80, 0xD2, 0x15, 0x6A, 0x70, 0x9B, 0x9E, 0x27, 0x51, 0xB0, 0x4B, 0x6A, 0xDB, 0x1E, 0xFF,
+    0x2D, 0xBC, 0x5B, 0x38, 0x6D, 0xCA, 0xAA, 0xF5, 0xCB, 0x42, 0x56, 0x96, 0x0F, 0x28, 0x6A, 0x59,
+    0x08, 0x08, 0xB5, 0x1B, 0xA4, 0xEC, 0xAD, 0x8B, 0x7E, 0xD9, 0x6F, 0x4E, 0x8E, 0x40, 0x92, 0xA2,
+    0x7F, 0x27, 0x88, 0x8D, 0x65, 0x5B, 0x3A, 0xF6, 0xA2, 0x41, 0xC8, 0xA6, 0x4A, 0xA6, 0xDE, 0x28,
+    0x8C, 0x48, 0xD7, 0x90, 0x9B, 0xAB, 0xAB, 0x73, 0x9A, 0x44, 0x6F, 0x73, 0xF7, 0xE4, 0x15, 0x04,
+    0xE6, 0xFF, 0xD8, 0xCD, 0xE7, 0xA9, 0x1D, 0x56, 0x4B, 0xEE, 0xD5, 0xBE, 0x32, 0x93, 0xCF, 0x7F,
+    0x39, 0x4D, 0xA8, 0x15, 0x32, 0xF7, 0x68, 0x67, 0x9B, 0x6F, 0xD6, 0x84, 0x30, 0xE3, 0xDE, 0x52,
+    0x5C, 0x09, 0x34, 0x54, 0xAB, 0x76, 0x03, 0x5A, 0xDE, 0xE0, 0xE4, 0xA0, 0x49, 0xAF, 0x0B, 0x7B,
+    0x54, 0x2C, 0x82, 0xBC, 0xF4, 0xCE, 0xA4, 0x0A, 0x99, 0x42, 0x5C, 0xD0, 0x67, 0xAC, 0x46, 0x1C,
+    0x05, 0x09, 0x4C, 0x87, 0x64, 0x61, 0x5D, 0x57, 0x4E, 0x13, 0x39, 0x67, 0x7D, 0x02, 0x26, 0x2A,
+    0x04, 0x49, 0xAE, 0xC3, 0x69, 0x7B, 0x52, 0xEC, 0x7D, 0xA7, 0x45, 0x0F, 0xB4, 0x12, 0xAE, 0x44,
+    0xA5, 0x1A, 0x22, 0xBC, 0xF3, 0xD0, 0x35, 0x3A, 0x50, 0x43, 0xEB, 0x94, 0x66, 0x9D, 0x85, 0xBB,
+    0x0C, 0x02, 0xD9, 0x0B, 0xC5, 0xE8, 0xB2, 0x1D, 0x1A, 0x20, 0x66, 0x25, 0xAE, 0xF9, 0x33, 0xEF,
+    0xD8, 0x1F, 0xA1, 0xDF, 0x68, 0xFB, 0xB0, 0xC5, 0x91, 0xA5, 0x24, 0x8D, 0xCC, 0xC1, 0x06, 0xA2,
+    0x5D, 0xFD, 0x50, 0x45, 0xE6, 0x75, 0x4A, 0x5E, 0x02, 0xB7, 0x6C, 0xFE, 0x81, 0xB8, 0x93, 0x8C,
+    0x4C, 0x29, 0x1F, 0xB1, 0x9A, 0xFA, 0x85, 0x23, 0xCE, 0x43, 0x04, 0x76, 0x5B, 0x14, 0xAF, 0x32,
+    0xDA, 0xF7, 0x48, 0xBF, 0x0C, 0x3F, 0x30, 0x6D, 0x86, 0x2C, 0xE2, 0xA3, 0x31, 0x0D, 0xFF, 0xD3,
+    0xF0, 0x76, 0x7B, 0x89, 0xED, 0xAA, 0x97, 0x0E, 0x45, 0xB3, 0x5B, 0xAB, 0x92, 0xB5, 0x79, 0xD3,
+    0xCC, 0xFA, 0x4C, 0x7E, 0xA0, 0x25, 0x1A, 0xED, 0x56, 0xF4, 0xBE, 0xCA, 0xF2, 0xD7, 0x00, 0x6A,
+    0xF3, 0x62, 0xC0, 0xBF, 0xB3, 0x56, 0x46, 0x50, 0x91, 0x69, 0x65, 0x5C, 0xC2, 0xDB, 0xE1, 0x8E,
+    0x79, 0x03, 0xD9, 0xAC, 0xED, 0x10, 0xBB, 0x5E, 0x8B, 0xC1, 0xE5, 0xD4, 0xF1, 0xD7, 0x78, 0x6E,
+    0x4D, 0x44, 0x01, 0x80, 0xDC, 0xB1, 0x0A, 0x97, 0x72, 0x1E, 0x5D, 0xB4, 0x99, 0xC0, 0xB0, 0x20,
+    0x04, 0x39, 0xF6, 0xC1, 0xF7, 0x2B, 0x3D, 0x45, 0x59, 0xEC, 0x53, 0x4F, 0x9A, 0x86, 0xE5, 0x06,
+    0x55, 0x80, 0x40, 0x9C, 0x86, 0x83, 0x90, 0x2D, 0x9B, 0xD1, 0xEF, 0xB9, 0x68, 0xDC, 0x93, 0x1D,
+    0xC2, 0x6C, 0xB6, 0xE2, 0xCF, 0x28, 0x10, 0xB6, 0x16, 0x52, 0x33, 0x60, 0xDB, 0x82, 0x41, 0xD0,
+    0xFD, 0x47, 0x1C, 0x35, 0x58, 0x94, 0x58, 0xE5, 0xBB, 0x58, 0xA2, 0x4C, 0x64, 0xCE, 0xCA, 0x94,
+    0x13, 0x47, 0x6F, 0xA7, 0xF8, 0x5E, 0xC4, 0x3B, 0x76, 0x85, 0x01, 0x61, 0x26, 0x13, 0x28, 0x23,
+    0x52, 0xFB, 0x19, 0x89, 0x35, 0xEA, 0xDD, 0x88, 0xE7, 0xBE, 0xA9, 0x2F, 0xD5, 0x82, 0x9C, 0x51,
+    0x7F, 0x51, 0x29, 0x31, 0xDA, 0xF9, 0xFA, 0x2A, 0x06, 0xC2, 0xE3, 0x14, 0xD5, 0xEC, 0x79, 0xE9,
+    0xDD, 0x75, 0x72, 0x37, 0x84, 0xFE, 0x7D, 0x72, 0x8E, 0xCD, 0xF5, 0xC3, 0xA8, 0xBA, 0xD8, 0xC6,
+    0xD4, 0x42, 0x0B, 0xEE, 0x1B, 0xE0, 0xFE, 0xF0, 0xD7, 0xA6, 0xC6, 0x22, 0x12, 0x25, 0x49, 0x81,
+    0xC0, 0x9F, 0x33, 0x9F, 0x38, 0xB7, 0x3C, 0x38, 0x42, 0xE2, 0xCB, 0xA9, 0x17, 0x03, 0x4E, 0xD3,
+    0xF5, 0xE9, 0x77, 0x0E, 0xA8, 0x08, 0x3D, 0x47, 0x86, 0x98, 0xDA, 0x20, 0xD6, 0x9F, 0x7D, 0xFC,
+    0xEE, 0xD6, 0x14, 0x5F, 0xFC, 0x5F, 0x46, 0xDF, 0xF2, 0x87, 0x39, 0x19, 0xA1, 0x62, 0x65, 0xA7,
+    0x51, 0x64, 0x20, 0x7E, 0x77, 0xFD, 0xC9, 0xA7, 0x80, 0x5F, 0xFB, 0x41, 0xA0, 0xBD, 0x9D, 0x83,
+    0xFC, 0xDF, 0xF0, 0xE3, 0x05, 0x60, 0xA6, 0xBB, 0xD9, 0xB0, 0x59, 0x43, 0xFF, 0xD1, 0x30, 0x7C,
+    0xB2, 0xAF, 0x8D, 0x8C, 0xC7, 0x53, 0xF1, 0x47, 0x92, 0x38, 0xA3, 0x57, 0x9C, 0x0A, 0x72, 0x95,
+    0x4A, 0x7C, 0xE4, 0xBF, 0x84, 0x49, 0x15, 0x2E, 0x2D, 0x0C, 0x68, 0x7C, 0x3B, 0x84, 0xEF, 0x05,
+    0x1C, 0x78, 0x78, 0x2E, 0x6B, 0x11, 0x1D, 0x27, 0x3D, 0x87, 0x61, 0x1B, 0x3C, 0x93, 0x5A, 0x99,
+    0xE8, 0x89, 0x95, 0xD2, 0xE1, 0x3B, 0x0C, 0x99, 0x11, 0x9F, 0x67, 0x3C, 0xC7, 0x98, 0xF2, 0x8B,
+    0xCD, 0x21, 0x66, 0x19, 0x73, 0x96, 0xEB, 0xAA, 0x89, 0xCB, 0x4B, 0x8F, 0x33, 0xA5, 0x74, 0x4B,
+    0x2F, 0xA1, 0x2A, 0x55, 0x48, 0x5A, 0xAD, 0x13, 0x0A, 0x43, 0xB4, 0xFA, 0xD5, 0x3F, 0xAC, 0x79,
+    0x85, 0xD6, 0x9E, 0x9D, 0x60, 0x73, 0x74, 0xB2, 0x98, 0xE1, 0xBA, 0xB2, 0x6B, 0x29, 0xEB, 0x71,
+    0x7B, 0xF2, 0x40, 0x37, 0x12, 0x90, 0x1F, 0x9D, 0x62, 0x1E, 0xAE, 0x6C, 0x3A, 0x8B, 0xCB, 0x1E,
+    0x71, 0xBD, 0x2C, 0x9C, 0xBE, 0x8E, 0x10, 0x98, 0x70, 0xD4, 0x11, 0x4F, 0x07, 0x2F, 0x8F, 0x09,
+    0x57, 0xE8, 0x0D, 0x3E, 0xC2, 0xDC, 0xA1, 0xF5, 0x6C, 0x46, 0x24, 0x26, 0x16, 0x8A, 0xEA, 0xED,
+    0x1F, 0x00, 0x74, 0x00, 0x18, 0x2F, 0xF6, 0xCC, 0xC6, 0x02, 0x78, 0xC7, 0xB8, 0x2C, 0x2A, 0xE3,
+    0x91, 0xD4, 0xFC, 0xF4, 0xCF, 0x7A, 0x3D, 0xC3, 0x14, 0xCC, 0x3F, 0xA0, 0xDE, 0x22, 0x0D, 0x1A,
+    0x6D, 0xBD, 0xD1, 0x27, 0xBA, 0x6D, 0x08, 0x4D, 0x50, 0x07, 0xE7, 0x6B, 0x71, 0x36, 0x8C, 0x35,
+    0x18, 0x2B, 0xFD, 0xA8, 0x74, 0x7A, 0xB9, 0x0E, 0x3B, 0x5D, 0x17, 0xBF, 0xB3, 0x3A, 0x00, 0x83,
+    0xEF, 0x26, 0xF6, 0x2D, 0x31, 0xF9, 0x9E, 0xC1, 0xC9, 0x58, 0xE7, 0xC7, 0x41, 0x71, 0x36, 0x07,
+    0xB7, 0x63, 0x29, 0xFB, 0x06, 0x63, 0x05, 0x0B, 0xF0, 0x88, 0x1B, 0x6B, 0x4F, 0xB8, 0xD8, 0x75,
+    0x34, 0x70, 0x21, 0x92, 0xE2, 0x95, 0xC6, 0x70, 0x63, 0xDA, 0xEA, 0xB4, 0x77, 0xBC, 0xCF, 0x3E,
+    0xC9, 0xE0, 0xC4, 0x65, 0xE0, 0xC5, 0x40, 0xEA, 0xC8, 0xF1, 0xF8, 0xD3, 0x34, 0x32, 0x54, 0x61,
+    0x5F, 0xAD, 0x6E, 0x34, 0x17, 0xA3, 0x2E, 0xB3, 0xA3, 0xC3, 0x5C, 0xD2, 0x54, 0x16, 0x5A, 0x55,
+    0xAC, 0x94, 0xAD, 0x53, 0xAA, 0x96, 0xCA, 0xEE, 0x64, 0x30, 0xB7, 0x87, 0x82, 0x36, 0xF1, 0xDD,
+    0x31, 0x69, 0xCD, 0xA9, 0x91, 0xA5, 0x18, 0xC5, 0xF9, 0xE4, 0x8F, 0x97, 0x90, 0xE6, 0xCE, 0x23,
+    0xB8, 0x59, 0x0D, 0xB1, 0x95, 0x4E, 0x2E, 0xB5, 0x8A, 0x57, 0x7C, 0xC9, 0x77, 0x88, 0xC4, 0x21,
+    0x7A, 0xFE, 0x75, 0x83, 0x81, 0x63, 0x10, 0x3E, 0xB6, 0x0F, 0x7E, 0x36, 0x5E, 0x62, 0x6E, 0xD2,
+    0x55, 0x19, 0x4F, 0x16, 0xC4, 0xBA, 0xA4, 0xAF, 0xE9, 0xB9, 0x0E, 0x25, 0x9A, 0xF8, 0x09, 0x0F,
+    0x03, 0x12, 0xBD, 0x7A, 0xB6, 0xE9, 0x3E, 0x37, 0x24, 0x8A, 0xF8, 0x97, 0xF3, 0xE8, 0x4D, 0xE5,
+    0x4A, 0x60, 0x3F, 0xA4, 0x2B, 0x24, 0xC0, 0xDC, 0x17, 0x66, 0x69, 0x21, 0xB5, 0x8F, 0x8D, 0x6E,
+    0x96, 0x07, 0xF4, 0xD1, 0x1C, 0xC8, 0xE1, 0x18, 0xDD, 0xEB, 0x22, 0x01, 0x9E, 0xDF, 0xC8, 0x8A,
+};
+
+u32 D_8003E320 = 0x00000000;
+
+u32 D_8003E324 = 0x00000001;
+
+u32 D_8003E328 = 0x00000000;
+
+u32 D_8003E32C = 0x00000000;
+
+u32 D_8003E330 = 0x00000032;
+
+u32 D_8003E334 = 0x00640096;
+
+u32 D_8003E338 = 0x00C900FB;
+
+u32 D_8003E33C = 0x012D015F;
+
+u32 D_8003E340 = 0x019201C4;
+
+u32 D_8003E344 = 0x01F60228;
+
+u32 D_8003E348 = 0x025B028D;
+
+u32 D_8003E34C = 0x02BF02F1;
+
+u32 D_8003E350 = 0x03240356;
+
+u32 D_8003E354 = 0x038803BA;
+
+u32 D_8003E358 = 0x03ED041F;
+
+u32 D_8003E35C = 0x04510483;
+
+u32 D_8003E360 = 0x04B604E8;
+
+u32 D_8003E364 = 0x051A054C;
+
+u32 D_8003E368 = 0x057F05B1;
+
+u32 D_8003E36C = 0x05E30615;
+
+u32 D_8003E370 = 0x0647067A;
+
+u32 D_8003E374 = 0x06AC06DE;
+
+u32 D_8003E378 = 0x07100742;
+
+u32 D_8003E37C = 0x077507A7;
+
+u32 D_8003E380 = 0x07D9080B;
+
+u32 D_8003E384 = 0x083D086F;
+
+u32 D_8003E388 = 0x08A208D4;
+
+u32 D_8003E38C = 0x09060938;
+
+u32 D_8003E390 = 0x096A099C;
+
+u32 D_8003E394 = 0x09CE0A00;
+
+u32 D_8003E398 = 0x0A330A65;
+
+u32 D_8003E39C = 0x0A970AC9;
+
+u32 D_8003E3A0 = 0x0AFB0B2D;
+
+u32 D_8003E3A4 = 0x0B5F0B91;
+
+u32 D_8003E3A8 = 0x0BC30BF5;
+
+u32 D_8003E3AC = 0x0C270C59;
+
+u32 D_8003E3B0 = 0x0C8B0CBD;
+
+u32 D_8003E3B4 = 0x0CEF0D21;
+
+u32 D_8003E3B8 = 0x0D530D85;
+
+u32 D_8003E3BC = 0x0DB70DE9;
+
+u32 D_8003E3C0 = 0x0E1B0E4D;
+
+u32 D_8003E3C4 = 0x0E7F0EB1;
+
+u32 D_8003E3C8 = 0x0EE30F15;
+
+u32 D_8003E3CC = 0x0F470F79;
+
+u32 D_8003E3D0 = 0x0FAB0FDD;
+
+u32 D_8003E3D4 = 0x100E1040;
+
+u32 D_8003E3D8 = 0x107210A4;
+
+u32 D_8003E3DC = 0x10D61108;
+
+u32 D_8003E3E0 = 0x1139116B;
+
+u32 D_8003E3E4 = 0x119D11CF;
+
+u32 D_8003E3E8 = 0x12011232;
+
+u32 D_8003E3EC = 0x12641296;
+
+u32 D_8003E3F0 = 0x12C812F9;
+
+u32 D_8003E3F4 = 0x132B135D;
+
+u32 D_8003E3F8 = 0x138E13C0;
+
+u32 D_8003E3FC = 0x13F21423;
+
+u32 D_8003E400 = 0x14551487;
+
+u32 D_8003E404 = 0x14B814EA;
+
+u32 D_8003E408 = 0x151B154D;
+
+u32 D_8003E40C = 0x157F15B0;
+
+u32 D_8003E410 = 0x15E21613;
+
+u32 D_8003E414 = 0x16451676;
+
+u32 D_8003E418 = 0x16A816D9;
+
+u32 D_8003E41C = 0x170A173C;
+
+u32 D_8003E420 = 0x176D179F;
+
+u32 D_8003E424 = 0x17D01802;
+
+u32 D_8003E428 = 0x18331864;
+
+u32 D_8003E42C = 0x189618C7;
+
+u32 D_8003E430 = 0x18F8192A;
+
+u32 D_8003E434 = 0x195B198C;
+
+u32 D_8003E438 = 0x19BD19EF;
+
+u32 D_8003E43C = 0x1A201A51;
+
+u32 D_8003E440 = 0x1A821AB3;
+
+u32 D_8003E444 = 0x1AE41B16;
+
+u32 D_8003E448 = 0x1B471B78;
+
+u32 D_8003E44C = 0x1BA91BDA;
+
+u32 D_8003E450 = 0x1C0B1C3C;
+
+u32 D_8003E454 = 0x1C6D1C9E;
+
+u32 D_8003E458 = 0x1CCF1D00;
+
+u32 D_8003E45C = 0x1D311D62;
+
+u32 D_8003E460 = 0x1D931DC4;
+
+u32 D_8003E464 = 0x1DF51E25;
+
+u32 D_8003E468 = 0x1E561E87;
+
+u32 D_8003E46C = 0x1EB81EE9;
+
+u32 D_8003E470 = 0x1F191F4A;
+
+u32 D_8003E474 = 0x1F7B1FAC;
+
+u32 D_8003E478 = 0x1FDC200D;
+
+u32 D_8003E47C = 0x203E206E;
+
+u32 D_8003E480 = 0x209F20D0;
+
+u32 D_8003E484 = 0x21002131;
+
+u32 D_8003E488 = 0x21612192;
+
+u32 D_8003E48C = 0x21C221F3;
+
+u32 D_8003E490 = 0x22232254;
+
+u32 D_8003E494 = 0x228422B4;
+
+u32 D_8003E498 = 0x22E52315;
+
+u32 D_8003E49C = 0x23452376;
+
+u32 D_8003E4A0 = 0x23A623D6;
+
+u32 D_8003E4A4 = 0x24072437;
+
+u32 D_8003E4A8 = 0x24672497;
+
+u32 D_8003E4AC = 0x24C724F7;
+
+u32 D_8003E4B0 = 0x25282558;
+
+u32 D_8003E4B4 = 0x258825B8;
+
+u32 D_8003E4B8 = 0x25E82618;
+
+u32 D_8003E4BC = 0x26482678;
+
+u32 D_8003E4C0 = 0x26A826D8;
+
+u32 D_8003E4C4 = 0x27072737;
+
+u32 D_8003E4C8 = 0x27672797;
+
+u32 D_8003E4CC = 0x27C727F6;
+
+u32 D_8003E4D0 = 0x28262856;
+
+u32 D_8003E4D4 = 0x288628B5;
+
+u32 D_8003E4D8 = 0x28E52915;
+
+u32 D_8003E4DC = 0x29442974;
+
+u32 D_8003E4E0 = 0x29A329D3;
+
+u32 D_8003E4E4 = 0x2A022A32;
+
+u32 D_8003E4E8 = 0x2A612A91;
+
+u32 D_8003E4EC = 0x2AC02AEF;
+
+u32 D_8003E4F0 = 0x2B1F2B4E;
+
+u32 D_8003E4F4 = 0x2B7D2BAD;
+
+u32 D_8003E4F8 = 0x2BDC2C0B;
+
+u32 D_8003E4FC = 0x2C3A2C69;
+
+u32 D_8003E500 = 0x2C982CC8;
+
+u32 D_8003E504 = 0x2CF72D26;
+
+u32 D_8003E508 = 0x2D552D84;
+
+u32 D_8003E50C = 0x2DB32DE2;
+
+u32 D_8003E510 = 0x2E112E3F;
+
+u32 D_8003E514 = 0x2E6E2E9D;
+
+u32 D_8003E518 = 0x2ECC2EFB;
+
+u32 D_8003E51C = 0x2F292F58;
+
+u32 D_8003E520 = 0x2F872FB5;
+
+u32 D_8003E524 = 0x2FE43013;
+
+u32 D_8003E528 = 0x30413070;
+
+u32 D_8003E52C = 0x309E30CD;
+
+u32 D_8003E530 = 0x30FB312A;
+
+u32 D_8003E534 = 0x31583186;
+
+u32 D_8003E538 = 0x31B531E3;
+
+u32 D_8003E53C = 0x32113240;
+
+u32 D_8003E540 = 0x326E329C;
+
+u32 D_8003E544 = 0x32CA32F8;
+
+u32 D_8003E548 = 0x33263354;
+
+u32 D_8003E54C = 0x338233B0;
+
+u32 D_8003E550 = 0x33DE340C;
+
+u32 D_8003E554 = 0x343A3468;
+
+u32 D_8003E558 = 0x349634C4;
+
+u32 D_8003E55C = 0x34F2351F;
+
+u32 D_8003E560 = 0x354D357B;
+
+u32 D_8003E564 = 0x35A835D6;
+
+u32 D_8003E568 = 0x36043631;
+
+u32 D_8003E56C = 0x365F368C;
+
+u32 D_8003E570 = 0x36BA36E7;
+
+u32 D_8003E574 = 0x37143742;
+
+u32 D_8003E578 = 0x376F379C;
+
+u32 D_8003E57C = 0x37CA37F7;
+
+u32 D_8003E580 = 0x38243851;
+
+u32 D_8003E584 = 0x387E38AB;
+
+u32 D_8003E588 = 0x38D83906;
+
+u32 D_8003E58C = 0x3932395F;
+
+u32 D_8003E590 = 0x398C39B9;
+
+u32 D_8003E594 = 0x39E63A13;
+
+u32 D_8003E598 = 0x3A403A6C;
+
+u32 D_8003E59C = 0x3A993AC6;
+
+u32 D_8003E5A0 = 0x3AF23B1F;
+
+u32 D_8003E5A4 = 0x3B4C3B78;
+
+u32 D_8003E5A8 = 0x3BA53BD1;
+
+u32 D_8003E5AC = 0x3BFD3C2A;
+
+u32 D_8003E5B0 = 0x3C563C83;
+
+u32 D_8003E5B4 = 0x3CAF3CDB;
+
+u32 D_8003E5B8 = 0x3D073D33;
+
+u32 D_8003E5BC = 0x3D603D8C;
+
+u32 D_8003E5C0 = 0x3DB83DE4;
+
+u32 D_8003E5C4 = 0x3E103E3C;
+
+u32 D_8003E5C8 = 0x3E683E93;
+
+u32 D_8003E5CC = 0x3EBF3EEB;
+
+u32 D_8003E5D0 = 0x3F173F43;
+
+u32 D_8003E5D4 = 0x3F6E3F9A;
+
+u32 D_8003E5D8 = 0x3FC53FF1;
+
+u32 D_8003E5DC = 0x401D4048;
+
+u32 D_8003E5E0 = 0x4073409F;
+
+u32 D_8003E5E4 = 0x40CA40F6;
+
+u32 D_8003E5E8 = 0x4121414C;
+
+u32 D_8003E5EC = 0x417741A2;
+
+u32 D_8003E5F0 = 0x41CE41F9;
+
+u32 D_8003E5F4 = 0x4224424F;
+
+u32 D_8003E5F8 = 0x427A42A5;
+
+u32 D_8003E5FC = 0x42D042FA;
+
+u32 D_8003E600 = 0x43254350;
+
+u32 D_8003E604 = 0x437B43A5;
+
+u32 D_8003E608 = 0x43D043FB;
+
+u32 D_8003E60C = 0x44254450;
+
+u32 D_8003E610 = 0x447A44A5;
+
+u32 D_8003E614 = 0x44CF44FA;
+
+u32 D_8003E618 = 0x4524454E;
+
+u32 D_8003E61C = 0x457845A3;
+
+u32 D_8003E620 = 0x45CD45F7;
+
+u32 D_8003E624 = 0x4621464B;
+
+u32 D_8003E628 = 0x4675469F;
+
+u32 D_8003E62C = 0x46C946F3;
+
+u32 D_8003E630 = 0x471C4746;
+
+u32 D_8003E634 = 0x4770479A;
+
+u32 D_8003E638 = 0x47C347ED;
+
+u32 D_8003E63C = 0x48164840;
+
+u32 D_8003E640 = 0x48694893;
+
+u32 D_8003E644 = 0x48BC48E6;
+
+u32 D_8003E648 = 0x490F4938;
+
+u32 D_8003E64C = 0x4961498A;
+
+u32 D_8003E650 = 0x49B449DD;
+
+u32 D_8003E654 = 0x4A064A2F;
+
+u32 D_8003E658 = 0x4A584A81;
+
+u32 D_8003E65C = 0x4AA94AD2;
+
+u32 D_8003E660 = 0x4AFB4B24;
+
+u32 D_8003E664 = 0x4B4C4B75;
+
+u32 D_8003E668 = 0x4B9E4BC6;
+
+u32 D_8003E66C = 0x4BEF4C17;
+
+u32 D_8003E670 = 0x4C3F4C68;
+
+u32 D_8003E674 = 0x4C904CB8;
+
+u32 D_8003E678 = 0x4CE14D09;
+
+u32 D_8003E67C = 0x4D314D59;
+
+u32 D_8003E680 = 0x4D814DA9;
+
+u32 D_8003E684 = 0x4DD14DF9;
+
+u32 D_8003E688 = 0x4E214E48;
+
+u32 D_8003E68C = 0x4E704E98;
+
+u32 D_8003E690 = 0x4EBF4EE7;
+
+u32 D_8003E694 = 0x4F0F4F36;
+
+u32 D_8003E698 = 0x4F5E4F85;
+
+u32 D_8003E69C = 0x4FAC4FD4;
+
+u32 D_8003E6A0 = 0x4FFB5022;
+
+u32 D_8003E6A4 = 0x50495070;
+
+u32 D_8003E6A8 = 0x509750BF;
+
+u32 D_8003E6AC = 0x50E5510C;
+
+u32 D_8003E6B0 = 0x5133515A;
+
+u32 D_8003E6B4 = 0x518151A8;
+
+u32 D_8003E6B8 = 0x51CE51F5;
+
+u32 D_8003E6BC = 0x521C5242;
+
+u32 D_8003E6C0 = 0x5269528F;
+
+u32 D_8003E6C4 = 0x52B552DC;
+
+u32 D_8003E6C8 = 0x53025328;
+
+u32 D_8003E6CC = 0x534E5375;
+
+u32 D_8003E6D0 = 0x539B53C1;
+
+u32 D_8003E6D4 = 0x53E7540D;
+
+u32 D_8003E6D8 = 0x54335458;
+
+u32 D_8003E6DC = 0x547E54A4;
+
+u32 D_8003E6E0 = 0x54CA54EF;
+
+u32 D_8003E6E4 = 0x5515553A;
+
+u32 D_8003E6E8 = 0x55605585;
+
+u32 D_8003E6EC = 0x55AB55D0;
+
+u32 D_8003E6F0 = 0x55F5561A;
+
+u32 D_8003E6F4 = 0x56405665;
+
+u32 D_8003E6F8 = 0x568A56AF;
+
+u32 D_8003E6FC = 0x56D456F9;
+
+u32 D_8003E700 = 0x571D5742;
+
+u32 D_8003E704 = 0x5767578C;
+
+u32 D_8003E708 = 0x57B057D5;
+
+u32 D_8003E70C = 0x57F9581E;
+
+u32 D_8003E710 = 0x58425867;
+
+u32 D_8003E714 = 0x588B58AF;
+
+u32 D_8003E718 = 0x58D458F8;
+
+u32 D_8003E71C = 0x591C5940;
+
+u32 D_8003E720 = 0x59645988;
+
+u32 D_8003E724 = 0x59AC59D0;
+
+u32 D_8003E728 = 0x59F35A17;
+
+u32 D_8003E72C = 0x5A3B5A5E;
+
+u32 D_8003E730 = 0x5A825AA5;
+
+u32 D_8003E734 = 0x5AC95AEC;
+
+u32 D_8003E738 = 0x5B105B33;
+
+u32 D_8003E73C = 0x5B565B79;
+
+u32 D_8003E740 = 0x5B9D5BC0;
+
+u32 D_8003E744 = 0x5BE35C06;
+
+u32 D_8003E748 = 0x5C295C4B;
+
+u32 D_8003E74C = 0x5C6E5C91;
+
+u32 D_8003E750 = 0x5CB45CD6;
+
+u32 D_8003E754 = 0x5CF95D1B;
+
+u32 D_8003E758 = 0x5D3E5D60;
+
+u32 D_8003E75C = 0x5D835DA5;
+
+u32 D_8003E760 = 0x5DC75DE9;
+
+u32 D_8003E764 = 0x5E0B5E2D;
+
+u32 D_8003E768 = 0x5E505E71;
+
+u32 D_8003E76C = 0x5E935EB5;
+
+u32 D_8003E770 = 0x5ED75EF9;
+
+u32 D_8003E774 = 0x5F1A5F3C;
+
+u32 D_8003E778 = 0x5F5E5F7F;
+
+u32 D_8003E77C = 0x5FA05FC2;
+
+u32 D_8003E780 = 0x5FE36004;
+
+u32 D_8003E784 = 0x60266047;
+
+u32 D_8003E788 = 0x60686089;
+
+u32 D_8003E78C = 0x60AA60CB;
+
+u32 D_8003E790 = 0x60EC610D;
+
+u32 D_8003E794 = 0x612D614E;
+
+u32 D_8003E798 = 0x616F618F;
+
+u32 D_8003E79C = 0x61B061D0;
+
+u32 D_8003E7A0 = 0x61F16211;
+
+u32 D_8003E7A4 = 0x62316251;
+
+u32 D_8003E7A8 = 0x62716292;
+
+u32 D_8003E7AC = 0x62B262D2;
+
+u32 D_8003E7B0 = 0x62F26311;
+
+u32 D_8003E7B4 = 0x63316351;
+
+u32 D_8003E7B8 = 0x63716390;
+
+u32 D_8003E7BC = 0x63B063CF;
+
+u32 D_8003E7C0 = 0x63EF640E;
+
+u32 D_8003E7C4 = 0x642D644D;
+
+u32 D_8003E7C8 = 0x646C648B;
+
+u32 D_8003E7CC = 0x64AA64C9;
+
+u32 D_8003E7D0 = 0x64E86507;
+
+u32 D_8003E7D4 = 0x65266545;
+
+u32 D_8003E7D8 = 0x65636582;
+
+u32 D_8003E7DC = 0x65A065BF;
+
+u32 D_8003E7E0 = 0x65DD65FC;
+
+u32 D_8003E7E4 = 0x661A6639;
+
+u32 D_8003E7E8 = 0x66576675;
+
+u32 D_8003E7EC = 0x669366B1;
+
+u32 D_8003E7F0 = 0x66CF66ED;
+
+u32 D_8003E7F4 = 0x670B6729;
+
+u32 D_8003E7F8 = 0x67466764;
+
+u32 D_8003E7FC = 0x6782679F;
+
+u32 D_8003E800 = 0x67BD67DA;
+
+u32 D_8003E804 = 0x67F76815;
+
+u32 D_8003E808 = 0x6832684F;
+
+u32 D_8003E80C = 0x686C6889;
+
+u32 D_8003E810 = 0x68A668C3;
+
+u32 D_8003E814 = 0x68E068FD;
+
+u32 D_8003E818 = 0x69196936;
+
+u32 D_8003E81C = 0x6953696F;
+
+u32 D_8003E820 = 0x698C69A8;
+
+u32 D_8003E824 = 0x69C469E1;
+
+u32 D_8003E828 = 0x69FD6A19;
+
+u32 D_8003E82C = 0x6A356A51;
+
+u32 D_8003E830 = 0x6A6D6A89;
+
+u32 D_8003E834 = 0x6AA56AC1;
+
+u32 D_8003E838 = 0x6ADC6AF8;
+
+u32 D_8003E83C = 0x6B136B2F;
+
+u32 D_8003E840 = 0x6B4A6B66;
+
+u32 D_8003E844 = 0x6B816B9C;
+
+u32 D_8003E848 = 0x6BB86BD3;
+
+u32 D_8003E84C = 0x6BEE6C09;
+
+u32 D_8003E850 = 0x6C246C3F;
+
+u32 D_8003E854 = 0x6C596C74;
+
+u32 D_8003E858 = 0x6C8F6CA9;
+
+u32 D_8003E85C = 0x6CC46CDE;
+
+u32 D_8003E860 = 0x6CF96D13;
+
+u32 D_8003E864 = 0x6D2D6D48;
+
+u32 D_8003E868 = 0x6D626D7C;
+
+u32 D_8003E86C = 0x6D966DB0;
+
+u32 D_8003E870 = 0x6DCA6DE3;
+
+u32 D_8003E874 = 0x6DFD6E17;
+
+u32 D_8003E878 = 0x6E306E4A;
+
+u32 D_8003E87C = 0x6E636E7D;
+
+u32 D_8003E880 = 0x6E966EAF;
+
+u32 D_8003E884 = 0x6EC96EE2;
+
+u32 D_8003E888 = 0x6EFB6F14;
+
+u32 D_8003E88C = 0x6F2D6F46;
+
+u32 D_8003E890 = 0x6F5F6F77;
+
+u32 D_8003E894 = 0x6F906FA9;
+
+u32 D_8003E898 = 0x6FC16FDA;
+
+u32 D_8003E89C = 0x6FF2700A;
+
+u32 D_8003E8A0 = 0x7023703B;
+
+u32 D_8003E8A4 = 0x7053706B;
+
+u32 D_8003E8A8 = 0x7083709B;
+
+u32 D_8003E8AC = 0x70B370CB;
+
+u32 D_8003E8B0 = 0x70E270FA;
+
+u32 D_8003E8B4 = 0x71127129;
+
+u32 D_8003E8B8 = 0x71417158;
+
+u32 D_8003E8BC = 0x716F7186;
+
+u32 D_8003E8C0 = 0x719E71B5;
+
+u32 D_8003E8C4 = 0x71CC71E3;
+
+u32 D_8003E8C8 = 0x71FA7211;
+
+u32 D_8003E8CC = 0x7227723E;
+
+u32 D_8003E8D0 = 0x7255726B;
+
+u32 D_8003E8D4 = 0x72827298;
+
+u32 D_8003E8D8 = 0x72AF72C5;
+
+u32 D_8003E8DC = 0x72DB72F1;
+
+u32 D_8003E8E0 = 0x7307731D;
+
+u32 D_8003E8E4 = 0x73337349;
+
+u32 D_8003E8E8 = 0x735F7375;
+
+u32 D_8003E8EC = 0x738A73A0;
+
+u32 D_8003E8F0 = 0x73B573CB;
+
+u32 D_8003E8F4 = 0x73E073F6;
+
+u32 D_8003E8F8 = 0x740B7420;
+
+u32 D_8003E8FC = 0x7435744A;
+
+u32 D_8003E900 = 0x745F7474;
+
+u32 D_8003E904 = 0x7489749E;
+
+u32 D_8003E908 = 0x74B274C7;
+
+u32 D_8003E90C = 0x74DB74F0;
+
+u32 D_8003E910 = 0x75047519;
+
+u32 D_8003E914 = 0x752D7541;
+
+u32 D_8003E918 = 0x75557569;
+
+u32 D_8003E91C = 0x757D7591;
+
+u32 D_8003E920 = 0x75A575B9;
+
+u32 D_8003E924 = 0x75CC75E0;
+
+u32 D_8003E928 = 0x75F47607;
+
+u32 D_8003E92C = 0x761B762E;
+
+u32 D_8003E930 = 0x76417654;
+
+u32 D_8003E934 = 0x7668767B;
+
+u32 D_8003E938 = 0x768E76A0;
+
+u32 D_8003E93C = 0x76B376C6;
+
+u32 D_8003E940 = 0x76D976EB;
+
+u32 D_8003E944 = 0x76FE7710;
+
+u32 D_8003E948 = 0x77237735;
+
+u32 D_8003E94C = 0x7747775A;
+
+u32 D_8003E950 = 0x776C777E;
+
+u32 D_8003E954 = 0x779077A2;
+
+u32 D_8003E958 = 0x77B477C5;
+
+u32 D_8003E95C = 0x77D777E9;
+
+u32 D_8003E960 = 0x77FA780C;
+
+u32 D_8003E964 = 0x781D782E;
+
+u32 D_8003E968 = 0x78407851;
+
+u32 D_8003E96C = 0x78627873;
+
+u32 D_8003E970 = 0x78847895;
+
+u32 D_8003E974 = 0x78A678B6;
+
+u32 D_8003E978 = 0x78C778D8;
+
+u32 D_8003E97C = 0x78E878F9;
+
+u32 D_8003E980 = 0x79097919;
+
+u32 D_8003E984 = 0x792A793A;
+
+u32 D_8003E988 = 0x794A795A;
+
+u32 D_8003E98C = 0x796A797A;
+
+u32 D_8003E990 = 0x798A7999;
+
+u32 D_8003E994 = 0x79A979B9;
+
+u32 D_8003E998 = 0x79C879D8;
+
+u32 D_8003E99C = 0x79E779F6;
+
+u32 D_8003E9A0 = 0x7A057A15;
+
+u32 D_8003E9A4 = 0x7A247A33;
+
+u32 D_8003E9A8 = 0x7A427A50;
+
+u32 D_8003E9AC = 0x7A5F7A6E;
+
+u32 D_8003E9B0 = 0x7A7D7A8B;
+
+u32 D_8003E9B4 = 0x7A9A7AA8;
+
+u32 D_8003E9B8 = 0x7AB67AC5;
+
+u32 D_8003E9BC = 0x7AD37AE1;
+
+u32 D_8003E9C0 = 0x7AEF7AFD;
+
+u32 D_8003E9C4 = 0x7B0B7B19;
+
+u32 D_8003E9C8 = 0x7B267B34;
+
+u32 D_8003E9CC = 0x7B427B4F;
+
+u32 D_8003E9D0 = 0x7B5D7B6A;
+
+u32 D_8003E9D4 = 0x7B777B84;
+
+u32 D_8003E9D8 = 0x7B927B9F;
+
+u32 D_8003E9DC = 0x7BAC7BB9;
+
+u32 D_8003E9E0 = 0x7BC57BD2;
+
+u32 D_8003E9E4 = 0x7BDF7BEB;
+
+u32 D_8003E9E8 = 0x7BF87C05;
+
+u32 D_8003E9EC = 0x7C117C1D;
+
+u32 D_8003E9F0 = 0x7C297C36;
+
+u32 D_8003E9F4 = 0x7C427C4E;
+
+u32 D_8003E9F8 = 0x7C5A7C66;
+
+u32 D_8003E9FC = 0x7C717C7D;
+
+u32 D_8003EA00 = 0x7C897C94;
+
+u32 D_8003EA04 = 0x7CA07CAB;
+
+u32 D_8003EA08 = 0x7CB77CC2;
+
+u32 D_8003EA0C = 0x7CCD7CD8;
+
+u32 D_8003EA10 = 0x7CE37CEE;
+
+u32 D_8003EA14 = 0x7CF97D04;
+
+u32 D_8003EA18 = 0x7D0F7D19;
+
+u32 D_8003EA1C = 0x7D247D2F;
+
+u32 D_8003EA20 = 0x7D397D43;
+
+u32 D_8003EA24 = 0x7D4E7D58;
+
+u32 D_8003EA28 = 0x7D627D6C;
+
+u32 D_8003EA2C = 0x7D767D80;
+
+u32 D_8003EA30 = 0x7D8A7D94;
+
+u32 D_8003EA34 = 0x7D9D7DA7;
+
+u32 D_8003EA38 = 0x7DB07DBA;
+
+u32 D_8003EA3C = 0x7DC37DCD;
+
+u32 D_8003EA40 = 0x7DD67DDF;
+
+u32 D_8003EA44 = 0x7DE87DF1;
+
+u32 D_8003EA48 = 0x7DFA7E03;
+
+u32 D_8003EA4C = 0x7E0C7E14;
+
+u32 D_8003EA50 = 0x7E1D7E26;
+
+u32 D_8003EA54 = 0x7E2E7E37;
+
+u32 D_8003EA58 = 0x7E3F7E47;
+
+u32 D_8003EA5C = 0x7E4F7E57;
+
+u32 D_8003EA60 = 0x7E5F7E67;
+
+u32 D_8003EA64 = 0x7E6F7E77;
+
+u32 D_8003EA68 = 0x7E7F7E86;
+
+u32 D_8003EA6C = 0x7E8E7E95;
+
+u32 D_8003EA70 = 0x7E9D7EA4;
+
+u32 D_8003EA74 = 0x7EAB7EB3;
+
+u32 D_8003EA78 = 0x7EBA7EC1;
+
+u32 D_8003EA7C = 0x7EC87ECF;
+
+u32 D_8003EA80 = 0x7ED57EDC;
+
+u32 D_8003EA84 = 0x7EE37EE9;
+
+u32 D_8003EA88 = 0x7EF07EF6;
+
+u32 D_8003EA8C = 0x7EFD7F03;
+
+u32 D_8003EA90 = 0x7F097F0F;
+
+u32 D_8003EA94 = 0x7F157F1B;
+
+u32 D_8003EA98 = 0x7F217F27;
+
+u32 D_8003EA9C = 0x7F2D7F32;
+
+u32 D_8003EAA0 = 0x7F387F3D;
+
+u32 D_8003EAA4 = 0x7F437F48;
+
+u32 D_8003EAA8 = 0x7F4D7F53;
+
+u32 D_8003EAAC = 0x7F587F5D;
+
+u32 D_8003EAB0 = 0x7F627F67;
+
+u32 D_8003EAB4 = 0x7F6B7F70;
+
+u32 D_8003EAB8 = 0x7F757F79;
+
+u32 D_8003EABC = 0x7F7E7F82;
+
+u32 D_8003EAC0 = 0x7F877F8B;
+
+u32 D_8003EAC4 = 0x7F8F7F93;
+
+u32 D_8003EAC8 = 0x7F977F9B;
+
+u32 D_8003EACC = 0x7F9F7FA3;
+
+u32 D_8003EAD0 = 0x7FA77FAA;
+
+u32 D_8003EAD4 = 0x7FAE7FB1;
+
+u32 D_8003EAD8 = 0x7FB57FB8;
+
+u32 D_8003EADC = 0x7FBC7FBF;
+
+u32 D_8003EAE0 = 0x7FC27FC5;
+
+u32 D_8003EAE4 = 0x7FC87FCB;
+
+u32 D_8003EAE8 = 0x7FCE7FD0;
+
+u32 D_8003EAEC = 0x7FD37FD6;
+
+u32 D_8003EAF0 = 0x7FD87FDA;
+
+u32 D_8003EAF4 = 0x7FDD7FDF;
+
+u32 D_8003EAF8 = 0x7FE17FE3;
+
+u32 D_8003EAFC = 0x7FE57FE7;
+
+u32 D_8003EB00 = 0x7FE97FEB;
+
+u32 D_8003EB04 = 0x7FED7FEE;
+
+u32 D_8003EB08 = 0x7FF07FF2;
+
+u32 D_8003EB0C = 0x7FF37FF4;
+
+u32 D_8003EB10 = 0x7FF67FF7;
+
+u32 D_8003EB14 = 0x7FF87FF9;
+
+u32 D_8003EB18 = 0x7FFA7FFB;
+
+u32 D_8003EB1C = 0x7FFC7FFC;
+
+u32 D_8003EB20 = 0x7FFD7FFE;
+
+u32 D_8003EB24 = 0x7FFE7FFF;
+
+u32 D_8003EB28 = 0x7FFF7FFF;
+
+u32 D_8003EB2C = 0x7FFF8000;
+
+u32 D_8003EB30 = 0x80007FFF;
+
+u32 D_8003EB34 = 0x7FFF7FFF;
+
+u32 D_8003EB38 = 0x7FFF7FFE;
+
+u32 D_8003EB3C = 0x7FFE7FFD;
+
+u32 D_8003EB40 = 0x7FFC7FFC;
+
+u32 D_8003EB44 = 0x7FFB7FFA;
+
+u32 D_8003EB48 = 0x7FF97FF8;
+
+u32 D_8003EB4C = 0x7FF77FF6;
+
+u32 D_8003EB50 = 0x7FF47FF3;
+
+u32 D_8003EB54 = 0x7FF27FF0;
+
+u32 D_8003EB58 = 0x7FEE7FED;
+
+u32 D_8003EB5C = 0x7FEB7FE9;
+
+u32 D_8003EB60 = 0x7FE77FE5;
+
+u32 D_8003EB64 = 0x7FE37FE1;
+
+u32 D_8003EB68 = 0x7FDF7FDD;
+
+u32 D_8003EB6C = 0x7FDA7FD8;
+
+u32 D_8003EB70 = 0x7FD67FD3;
+
+u32 D_8003EB74 = 0x7FD07FCE;
+
+u32 D_8003EB78 = 0x7FCB7FC8;
+
+u32 D_8003EB7C = 0x7FC57FC2;
+
+u32 D_8003EB80 = 0x7FBF7FBC;
+
+u32 D_8003EB84 = 0x7FB87FB5;
+
+u32 D_8003EB88 = 0x7FB17FAE;
+
+u32 D_8003EB8C = 0x7FAA7FA7;
+
+u32 D_8003EB90 = 0x7FA37F9F;
+
+u32 D_8003EB94 = 0x7F9B7F97;
+
+u32 D_8003EB98 = 0x7F937F8F;
+
+u32 D_8003EB9C = 0x7F8B7F87;
+
+u32 D_8003EBA0 = 0x7F827F7E;
+
+u32 D_8003EBA4 = 0x7F797F75;
+
+u32 D_8003EBA8 = 0x7F707F6B;
+
+u32 D_8003EBAC = 0x7F677F62;
+
+u32 D_8003EBB0 = 0x7F5D7F58;
+
+u32 D_8003EBB4 = 0x7F537F4D;
+
+u32 D_8003EBB8 = 0x7F487F43;
+
+u32 D_8003EBBC = 0x7F3D7F38;
+
+u32 D_8003EBC0 = 0x7F327F2D;
+
+u32 D_8003EBC4 = 0x7F277F21;
+
+u32 D_8003EBC8 = 0x7F1B7F15;
+
+u32 D_8003EBCC = 0x7F0F7F09;
+
+u32 D_8003EBD0 = 0x7F037EFD;
+
+u32 D_8003EBD4 = 0x7EF67EF0;
+
+u32 D_8003EBD8 = 0x7EE97EE3;
+
+u32 D_8003EBDC = 0x7EDC7ED5;
+
+u32 D_8003EBE0 = 0x7ECF7EC8;
+
+u32 D_8003EBE4 = 0x7EC17EBA;
+
+u32 D_8003EBE8 = 0x7EB37EAB;
+
+u32 D_8003EBEC = 0x7EA47E9D;
+
+u32 D_8003EBF0 = 0x7E957E8E;
+
+u32 D_8003EBF4 = 0x7E867E7F;
+
+u32 D_8003EBF8 = 0x7E777E6F;
+
+u32 D_8003EBFC = 0x7E677E5F;
+
+u32 D_8003EC00 = 0x7E577E4F;
+
+u32 D_8003EC04 = 0x7E477E3F;
+
+u32 D_8003EC08 = 0x7E377E2E;
+
+u32 D_8003EC0C = 0x7E267E1D;
+
+u32 D_8003EC10 = 0x7E147E0C;
+
+u32 D_8003EC14 = 0x7E037DFA;
+
+u32 D_8003EC18 = 0x7DF17DE8;
+
+u32 D_8003EC1C = 0x7DDF7DD6;
+
+u32 D_8003EC20 = 0x7DCD7DC3;
+
+u32 D_8003EC24 = 0x7DBA7DB0;
+
+u32 D_8003EC28 = 0x7DA77D9D;
+
+u32 D_8003EC2C = 0x7D947D8A;
+
+u32 D_8003EC30 = 0x7D807D76;
+
+u32 D_8003EC34 = 0x7D6C7D62;
+
+u32 D_8003EC38 = 0x7D587D4E;
+
+u32 D_8003EC3C = 0x7D437D39;
+
+u32 D_8003EC40 = 0x7D2F7D24;
+
+u32 D_8003EC44 = 0x7D197D0F;
+
+u32 D_8003EC48 = 0x7D047CF9;
+
+u32 D_8003EC4C = 0x7CEE7CE3;
+
+u32 D_8003EC50 = 0x7CD87CCD;
+
+u32 D_8003EC54 = 0x7CC27CB7;
+
+u32 D_8003EC58 = 0x7CAB7CA0;
+
+u32 D_8003EC5C = 0x7C947C89;
+
+u32 D_8003EC60 = 0x7C7D7C71;
+
+u32 D_8003EC64 = 0x7C667C5A;
+
+u32 D_8003EC68 = 0x7C4E7C42;
+
+u32 D_8003EC6C = 0x7C367C29;
+
+u32 D_8003EC70 = 0x7C1D7C11;
+
+u32 D_8003EC74 = 0x7C057BF8;
+
+u32 D_8003EC78 = 0x7BEB7BDF;
+
+u32 D_8003EC7C = 0x7BD27BC5;
+
+u32 D_8003EC80 = 0x7BB97BAC;
+
+u32 D_8003EC84 = 0x7B9F7B92;
+
+u32 D_8003EC88 = 0x7B847B77;
+
+u32 D_8003EC8C = 0x7B6A7B5D;
+
+u32 D_8003EC90 = 0x7B4F7B42;
+
+u32 D_8003EC94 = 0x7B347B26;
+
+u32 D_8003EC98 = 0x7B197B0B;
+
+u32 D_8003EC9C = 0x7AFD7AEF;
+
+u32 D_8003ECA0 = 0x7AE17AD3;
+
+u32 D_8003ECA4 = 0x7AC57AB6;
+
+u32 D_8003ECA8 = 0x7AA87A9A;
+
+u32 D_8003ECAC = 0x7A8B7A7D;
+
+u32 D_8003ECB0 = 0x7A6E7A5F;
+
+u32 D_8003ECB4 = 0x7A507A42;
+
+u32 D_8003ECB8 = 0x7A337A24;
+
+u32 D_8003ECBC = 0x7A157A05;
+
+u32 D_8003ECC0 = 0x79F679E7;
+
+u32 D_8003ECC4 = 0x79D879C8;
+
+u32 D_8003ECC8 = 0x79B979A9;
+
+u32 D_8003ECCC = 0x7999798A;
+
+u32 D_8003ECD0 = 0x797A796A;
+
+u32 D_8003ECD4 = 0x795A794A;
+
+u32 D_8003ECD8 = 0x793A792A;
+
+u32 D_8003ECDC = 0x79197909;
+
+u32 D_8003ECE0 = 0x78F978E8;
+
+u32 D_8003ECE4 = 0x78D878C7;
+
+u32 D_8003ECE8 = 0x78B678A6;
+
+u32 D_8003ECEC = 0x78957884;
+
+u32 D_8003ECF0 = 0x78737862;
+
+u32 D_8003ECF4 = 0x78517840;
+
+u32 D_8003ECF8 = 0x782E781D;
+
+u32 D_8003ECFC = 0x780C77FA;
+
+u32 D_8003ED00 = 0x77E977D7;
+
+u32 D_8003ED04 = 0x77C577B4;
+
+u32 D_8003ED08 = 0x77A27790;
+
+u32 D_8003ED0C = 0x777E776C;
+
+u32 D_8003ED10 = 0x775A7747;
+
+u32 D_8003ED14 = 0x77357723;
+
+u32 D_8003ED18 = 0x771076FE;
+
+u32 D_8003ED1C = 0x76EB76D9;
+
+u32 D_8003ED20 = 0x76C676B3;
+
+u32 D_8003ED24 = 0x76A0768E;
+
+u32 D_8003ED28 = 0x767B7668;
+
+u32 D_8003ED2C = 0x76547641;
+
+u32 D_8003ED30 = 0x762E761B;
+
+u32 D_8003ED34 = 0x760775F4;
+
+u32 D_8003ED38 = 0x75E075CC;
+
+u32 D_8003ED3C = 0x75B975A5;
+
+u32 D_8003ED40 = 0x7591757D;
+
+u32 D_8003ED44 = 0x75697555;
+
+u32 D_8003ED48 = 0x7541752D;
+
+u32 D_8003ED4C = 0x75197504;
+
+u32 D_8003ED50 = 0x74F074DB;
+
+u32 D_8003ED54 = 0x74C774B2;
+
+u32 D_8003ED58 = 0x749E7489;
+
+u32 D_8003ED5C = 0x7474745F;
+
+u32 D_8003ED60 = 0x744A7435;
+
+u32 D_8003ED64 = 0x7420740B;
+
+u32 D_8003ED68 = 0x73F673E0;
+
+u32 D_8003ED6C = 0x73CB73B5;
+
+u32 D_8003ED70 = 0x73A0738A;
+
+u32 D_8003ED74 = 0x7375735F;
+
+u32 D_8003ED78 = 0x73497333;
+
+u32 D_8003ED7C = 0x731D7307;
+
+u32 D_8003ED80 = 0x72F172DB;
+
+u32 D_8003ED84 = 0x72C572AF;
+
+u32 D_8003ED88 = 0x72987282;
+
+u32 D_8003ED8C = 0x726B7255;
+
+u32 D_8003ED90 = 0x723E7227;
+
+u32 D_8003ED94 = 0x721171FA;
+
+u32 D_8003ED98 = 0x71E371CC;
+
+u32 D_8003ED9C = 0x71B5719E;
+
+u32 D_8003EDA0 = 0x7186716F;
+
+u32 D_8003EDA4 = 0x71587141;
+
+u32 D_8003EDA8 = 0x71297112;
+
+u32 D_8003EDAC = 0x70FA70E2;
+
+u32 D_8003EDB0 = 0x70CB70B3;
+
+u32 D_8003EDB4 = 0x709B7083;
+
+u32 D_8003EDB8 = 0x706B7053;
+
+u32 D_8003EDBC = 0x703B7023;
+
+u32 D_8003EDC0 = 0x700A6FF2;
+
+u32 D_8003EDC4 = 0x6FDA6FC1;
+
+u32 D_8003EDC8 = 0x6FA96F90;
+
+u32 D_8003EDCC = 0x6F776F5F;
+
+u32 D_8003EDD0 = 0x6F466F2D;
+
+u32 D_8003EDD4 = 0x6F146EFB;
+
+u32 D_8003EDD8 = 0x6EE26EC9;
+
+u32 D_8003EDDC = 0x6EAF6E96;
+
+u32 D_8003EDE0 = 0x6E7D6E63;
+
+u32 D_8003EDE4 = 0x6E4A6E30;
+
+u32 D_8003EDE8 = 0x6E176DFD;
+
+u32 D_8003EDEC = 0x6DE36DCA;
+
+u32 D_8003EDF0 = 0x6DB06D96;
+
+u32 D_8003EDF4 = 0x6D7C6D62;
+
+u32 D_8003EDF8 = 0x6D486D2D;
+
+u32 D_8003EDFC = 0x6D136CF9;
+
+u32 D_8003EE00 = 0x6CDE6CC4;
+
+u32 D_8003EE04 = 0x6CA96C8F;
+
+u32 D_8003EE08 = 0x6C746C59;
+
+u32 D_8003EE0C = 0x6C3F6C24;
+
+u32 D_8003EE10 = 0x6C096BEE;
+
+u32 D_8003EE14 = 0x6BD36BB8;
+
+u32 D_8003EE18 = 0x6B9C6B81;
+
+u32 D_8003EE1C = 0x6B666B4A;
+
+u32 D_8003EE20 = 0x6B2F6B13;
+
+u32 D_8003EE24 = 0x6AF86ADC;
+
+u32 D_8003EE28 = 0x6AC16AA5;
+
+u32 D_8003EE2C = 0x6A896A6D;
+
+u32 D_8003EE30 = 0x6A516A35;
+
+u32 D_8003EE34 = 0x6A1969FD;
+
+u32 D_8003EE38 = 0x69E169C4;
+
+u32 D_8003EE3C = 0x69A8698C;
+
+u32 D_8003EE40 = 0x696F6953;
+
+u32 D_8003EE44 = 0x69366919;
+
+u32 D_8003EE48 = 0x68FD68E0;
+
+u32 D_8003EE4C = 0x68C368A6;
+
+u32 D_8003EE50 = 0x6889686C;
+
+u32 D_8003EE54 = 0x684F6832;
+
+u32 D_8003EE58 = 0x681567F7;
+
+u32 D_8003EE5C = 0x67DA67BD;
+
+u32 D_8003EE60 = 0x679F6782;
+
+u32 D_8003EE64 = 0x67646746;
+
+u32 D_8003EE68 = 0x6729670B;
+
+u32 D_8003EE6C = 0x66ED66CF;
+
+u32 D_8003EE70 = 0x66B16693;
+
+u32 D_8003EE74 = 0x66756657;
+
+u32 D_8003EE78 = 0x6639661A;
+
+u32 D_8003EE7C = 0x65FC65DD;
+
+u32 D_8003EE80 = 0x65BF65A0;
+
+u32 D_8003EE84 = 0x65826563;
+
+u32 D_8003EE88 = 0x65456526;
+
+u32 D_8003EE8C = 0x650764E8;
+
+u32 D_8003EE90 = 0x64C964AA;
+
+u32 D_8003EE94 = 0x648B646C;
+
+u32 D_8003EE98 = 0x644D642D;
+
+u32 D_8003EE9C = 0x640E63EF;
+
+u32 D_8003EEA0 = 0x63CF63B0;
+
+u32 D_8003EEA4 = 0x63906371;
+
+u32 D_8003EEA8 = 0x63516331;
+
+u32 D_8003EEAC = 0x631162F2;
+
+u32 D_8003EEB0 = 0x62D262B2;
+
+u32 D_8003EEB4 = 0x62926271;
+
+u32 D_8003EEB8 = 0x62516231;
+
+u32 D_8003EEBC = 0x621161F1;
+
+u32 D_8003EEC0 = 0x61D061B0;
+
+u32 D_8003EEC4 = 0x618F616F;
+
+u32 D_8003EEC8 = 0x614E612D;
+
+u32 D_8003EECC = 0x610D60EC;
+
+u32 D_8003EED0 = 0x60CB60AA;
+
+u32 D_8003EED4 = 0x60896068;
+
+u32 D_8003EED8 = 0x60476026;
+
+u32 D_8003EEDC = 0x60045FE3;
+
+u32 D_8003EEE0 = 0x5FC25FA0;
+
+u32 D_8003EEE4 = 0x5F7F5F5E;
+
+u32 D_8003EEE8 = 0x5F3C5F1A;
+
+u32 D_8003EEEC = 0x5EF95ED7;
+
+u32 D_8003EEF0 = 0x5EB55E93;
+
+u32 D_8003EEF4 = 0x5E715E50;
+
+u32 D_8003EEF8 = 0x5E2D5E0B;
+
+u32 D_8003EEFC = 0x5DE95DC7;
+
+u32 D_8003EF00 = 0x5DA55D83;
+
+u32 D_8003EF04 = 0x5D605D3E;
+
+u32 D_8003EF08 = 0x5D1B5CF9;
+
+u32 D_8003EF0C = 0x5CD65CB4;
+
+u32 D_8003EF10 = 0x5C915C6E;
+
+u32 D_8003EF14 = 0x5C4B5C29;
+
+u32 D_8003EF18 = 0x5C065BE3;
+
+u32 D_8003EF1C = 0x5BC05B9D;
+
+u32 D_8003EF20 = 0x5B795B56;
+
+u32 D_8003EF24 = 0x5B335B10;
+
+u32 D_8003EF28 = 0x5AEC5AC9;
+
+u32 D_8003EF2C = 0x5AA55A82;
+
+u32 D_8003EF30 = 0x5A5E5A3B;
+
+u32 D_8003EF34 = 0x5A1759F3;
+
+u32 D_8003EF38 = 0x59D059AC;
+
+u32 D_8003EF3C = 0x59885964;
+
+u32 D_8003EF40 = 0x5940591C;
+
+u32 D_8003EF44 = 0x58F858D4;
+
+u32 D_8003EF48 = 0x58AF588B;
+
+u32 D_8003EF4C = 0x58675842;
+
+u32 D_8003EF50 = 0x581E57F9;
+
+u32 D_8003EF54 = 0x57D557B0;
+
+u32 D_8003EF58 = 0x578C5767;
+
+u32 D_8003EF5C = 0x5742571D;
+
+u32 D_8003EF60 = 0x56F956D4;
+
+u32 D_8003EF64 = 0x56AF568A;
+
+u32 D_8003EF68 = 0x56655640;
+
+u32 D_8003EF6C = 0x561A55F5;
+
+u32 D_8003EF70 = 0x55D055AB;
+
+u32 D_8003EF74 = 0x55855560;
+
+u32 D_8003EF78 = 0x553A5515;
+
+u32 D_8003EF7C = 0x54EF54CA;
+
+u32 D_8003EF80 = 0x54A4547E;
+
+u32 D_8003EF84 = 0x54585433;
+
+u32 D_8003EF88 = 0x540D53E7;
+
+u32 D_8003EF8C = 0x53C1539B;
+
+u32 D_8003EF90 = 0x5375534E;
+
+u32 D_8003EF94 = 0x53285302;
+
+u32 D_8003EF98 = 0x52DC52B5;
+
+u32 D_8003EF9C = 0x528F5269;
+
+u32 D_8003EFA0 = 0x5242521C;
+
+u32 D_8003EFA4 = 0x51F551CE;
+
+u32 D_8003EFA8 = 0x51A85181;
+
+u32 D_8003EFAC = 0x515A5133;
+
+u32 D_8003EFB0 = 0x510C50E5;
+
+u32 D_8003EFB4 = 0x50BF5097;
+
+u32 D_8003EFB8 = 0x50705049;
+
+u32 D_8003EFBC = 0x50224FFB;
+
+u32 D_8003EFC0 = 0x4FD44FAC;
+
+u32 D_8003EFC4 = 0x4F854F5E;
+
+u32 D_8003EFC8 = 0x4F364F0F;
+
+u32 D_8003EFCC = 0x4EE74EBF;
+
+u32 D_8003EFD0 = 0x4E984E70;
+
+u32 D_8003EFD4 = 0x4E484E21;
+
+u32 D_8003EFD8 = 0x4DF94DD1;
+
+u32 D_8003EFDC = 0x4DA94D81;
+
+u32 D_8003EFE0 = 0x4D594D31;
+
+u32 D_8003EFE4 = 0x4D094CE1;
+
+u32 D_8003EFE8 = 0x4CB84C90;
+
+u32 D_8003EFEC = 0x4C684C3F;
+
+u32 D_8003EFF0 = 0x4C174BEF;
+
+u32 D_8003EFF4 = 0x4BC64B9E;
+
+u32 D_8003EFF8 = 0x4B754B4C;
+
+u32 D_8003EFFC = 0x4B244AFB;
+
+u32 D_8003F000 = 0x4AD24AA9;
+
+u32 D_8003F004 = 0x4A814A58;
+
+u32 D_8003F008 = 0x4A2F4A06;
+
+u32 D_8003F00C = 0x49DD49B4;
+
+u32 D_8003F010 = 0x498A4961;
+
+u32 D_8003F014 = 0x4938490F;
+
+u32 D_8003F018 = 0x48E648BC;
+
+u32 D_8003F01C = 0x48934869;
+
+u32 D_8003F020 = 0x48404816;
+
+u32 D_8003F024 = 0x47ED47C3;
+
+u32 D_8003F028 = 0x479A4770;
+
+u32 D_8003F02C = 0x4746471C;
+
+u32 D_8003F030 = 0x46F346C9;
+
+u32 D_8003F034 = 0x469F4675;
+
+u32 D_8003F038 = 0x464B4621;
+
+u32 D_8003F03C = 0x45F745CD;
+
+u32 D_8003F040 = 0x45A34578;
+
+u32 D_8003F044 = 0x454E4524;
+
+u32 D_8003F048 = 0x44FA44CF;
+
+u32 D_8003F04C = 0x44A5447A;
+
+u32 D_8003F050 = 0x44504425;
+
+u32 D_8003F054 = 0x43FB43D0;
+
+u32 D_8003F058 = 0x43A5437B;
+
+u32 D_8003F05C = 0x43504325;
+
+u32 D_8003F060 = 0x42FA42D0;
+
+u32 D_8003F064 = 0x42A5427A;
+
+u32 D_8003F068 = 0x424F4224;
+
+u32 D_8003F06C = 0x41F941CE;
+
+u32 D_8003F070 = 0x41A24177;
+
+u32 D_8003F074 = 0x414C4121;
+
+u32 D_8003F078 = 0x40F640CA;
+
+u32 D_8003F07C = 0x409F4073;
+
+u32 D_8003F080 = 0x4048401D;
+
+u32 D_8003F084 = 0x3FF13FC5;
+
+u32 D_8003F088 = 0x3F9A3F6E;
+
+u32 D_8003F08C = 0x3F433F17;
+
+u32 D_8003F090 = 0x3EEB3EBF;
+
+u32 D_8003F094 = 0x3E933E68;
+
+u32 D_8003F098 = 0x3E3C3E10;
+
+u32 D_8003F09C = 0x3DE43DB8;
+
+u32 D_8003F0A0 = 0x3D8C3D60;
+
+u32 D_8003F0A4 = 0x3D333D07;
+
+u32 D_8003F0A8 = 0x3CDB3CAF;
+
+u32 D_8003F0AC = 0x3C833C56;
+
+u32 D_8003F0B0 = 0x3C2A3BFD;
+
+u32 D_8003F0B4 = 0x3BD13BA5;
+
+u32 D_8003F0B8 = 0x3B783B4C;
+
+u32 D_8003F0BC = 0x3B1F3AF2;
+
+u32 D_8003F0C0 = 0x3AC63A99;
+
+u32 D_8003F0C4 = 0x3A6C3A40;
+
+u32 D_8003F0C8 = 0x3A1339E6;
+
+u32 D_8003F0CC = 0x39B9398C;
+
+u32 D_8003F0D0 = 0x395F3932;
+
+u32 D_8003F0D4 = 0x390638D8;
+
+u32 D_8003F0D8 = 0x38AB387E;
+
+u32 D_8003F0DC = 0x38513824;
+
+u32 D_8003F0E0 = 0x37F737CA;
+
+u32 D_8003F0E4 = 0x379C376F;
+
+u32 D_8003F0E8 = 0x37423714;
+
+u32 D_8003F0EC = 0x36E736BA;
+
+u32 D_8003F0F0 = 0x368C365F;
+
+u32 D_8003F0F4 = 0x36313604;
+
+u32 D_8003F0F8 = 0x35D635A8;
+
+u32 D_8003F0FC = 0x357B354D;
+
+u32 D_8003F100 = 0x351F34F2;
+
+u32 D_8003F104 = 0x34C43496;
+
+u32 D_8003F108 = 0x3468343A;
+
+u32 D_8003F10C = 0x340C33DE;
+
+u32 D_8003F110 = 0x33B03382;
+
+u32 D_8003F114 = 0x33543326;
+
+u32 D_8003F118 = 0x32F832CA;
+
+u32 D_8003F11C = 0x329C326E;
+
+u32 D_8003F120 = 0x32403211;
+
+u32 D_8003F124 = 0x31E331B5;
+
+u32 D_8003F128 = 0x31863158;
+
+u32 D_8003F12C = 0x312A30FB;
+
+u32 D_8003F130 = 0x30CD309E;
+
+u32 D_8003F134 = 0x30703041;
+
+u32 D_8003F138 = 0x30132FE4;
+
+u32 D_8003F13C = 0x2FB52F87;
+
+u32 D_8003F140 = 0x2F582F29;
+
+u32 D_8003F144 = 0x2EFB2ECC;
+
+u32 D_8003F148 = 0x2E9D2E6E;
+
+u32 D_8003F14C = 0x2E3F2E11;
+
+u32 D_8003F150 = 0x2DE22DB3;
+
+u32 D_8003F154 = 0x2D842D55;
+
+u32 D_8003F158 = 0x2D262CF7;
+
+u32 D_8003F15C = 0x2CC82C98;
+
+u32 D_8003F160 = 0x2C692C3A;
+
+u32 D_8003F164 = 0x2C0B2BDC;
+
+u32 D_8003F168 = 0x2BAD2B7D;
+
+u32 D_8003F16C = 0x2B4E2B1F;
+
+u32 D_8003F170 = 0x2AEF2AC0;
+
+u32 D_8003F174 = 0x2A912A61;
+
+u32 D_8003F178 = 0x2A322A02;
+
+u32 D_8003F17C = 0x29D329A3;
+
+u32 D_8003F180 = 0x29742944;
+
+u32 D_8003F184 = 0x291528E5;
+
+u32 D_8003F188 = 0x28B52886;
+
+u32 D_8003F18C = 0x28562826;
+
+u32 D_8003F190 = 0x27F627C7;
+
+u32 D_8003F194 = 0x27972767;
+
+u32 D_8003F198 = 0x27372707;
+
+u32 D_8003F19C = 0x26D826A8;
+
+u32 D_8003F1A0 = 0x26782648;
+
+u32 D_8003F1A4 = 0x261825E8;
+
+u32 D_8003F1A8 = 0x25B82588;
+
+u32 D_8003F1AC = 0x25582528;
+
+u32 D_8003F1B0 = 0x24F724C7;
+
+u32 D_8003F1B4 = 0x24972467;
+
+u32 D_8003F1B8 = 0x24372407;
+
+u32 D_8003F1BC = 0x23D623A6;
+
+u32 D_8003F1C0 = 0x23762345;
+
+u32 D_8003F1C4 = 0x231522E5;
+
+u32 D_8003F1C8 = 0x22B42284;
+
+u32 D_8003F1CC = 0x22542223;
+
+u32 D_8003F1D0 = 0x21F321C2;
+
+u32 D_8003F1D4 = 0x21922161;
+
+u32 D_8003F1D8 = 0x21312100;
+
+u32 D_8003F1DC = 0x20D0209F;
+
+u32 D_8003F1E0 = 0x206E203E;
+
+u32 D_8003F1E4 = 0x200D1FDC;
+
+u32 D_8003F1E8 = 0x1FAC1F7B;
+
+u32 D_8003F1EC = 0x1F4A1F19;
+
+u32 D_8003F1F0 = 0x1EE91EB8;
+
+u32 D_8003F1F4 = 0x1E871E56;
+
+u32 D_8003F1F8 = 0x1E251DF5;
+
+u32 D_8003F1FC = 0x1DC41D93;
+
+u32 D_8003F200 = 0x1D621D31;
+
+u32 D_8003F204 = 0x1D001CCF;
+
+u32 D_8003F208 = 0x1C9E1C6D;
+
+u32 D_8003F20C = 0x1C3C1C0B;
+
+u32 D_8003F210 = 0x1BDA1BA9;
+
+u32 D_8003F214 = 0x1B781B47;
+
+u32 D_8003F218 = 0x1B161AE4;
+
+u32 D_8003F21C = 0x1AB31A82;
+
+u32 D_8003F220 = 0x1A511A20;
+
+u32 D_8003F224 = 0x19EF19BD;
+
+u32 D_8003F228 = 0x198C195B;
+
+u32 D_8003F22C = 0x192A18F8;
+
+u32 D_8003F230 = 0x18C71896;
+
+u32 D_8003F234 = 0x18641833;
+
+u32 D_8003F238 = 0x180217D0;
+
+u32 D_8003F23C = 0x179F176D;
+
+u32 D_8003F240 = 0x173C170A;
+
+u32 D_8003F244 = 0x16D916A8;
+
+u32 D_8003F248 = 0x16761645;
+
+u32 D_8003F24C = 0x161315E2;
+
+u32 D_8003F250 = 0x15B0157F;
+
+u32 D_8003F254 = 0x154D151B;
+
+u32 D_8003F258 = 0x14EA14B8;
+
+u32 D_8003F25C = 0x14871455;
+
+u32 D_8003F260 = 0x142313F2;
+
+u32 D_8003F264 = 0x13C0138E;
+
+u32 D_8003F268 = 0x135D132B;
+
+u32 D_8003F26C = 0x12F912C8;
+
+u32 D_8003F270 = 0x12961264;
+
+u32 D_8003F274 = 0x12321201;
+
+u32 D_8003F278 = 0x11CF119D;
+
+u32 D_8003F27C = 0x116B1139;
+
+u32 D_8003F280 = 0x110810D6;
+
+u32 D_8003F284 = 0x10A41072;
+
+u32 D_8003F288 = 0x1040100E;
+
+u32 D_8003F28C = 0x0FDD0FAB;
+
+u32 D_8003F290 = 0x0F790F47;
+
+u32 D_8003F294 = 0x0F150EE3;
+
+u32 D_8003F298 = 0x0EB10E7F;
+
+u32 D_8003F29C = 0x0E4D0E1B;
+
+u32 D_8003F2A0 = 0x0DE90DB7;
+
+u32 D_8003F2A4 = 0x0D850D53;
+
+u32 D_8003F2A8 = 0x0D210CEF;
+
+u32 D_8003F2AC = 0x0CBD0C8B;
+
+u32 D_8003F2B0 = 0x0C590C27;
+
+u32 D_8003F2B4 = 0x0BF50BC3;
+
+u32 D_8003F2B8 = 0x0B910B5F;
+
+u32 D_8003F2BC = 0x0B2D0AFB;
+
+u32 D_8003F2C0 = 0x0AC90A97;
+
+u32 D_8003F2C4 = 0x0A650A33;
+
+u32 D_8003F2C8 = 0x0A0009CE;
+
+u32 D_8003F2CC = 0x099C096A;
+
+u32 D_8003F2D0 = 0x09380906;
+
+u32 D_8003F2D4 = 0x08D408A2;
+
+u32 D_8003F2D8 = 0x086F083D;
+
+u32 D_8003F2DC = 0x080B07D9;
+
+u32 D_8003F2E0 = 0x07A70775;
+
+u32 D_8003F2E4 = 0x07420710;
+
+u32 D_8003F2E8 = 0x06DE06AC;
+
+u32 D_8003F2EC = 0x067A0647;
+
+u32 D_8003F2F0 = 0x061505E3;
+
+u32 D_8003F2F4 = 0x05B1057F;
+
+u32 D_8003F2F8 = 0x054C051A;
+
+u32 D_8003F2FC = 0x04E804B6;
+
+u32 D_8003F300 = 0x04830451;
+
+u32 D_8003F304 = 0x041F03ED;
+
+u32 D_8003F308 = 0x03BA0388;
+
+u32 D_8003F30C = 0x03560324;
+
+u32 D_8003F310 = 0x02F102BF;
+
+u32 D_8003F314 = 0x028D025B;
+
+u32 D_8003F318 = 0x022801F6;
+
+u32 D_8003F31C = 0x01C40192;
+
+u32 D_8003F320 = 0x015F012D;
+
+u32 D_8003F324 = 0x00FB00C9;
+
+u32 D_8003F328 = 0x00960064;
+
+u32 D_8003F32C = 0x00320000;
+
+u32 D_8003F330 = 0x00000000;
+
+u32 D_8003F334 = 0x00000000;
+
+
+
+u32 D_8003F338 = 0x00000001;
+
+u32 D_8003F33C = 0x00000000;
+
+u32 D_8003F340 = 0x00000000;
+
+u32 D_8003F344 = 0x00000001;
+
+u32 D_8003F348 = 0x00000000;
+
+u32 D_8003F34C = 0x00000000;
+
+
+u32 D_8003F350 = 0x8004ADC0;
+u32 D_8003F354 = 0x0004B000;
+u32 D_8003F358 = 0x7D001018;
+u32 D_8003F35C = 0x40400818;
+u32 D_8003F360 = 0x00003214;
+
+u32 *D_8003F364 = _data_3E1400SegmentRomStart;
+u32 *D_8003F368 = _data_3E6BC0SegmentRomStart;
+u32 *D_8003F36C = _data_3E6BC0SegmentRomStart;
+u32 *D_8003F370 = _data_2A8CB0SegmentRomStart;
+u32 *D_8003F374 = _data_2B1510SegmentRomStart;
+u32 *D_8003F378 = _data_2B1510SegmentRomStart;
+u32 *D_8003F37C = _data_250320SegmentRomStart;
+
+u32 D_8003F380 = 0x00100808;
+u32 D_8003F384 = 0x00000000;
+u32 D_8003F388 = 0x00000000;
+u32 D_8003F38C = 0x00000000;
+
+u32 D_8003F390 = 0x00000000;
+u32 D_8003F394 = 0x00000000;
+u32 D_8003F398 = 0x00000000;
+u32 D_8003F39C = 0x00000000;
+
+u32 *D_8003F3A0 = _data_49F590SegmentRomStart;
+u32 *D_8003F3A4 = _data_49F590SegmentRomEnd;
+
+u32 *D_8003F3A8 = _data_4A0340SegmentRomStart;
+u32 *D_8003F3AC = _data_4A0340SegmentRomEnd;
+
+u32 *D_8003F3B0 = _data_4A3B60SegmentRomStart;
+u32 *D_8003F3B4 = _data_4A3B60SegmentRomEnd;
+
+u32 D_8003F3B8 = 0x00000000;
+u32 D_8003F3BC = 0x00000000;
+
+u32 D_8003F3C0 = 0x00000000;
+
+u32 D_8003F3C4 = 0x00000000;
+
+u32 D_8003F3C8 = 0x00000003;
+
+u32 D_8003F3CC = 0x00000000;
+
+u32 D_8003F3D0 = 0x00000000;
+
+u32 D_8003F3D4 = 0x00000000;
+
+u32 D_8003F3D8 = 0x00000000;
+
+u32 D_8003F3DC = 0x00000003;
+
+u32 D_8003F3E0 = 0x00000003;
+
+u32 D_8003F3E4 = 0x00000000;
+
+u32 D_8003F3E8 = 0x00000003;
+
+u32 D_8003F3EC = 0x00000003;
+
+u32 D_8003F3F0 = 0x00000000;
+
+u32 D_8003F3F4 = 0x00000003;
+
+u32 D_8003F3F8 = 0x00000003;
+
+u32 D_8003F3FC = 0x00000003;
+
+u32 D_8003F400 = 0x00000000;
+
+u32 D_8003F404 = 0x00000003;
+
+u32 D_8003F408 = 0x00000000;
+
+u32 D_8003F40C = 0x00000006;
+
+u32 D_8003F410 = 0x00000003;
+
+u32 D_8003F414 = 0x00000003;
+
+u32 D_8003F418 = 0x00000003;
+
+u32 D_8003F41C = 0x00000006;
+
+u32 D_8003F420 = 0x00000000;
+
+u32 D_8003F424 = 0x00000006;
+
+u32 D_8003F428 = 0x00000003;
+
+u32 D_8003F42C = 0x00000006;
+
+u32 D_8003F430 = 0x00000005;
+
+u32 D_8003F434 = 0x00000006;
+
+u32 D_8003F438 = 0x00000005;
+
+u32 D_8003F43C = 0x00000006;
+
+u32 D_8003F440 = 0x00000001;
+
+u32 D_8003F444 = 0x00000001;
+
+u32 D_8003F448 = 0x00000001;
+
+u32 D_8003F44C = 0x00000000;
+
+u32 D_8003F450 = 0x00000001;
+
+u32 D_8003F454 = 0x00000001;
+
+u32 D_8003F458 = 0x00000001;
+
+u32 D_8003F45C = 0x00000000;
+
+u32 D_8003F460 = 0x00000000;
+
+u32 D_8003F464 = 0x00000000;
+
+u32 D_8003F468 = 0x00000001;
+
+u32 D_8003F46C = 0x00000000;
+
+u32 D_8003F470 = 0x00000000;
+
+u32 D_8003F474 = 0x00000001;
+
+u32 D_8003F478 = 0x00000000;
+
+u32 D_8003F47C = 0x00000000;
+
+u32 D_8003F480 = 0x00000001;
+
+u32 D_8003F484 = 0x00000000;
+
+u32 D_8003F488 = 0x00000001;
+
+u32 D_8003F48C = 0x00000001;
+
+u32 D_8003F490 = 0x00000001;
+
+u32 D_8003F494 = 0x00000000;
+
+u32 D_8003F498 = 0x00000001;
+
+u32 D_8003F49C = 0x00000000;
+
+u32 D_8003F4A0 = 0x00000001;
+
+u32 D_8003F4A4 = 0x00000000;
+
+u32 D_8003F4A8 = 0x00000001;
+
+u32 D_8003F4AC = 0x00000001;
+
+u32 D_8003F4B0 = 0x00000000;
+
+u32 D_8003F4B4 = 0x00000001;
+
+u32 D_8003F4B8 = 0x00000001;
+
+u32 D_8003F4BC = 0x00000000;
+
+u32 D_8003F4C0 = 0x00000000;
+
+u32 D_8003F4C4 = 0x00000001;
+
+u32 D_8003F4C8 = 0x00000001;
+
+u32 D_8003F4CC = 0x00000001;
+
+u32 D_8003F4D0 = 0x00000000;
+
+u32 D_8003F4D4 = 0x00000001;
+
+u32 D_8003F4D8 = 0x00000000;
+
+u32 D_8003F4DC = 0x00000000;
+
+u32 D_8003F4E0 = 0x00000001;
+
+u32 D_8003F4E4 = 0x00000001;
+
+u32 D_8003F4E8 = 0x00000000;
+
+u32 D_8003F4EC = 0x00000001;
+
+u32 D_8003F4F0 = 0x00000000;
+
+u32 D_8003F4F4 = 0x00000001;
+
+u32 D_8003F4F8 = 0x00000001;
+
+u32 D_8003F4FC = 0x00000000;
+
+u32 D_8003F500 = 0x00000001;
+
+u32 D_8003F504 = 0x00000001;
+
+u32 D_8003F508 = 0x00000000;
+
+u32 D_8003F50C = 0x00000001;
+
+u32 D_8003F510 = 0x00000001;
+
+u32 D_8003F514 = 0x00000001;
+
+u32 D_8003F518 = 0x00000001;
+
+u32 D_8003F51C = 0x00000000;
+
+u32 D_8003F520 = 0x00000001;
+
+u32 D_8003F524 = 0x00000000;
+
+u32 D_8003F528 = 0x00000001;
+
+u32 D_8003F52C = 0x00000000;
+
+u32 D_8003F530 = 0x00000000;
+
+u32 D_8003F534 = 0x00000001;
+
+u32 D_8003F538 = 0x00000000;
+
+u32 D_8003F53C = 0x00000000;
+
+u32 D_8003F540 = 0x00000001;
+
+u32 D_8003F544 = 0x00000001;
+
+u32 D_8003F548 = 0x00000001;
+
+u32 D_8003F54C = 0x00000001;
+
+u32 D_8003F550 = 0x00000001;
+
+u32 D_8003F554 = 0x00000001;
+
+u32 D_8003F558 = 0x00000001;
+
+u32 D_8003F55C = 0x00000000;
+
+u32 D_8003F560 = 0x00000001;
+
+u32 D_8003F564 = 0x00000001;
+
+u32 D_8003F568 = 0x00000001;
+
+u32 D_8003F56C = 0x00000001;
+
+u32 D_8003F570 = 0x00000000;
+
+u32 D_8003F574 = 0x00000001;
+
+u32 D_8003F578 = 0x00000001;
+
+u32 D_8003F57C = 0x00000000;
+
+u32 D_8003F580 = 0x00000001;
+
+u32 D_8003F584 = 0x00000001;
+
+u32 D_8003F588 = 0x00000001;
+
+u32 D_8003F58C = 0x00000001;
+
+u32 D_8003F590 = 0x00000001;
+
+u32 D_8003F594 = 0x00000001;
+
+u32 D_8003F598 = 0x00000000;
+
+u32 D_8003F59C = 0x00000000;
+
+u32 D_8003F5A0 = 0x00000000;
+
+u32 D_8003F5A4 = 0x00000001;
+
+u32 D_8003F5A8 = 0x00000000;
+
+u32 D_8003F5AC = 0x00000001;
+
+u32 D_8003F5B0 = 0x00000001;
+
+u32 D_8003F5B4 = 0x00000001;
+
+u32 D_8003F5B8 = 0x00000001;
+
+u32 D_8003F5BC = 0x00000000;
+
+u32 D_8003F5C0 = 0x00000000;
+
+u32 D_8003F5C4 = 0x00000000;
+
+u32 D_8003F5C8 = 0x00000000;
+
+u32 D_8003F5CC = 0x00000001;
+
+u32 D_8003F5D0 = 0x00000001;
+
+u32 D_8003F5D4 = 0x00000000;
+
+u32 D_8003F5D8 = 0x00000001;
+
+u32 D_8003F5DC = 0x00000000;
+
+u32 D_8003F5E0 = 0x00000000;
+
+u32 D_8003F5E4 = 0x00000000;
+
+u32 D_8003F5E8 = 0x00000001;
+
+u32 D_8003F5EC = 0x00000001;
+
+u32 D_8003F5F0 = 0x00000001;
+
+u32 D_8003F5F4 = 0x00000001;
+
+u32 D_8003F5F8 = 0x00000001;
+
+u32 D_8003F5FC = 0x00000000;
+
+u32 D_8003F600 = 0x00000001;
+
+u32 D_8003F604 = 0x00000001;
+
+u32 D_8003F608 = 0x00000000;
+
+u32 D_8003F60C = 0x00000001;
+
+u32 D_8003F610 = 0x00000001;
+
+u32 D_8003F614 = 0x00000000;
+
+u32 D_8003F618 = 0x00000001;
+
+u32 D_8003F61C = 0x00000000;
+
+u32 D_8003F620 = 0x00000001;
+
+u32 D_8003F624 = 0x00000001;
+
+u32 D_8003F628 = 0x00000000;
+
+u32 D_8003F62C = 0x00000001;
+
+u32 D_8003F630 = 0x00000001;
+
+u32 D_8003F634 = 0x00000000;
+
+u32 D_8003F638 = 0x00000000;
+
+u32 D_8003F63C = 0x00000000;
+
+u32 D_8003F640 = 0x00000000;
+
+u32 D_8003F644 = 0x00000000;
+
+u32 D_8003F648 = 0x00000000;
+
+u32 D_8003F64C = 0x00000001;
+
+u32 D_8003F650 = 0x00000000;
+
+u32 D_8003F654 = 0x00000000;
+
+u32 D_8003F658 = 0x00000000;
+
+u32 D_8003F65C = 0x00000000;
+
+f32 D_8003F660 = 1.0f;
+
+f32 D_8003F664 = 10.0f;
+
+f32 D_8003F668 = 100.0f;
+
+f32 D_8003F66C = 1000.0f;
+
+f32 D_8003F670 = 10000.0f;
+
+f32 D_8003F674 = 100000.0f;
+
+f32 D_8003F678 = 1000000.0f;
+
+f32 D_8003F67C = 10000000.0f;
+
+f32 D_8003F680 = 100000000.0f;
+
+f32 D_8003F684 = 1000000000.0f;
+
+u32 D_8003F688 = 0x00000000;
+
+u32 D_8003F68C = 0xFFFFFFFF;
+
+u32 D_8003F690 = 0xFFFFFFFF;
+
+u32 D_8003F694 = 0xFFFFFFFF;
+
+u32 D_8003F698 = 0xFFFFFFFF;
+
+u32 D_8003F69C = 0xFFFFFFFF;
+
+u32 D_8003F6A0 = 0xFFFFFFFF;
+
+u32 D_8003F6A4 = 0xFFFFFFFF;
+
+u32 D_8003F6A8 = 0xFFFFFFFF;
+
+u32 D_8003F6AC = 0xFF29FFFF;
+
+u32 D_8003F6B0 = 0xFF2BFFFF;
+
+u32 D_8003F6B4 = 0x2526FF2A;
+
+u32 D_8003F6B8 = 0xFF272CFF;
+
+u32 D_8003F6BC = 0x00010203;
+
+u32 D_8003F6C0 = 0x04050607;
+
+u32 D_8003F6C4 = 0x080924FF;
+
+u32 D_8003F6C8 = 0xFFFFFF28;
+
+u32 D_8003F6CC = 0xFF0A0B0C;
+
+u32 D_8003F6D0 = 0x0D0E0F10;
+
+u32 D_8003F6D4 = 0x11121314;
+
+u32 D_8003F6D8 = 0x15161718;
+
+u32 D_8003F6DC = 0x191A1B1C;
+
+u32 D_8003F6E0 = 0x1D1E1F20;
+
+u32 D_8003F6E4 = 0x212223FF;
+
+u32 D_8003F6E8 = 0xFFFFFFFF;
+
+u32 D_8003F6EC = 0xFF0A0B0C;
+
+u32 D_8003F6F0 = 0x0D0E0F10;
+
+u32 D_8003F6F4 = 0x11121314;
+
+u32 D_8003F6F8 = 0x15161718;
+
+u32 D_8003F6FC = 0x191A1B1C;
+
+u32 D_8003F700 = 0x1D1E1F20;
+
+u32 D_8003F704 = 0x212223FF;
+
+u32 D_8003F708 = 0xFFFFFFFF;
+
+u32 D_8003F70C = 0x70871C30;
+
+u32 D_8003F710 = 0x8988A250;
+
+u32 D_8003F714 = 0x88808290;
+
+u32 D_8003F718 = 0x88831C90;
+
+u32 D_8003F71C = 0x888402F8;
+
+u32 D_8003F720 = 0x88882210;
+
+u32 D_8003F724 = 0x71CF9C10;
+
+u32 D_8003F728 = 0xF9CF9C70;
+
+u32 D_8003F72C = 0x8228A288;
+
+u32 D_8003F730 = 0xF200A288;
+
+u32 D_8003F734 = 0x0BC11C78;
+
+u32 D_8003F738 = 0x0A222208;
+
+u32 D_8003F73C = 0x8A222288;
+
+u32 D_8003F740 = 0x71C21C70;
+
+u32 D_8003F744 = 0x23C738F8;
+
+u32 D_8003F748 = 0x5228A480;
+
+u32 D_8003F74C = 0x8A282280;
+
+u32 D_8003F750 = 0x8BC822F0;
+
+u32 D_8003F754 = 0xFA282280;
+
+u32 D_8003F758 = 0x8A28A480;
+
+u32 D_8003F75C = 0x8BC738F8;
+
+u32 D_8003F760 = 0xF9C89C08;
+
+u32 D_8003F764 = 0x82288808;
+
+u32 D_8003F768 = 0x82088808;
+
+u32 D_8003F76C = 0xF2EF8808;
+
+u32 D_8003F770 = 0x82288888;
+
+u32 D_8003F774 = 0x82288888;
+
+u32 D_8003F778 = 0x81C89C70;
+
+u32 D_8003F77C = 0x8A08A270;
+
+u32 D_8003F780 = 0x920DA288;
+
+u32 D_8003F784 = 0xA20AB288;
+
+u32 D_8003F788 = 0xC20AAA88;
+
+u32 D_8003F78C = 0xA208A688;
+
+u32 D_8003F790 = 0x9208A288;
+
+u32 D_8003F794 = 0x8BE8A270;
+
+u32 D_8003F798 = 0xF1CF1CF8;
+
+u32 D_8003F79C = 0x8A28A220;
+
+u32 D_8003F7A0 = 0x8A28A020;
+
+u32 D_8003F7A4 = 0xF22F1C20;
+
+u32 D_8003F7A8 = 0x82AA0220;
+
+u32 D_8003F7AC = 0x82492220;
+
+u32 D_8003F7B0 = 0x81A89C20;
+
+u32 D_8003F7B4 = 0x8A28A288;
+
+u32 D_8003F7B8 = 0x8A28A288;
+
+u32 D_8003F7BC = 0x8A289488;
+
+u32 D_8003F7C0 = 0x8A2A8850;
+
+u32 D_8003F7C4 = 0x894A9420;
+
+u32 D_8003F7C8 = 0x894AA220;
+
+u32 D_8003F7CC = 0x70852220;
+
+u32 D_8003F7D0 = 0xF8011000;
+
+u32 D_8003F7D4 = 0x08020800;
+
+u32 D_8003F7D8 = 0x10840400;
+
+u32 D_8003F7DC = 0x20040470;
+
+u32 D_8003F7E0 = 0x40840400;
+
+u32 D_8003F7E4 = 0x80020800;
+
+u32 D_8003F7E8 = 0xF8011000;
+
+u32 D_8003F7EC = 0x70800000;
+
+u32 D_8003F7F0 = 0x88822200;
+
+u32 D_8003F7F4 = 0x08820400;
+
+u32 D_8003F7F8 = 0x108F8800;
+
+u32 D_8003F7FC = 0x20821000;
+
+u32 D_8003F800 = 0x00022200;
+
+u32 D_8003F804 = 0x20800020;
+
+u32 D_8003F808 = 0x00000000;
+
+u32 D_8003F80C = 0x80040D10;
+
+u32 D_8003F810 = 0x80040D1C;
+
+u32 D_8003F814 = 0x80040D30;
+
+u32 D_8003F818 = 0x80040D48;
+
+u32 D_8003F81C = 0x80040D60;
+
+u32 D_8003F820 = 0x80040D78;
+
+u32 D_8003F824 = 0x80040D90;
+
+u32 D_8003F828 = 0x80040DA4;
+
+u32 D_8003F82C = 0x80040DB8;
+
+u32 D_8003F830 = 0x80040DD0;
+
+u32 D_8003F834 = 0x80040DE8;
+
+u32 D_8003F838 = 0x80040E00;
+
+u32 D_8003F83C = 0x80040E18;
+
+u32 D_8003F840 = 0x80040E2C;
+
+u32 D_8003F844 = 0x80040E3C;
+
+u32 D_8003F848 = 0x80040E58;
+
+u32 D_8003F84C = 0x80040E74;
+
+u32 D_8003F850 = 0x80040E8C;
+
+char *D_8003F854[] = {
+	0x80040EA8,
+	0x80040EC0,
+	0x80040ED4,
+	0x80040EE8,
+	0x80040EF4,
+	0x80040F00,
+};
+
+u32 D_8003F86C = 0x0000001E;
+
+u32 D_8003F870 = 0x00000019;
+
+u32 D_8003F874 = 0x00000000;
+
+u32 D_8003F878 = 0x00000000;
+
+u32 D_8003F87C = 0x00000000;
+
+u32 D_8003F880 = 0x00000003;
+
+u32 D_8003F884 = 0x00000FA0;
+
+u32 D_8003F888 = 0x00000000;
+
+u32 D_8003F88C = 0x00000870;
+
+u32 D_8003F890 = 0x00002666;
+
+u32 D_8003F894 = 0xFFFFD99A;
+
+u32 D_8003F898 = 0x00000000;
+
+u32 D_8003F89C = 0x00000000;
+
+u32 D_8003F8A0 = 0x00000000;
+
+u32 D_8003F8A4 = 0x00000000;
+
+u32 D_8003F8A8 = 0x000002F8;
+
+u32 D_8003F8AC = 0x000005F0;
+
+u32 D_8003F8B0 = 0x00000CCC;
+
+u32 D_8003F8B4 = 0xFFFFF334;
+
+u32 D_8003F8B8 = 0x00003FFF;
+
+u32 D_8003F8BC = 0x00000000;
+
+u32 D_8003F8C0 = 0x00000000;
+
+u32 D_8003F8C4 = 0x00000000;
+
+u32 D_8003F8C8 = 0x00000000;
+
+u32 D_8003F8CC = 0x00000960;
+
+u32 D_8003F8D0 = 0x00001388;
+
+u32 D_8003F8D4 = 0x00000000;
+
+u32 D_8003F8D8 = 0x00000000;
+
+u32 D_8003F8DC = 0x00000000;
+
+u32 D_8003F8E0 = 0x00000000;
+
+u32 D_8003F8E4 = 0x00005000;
+
+u32 D_8003F8E8 = 0x00000004;
+
+u32 D_8003F8EC = 0x00000FA0;
+
+u32 D_8003F8F0 = 0x00000000;
+
+u32 D_8003F8F4 = 0x00000A50;
+
+u32 D_8003F8F8 = 0x00002666;
+
+u32 D_8003F8FC = 0xFFFFD99A;
+
+u32 D_8003F900 = 0x00000000;
+
+u32 D_8003F904 = 0x00000000;
+
+u32 D_8003F908 = 0x00000000;
+
+u32 D_8003F90C = 0x00000000;
+
+u32 D_8003F910 = 0x00000370;
+
+u32 D_8003F914 = 0x00000870;
+
+u32 D_8003F918 = 0x00000CCC;
+
+u32 D_8003F91C = 0xFFFFF334;
+
+u32 D_8003F920 = 0x00003FFF;
+
+u32 D_8003F924 = 0x00000000;
+
+u32 D_8003F928 = 0x00000000;
+
+u32 D_8003F92C = 0x00000000;
+
+u32 D_8003F930 = 0x00000A50;
+
+u32 D_8003F934 = 0x00000E38;
+
+u32 D_8003F938 = 0x00000CCC;
+
+u32 D_8003F93C = 0xFFFFF334;
+
+u32 D_8003F940 = 0x00003FFF;
+
+u32 D_8003F944 = 0x00000000;
+
+u32 D_8003F948 = 0x00000000;
+
+u32 D_8003F94C = 0x00000000;
+
+u32 D_8003F950 = 0x00000000;
+
+u32 D_8003F954 = 0x00000EB0;
+
+u32 D_8003F958 = 0x00001F40;
+
+u32 D_8003F95C = 0x00000000;
+
+u32 D_8003F960 = 0x00000000;
+
+u32 D_8003F964 = 0x00000000;
+
+u32 D_8003F968 = 0x00000000;
+
+u32 D_8003F96C = 0x00005000;
+
+u32 D_8003F970 = 0x00000001;
+
+u32 D_8003F974 = 0x00001F40;
+
+u32 D_8003F978 = 0x00000000;
+
+u32 D_8003F97C = 0x00001BF8;
+
+u32 D_8003F980 = 0x00002EE0;
+
+u32 D_8003F984 = 0x00000000;
+
+u32 D_8003F988 = 0x00007FFF;
+
+u32 D_8003F98C = 0x00000000;
+
+u32 D_8003F990 = 0x00000000;
+
+u32 D_8003F994 = 0x00000000;
+
+u32 D_8003F998 = 0x00000001;
+
+u32 D_8003F99C = 0x00000320;
+
+u32 D_8003F9A0 = 0x00000000;
+
+u32 D_8003F9A4 = 0x000000C8;
+
+u32 D_8003F9A8 = 0x00004000;
+
+u32 D_8003F9AC = 0x00000000;
+
+u32 D_8003F9B0 = 0x00007FFF;
+
+u32 D_8003F9B4 = 0x00001DB0;
+
+u32 D_8003F9B8 = 0x000002BC;
+
+u32 D_8003F9BC = 0x00000000;
+
+u32 D_8003F9C0 = 0x00000001;
+
+u32 D_8003F9C4 = 0x00000320;
+
+u32 D_8003F9C8 = 0x00000000;
+
+u32 D_8003F9CC = 0x000000C8;
+
+u32 D_8003F9D0 = 0x00000000;
+
+u32 D_8003F9D4 = 0x00005FFF;
+
+u32 D_8003F9D8 = 0x00007FFF;
+
+u32 D_8003F9DC = 0x0000017C;
+
+u32 D_8003F9E0 = 0x000001F4;
+
+u32 D_8003F9E4 = 0x00000000;
+
+u32 D_8003F9E8 = 0x00000000;
+
+u32 D_8003F9EC = 0x00000000;
+
+u32 D_8003F9F0 = 0x00000000;
+
+u32 D_8003F9F4 = 0x00000000;
+
+u32 D_8003F9F8 = 0x00000000;
+
+u32 D_8003F9FC = 0x00000000;
+
+u32 D_8003FA00 = 0x00000000;
+
+u32 D_8003FA04 = 0x00000000;
+
+u32 D_8003FA08 = 0x00000000;
+
+u32 D_8003FA0C = 0x00000000;
+
+u32 D_8003FA10 = 0x7FFF7FFC;
+
+u32 D_8003FA14 = 0x7FF57FE8;
+
+u32 D_8003FA18 = 0x7FD77FC0;
+
+u32 D_8003FA1C = 0x7FA57F84;
+
+u32 D_8003FA20 = 0x7F5F7F34;
+
+u32 D_8003FA24 = 0x7F057ED0;
+
+u32 D_8003FA28 = 0x7E977E58;
+
+u32 D_8003FA2C = 0x7E157DCD;
+
+u32 D_8003FA30 = 0x7D7F7D2D;
+
+u32 D_8003FA34 = 0x7CD67C7A;
+
+u32 D_8003FA38 = 0x7C1A7BB4;
+
+u32 D_8003FA3C = 0x7B497ADA;
+
+u32 D_8003FA40 = 0x7A6679ED;
+
+u32 D_8003FA44 = 0x796F78ED;
+
+u32 D_8003FA48 = 0x786677DA;
+
+u32 D_8003FA4C = 0x774976B4;
+
+u32 D_8003FA50 = 0x761A757B;
+
+u32 D_8003FA54 = 0x74D87430;
+
+u32 D_8003FA58 = 0x738472D3;
+
+u32 D_8003FA5C = 0x721E7164;
+
+u32 D_8003FA60 = 0x70A66FE3;
+
+u32 D_8003FA64 = 0x6F1C6E51;
+
+u32 D_8003FA68 = 0x6D816CAD;
+
+u32 D_8003FA6C = 0x6BD56AF9;
+
+u32 D_8003FA70 = 0x6A186933;
+
+u32 D_8003FA74 = 0x684A675D;
+
+u32 D_8003FA78 = 0x666C6577;
+
+u32 D_8003FA7C = 0x647E6381;
+
+u32 D_8003FA80 = 0x6280617C;
+
+u32 D_8003FA84 = 0x60735F67;
+
+u32 D_8003FA88 = 0x5E575D43;
+
+u32 D_8003FA8C = 0x5C2C5B11;
+
+u32 D_8003FA90 = 0x59F258D0;
+
+u32 D_8003FA94 = 0x57AA5681;
+
+u32 D_8003FA98 = 0x55555425;
+
+u32 D_8003FA9C = 0x52F251BC;
+
+u32 D_8003FAA0 = 0x50824F46;
+
+u32 D_8003FAA4 = 0x4E064CC3;
+
+u32 D_8003FAA8 = 0x4B7D4A35;
+
+u32 D_8003FAAC = 0x48E9479B;
+
+u32 D_8003FAB0 = 0x464944F5;
+
+u32 D_8003FAB4 = 0x439E4245;
+
+u32 D_8003FAB8 = 0x40E93F8A;
+
+u32 D_8003FABC = 0x3E293CC6;
+
+u32 D_8003FAC0 = 0x3B6039F8;
+
+u32 D_8003FAC4 = 0x388D3721;
+
+u32 D_8003FAC8 = 0x35B23441;
+
+u32 D_8003FACC = 0x32CE3159;
+
+u32 D_8003FAD0 = 0x2FE22E69;
+
+u32 D_8003FAD4 = 0x2CEF2B72;
+
+u32 D_8003FAD8 = 0x29F42875;
+
+u32 D_8003FADC = 0x26F32570;
+
+u32 D_8003FAE0 = 0x23EC2266;
+
+u32 D_8003FAE4 = 0x20DF1F57;
+
+u32 D_8003FAE8 = 0x1DCE1C43;
+
+u32 D_8003FAEC = 0x1AB7192A;
+
+u32 D_8003FAF0 = 0x179C160E;
+
+u32 D_8003FAF4 = 0x147E12ED;
+
+u32 D_8003FAF8 = 0x115C0FCA;
+
+u32 D_8003FAFC = 0x0E380CA5;
+
+u32 D_8003FB00 = 0x0B11097D;
+
+u32 D_8003FB04 = 0x07E90654;
+
+u32 D_8003FB08 = 0x04C0032A;
+
+u32 D_8003FB0C = 0x01950000;
+
+u32 D_8003FB10 = 0x00000000;
+
+u32 D_8003FB14 = 0x00000000;
+
+f32 D_8003FB18 = 1.0f;
+
+u32 D_8003FB1C = 0xFFFF0000;
+
+u32 D_8003FB20 = 0x00000001;
+
+u32 D_8003FB24 = 0x00000001;
+
+u32 D_8003FB28 = 0x00000000;
+
+u32 D_8003FB2C = 0x00000000;
+
+u32 __osHwIntTable = 0x00000000;
+
+u32 D_8003FB34 = 0x00000000;
+
+u32 D_8003FB38 = 0x00000000;
+
+u32 D_8003FB3C = 0x00000000;
+
+u32 D_8003FB40 = 0x00000000;
+
+u32 D_8003FB44 = 0x00000000;
+
+u32 D_8003FB48 = 0x00000000;
+
+u32 D_8003FB4C = 0x00000000;
+
+u32 __osThreadTail = 0x00000000;
+
+u32 D_8003FB54 = 0xFFFFFFFF;
+
+u32 __osRunQueue = 0x8003FB50;
+
+u32 __osActiveQueue = 0x8003FB50;
+
+u32 __osRunningThread = 0x00000000;
+
+u32 __osFaultedThread = 0x00000000;
+
+u32 D_8003FB68 = 0x00000000;
+
+u32 D_8003FB6C = 0x00000000;
+
+u32 D_8003FB70 = 0x00000000;
+
+u32 D_8003FB74 = 0x00000000;
+
+u32 D_8003FB78 = 0x00000000;
+
+u32 D_8003FB7C = 0x00000000;
+
+u32 D_8003FB80 = 0x00000000;
+
+u32 D_8003FB84 = 0x00000000;
+
+u32 D_8003FB88 = 0x00000000;
+
+u32 D_8003FB8C = 0x00000000;
+
+u64 osClockRate = 0x0000000003B9ACA0;
+
+u32 osViClock = 0x02E6D354;
+
+u32 __osShutdown = 0x00000000;
+
+u32 __OSGlobalIntMask = 0x003FFF01;
+
+u32 D_8003FBA4 = 0x00000000;
+
+u32 D_8003FBA8 = 0x00000000;
+
+u32 D_8003FBAC = 0x00000000;
+
+u32 D_8003FBB0 = 0xFFFFFFFF;
+
+u32 D_8003FBB4 = 0xFA000000;
+
+u32 D_8003FBB8 = 0x00000000;
+
+u32 D_8003FBBC = 0x00000000;
+
+OSDevMgr __osPiDevMgr = {0};
+u32 __osPiTable = 0x00000000;
+
+u32 __osCurrentHandle = 0x80098E00;
+
+u32 D_8003FBE4 = 0x80098E78;
+
+u32 D_8003FBE8 = 0x00000000;
+
+u32 D_8003FBEC = 0x00000000;
+
+u32 D_8003FBF0 = 0x00000000;
+
+u32 D_8003FBF4 = 0x00000000;
+
+u32 D_8003FBF8 = 0x00000000;
+
+u32 D_8003FBFC = 0x00000000;
+
+u32 D_8003FC00 = 0x20202020;
+
+u32 D_8003FC04 = 0x20202020;
+
+u32 D_8003FC08 = 0x20202020;
+
+u32 D_8003FC0C = 0x20202020;
+
+u32 D_8003FC10 = 0x20202020;
+
+u32 D_8003FC14 = 0x20202020;
+
+u32 D_8003FC18 = 0x20202020;
+
+u32 D_8003FC1C = 0x20202020;
+
+u32 D_8003FC20 = 0x00000000;
+
+u32 D_8003FC24 = 0x30303030;
+
+u32 D_8003FC28 = 0x30303030;
+
+u32 D_8003FC2C = 0x30303030;
+
+u32 D_8003FC30 = 0x30303030;
+
+u32 D_8003FC34 = 0x30303030;
+
+u32 D_8003FC38 = 0x30303030;
+
+u32 D_8003FC3C = 0x30303030;
+
+u32 D_8003FC40 = 0x30303030;
+
+u32 D_8003FC44 = 0x00000000;
+
+u32 D_8003FC48 = 0x00000000;
+
+u32 D_8003FC4C = 0x00000000;
+
