+++ data/kirby.039DF0.3.c
@@ -350,1492 +350,3 @@ u32 D_80041FE4 = 0x66960B39;
 u32 D_80041FE8 = 0xFFDF0D46;
 
 u32 D_80041FEC = 0x66AD0C39;
-
-u32 D_80041FF0 = 0x00000000;
-
-u32 D_80041FF4 = 0x00000000;
-
-u32 D_80041FF8 = 0x00000000;
-
-u32 D_80041FFC = 0x00000000;
-
-u32 D_80042000 = 0x00000000;
-
-u32 D_80042004 = 0x00000000;
-
-u32 D_80042008 = 0x00000000;
-
-u32 D_8004200C = 0x00000000;
-
-u32 D_80042010 = 0x00000000;
-
-u32 D_80042014 = 0x00000000;
-
-u32 D_80042018 = 0x00000000;
-
-u32 D_8004201C = 0x00000000;
-
-u32 D_80042020 = 0x00000000;
-
-u32 D_80042024 = 0x00000000;
-
-u32 D_80042028 = 0x00000000;
-
-u32 D_8004202C = 0x00000000;
-
-u32 D_80042030 = 0x00000000;
-
-u32 D_80042034 = 0x00000000;
-
-u32 D_80042038 = 0x00000000;
-
-u32 D_8004203C = 0x00000000;
-
-u32 D_80042040 = 0x00000000;
-
-u32 D_80042044 = 0x00000000;
-
-u32 D_80042048 = 0x00000000;
-
-u32 D_8004204C = 0x00000000;
-
-u32 D_80042050 = 0x00000000;
-
-u32 D_80042054 = 0x00000000;
-
-u32 D_80042058 = 0x00000000;
-
-u32 D_8004205C = 0x00000000;
-
-u32 D_80042060 = 0x00000000;
-
-u32 D_80042064 = 0x00000000;
-
-u32 D_80042068 = 0x00000000;
-
-u32 D_8004206C = 0x00000000;
-
-u32 D_80042070 = 0x00000000;
-
-u32 D_80042074 = 0x00000000;
-
-u32 D_80042078 = 0x00000000;
-
-u32 D_8004207C = 0x00000000;
-
-u32 D_80042080 = 0x00000000;
-
-u32 D_80042084 = 0x00000000;
-
-u32 D_80042088 = 0x00000000;
-
-u32 D_8004208C = 0x00000000;
-
-u32 D_80042090 = 0x00000000;
-
-u32 D_80042094 = 0x00000000;
-
-u32 D_80042098 = 0x00000000;
-
-u32 D_8004209C = 0x00000000;
-
-u32 D_800420A0 = 0x00000000;
-
-u32 D_800420A4 = 0x00000000;
-
-u32 D_800420A8 = 0x00000000;
-
-u32 D_800420AC = 0x00000000;
-
-u32 D_800420B0 = 0xED000000;
-
-u32 D_800420B4 = 0x005003C0;
-
-u32 D_800420B8 = 0xEF080CFF;
-
-u32 D_800420BC = 0x00000000;
-
-u32 D_800420C0 = 0x00000000;
-
-u32 D_800420C4 = 0x00000000;
-
-u32 D_800420C8 = 0x00000000;
-
-u32 D_800420CC = 0xFFFF0048;
-
-u32 D_800420D0 = 0x00000000;
-
-u32 D_800420D4 = 0x00000000;
-
-u32 D_800420D8 = 0x00000000;
-
-u32 D_800420DC = 0x00000000;
-
-u32 D_800420E0 = 0x00000000;
-
-u32 D_800420E4 = 0x00000000;
-
-u32 D_800420E8 = 0x00000000;
-
-u32 D_800420EC = 0x00000000;
-
-u32 D_800420F0 = 0x00000000;
-
-u32 D_800420F4 = 0x00000000;
-
-u32 D_800420F8 = 0x00000000;
-
-u32 D_800420FC = 0x00000000;
-
-u32 D_80042100 = 0x00000000;
-
-u32 D_80042104 = 0x00000000;
-
-u32 D_80042108 = 0x00000000;
-
-u32 D_8004210C = 0x00000000;
-
-u32 D_80042110 = 0x00000000;
-
-u32 D_80042114 = 0x00000000;
-
-u32 D_80042118 = 0x00000000;
-
-u32 D_8004211C = 0x00000000;
-
-u32 D_80042120 = 0x00000000;
-
-u32 D_80042124 = 0x00000000;
-
-u32 D_80042128 = 0x0A525350;
-
-u32 D_8004212C = 0x20476678;
-
-u32 D_80042130 = 0x2075636F;
-
-u32 D_80042134 = 0x64652046;
-
-u32 D_80042138 = 0x33444558;
-
-u32 D_8004213C = 0x20202020;
-
-u32 D_80042140 = 0x20202066;
-
-u32 D_80042144 = 0x69666F20;
-
-u32 D_80042148 = 0x322E3034;
-
-u32 D_8004214C = 0x4820596F;
-
-u32 D_80042150 = 0x73686974;
-
-u32 D_80042154 = 0x616B6120;
-
-u32 D_80042158 = 0x59617375;
-
-u32 D_8004215C = 0x6D6F746F;
-
-u32 D_80042160 = 0x20313939;
-
-u32 D_80042164 = 0x38204E69;
-
-u32 D_80042168 = 0x6E74656E;
-
-u32 D_8004216C = 0x646F2E0A;
-
-u32 D_80042170 = 0x00010000;
-
-u32 D_80042174 = 0x00000002;
-
-u32 D_80042178 = 0x00000001;
-
-u32 D_8004217C = 0x00000002;
-
-u32 D_80042180 = 0x00010000;
-
-u32 D_80042184 = 0x0000FFFE;
-
-u32 D_80042188 = 0x00000001;
-
-u32 D_8004218C = 0x0000FFFE;
-
-u32 D_80042190 = 0x00000000;
-
-u32 D_80042194 = 0x0001FFFF;
-
-u32 D_80042198 = 0x00000000;
-
-u32 D_8004219C = 0x00010001;
-
-u32 D_800421A0 = 0xFFFF0004;
-
-u32 D_800421A4 = 0x00087F00;
-
-u32 D_800421A8 = 0xFFFC4000;
-
-u32 D_800421AC = 0x04207FFF;
-
-u32 D_800421B0 = 0x7FFC1400;
-
-u32 D_800421B4 = 0x01CC0200;
-
-u32 D_800421B8 = 0xFFF00010;
-
-u32 D_800421BC = 0x00200100;
-
-u32 D_800421C0 = 0xC00044D3;
-
-u32 D_800421C4 = 0x6CB30002;
-
-u32 D_800421C8 = 0x00010000;
-
-u32 D_800421CC = 0x01000BA8;
-
-u32 D_800421D0 = 0x00000000;
-
-u32 D_800421D4 = 0x00000000;
-
-u32 D_800421D8 = 0x00000000;
-
-u32 D_800421DC = 0x00800000;
-
-u32 D_800421E0 = 0x00000000;
-
-u32 D_800421E4 = 0x00000000;
-
-u32 D_800421E8 = 0x00000000;
-
-u32 D_800421EC = 0x00000000;
-
-u32 D_800421F0 = 0x00000000;
-
-u32 D_800421F4 = 0x00000000;
-
-u32 D_800421F8 = 0x00000000;
-
-u32 D_800421FC = 0x00000000;
-
-u32 D_80042200 = 0x00000000;
-
-u32 D_80042204 = 0x00000000;
-
-u32 D_80042208 = 0x00000000;
-
-u32 D_8004220C = 0x00000000;
-
-u32 D_80042210 = 0x00000000;
-
-u32 D_80042214 = 0x00000000;
-
-u32 D_80042218 = 0x00000000;
-
-u32 D_8004221C = 0x00000000;
-
-u32 D_80042220 = 0x00000000;
-
-u32 D_80042224 = 0x00000000;
-
-u32 D_80042228 = 0x00000000;
-
-u32 D_8004222C = 0x00000000;
-
-u32 D_80042230 = 0x00000000;
-
-u32 D_80042234 = 0x00000000;
-
-u32 D_80042238 = 0x00000000;
-
-u32 D_8004223C = 0x00000000;
-
-u32 D_80042240 = 0x00000000;
-
-u32 D_80042244 = 0x00000000;
-
-u32 D_80042248 = 0x00000000;
-
-u32 D_8004224C = 0x00000000;
-
-u32 D_80042250 = 0x00000000;
-
-u32 D_80042254 = 0x00000000;
-
-u32 D_80042258 = 0x00000000;
-
-u32 D_8004225C = 0x00000000;
-
-u32 D_80042260 = 0x00000000;
-
-u32 D_80042264 = 0x00000000;
-
-u32 D_80042268 = 0x00000000;
-
-u32 D_8004226C = 0x00000000;
-
-u32 D_80042270 = 0x00000000;
-
-u32 D_80042274 = 0x00000000;
-
-u32 D_80042278 = 0x00000000;
-
-u32 D_8004227C = 0x00000000;
-
-u32 D_80042280 = 0x00000000;
-
-u32 D_80042284 = 0x00000000;
-
-u32 D_80042288 = 0x00000000;
-
-u32 D_8004228C = 0x00000000;
-
-u32 D_80042290 = 0x00000000;
-
-u32 D_80042294 = 0x00000000;
-
-u32 D_80042298 = 0x00000000;
-
-u32 D_8004229C = 0x00000000;
-
-u32 D_800422A0 = 0x00000000;
-
-u32 D_800422A4 = 0x00000000;
-
-u32 D_800422A8 = 0x00000000;
-
-u32 D_800422AC = 0x00000000;
-
-u32 D_800422B0 = 0x00000000;
-
-u32 D_800422B4 = 0x00000000;
-
-u32 D_800422B8 = 0x00000000;
-
-u32 D_800422BC = 0x00000000;
-
-u32 D_800422C0 = 0x00000000;
-
-u32 D_800422C4 = 0x00000000;
-
-u32 D_800422C8 = 0x00000000;
-
-u32 D_800422CC = 0x00000000;
-
-u32 D_800422D0 = 0x00000F80;
-
-u32 D_800422D4 = 0x00971000;
-
-u32 D_800422D8 = 0x00001018;
-
-u32 D_800422DC = 0x016F1000;
-
-u32 D_800422E0 = 0x09D00000;
-
-u32 D_800422E4 = 0x09D00040;
-
-u32 D_800422E8 = 0x00E001F0;
-
-u32 D_800422EC = 0x04200080;
-
-u32 D_800422F0 = 0x01DA0180;
-
-u32 D_800422F4 = 0x00F801E0;
-
-u32 D_800422F8 = 0x022001D8;
-
-u32 D_800422FC = 0x00DA1188;
-
-u32 D_80042300 = 0x10781188;
-
-u32 D_80042304 = 0x118C118C;
-
-u32 D_80042308 = 0x11B811D0;
-
-u32 D_8004230C = 0x1028104C;
-
-u32 D_80042310 = 0x11E810F0;
-
-u32 D_80042314 = 0x10381120;
-
-u32 D_80042318 = 0x1FAC1000;
-
-u32 D_8004231C = 0x11FC118C;
-
-u32 D_80042320 = 0x10301140;
-
-u32 D_80042324 = 0x113C102C;
-
-u32 D_80042328 = 0x102C1224;
-
-u32 D_8004232C = 0x12241224;
-
-u32 D_80042330 = 0x12241220;
-
-u32 D_80042334 = 0x12201220;
-
-u32 D_80042338 = 0x125C1220;
-
-u32 D_8004233C = 0x12501220;
-
-u32 D_80042340 = 0x12101220;
-
-u32 D_80042344 = 0x12201220;
-
-u32 D_80042348 = 0x12201220;
-
-u32 D_8004234C = 0x12201220;
-
-u32 D_80042350 = 0x12201220;
-
-u32 D_80042354 = 0x12201220;
-
-u32 D_80042358 = 0x12301230;
-
-u32 D_8004235C = 0x12301224;
-
-u32 D_80042360 = 0x15001C6C;
-
-u32 D_80042364 = 0x1C2C1C54;
-
-u32 D_80042368 = 0x17AC17A4;
-
-u32 D_8004236C = 0x17A4118C;
-
-u32 D_80042370 = 0x04200448;
-
-u32 D_80042374 = 0x04700498;
-
-u32 D_80042378 = 0x04C004E8;
-
-u32 D_8004237C = 0x05100538;
-
-u32 D_80042380 = 0x05600588;
-
-u32 D_80042384 = 0x05B005D8;
-
-u32 D_80042388 = 0x06000628;
-
-u32 D_8004238C = 0x06500678;
-
-u32 D_80042390 = 0x06A006C8;
-
-u32 D_80042394 = 0x06F00718;
-
-u32 D_80042398 = 0x07400768;
-
-u32 D_8004239C = 0x079007B8;
-
-u32 D_800423A0 = 0x07E00808;
-
-u32 D_800423A4 = 0x08300858;
-
-u32 D_800423A8 = 0x088008A8;
-
-u32 D_800423AC = 0x08D008F8;
-
-u32 D_800423B0 = 0x0920FFFF;
-
-u32 D_800423B4 = 0x80000000;
-
-u32 D_800423B8 = 0x00008000;
-
-u32 D_800423BC = 0x30304040;
-
-u32 D_800423C0 = 0x00000000;
-
-u32 D_800423C4 = 0x00000000;
-
-u32 D_800423C8 = 0x00000000;
-
-u32 D_800423CC = 0x00000000;
-
-u32 D_800423D0 = 0x00000000;
-
-u32 D_800423D4 = 0x00000000;
-
-u32 D_800423D8 = 0x00000000;
-
-u32 D_800423DC = 0x00000000;
-
-u32 D_800423E0 = 0x00000000;
-
-u32 D_800423E4 = 0x00000000;
-
-u32 D_800423E8 = 0x00100000;
-
-u32 D_800423EC = 0x00200000;
-
-u32 D_800423F0 = 0x10000000;
-
-u32 D_800423F4 = 0x20000000;
-
-u32 D_800423F8 = 0x00004000;
-
-u32 D_800423FC = 0x00000040;
-
-u32 D_80042400 = 0x00001188;
-
-u32 D_80042404 = 0x020712E8;
-
-u32 D_80042408 = 0x00000268;
-
-u32 D_8004240C = 0x021712E8;
-
-u32 D_80042410 = 0x00000000;
-
-u32 D_80042414 = 0x00000000;
-
-u32 D_80042418 = 0x00000000;
-
-u32 D_8004241C = 0x00000000;
-
-u32 D_80042420 = 0x00000000;
-
-u32 D_80042424 = 0x00000000;
-
-u32 D_80042428 = 0x00000000;
-
-u32 D_8004242C = 0x00000000;
-
-u32 D_80042430 = 0x00000000;
-
-u32 D_80042434 = 0x00000000;
-
-u32 D_80042438 = 0x00000000;
-
-u32 D_8004243C = 0x00000000;
-
-u32 D_80042440 = 0x00000000;
-
-u32 D_80042444 = 0x00000000;
-
-u32 D_80042448 = 0x00000000;
-
-u32 D_8004244C = 0x00000000;
-
-u32 D_80042450 = 0x00000000;
-
-u32 D_80042454 = 0x00000000;
-
-u32 D_80042458 = 0x00000000;
-
-u32 D_8004245C = 0x00000000;
-
-u32 D_80042460 = 0x00000000;
-
-u32 D_80042464 = 0x00000000;
-
-u32 D_80042468 = 0x00000000;
-
-u32 D_8004246C = 0x00000000;
-
-u32 D_80042470 = 0x00000000;
-
-u32 D_80042474 = 0x00000000;
-
-u32 D_80042478 = 0x00000000;
-
-u32 D_8004247C = 0x00000000;
-
-u32 D_80042480 = 0x00000000;
-
-u32 D_80042484 = 0x00000000;
-
-u32 D_80042488 = 0x00000000;
-
-u32 D_8004248C = 0x00000000;
-
-u32 D_80042490 = 0x00000000;
-
-u32 D_80042494 = 0x00000000;
-
-u32 D_80042498 = 0x00000000;
-
-u32 D_8004249C = 0x00000000;
-
-u32 D_800424A0 = 0x00000000;
-
-u32 D_800424A4 = 0x00000000;
-
-u32 D_800424A8 = 0x00000000;
-
-u32 D_800424AC = 0x00000000;
-
-u32 D_800424B0 = 0x00000000;
-
-u32 D_800424B4 = 0x00000000;
-
-u32 D_800424B8 = 0x00000000;
-
-u32 D_800424BC = 0x00000000;
-
-u32 D_800424C0 = 0x00000000;
-
-u32 D_800424C4 = 0x00000000;
-
-u32 D_800424C8 = 0x00000000;
-
-u32 D_800424CC = 0x00000000;
-
-u32 D_800424D0 = 0xED000000;
-
-u32 D_800424D4 = 0x005003C0;
-
-u32 D_800424D8 = 0xEF080CFF;
-
-u32 D_800424DC = 0x00000000;
-
-u32 D_800424E0 = 0x00000000;
-
-u32 D_800424E4 = 0x00000000;
-
-u32 D_800424E8 = 0x00000000;
-
-u32 D_800424EC = 0xFFFF0048;
-
-u32 D_800424F0 = 0x00000000;
-
-u32 D_800424F4 = 0x00000000;
-
-u32 D_800424F8 = 0x00000000;
-
-u32 D_800424FC = 0x00000000;
-
-u32 D_80042500 = 0x00000000;
-
-u32 D_80042504 = 0x00000000;
-
-u32 D_80042508 = 0x00000000;
-
-u32 D_8004250C = 0x00000000;
-
-u32 D_80042510 = 0x00000000;
-
-u32 D_80042514 = 0x00000000;
-
-u32 D_80042518 = 0x00000000;
-
-u32 D_8004251C = 0x00000000;
-
-u32 D_80042520 = 0x00000000;
-
-u32 D_80042524 = 0x00000000;
-
-u32 D_80042528 = 0x00000000;
-
-u32 D_8004252C = 0x00000000;
-
-u32 D_80042530 = 0x00000000;
-
-u32 D_80042534 = 0x00000000;
-
-u32 D_80042538 = 0x00000000;
-
-u32 D_8004253C = 0x00000000;
-
-u32 D_80042540 = 0x00000000;
-
-u32 D_80042544 = 0x00000000;
-
-u32 D_80042548 = 0x0A525350;
-
-u32 D_8004254C = 0x20476678;
-
-u32 D_80042550 = 0x2075636F;
-
-u32 D_80042554 = 0x6465204C;
-
-u32 D_80042558 = 0x33444558;
-
-u32 D_8004255C = 0x20202020;
-
-u32 D_80042560 = 0x20202066;
-
-u32 D_80042564 = 0x69666F20;
-
-u32 D_80042568 = 0x322E3034;
-
-u32 D_8004256C = 0x4820596F;
-
-u32 D_80042570 = 0x73686974;
-
-u32 D_80042574 = 0x616B6120;
-
-u32 D_80042578 = 0x59617375;
-
-u32 D_8004257C = 0x6D6F746F;
-
-u32 D_80042580 = 0x20313939;
-
-u32 D_80042584 = 0x38204E69;
-
-u32 D_80042588 = 0x6E74656E;
-
-u32 D_8004258C = 0x646F2E0A;
-
-u32 D_80042590 = 0x00010000;
-
-u32 D_80042594 = 0x00000002;
-
-u32 D_80042598 = 0x00000001;
-
-u32 D_8004259C = 0x00000002;
-
-u32 D_800425A0 = 0x00010000;
-
-u32 D_800425A4 = 0x0000FFFE;
-
-u32 D_800425A8 = 0x00000001;
-
-u32 D_800425AC = 0x0000FFFE;
-
-u32 D_800425B0 = 0x00000000;
-
-u32 D_800425B4 = 0x0001FFFF;
-
-u32 D_800425B8 = 0x00000000;
-
-u32 D_800425BC = 0x00010001;
-
-u32 D_800425C0 = 0xFFFF0004;
-
-u32 D_800425C4 = 0x00087F00;
-
-u32 D_800425C8 = 0xFFFC4000;
-
-u32 D_800425CC = 0x03E87FFF;
-
-u32 D_800425D0 = 0x7FFC1400;
-
-u32 D_800425D4 = 0x01CC0200;
-
-u32 D_800425D8 = 0xFFF00010;
-
-u32 D_800425DC = 0x00200100;
-
-u32 D_800425E0 = 0xC00044D3;
-
-u32 D_800425E4 = 0x6CB30002;
-
-u32 D_800425E8 = 0x00010000;
-
-u32 D_800425EC = 0x01000A80;
-
-u32 D_800425F0 = 0x00000000;
-
-u32 D_800425F4 = 0x00000000;
-
-u32 D_800425F8 = 0x00000000;
-
-u32 D_800425FC = 0x00800000;
-
-u32 D_80042600 = 0x00000000;
-
-u32 D_80042604 = 0x00000000;
-
-u32 D_80042608 = 0x00000000;
-
-u32 D_8004260C = 0x00000000;
-
-u32 D_80042610 = 0x00000000;
-
-u32 D_80042614 = 0x00000000;
-
-u32 D_80042618 = 0x00000000;
-
-u32 D_8004261C = 0x00000000;
-
-u32 D_80042620 = 0x00000000;
-
-u32 D_80042624 = 0x00000000;
-
-u32 D_80042628 = 0x00000000;
-
-u32 D_8004262C = 0x00000000;
-
-u32 D_80042630 = 0x00000000;
-
-u32 D_80042634 = 0x00000000;
-
-u32 D_80042638 = 0x00000000;
-
-u32 D_8004263C = 0x00000000;
-
-u32 D_80042640 = 0x00000000;
-
-u32 D_80042644 = 0x00000000;
-
-u32 D_80042648 = 0x00000000;
-
-u32 D_8004264C = 0x00000000;
-
-u32 D_80042650 = 0x00000000;
-
-u32 D_80042654 = 0x00000000;
-
-u32 D_80042658 = 0x00000000;
-
-u32 D_8004265C = 0x00000000;
-
-u32 D_80042660 = 0x00000000;
-
-u32 D_80042664 = 0x00000000;
-
-u32 D_80042668 = 0x00000000;
-
-u32 D_8004266C = 0x00000000;
-
-u32 D_80042670 = 0x00000000;
-
-u32 D_80042674 = 0x00000000;
-
-u32 D_80042678 = 0x00000000;
-
-u32 D_8004267C = 0x00000000;
-
-u32 D_80042680 = 0x00000000;
-
-u32 D_80042684 = 0x00000000;
-
-u32 D_80042688 = 0x00000000;
-
-u32 D_8004268C = 0x00000000;
-
-u32 D_80042690 = 0x00000000;
-
-u32 D_80042694 = 0x00000000;
-
-u32 D_80042698 = 0x00000000;
-
-u32 D_8004269C = 0x00000000;
-
-u32 D_800426A0 = 0x00000000;
-
-u32 D_800426A4 = 0x00000000;
-
-u32 D_800426A8 = 0x00000000;
-
-u32 D_800426AC = 0x00000000;
-
-u32 D_800426B0 = 0x00000000;
-
-u32 D_800426B4 = 0x00000000;
-
-u32 D_800426B8 = 0x00000000;
-
-u32 D_800426BC = 0x00000000;
-
-u32 D_800426C0 = 0x00000000;
-
-u32 D_800426C4 = 0x00000000;
-
-u32 D_800426C8 = 0x00000000;
-
-u32 D_800426CC = 0x00000000;
-
-u32 D_800426D0 = 0x00000000;
-
-u32 D_800426D4 = 0x00000000;
-
-u32 D_800426D8 = 0x00000000;
-
-u32 D_800426DC = 0x00000000;
-
-u32 D_800426E0 = 0x00000000;
-
-u32 D_800426E4 = 0x00000000;
-
-u32 D_800426E8 = 0x00000000;
-
-u32 D_800426EC = 0x00000000;
-
-u32 D_800426F0 = 0x00000F80;
-
-u32 D_800426F4 = 0x009F1000;
-
-u32 D_800426F8 = 0x00001020;
-
-u32 D_800426FC = 0x016F1000;
-
-u32 D_80042700 = 0x09900000;
-
-u32 D_80042704 = 0x09900040;
-
-u32 D_80042708 = 0x00E001F0;
-
-u32 D_8004270C = 0x03E80080;
-
-u32 D_80042710 = 0x01DA0180;
-
-u32 D_80042714 = 0x00F801E0;
-
-u32 D_80042718 = 0x022001D8;
-
-u32 D_8004271C = 0x00DA1188;
-
-u32 D_80042720 = 0x10781188;
-
-u32 D_80042724 = 0x118C118C;
-
-u32 D_80042728 = 0x11B811D0;
-
-u32 D_8004272C = 0x1028104C;
-
-u32 D_80042730 = 0x11E810F0;
-
-u32 D_80042734 = 0x10381120;
-
-u32 D_80042738 = 0x1FAC1000;
-
-u32 D_8004273C = 0x11FC118C;
-
-u32 D_80042740 = 0x10301140;
-
-u32 D_80042744 = 0x113C102C;
-
-u32 D_80042748 = 0x102C122C;
-
-u32 D_8004274C = 0x122C122C;
-
-u32 D_80042750 = 0x122C1228;
-
-u32 D_80042754 = 0x12281228;
-
-u32 D_80042758 = 0x12641228;
-
-u32 D_8004275C = 0x12581228;
-
-u32 D_80042760 = 0x12101228;
-
-u32 D_80042764 = 0x12281228;
-
-u32 D_80042768 = 0x12281228;
-
-u32 D_8004276C = 0x12281228;
-
-u32 D_80042770 = 0x12281228;
-
-u32 D_80042774 = 0x12281228;
-
-u32 D_80042778 = 0x12381238;
-
-u32 D_8004277C = 0x1238122C;
-
-u32 D_80042780 = 0x13201B60;
-
-u32 D_80042784 = 0x1B201B48;
-
-u32 D_80042788 = 0x15C015AC;
-
-u32 D_8004278C = 0x15AC1610;
-
-u32 D_80042790 = 0x03E80410;
-
-u32 D_80042794 = 0x04380460;
-
-u32 D_80042798 = 0x048804B0;
-
-u32 D_8004279C = 0x04D80500;
-
-u32 D_800427A0 = 0x05280550;
-
-u32 D_800427A4 = 0x057805A0;
-
-u32 D_800427A8 = 0x05C805F0;
-
-u32 D_800427AC = 0x06180640;
-
-u32 D_800427B0 = 0x06680690;
-
-u32 D_800427B4 = 0x06B806E0;
-
-u32 D_800427B8 = 0x07080730;
-
-u32 D_800427BC = 0x07580780;
-
-u32 D_800427C0 = 0x07A807D0;
-
-u32 D_800427C4 = 0x07F80820;
-
-u32 D_800427C8 = 0x08480870;
-
-u32 D_800427CC = 0x089808C0;
-
-u32 D_800427D0 = 0x08E80000;
-
-u32 D_800427D4 = 0x30304040;
-
-u32 D_800427D8 = 0x00100000;
-
-u32 D_800427DC = 0x00200000;
-
-u32 D_800427E0 = 0x10000000;
-
-u32 D_800427E4 = 0x20000000;
-
-u32 D_800427E8 = 0x00004000;
-
-u32 D_800427EC = 0x00000040;
-
-u32 D_800427F0 = 0x00000000;
-
-u32 D_800427F4 = 0x050003C0;
-
-u32 D_800427F8 = 0x00000000;
-
-u32 D_800427FC = 0x00000000;
-
-u32 D_80042800 = 0x00000000;
-
-u32 D_80042804 = 0x00000000;
-
-u32 D_80042808 = 0x00000000;
-
-u32 D_8004280C = 0x00000000;
-
-u32 D_80042810 = 0x00000000;
-
-u32 D_80042814 = 0x00000000;
-
-u32 D_80042818 = 0x00000000;
-
-u32 D_8004281C = 0x00000000;
-
-u32 D_80042820 = 0x00000000;
-
-u32 D_80042824 = 0x00000000;
-
-u32 D_80042828 = 0x00000000;
-
-u32 D_8004282C = 0x00000000;
-
-u32 D_80042830 = 0x00000000;
-
-u32 D_80042834 = 0x00000000;
-
-u32 D_80042838 = 0x00000000;
-
-u32 D_8004283C = 0x00000000;
-
-u32 D_80042840 = 0x00000000;
-
-u32 D_80042844 = 0x00000000;
-
-u32 D_80042848 = 0x00000000;
-
-u32 D_8004284C = 0x00000000;
-
-u32 D_80042850 = 0x00000000;
-
-u32 D_80042854 = 0x00000000;
-
-u32 D_80042858 = 0x00000000;
-
-u32 D_8004285C = 0x00000000;
-
-u32 D_80042860 = 0x00000000;
-
-u32 D_80042864 = 0x00000000;
-
-u32 D_80042868 = 0x00000000;
-
-u32 D_8004286C = 0x00000000;
-
-u32 D_80042870 = 0x00000000;
-
-u32 D_80042874 = 0x00000000;
-
-u32 D_80042878 = 0x00000000;
-
-u32 D_8004287C = 0x00000000;
-
-u32 D_80042880 = 0x00000000;
-
-u32 D_80042884 = 0x00000000;
-
-u32 D_80042888 = 0x00000000;
-
-u32 D_8004288C = 0x00000000;
-
-u32 D_80042890 = 0x00000000;
-
-u32 D_80042894 = 0x00000000;
-
-u32 D_80042898 = 0x00000000;
-
-u32 D_8004289C = 0x00000000;
-
-u32 D_800428A0 = 0x00000000;
-
-u32 D_800428A4 = 0x00000000;
-
-u32 D_800428A8 = 0x00000000;
-
-u32 D_800428AC = 0x00000000;
-
-u32 D_800428B0 = 0x00000000;
-
-u32 D_800428B4 = 0x00000000;
-
-u32 D_800428B8 = 0x00000000;
-
-u32 D_800428BC = 0x00000000;
-
-u32 D_800428C0 = 0xED000000;
-
-u32 D_800428C4 = 0x005003C0;
-
-u32 D_800428C8 = 0xEF080CFF;
-
-u32 D_800428CC = 0x00000000;
-
-u32 D_800428D0 = 0x00000000;
-
-u32 D_800428D4 = 0x00000000;
-
-u32 D_800428D8 = 0x00000000;
-
-u32 D_800428DC = 0xFFFF0048;
-
-u32 D_800428E0 = 0x00000000;
-
-u32 D_800428E4 = 0x00000000;
-
-u32 D_800428E8 = 0x00000000;
-
-u32 D_800428EC = 0x00000000;
-
-u32 D_800428F0 = 0x00000000;
-
-u32 D_800428F4 = 0x00000000;
-
-u32 D_800428F8 = 0x00000000;
-
-u32 D_800428FC = 0x00000000;
-
-u32 D_80042900 = 0x00000000;
-
-u32 D_80042904 = 0x00000000;
-
-u32 D_80042908 = 0x00000000;
-
-u32 D_8004290C = 0x00000000;
-
-u32 D_80042910 = 0x00000000;
-
-u32 D_80042914 = 0x00000000;
-
-u32 D_80042918 = 0x00000000;
-
-u32 D_8004291C = 0x00000000;
-
-u32 D_80042920 = 0x00000000;
-
-u32 D_80042924 = 0x00000000;
-
-u32 D_80042928 = 0x00000000;
-
-u32 D_8004292C = 0x00000000;
-
-u32 D_80042930 = 0x00000000;
-
-u32 D_80042934 = 0x00000000;
-
-u32 D_80042938 = 0x0A525350;
-
-u32 D_8004293C = 0x20476678;
-
-u32 D_80042940 = 0x2075636F;
-
-u32 D_80042944 = 0x64652053;
-
-u32 D_80042948 = 0x32444558;
-
-u32 D_8004294C = 0x20202020;
-
-u32 D_80042950 = 0x20202066;
-
-u32 D_80042954 = 0x69666F20;
-
-u32 D_80042958 = 0x322E3034;
-
-u32 D_8004295C = 0x2020596F;
-
-u32 D_80042960 = 0x73686974;
-
-u32 D_80042964 = 0x616B6120;
-
-u32 D_80042968 = 0x59617375;
-
-u32 D_8004296C = 0x6D6F746F;
-
-u32 D_80042970 = 0x20313939;
-
-u32 D_80042974 = 0x38204E69;
-
-u32 D_80042978 = 0x6E74656E;
-
-u32 D_8004297C = 0x646F2E0A;
-
-u32 D_80042980 = 0x08000040;
-
-u32 D_80042984 = 0x0002FFFF;
-
-u32 D_80042988 = 0x40000003;
-
-u32 D_8004298C = 0x0008FFF8;
-
-u32 D_80042990 = 0x01008000;
-
-u32 D_80042994 = 0xFE000000;
-
-u32 D_80042998 = 0x10000400;
-
-u32 D_8004299C = 0x20000200;
-
-u32 D_800429A0 = 0x3D100001;
-
-u32 D_800429A4 = 0x32000000;
-
-u32 D_800429A8 = 0x27000000;
-
-u32 D_800429AC = 0x35242628;
-
-u32 D_800429B0 = 0xFFFC0400;
-
-u32 D_800429B4 = 0x00200010;
-
-u32 D_800429B8 = 0xFC0003FF;
-
-u32 D_800429BC = 0x001FFFE0;
-
-u32 D_800429C0 = 0x00000000;
-
-u32 D_800429C4 = 0x00000000;
-
-u32 D_800429C8 = 0x00000000;
-
-u32 D_800429CC = 0x00000000;
-
-u32 D_800429D0 = 0x00000000;
-
-u32 D_800429D4 = 0x00000000;
-
-u32 D_800429D8 = 0x00000000;
-
-u32 D_800429DC = 0x00000000;
-
-u32 D_800429E0 = 0x00010000;
-
-u32 D_800429E4 = 0x00000000;
-
-u32 D_800429E8 = 0x00000000;
-
-u32 D_800429EC = 0x00010000;
-
-u32 D_800429F0 = 0x00000000;
-
-u32 D_800429F4 = 0x04000400;
-
-u32 D_800429F8 = 0x00000000;
-
-u32 D_800429FC = 0x0001FFFE;
-
-u32 D_80042A00 = 0x0020FFFF;
-
-u32 D_80042A04 = 0x00000500;
-
-u32 D_80042A08 = 0x000003C0;
-
-u32 D_80042A0C = 0x11A416E0;
-
-u32 D_80042A10 = 0x1BA09BA0;
-
-u32 D_80042A14 = 0x00000000;
-
-u32 D_80042A18 = 0x00100020;
-
-u32 D_80042A1C = 0x00200040;
-
-u32 D_80042A20 = 0x00300060;
-
-u32 D_80042A24 = 0x0000FFF4;
-
-u32 D_80042A28 = 0x00100014;
-
-u32 D_80042A2C = 0x00200034;
-
-u32 D_80042A30 = 0x00300054;
-
-u32 D_80042A34 = 0x0001FFFE;
-
-u32 D_80042A38 = 0xFFFEFFFE;
-
-u32 D_80042A3C = 0x00010000;
-
-u32 D_80042A40 = 0x00000000;
-
-u32 D_80042A44 = 0x02480258;
-
-u32 D_80042A48 = 0xFFFC0000;
-
-u32 D_80042A4C = 0x00000001;
-
-u32 D_80042A50 = 0xFFFF0003;
-
-u32 D_80042A54 = 0xFFF00000;
-
-u32 D_80042A58 = 0x0001FFFF;
-
-u32 D_80042A5C = 0x00080200;
-
-u32 D_80042A60 = 0xFFFF0001;
-
-u32 D_80042A64 = 0x0007C1F0;
-
-u32 D_80042A68 = 0x00000800;
-
-u32 D_80042A6C = 0x08000400;
-
-u32 D_80042A70 = 0x08000800;
-
-u32 D_80042A74 = 0x003F0800;
-
-u32 D_80042A78 = 0x10000080;
-
-u32 D_80042A7C = 0x001F1000;
-
-u32 D_80042A80 = 0x20000100;
-
-u32 D_80042A84 = 0x000F2000;
-
-u32 D_80042A88 = 0x40000200;
-
-u32 D_80042A8C = 0x00074000;
-
-u32 D_80042A90 = 0x80000400;
-
-u32 D_80042A94 = 0x01FF0080;
-
-u32 D_80042A98 = 0x00FF0100;
-
-u32 D_80042A9C = 0x007F0200;
-
-u32 D_80042AA0 = 0x003F0400;
-
-u32 D_80042AA4 = 0x04000400;
-
-u32 D_80042AA8 = 0x02000400;
-
-u32 D_80042AAC = 0x04008002;
-
-u32 D_80042AB0 = 0x00000000;
-
-u32 D_80042AB4 = 0x00000000;
-
-u32 D_80042AB8 = 0x00000000;
-
-u32 D_80042ABC = 0x00000000;
-
-u32 D_80042AC0 = 0x00010000;
-
-u32 D_80042AC4 = 0x00000001;
-
-u32 D_80042AC8 = 0x00000001;
-
-u32 D_80042ACC = 0x00000001;
-
-u32 D_80042AD0 = 0x00000001;
-
-u32 D_80042AD4 = 0x00010000;
-
-u32 D_80042AD8 = 0x00000001;
-
-u32 D_80042ADC = 0x00000001;
-
-u32 D_80042AE0 = 0x00000001;
-
-u32 D_80042AE4 = 0x00010000;
-
-u32 D_80042AE8 = 0x00010000;
-
-u32 D_80042AEC = 0x00010000;
-
-u32 D_80042AF0 = 0x00010000;
-
-u32 D_80042AF4 = 0x00000001;
-
-u32 D_80042AF8 = 0x00010000;
-
-u32 D_80042AFC = 0x00010000;
-
-u32 D_80042B00 = 0x00000000;
-
-u32 D_80042B04 = 0x00000F80;
-
-u32 D_80042B08 = 0x00971000;
-
-u32 D_80042B0C = 0x00001018;
-
-u32 D_80042B10 = 0x017F1000;
-
-u32 D_80042B14 = 0x00000640;
-
-u32 D_80042B18 = 0x072F16C0;
-
-u32 D_80042B1C = 0x00001198;
-
-u32 D_80042B20 = 0x072716C0;
-
-u32 D_80042B24 = 0x1B801090;
-
-u32 D_80042B28 = 0x10A41FAC;
-
-u32 D_80042B2C = 0x1054107C;
-
-u32 D_80042B30 = 0x11A4101C;
-
-u32 D_80042B34 = 0x10B810B4;
-
-u32 D_80042B38 = 0x10181018;
-
-u32 D_80042B3C = 0x11441144;
-
-u32 D_80042B40 = 0x11441144;
-
-u32 D_80042B44 = 0x11401140;
-
-u32 D_80042B48 = 0x114011D0;
-
-u32 D_80042B4C = 0x11401170;
-
-u32 D_80042B50 = 0x11401128;
-
-u32 D_80042B54 = 0x11401140;
-
-u32 D_80042B58 = 0x11401140;
-
-u32 D_80042B5C = 0x11401140;
-
-u32 D_80042B60 = 0x11401140;
-
-u32 D_80042B64 = 0x11401140;
-
-u32 D_80042B68 = 0x11401150;
-
-u32 D_80042B6C = 0x11501150;
-
-u32 D_80042B70 = 0x11441B80;
-
-u32 D_80042B74 = 0x16C011A4;
-
-u32 D_80042B78 = 0x10241550;
-
-u32 D_80042B7C = 0x15501550;
-
-u32 D_80042B80 = 0x15501B74;
-
-u32 D_80042B84 = 0x120810E8;
-
-u32 D_80042B88 = 0x01E001F0;
-
-u32 D_80042B8C = 0x00F802B0;
-
