+++ data/kirby.039DF0.c
@@ -5,18 +5,6 @@
 #include "banks.h"
 #include "src/ovl0/main.h"
 
-typedef struct {
-        s32             active;		/* Status flag */
-	OSThread	*thread;	/* Calling thread */
-        OSMesgQueue  	*cmdQueue;	/* Command queue */
-        OSMesgQueue  	*evtQueue;	/* Event queue */
-        OSMesgQueue  	*acsQueue;	/* Access queue */
-					/* Raw DMA routine */
-        s32             (*dma)(s32, u32, void *, u32);
-        s32             (*edma)(OSPiHandle *, s32, u32, void *, u32);
-} OSDevMgr; // 0x1C
-
-
 #define FLT_MAX (3.40282346639e+38)
 
 u32 D_80039230 = 0x40055800;
@@ -1602,11421 +1590,3 @@ u32 D_80039E84 = 0x00000000;
 u32 D_80039E88 = 0x00000000;
 
 u32 D_80039E8C = 0x00000000;
-
-
-// F3DEX2 start? (0x3AAB0)
-u32 D_80039E90 = 0x4A00002C;
-
-u32 D_80039E94 = 0xC81F201B;
-
-u32 D_80039E98 = 0xC81E201C;
-
-u32 D_80039E9C = 0x20170BA8;
-
-u32 D_80039EA0 = 0x20160D00;
-
-u32 D_80039EA4 = 0x4B1F0051;
-
-u32 D_80039EA8 = 0x8C0B00F0;
-
-u32 D_80039EAC = 0x8C0C0FC4;
-
-u32 D_80039EB0 = 0x20012800;
-
-u32 D_80039EB4 = 0x11600006;
-
-u32 D_80039EB8 = 0x40812000;
-
-u32 D_80039EBC = 0x318C0001;
-
-u32 D_80039EC0 = 0x1180001E;
-
-u32 D_80039EC4 = 0xAC000FC4;
-
-u32 D_80039EC8 = 0x08000459;
-
-u32 D_80039ECC = 0x8C1A0BF8;
-
-u32 D_80039ED0 = 0x400B5800;
-
-u32 D_80039ED4 = 0x316B0001;
-
-u32 D_80039ED8 = 0x1560000B;
-
-u32 D_80039EDC = 0x40024800;
-
-u32 D_80039EE0 = 0x8C030FE8;
-
-u32 D_80039EE4 = 0x00625822;
-
-u32 D_80039EE8 = 0x1D600007;
-
-u32 D_80039EEC = 0x40015000;
-
-u32 D_80039EF0 = 0x8C040FEC;
-
-u32 D_80039EF4 = 0x10200004;
-
-u32 D_80039EF8 = 0x00245822;
-
-u32 D_80039EFC = 0x05610002;
-
-u32 D_80039F00 = 0x00000000;
-
-u32 D_80039F04 = 0x14220008;
-
-u32 D_80039F08 = 0x400B5800;
-
-u32 D_80039F0C = 0x316B0400;
-
-u32 D_80039F10 = 0x1560FFFD;
-
-u32 D_80039F14 = 0x200B0001;
-
-u32 D_80039F18 = 0x408B5800;
-
-u32 D_80039F1C = 0x8C020FEC;
-
-u32 D_80039F20 = 0x40824000;
-
-u32 D_80039F24 = 0x40824800;
-
-u32 D_80039F28 = 0xAC0200F0;
-
-u32 D_80039F2C = 0x8C0B00F4;
-
-u32 D_80039F30 = 0x15600002;
-
-u32 D_80039F34 = 0x8C0B0FE0;
-
-u32 D_80039F38 = 0xAC0B00F4;
-
-u32 D_80039F3C = 0x8C010FD0;
-
-u32 D_80039F40 = 0x8C0202E0;
-
-u32 D_80039F44 = 0x8C0302E8;
-
-u32 D_80039F48 = 0x8C040410;
-
-u32 D_80039F4C = 0x8C050418;
-
-u32 D_80039F50 = 0x00411020;
-
-u32 D_80039F54 = 0x00611820;
-
-u32 D_80039F58 = 0xAC0202E0;
-
-u32 D_80039F5C = 0xAC0302E8;
-
-u32 D_80039F60 = 0x00812020;
-
-u32 D_80039F64 = 0x00A12820;
-
-u32 D_80039F68 = 0xAC040410;
-
-u32 D_80039F6C = 0xAC050418;
-
-u32 D_80039F70 = 0x8C1A0FF0;
-
-u32 D_80039F74 = 0x200B02E8;
-
-u32 D_80039F78 = 0x0C0007ED;
-
-u32 D_80039F7C = 0x37EC0000;
-
-u32 D_80039F80 = 0x201300A7;
-
-u32 D_80039F84 = 0x37580000;
-
-u32 D_80039F88 = 0x0C0007F6;
-
-u32 D_80039F8C = 0x24140920;
-
-u32 D_80039F90 = 0x275A00A8;
-
-u32 D_80039F94 = 0x201BFF58;
-
-u32 D_80039F98 = 0x0C0007F2;
-
-u32 D_80039F9C = 0x40012000;
-
-u32 D_80039FA0 = 0x8F7909C8;
-
-u32 D_80039FA4 = 0x1360FFF6;
-
-u32 D_80039FA8 = 0x30210080;
-
-u32 D_80039FAC = 0x00196603;
-
-u32 D_80039FB0 = 0x000C5840;
-
-u32 D_80039FB4 = 0x956B036E;
-
-u32 D_80039FB8 = 0x14200380;
-
-u32 D_80039FBC = 0x8F7809CC;
-
-u32 D_80039FC0 = 0x01600008;
-
-u32 D_80039FC4 = 0x277B0008;
-
-u32 D_80039FC8 = 0x201F118C;
-
-u32 D_80039FCC = 0x20150040;
-
-u32 D_80039FD0 = 0x20140000;
-
-u32 D_80039FD4 = 0x20130080;
-
-u32 D_80039FD8 = 0x08000422;
-
-u32 D_80039FDC = 0xA01901D9;
-
-u32 D_80039FE0 = 0x0C00048D;
-
-u32 D_80039FE4 = 0x877409C1;
-
-u32 D_80039FE8 = 0x33330FF8;
-
-u32 D_80039FEC = 0x0014A083;
-
-u32 D_80039FF0 = 0x080007F6;
-
-u32 D_80039FF4 = 0x201F1188;
-
-u32 D_80039FF8 = 0x8C0B01EC;
-
-u32 D_80039FFC = 0x01795824;
-
-u32 D_8003A000 = 0x01785825;
-
-u32 D_8003A004 = 0x08000463;
-
-u32 D_8003A008 = 0xAC0B01EC;
-
-u32 D_8003A00C = 0x900100DE;
-
-u32 D_8003A010 = 0x1020036A;
-
-u32 D_8003A014 = 0x2021FFFC;
-
-u32 D_8003A018 = 0x08000408;
-
-u32 D_8003A01C = 0x8C3A0138;
-
-u32 D_8003A020 = 0xC81D181A;
-
-u32 D_8003A024 = 0x8C1900D8;
-
-u32 D_8003A028 = 0x22F70008;
-
-u32 D_8003A02C = 0xEAFD187F;
-
-u32 D_8003A030 = 0xAEF80004;
-
-u32 D_8003A034 = 0xAEF90000;
-
-u32 D_8003A038 = 0x0800049A;
-
-u32 D_8003A03C = 0x22F70008;
-
-u32 D_8003A040 = 0x201F1220;
-
-u32 D_8003A044 = 0x00185D82;
-
-u32 D_8003A048 = 0x316B003C;
-
-u32 D_8003A04C = 0x8D6B00F8;
-
-u32 D_8003A050 = 0x0018C200;
-
-u32 D_8003A054 = 0x0018C202;
-
-u32 D_8003A058 = 0x03E00008;
-
-u32 D_8003A05C = 0x030BC020;
-
-u32 D_8003A060 = 0xAC1900C8;
-
-u32 D_8003A064 = 0x08000488;
-
-u32 D_8003A068 = 0xAC1800CC;
-
-u32 D_8003A06C = 0xAC1900C0;
-
-u32 D_8003A070 = 0x08000488;
-
-u32 D_8003A074 = 0xAC1800C4;
-
-u32 D_8003A078 = 0x201F118C;
-
-u32 D_8003A07C = 0x02F65822;
-
-u32 D_8003A080 = 0x19600358;
-
-u32 D_8003A084 = 0x400C3000;
-
-u32 D_8003A088 = 0x8C1800F0;
-
-u32 D_8003A08C = 0x25730158;
-
-u32 D_8003A090 = 0x1580FFFC;
-
-u32 D_8003A094 = 0x8C0C0FEC;
-
-u32 D_8003A098 = 0x40984800;
-
-u32 D_8003A09C = 0x03135820;
-
-u32 D_8003A0A0 = 0x018B6022;
-
-u32 D_8003A0A4 = 0x05810008;
-
-u32 D_8003A0A8 = 0x400B5800;
-
-u32 D_8003A0AC = 0x316B0400;
-
-u32 D_8003A0B0 = 0x1560FFFD;
-
-u32 D_8003A0B4 = 0x8C180FE8;
-
-u32 D_8003A0B8 = 0x400B5000;
-
-u32 D_8003A0BC = 0x1178FFFE;
-
-u32 D_8003A0C0 = 0x00000000;
-
-u32 D_8003A0C4 = 0x40984000;
-
-u32 D_8003A0C8 = 0x400B5000;
-
-u32 D_8003A0CC = 0x01785822;
-
-u32 D_8003A0D0 = 0x19600002;
-
-u32 D_8003A0D4 = 0x01735822;
-
-u32 D_8003A0D8 = 0x1960FFFB;
-
-u32 D_8003A0DC = 0x03135820;
-
-u32 D_8003A0E0 = 0xAC0B00F0;
-
-u32 D_8003A0E4 = 0x2273FFFF;
-
-u32 D_8003A0E8 = 0x22D4DEA8;
-
-u32 D_8003A0EC = 0x3AD60208;
-
-u32 D_8003A0F0 = 0x080007F6;
-
-u32 D_8003A0F4 = 0x22D7FEA8;
-
-u32 D_8003A0F8 = 0x200B0410;
-
-u32 D_8003A0FC = 0x080007ED;
-
-u32 D_8003A100 = 0x200C12E8;
-
-u32 D_8003A104 = 0x37FE0000;
-
-u32 D_8003A108 = 0x24050014;
-
-u32 D_8003A10C = 0x24120006;
-
-u32 D_8003A110 = 0x240F09C8;
-
-u32 D_8003A114 = 0xA64103CA;
-
-u32 D_8003A118 = 0xA64203CC;
-
-u32 D_8003A11C = 0xA64303CE;
-
-u32 D_8003A120 = 0xA64003D0;
-
-u32 D_8003A124 = 0x8C1D03CC;
-
-u32 D_8003A128 = 0x8CA903F8;
-
-u32 D_8003A12C = 0x8C700024;
-
-u32 D_8003A130 = 0x02098024;
-
-u32 D_8003A134 = 0x2251FFFA;
-
-u32 D_8003A138 = 0x3A52001C;
-
-u32 D_8003A13C = 0x2255FFFA;
-
-u32 D_8003A140 = 0x962203D0;
-
-u32 D_8003A144 = 0x22310002;
-
-u32 D_8003A148 = 0x1040005D;
-
-u32 D_8003A14C = 0x8C4B0024;
-
-u32 D_8003A150 = 0x01695824;
-
-u32 D_8003A154 = 0x11700055;
-
-u32 D_8003A158 = 0x35700000;
-
-u32 D_8003A15C = 0x12000003;
-
-u32 D_8003A160 = 0x34530000;
-
-u32 D_8003A164 = 0x34730000;
-
-u32 D_8003A168 = 0x34430000;
-
-u32 D_8003A16C = 0x00055840;
-
-u32 D_8003A170 = 0xC9621830;
-
-u32 D_8003A174 = 0xCA641801;
-
-u32 D_8003A178 = 0xCA651800;
-
-u32 D_8003A17C = 0xC8661801;
-
-u32 D_8003A180 = 0xC8671800;
-
-u32 D_8003A184 = 0x4B1F10C7;
-
-u32 D_8003A188 = 0x4A022206;
-
-u32 D_8003A18C = 0x4A022A4F;
-
-u32 D_8003A190 = 0x4A03328E;
-
-u32 D_8003A194 = 0x4A033ACF;
-
-u32 D_8003A198 = 0x4A484214;
-
-u32 D_8003A19C = 0xC819201D;
-
-u32 D_8003A1A0 = 0x4A494A50;
-
-u32 D_8003A1A4 = 0x4A4A5294;
-
-u32 D_8003A1A8 = 0x4A4B5AD0;
-
-u32 D_8003A1AC = 0x4AA84214;
-
-u32 D_8003A1B0 = 0x4AA94A50;
-
-u32 D_8003A1B4 = 0x4AAA5294;
-
-u32 D_8003A1B8 = 0x4AAB5AD0;
-
-u32 D_8003A1BC = 0x4B6B4772;
-
-u32 D_8003A1C0 = 0x4B6A58B1;
-
-u32 D_8003A1C4 = 0x4B0058F2;
-
-u32 D_8003A1C8 = 0x4B795F53;
-
-u32 D_8003A1CC = 0x4B7D1086;
-
-u32 D_8003A1D0 = 0x4B7D18CF;
-
-u32 D_8003A1D4 = 0x4B0018E1;
-
-u32 D_8003A1D8 = 0x4B1F10A7;
-
-u32 D_8003A1DC = 0x4B625744;
-
-u32 D_8003A1E0 = 0x4B625ACD;
-
-u32 D_8003A1E4 = 0x4B00028E;
-
-u32 D_8003A1E8 = 0x4B6B5B72;
-
-u32 D_8003A1EC = 0x4B6A5B31;
-
-u32 D_8003A1F0 = 0x4B005B72;
-
-u32 D_8003A1F4 = 0x4A0A6744;
-
-u32 D_8003A1F8 = 0x4A0A6F4D;
-
-u32 D_8003A1FC = 0x4A0B628E;
-
-u32 D_8003A200 = 0x4A0B6ACF;
-
-u32 D_8003A204 = 0x4B3F0F47;
-
-u32 D_8003A208 = 0x4B9F528E;
-
-u32 D_8003A20C = 0x4B9F5ACF;
-
-u32 D_8003A210 = 0x4A0A6744;
-
-u32 D_8003A214 = 0x4A0A6F4D;
-
-u32 D_8003A218 = 0x4A0B630E;
-
-u32 D_8003A21C = 0x4A0B6B4F;
-
-u32 D_8003A220 = 0x4A0C4744;
-
-u32 D_8003A224 = 0xC87A3802;
-
-u32 D_8003A228 = 0x4A0C4F4D;
-
-u32 D_8003A22C = 0xC87A1405;
-
-u32 D_8003A230 = 0x4A0D428E;
-
-u32 D_8003A234 = 0xCA793802;
-
-u32 D_8003A238 = 0x4A0D4ACF;
-
-u32 D_8003A23C = 0xCA791405;
-
-u32 D_8003A240 = 0x4B625744;
-
-u32 D_8003A244 = 0x4B625ACD;
-
-u32 D_8003A248 = 0x4B00528E;
-
-u32 D_8003A24C = 0x4B015AE0;
-
-u32 D_8003A250 = 0x4B1F52A7;
-
-u32 D_8003A254 = 0x4B015755;
-
-u32 D_8003A258 = 0x4B005AE3;
-
-u32 D_8003A25C = 0x4B0152A7;
-
-u32 D_8003A260 = 0x4B1F5086;
-
-u32 D_8003A264 = 0x4B6A3744;
-
-u32 D_8003A268 = 0x4B6A3F4D;
-
-u32 D_8003A26C = 0x4B62274C;
-
-u32 D_8003A270 = 0x4B622E0D;
-
-u32 D_8003A274 = 0x4B0005CE;
-
-u32 D_8003A278 = 0x4B6AD745;
-
-u32 D_8003A27C = 0x4B62CD8D;
-
-u32 D_8003A280 = 0x20070000;
-
-u32 D_8003A284 = 0x20010002;
-
-u32 D_8003A288 = 0xA6AF03D0;
-
-u32 D_8003A28C = 0x080005D7;
-
-u32 D_8003A290 = 0x201F95F0;
-
-u32 D_8003A294 = 0xE9F91879;
-
-u32 D_8003A298 = 0xE9FA0A67;
-
-u32 D_8003A29C = 0xE9F63878;
-
-u32 D_8003A2A0 = 0xE9F61471;
-
-u32 D_8003A2A4 = 0x21EFFFD8;
-
-u32 D_8003A2A8 = 0x22B50002;
-
-u32 D_8003A2AC = 0x1600FFA4;
-
-u32 D_8003A2B0 = 0x34430000;
-
-u32 D_8003A2B4 = 0xA6A303D0;
-
-u32 D_8003A2B8 = 0x080004CC;
-
-u32 D_8003A2BC = 0x22B50002;
-
-u32 D_8003A2C0 = 0x02B25822;
-
-u32 D_8003A2C4 = 0x0560000F;
-
-u32 D_8003A2C8 = 0xA6A003D0;
-
-u32 D_8003A2CC = 0x96A303CE;
-
-u32 D_8003A2D0 = 0x14A0FF95;
-
-u32 D_8003A2D4 = 0x20A5FFFC;
-
-u32 D_8003A2D8 = 0xAC0003CC;
-
-u32 D_8003A2DC = 0x964103CA;
-
-u32 D_8003A2E0 = 0x96A203CC;
-
-u32 D_8003A2E4 = 0x96A303CE;
-
-u32 D_8003A2E8 = 0x48811500;
-
-u32 D_8003A2EC = 0x4BBF00EA;
-
-u32 D_8003A2F0 = 0x48822600;
-
-u32 D_8003A2F4 = 0x0C0005F9;
-
-u32 D_8003A2F8 = 0x48831700;
-
-u32 D_8003A2FC = 0x16B2FFF7;
-
-u32 D_8003A300 = 0x22B5FFFE;
-
-u32 D_8003A304 = 0x03C00008;
-
-u32 D_8003A308 = 0xAC1D03CC;
-
-u32 D_8003A30C = 0x00000000;
-
-u32 D_8003A310 = 0x97340380;
-
-u32 D_8003A314 = 0x0C00048D;
-
-u32 D_8003A318 = 0x976109C1;
-
-u32 D_8003A31C = 0x0281A022;
-
-u32 D_8003A320 = 0x0C0007F6;
-
-u32 D_8003A324 = 0x2033FFFF;
-
-u32 D_8003A328 = 0x940501EC;
-
-u32 D_8003A32C = 0x000108C2;
-
-u32 D_8003A330 = 0x03217822;
-
-u32 D_8003A334 = 0x95EF0380;
-
-u32 D_8003A338 = 0x368E0000;
-
-u32 D_8003A33C = 0x900801D9;
-
-u32 D_8003A340 = 0x30A60002;
-
-u32 D_8003A344 = 0x14C0FF6C;
-
-u32 D_8003A348 = 0x30A70001;
-
-u32 D_8003A34C = 0x15000006;
-
-u32 D_8003A350 = 0x000738C0;
-
-u32 D_8003A354 = 0xA01901D9;
-
-u32 D_8003A358 = 0x20150040;
-
-u32 D_8003A35C = 0x20140000;
-
-u32 D_8003A360 = 0x0C000422;
-
-u32 D_8003A364 = 0x20130080;
-
-u32 D_8003A368 = 0xC8082008;
-
-u32 D_8003A36C = 0xC80A2009;
-
-u32 D_8003A370 = 0xC80C200A;
-
-u32 D_8003A374 = 0xC80E200B;
-
-u32 D_8003A378 = 0x4B004250;
-
-u32 D_8003A37C = 0xC8091811;
-
-u32 D_8003A380 = 0x4B0052D0;
-
-u32 D_8003A384 = 0xC80B1813;
-
-u32 D_8003A388 = 0x4B006350;
-
-u32 D_8003A38C = 0xC80D1815;
-
-u32 D_8003A390 = 0x4B0073D0;
-
-u32 D_8003A394 = 0xC80F1817;
-
-u32 D_8003A398 = 0xC8081C10;
-
-u32 D_8003A39C = 0xC80A1C12;
-
-u32 D_8003A3A0 = 0x0C0005D7;
-
-u32 D_8003A3A4 = 0xC80C1C14;
-
-u32 D_8003A3A8 = 0x0C0007F2;
-
-u32 D_8003A3AC = 0xC80E1C16;
-
-u32 D_8003A3B0 = 0xC9D41800;
-
-u32 D_8003A3B4 = 0x4B356C33;
-
-u32 D_8003A3B8 = 0xC9D41C02;
-
-u32 D_8003A3BC = 0x4B017F46;
-
-u32 D_8003A3C0 = 0x8DCB001C;
-
-u32 D_8003A3C4 = 0x4B015F4F;
-
-u32 D_8003A3C8 = 0xC9D61602;
-
-u32 D_8003A3CC = 0x4A94674E;
-
-u32 D_8003A3D0 = 0x34C90000;
-
-u32 D_8003A3D4 = 0x4A94474F;
-
-u32 D_8003A3D8 = 0xC9223016;
-
-u32 D_8003A3DC = 0x4AB46F4E;
-
-u32 D_8003A3E0 = 0xADCB0008;
-
-u32 D_8003A3E4 = 0x4AB44F4F;
-
-u32 D_8003A3E8 = 0xC9C73001;
-
-u32 D_8003A3EC = 0x4AD475CE;
-
-u32 D_8003A3F0 = 0x14C0FF81;
-
-u32 D_8003A3F4 = 0x4AD4560F;
-
-u32 D_8003A3F8 = 0x4B7FCEE3;
-
-u32 D_8003A3FC = 0xC9D61206;
-
-u32 D_8003A400 = 0x2021FFFC;
-
-u32 D_8003A404 = 0x4B92BF44;
-
-u32 D_8003A408 = 0x01075822;
-
-u32 D_8003A40C = 0x4B92C08D;
-
-u32 D_8003A410 = 0xE97B07F3;
-
-u32 D_8003A414 = 0x4B00054E;
-
-u32 D_8003A418 = 0xE97B03CB;
-
-u32 D_8003A41C = 0x4BB2B9C6;
-
-u32 D_8003A420 = 0xE9191C7E;
-
-u32 D_8003A424 = 0x4BB2C18F;
-
-u32 D_8003A428 = 0xE9191879;
-
-u32 D_8003A42C = 0x4B624772;
-
-u32 D_8003A430 = 0xE91A0E7B;
-
-u32 D_8003A434 = 0x4B755971;
-
-u32 D_8003A438 = 0xE91A0A67;
-
-u32 D_8003A43C = 0x4BE25932;
-
-u32 D_8003A440 = 0xC9C31801;
-
-u32 D_8003A444 = 0x4BF57971;
-
-u32 D_8003A448 = 0x00015FC3;
-
-u32 D_8003A44C = 0x4B007932;
-
-u32 D_8003A450 = 0x316B0028;
-
-u32 D_8003A454 = 0x4AF8C765;
-
-u32 D_8003A458 = 0x21EF0050;
-
-u32 D_8003A45C = 0x4AF7BF64;
-
-u32 D_8003A460 = 0x01EB4022;
-
-u32 D_8003A464 = 0x4A05AF44;
-
-u32 D_8003A468 = 0x484A0800;
-
-u32 D_8003A46C = 0x4A05174D;
-
-u32 D_8003A470 = 0xE9171C7C;
-
-u32 D_8003A474 = 0x4A04AD4E;
-
-u32 D_8003A478 = 0xC9D41804;
-
-u32 D_8003A47C = 0x4A04108F;
-
-u32 D_8003A480 = 0xE9F71877;
-
-u32 D_8003A484 = 0x4B00C763;
-
-u32 D_8003A488 = 0xC9170F72;
-
-u32 D_8003A48C = 0x4B3F0F47;
-
-u32 D_8003A490 = 0xE9181C7B;
-
-u32 D_8003A494 = 0x4B9FAE8E;
-
-u32 D_8003A498 = 0xC9F70B5E;
-
-u32 D_8003A49C = 0x4B9F164F;
-
-u32 D_8003A4A0 = 0xE9F81876;
-
-u32 D_8003A4A4 = 0x4BFF00A7;
-
-u32 D_8003A4A8 = 0xC9D41C06;
-
-u32 D_8003A4AC = 0x4AE6C765;
-
-u32 D_8003A4B0 = 0xE903107A;
-
-u32 D_8003A4B4 = 0x4A05D744;
-
-u32 D_8003A4B8 = 0xC9180F6E;
-
-u32 D_8003A4BC = 0x4A05CF4D;
-
-u32 D_8003A4C0 = 0xE9E31270;
-
-u32 D_8003A4C4 = 0x4A04D14E;
-
-u32 D_8003A4C8 = 0xC9F80B5A;
-
-u32 D_8003A4CC = 0x4A04C90F;
-
-u32 D_8003A4D0 = 0xA50AFFFE;
-
-u32 D_8003A4D4 = 0x4BFF108F;
-
-u32 D_8003A4D8 = 0x000A5900;
-
-u32 D_8003A4DC = 0x4AE7BF64;
-
-u32 D_8003A4E0 = 0x484A0800;
-
-u32 D_8003A4E4 = 0x4AE5BF44;
-
-u32 D_8003A4E8 = 0xE9050F7D;
-
-u32 D_8003A4EC = 0x4AE5C74D;
-
-u32 D_8003A4F0 = 0x21CE0020;
-
-u32 D_8003A4F4 = 0x4AE2BE8E;
-
-u32 D_8003A4F8 = 0xA50AFFFC;
-
-u32 D_8003A4FC = 0x4AE2C64F;
-
-u32 D_8003A500 = 0x000A5100;
-
-u32 D_8003A504 = 0x4A12B0C5;
-
-u32 D_8003A508 = 0xA5EBFFD6;
-
-u32 D_8003A50C = 0xA5EAFFD4;
-
-u32 D_8003A510 = 0x4B92D744;
-
-u32 D_8003A514 = 0xE9E50B69;
-
-u32 D_8003A518 = 0x4B92CE4D;
-
-u32 D_8003A51C = 0xE9040F7C;
-
-u32 D_8003A520 = 0x4B00068E;
-
-u32 D_8003A524 = 0xE9E40B68;
-
-u32 D_8003A528 = 0xE903127B;
-
-u32 D_8003A52C = 0x4B018F47;
-
-u32 D_8003A530 = 0xE9E31671;
-
-u32 D_8003A534 = 0x4B7F9F4F;
-
-u32 D_8003A538 = 0x4A10D68E;
-
-u32 D_8003A53C = 0x1C20FF9F;
-
-u32 D_8003A540 = 0x4A10CE4F;
-
-u32 D_8003A544 = 0x07E0FF53;
-
-u32 D_8003A548 = 0x4B7FCEE3;
-
-u32 D_8003A54C = 0xE9191C7E;
-
-u32 D_8003A550 = 0xE9F91879;
-
-u32 D_8003A554 = 0xE91A0E7B;
-
-u32 D_8003A558 = 0x10E0FE90;
-
-u32 D_8003A55C = 0xE9FA0A67;
-
-u32 D_8003A560 = 0xE91B07EB;
-
-u32 D_8003A564 = 0x08000463;
-
-u32 D_8003A568 = 0xE9FB03C3;
-
-u32 D_8003A56C = 0x200D0180;
-
-u32 D_8003A570 = 0xC810181C;
-
-u32 D_8003A574 = 0xC8101C1C;
-
-u32 D_8003A578 = 0xC9BD1018;
-
-u32 D_8003A57C = 0xC811181D;
-
-u32 D_8003A580 = 0xC8111C1D;
-
-u32 D_8003A584 = 0x4B7FFCE0;
-
-u32 D_8003A588 = 0x4A100551;
-
-u32 D_8003A58C = 0xC9B2121A;
-
-u32 D_8003A590 = 0x4B1D8427;
-
-u32 D_8003A594 = 0xC9B2161A;
-
-u32 D_8003A598 = 0x4B0104E7;
-
-u32 D_8003A59C = 0xC8121437;
-
-u32 D_8003A5A0 = 0x4B3D8C67;
-
-u32 D_8003A5A4 = 0xC9B20D03;
-
-u32 D_8003A5A8 = 0x4B354C33;
-
-u32 D_8003A5AC = 0x03E00008;
-
-u32 D_8003A5B0 = 0x22E80050;
-
-u32 D_8003A5B4 = 0x0C0005ED;
-
-u32 D_8003A5B8 = 0xAEF80004;
-
-u32 D_8003A5BC = 0x201F118C;
-
-u32 D_8003A5C0 = 0xAEF90004;
-
-u32 D_8003A5C4 = 0xCAE23000;
-
-u32 D_8003A5C8 = 0x92E10005;
-
-u32 D_8003A5CC = 0x92E20006;
-
-u32 D_8003A5D0 = 0x92E30007;
-
-u32 D_8003A5D4 = 0x4BBF00EA;
-
-u32 D_8003A5D8 = 0x94210380;
-
-u32 D_8003A5DC = 0x4BDF0906;
-
-u32 D_8003A5E0 = 0x94420380;
-
-u32 D_8003A5E4 = 0x4B3E108C;
-
-u32 D_8003A5E8 = 0x94630380;
-
-u32 D_8003A5EC = 0x4B00010E;
-
-u32 D_8003A5F0 = 0x34240000;
-
-u32 D_8003A5F4 = 0x4BFF016D;
-
-u32 D_8003A5F8 = 0xC8261006;
-
-u32 D_8003A5FC = 0x4BFF01ED;
-
-u32 D_8003A600 = 0xC8441006;
-
-u32 D_8003A604 = 0x4BA271B3;
-
-u32 D_8003A608 = 0xC8681006;
-
-u32 D_8003A60C = 0x4BFF026D;
-
-u32 D_8003A610 = 0x8C250024;
-
-u32 D_8003A614 = 0x4BE27233;
-
-u32 D_8003A618 = 0x8C460024;
-
-u32 D_8003A61C = 0x4B260090;
-
-u32 D_8003A620 = 0x8C670024;
-
-u32 D_8003A624 = 0x4A043291;
-
-u32 D_8003A628 = 0x30AB7070;
-
-u32 D_8003A62C = 0x4A0622D1;
-
-u32 D_8003A630 = 0x00CB5824;
-
-u32 D_8003A634 = 0x4A083311;
-
-u32 D_8003A638 = 0x00EB5824;
-
-u32 D_8003A63C = 0x4B241360;
-
-u32 D_8003A640 = 0x4A0433A7;
-
-u32 D_8003A644 = 0x156001E7;
-
-u32 D_8003A648 = 0x900B01EE;
-
-u32 D_8003A64C = 0x4B2C5747;
-
-u32 D_8003A650 = 0x8C0C03CC;
-
-u32 D_8003A654 = 0x4B2B674F;
-
-u32 D_8003A658 = 0x00A62825;
-
-u32 D_8003A65C = 0x4B2410A3;
-
-u32 D_8003A660 = 0x00A72825;
-
-u32 D_8003A664 = 0x4A0432A7;
-
-u32 D_8003A668 = 0x8D6B03C2;
-
-u32 D_8003A66C = 0x4B2869A3;
-
-u32 D_8003A670 = 0x4806E800;
-
-u32 D_8003A674 = 0x4A087127;
-
-u32 D_8003A678 = 0x00AC2824;
-
-u32 D_8003A67C = 0x4A0E43A7;
-
-u32 D_8003A680 = 0x14A0FEA0;
-
-u32 D_8003A684 = 0x00CB5820;
-
-u32 D_8003A688 = 0x4A0231A0;
-
-u32 D_8003A68C = 0x056101D5;
-
-u32 D_8003A690 = 0x4A0A20A7;
-
-u32 D_8003A694 = 0x4A0452A7;
-
-u32 D_8003A698 = 0x48017600;
-
-u32 D_8003A69C = 0x4BBF7106;
-
-u32 D_8003A6A0 = 0x10C001D0;
-
-u32 D_8003A6A4 = 0x4A0E1191;
-
-u32 D_8003A6A8 = 0x48021600;
-
-u32 D_8003A6AC = 0x4A0E5211;
-
-u32 D_8003A6B0 = 0x48035600;
-
-u32 D_8003A6B4 = 0x4A0272D1;
-
-u32 D_8003A6B8 = 0x8C0601EC;
-
-u32 D_8003A6BC = 0x4A0A7311;
-
-u32 D_8003A6C0 = 0xC82D1008;
-
-u32 D_8003A6C4 = 0x4A0253D1;
-
-u32 D_8003A6C8 = 0xC84D1408;
-
-u32 D_8003A6CC = 0x4B083407;
-
-u32 D_8003A6D0 = 0xC86D1608;
-
-u32 D_8003A6D4 = 0x4B0B440F;
-
-u32 D_8003A6D8 = 0x00065A80;
-
-u32 D_8003A6DC = 0x4B118C5D;
-
-u32 D_8003A6E0 = 0x0561000A;
-
-u32 D_8003A6E4 = 0x4B30841D;
-
-u32 D_8003A6E8 = 0xC8323002;
-
-u32 D_8003A6EC = 0x4B0653F3;
-
-u32 D_8003A6F0 = 0xC8533002;
-
-u32 D_8003A6F4 = 0x4B2F4530;
-
-u32 D_8003A6F8 = 0xC8753002;
-
-u32 D_8003A6FC = 0x4B3145B2;
-
-u32 D_8003A700 = 0x4B304DF1;
-
-u32 D_8003A704 = 0x0800064A;
-
-u32 D_8003A708 = 0x4B004E32;
-
-u32 D_8003A70C = 0xC8923002;
-
-u32 D_8003A710 = 0x4B2F4530;
-
-u32 D_8003A714 = 0xC8320313;
-
-u32 D_8003A718 = 0x4B3145B2;
-
-u32 D_8003A71C = 0xC8933002;
-
-u32 D_8003A720 = 0x4B304DF1;
-
-u32 D_8003A724 = 0xC8530313;
-
-u32 D_8003A728 = 0x4B004E32;
-
-u32 D_8003A72C = 0xC8953002;
-
-u32 D_8003A730 = 0x4B0653F3;
-
-u32 D_8003A734 = 0xC8750313;
-
-u32 D_8003A738 = 0x4B265530;
-
-u32 D_8003A73C = 0x4B2655B2;
-
-u32 D_8003A740 = 0x8C250020;
-
-u32 D_8003A744 = 0x4B285D30;
-
-u32 D_8003A748 = 0x8C470020;
-
-u32 D_8003A74C = 0x4B285DB2;
-
-u32 D_8003A750 = 0x8C680020;
-
-u32 D_8003A754 = 0x4BFE9484;
-
-u32 D_8003A758 = 0x900901E7;
-
-u32 D_8003A75C = 0x4BFE9CC4;
-
-u32 D_8003A760 = 0x00A75822;
-
-u32 D_8003A764 = 0x4BFEAD44;
-
-u32 D_8003A768 = 0x000B67C3;
-
-u32 D_8003A76C = 0x4B085BF3;
-
-u32 D_8003A770 = 0x016C5824;
-
-u32 D_8003A774 = 0x4B5FA744;
-
-u32 D_8003A778 = 0x00AB2822;
-
-u32 D_8003A77C = 0x4B5FB58D;
-
-u32 D_8003A780 = 0x00A85822;
-
-u32 D_8003A784 = 0x4B00050E;
-
-u32 D_8003A788 = 0x000B67C3;
-
-u32 D_8003A78C = 0x4BBF7E45;
-
-u32 D_8003A790 = 0x016C5824;
-
-u32 D_8003A794 = 0x4B0003CE;
-
-u32 D_8003A798 = 0x00AB2822;
-
-u32 D_8003A79C = 0x4A040115;
-
-u32 D_8003A7A0 = 0xAEE50010;
-
-u32 D_8003A7A4 = 0x4A000691;
-
-u32 D_8003A7A8 = 0xCAFB1004;
-
-u32 D_8003A7AC = 0x4A14CF45;
-
-u32 D_8003A7B0 = 0x48058880;
-
-u32 D_8003A7B4 = 0x4A147F4C;
-
-u32 D_8003A7B8 = 0x900701E6;
-
-u32 D_8003A7BC = 0x4A167D0E;
-
-u32 D_8003A7C0 = 0xC8530F0E;
-
-u32 D_8003A7C4 = 0x4A16CBCF;
-
-u32 D_8003A7C8 = 0xC8750F0E;
-
-u32 D_8003A7CC = 0x4A10BF44;
-
-u32 D_8003A7D0 = 0xC8470F0F;
-
-u32 D_8003A7D4 = 0x4A10C74D;
-
-u32 D_8003A7D8 = 0xC8690F0F;
-
-u32 D_8003A7DC = 0x4A11BC0E;
-
-u32 D_8003A7E0 = 0x34CB00C8;
-
-u32 D_8003A7E4 = 0x4A11C44F;
-
-u32 D_8003A7E8 = 0x01695825;
-
-u32 D_8003A7EC = 0x4B5E7BE6;
-
-u32 D_8003A7F0 = 0xA2EB0000;
-
-u32 D_8003A7F4 = 0x4BBE0F47;
-
-u32 D_8003A7F8 = 0xEAEA0901;
-
-u32 D_8003A7FC = 0x4B9E840E;
-
-u32 D_8003A800 = 0xEAE20902;
-
-u32 D_8003A804 = 0x4B9E8C4F;
-
-u32 D_8003A808 = 0xEAEE0903;
-
-u32 D_8003A80C = 0x4B0E1F46;
-
-u32 D_8003A810 = 0x30AC0080;
-
-u32 D_8003A814 = 0x4B24A74C;
-
-u32 D_8003A818 = 0x01876025;
-
-u32 D_8003A81C = 0x4B247F4D;
-
-u32 D_8003A820 = 0xA2EC0001;
-
-u32 D_8003A824 = 0x4B3AA08E;
-
-u32 D_8003A828 = 0x1120001D;
-
-u32 D_8003A82C = 0x4B3A78CF;
-
-u32 D_8003A830 = 0x4B1B4772;
-
-u32 D_8003A834 = 0x4B3B42B1;
-
-u32 D_8003A838 = 0x4A6D0390;
-
-u32 D_8003A83C = 0x4B0046F2;
-
-u32 D_8003A840 = 0x4BFF05AA;
-
-u32 D_8003A844 = 0x4B0A6F45;
-
-u32 D_8003A848 = 0x4B0A774C;
-
-u32 D_8003A84C = 0xC8361005;
-
-u32 D_8003A850 = 0x4B1B738E;
-
-u32 D_8003A854 = 0xC8561405;
-
-u32 D_8003A858 = 0x4B1B6B4F;
-
-u32 D_8003A85C = 0x4BFF02AA;
-
-u32 D_8003A860 = 0x4BFEF763;
-
-u32 D_8003A864 = 0xC86A1405;
-
-u32 D_8003A868 = 0x4A8EB745;
-
-u32 D_8003A86C = 0x4A8DB58F;
-
-u32 D_8003A870 = 0x4B00064E;
-
-u32 D_8003A874 = 0x4BCE5745;
-
-u32 D_8003A878 = 0x4BCD528F;
-
-u32 D_8003A87C = 0x4B00034E;
-
-u32 D_8003A880 = 0xEAF61804;
-
-u32 D_8003A884 = 0x4A169CE7;
-
-u32 D_8003A888 = 0xEAF91805;
-
-u32 D_8003A88C = 0x4A1939E7;
-
-u32 D_8003A890 = 0xCAF21C04;
-
-u32 D_8003A894 = 0x4A0AAD67;
-
-u32 D_8003A898 = 0xCAE51C05;
-
-u32 D_8003A89C = 0x4A0D4A67;
-
-u32 D_8003A8A0 = 0x4A178744;
-
-u32 D_8003A8A4 = 0xC8250F0F;
-
-u32 D_8003A8A8 = 0x4A178F4D;
-
-u32 D_8003A8AC = 0xC8320F0E;
-
-u32 D_8003A8B0 = 0x4A1885CE;
-
-u32 D_8003A8B4 = 0x84410018;
-
-u32 D_8003A8B8 = 0x4A188E0F;
-
-u32 D_8003A8BC = 0x26E20020;
-
-u32 D_8003A8C0 = 0x4A054A95;
-
-u32 D_8003A8C4 = 0x30C30004;
-
-u32 D_8003A8C8 = 0x4A12AA51;
-
-u32 D_8003A8CC = 0x00010B80;
-
-u32 D_8003A8D0 = 0x4A053B55;
-
-u32 D_8003A8D4 = 0xAEE10008;
-
-u32 D_8003A8D8 = 0x4A1299D1;
-
-u32 D_8003A8DC = 0xEAE30B08;
-
-u32 D_8003A8E0 = 0x4B265746;
-
-u32 D_8003A8E4 = 0xEAE20B09;
-
-u32 D_8003A8E8 = 0x4B264F4F;
-
-u32 D_8003A8EC = 0xEAE30A0C;
-
-u32 D_8003A8F0 = 0x4B2C6F4E;
-
-u32 D_8003A8F4 = 0xEAE20A0D;
-
-u32 D_8003A8F8 = 0x4B2C3F4F;
-
-u32 D_8003A8FC = 0xEAEF0806;
-
-u32 D_8003A900 = 0x4B22109D;
-
-u32 D_8003A904 = 0xEAF40807;
-
-u32 D_8003A908 = 0x4B0318DD;
-
-u32 D_8003A90C = 0xEAEF0B0A;
-
-u32 D_8003A910 = 0x4B086F46;
-
-u32 D_8003A914 = 0xEAF40B0B;
-
-u32 D_8003A918 = 0x4B083F4F;
-
-u32 D_8003A91C = 0xEAEF0A0E;
-
-u32 D_8003A920 = 0x4B0B574E;
-
-u32 D_8003A924 = 0xEAF40A0F;
-
-u32 D_8003A928 = 0x4B0B4F4F;
-
-u32 D_8003A92C = 0x00035900;
-
-u32 D_8003A930 = 0x4B26319D;
-
-u32 D_8003A934 = 0x004B0820;
-
-u32 D_8003A938 = 0x4B0739DD;
-
-u32 D_8003A93C = 0x00095940;
-
-u32 D_8003A940 = 0x4B371744;
-
-u32 D_8003A944 = 0x002BB820;
-
-u32 D_8003A948 = 0x4B371F4D;
-
-u32 D_8003A94C = 0x30C60001;
-
-u32 D_8003A950 = 0x4B38108E;
-
-u32 D_8003A954 = 0x00065900;
-
-u32 D_8003A958 = 0x4B3818CF;
-
-u32 D_8003A95C = 0x02EBB820;
-
-u32 D_8003A960 = 0x4B373744;
-
-u32 D_8003A964 = 0x4B373F4D;
-
-u32 D_8003A968 = 0x4B38318E;
-
-u32 D_8003A96C = 0xE8421803;
-
-u32 D_8003A970 = 0x4B3839CF;
-
-u32 D_8003A974 = 0xE8431801;
-
-u32 D_8003A978 = 0x4B74174C;
-
-u32 D_8003A97C = 0xE8221C03;
-
-u32 D_8003A980 = 0x4B741F4D;
-
-u32 D_8003A984 = 0xE8231C01;
-
-u32 D_8003A988 = 0x4B6F120E;
-
-u32 D_8003A98C = 0xE8461807;
-
-u32 D_8003A990 = 0x4B6F1A4F;
-
-u32 D_8003A994 = 0xE8471805;
-
-u32 D_8003A998 = 0x4B012F46;
-
-u32 D_8003A99C = 0xE8261C07;
-
-u32 D_8003A9A0 = 0x4B01974F;
-
-u32 D_8003A9A4 = 0xE8271C05;
-
-u32 D_8003A9A8 = 0x4B24474C;
-
-u32 D_8003A9AC = 0xE8481806;
-
-u32 D_8003A9B0 = 0x4B244F4D;
-
-u32 D_8003A9B4 = 0xE8491804;
-
-u32 D_8003A9B8 = 0x4B3A414E;
-
-u32 D_8003A9BC = 0xE8281C06;
-
-u32 D_8003A9C0 = 0x4B3A4C8F;
-
-u32 D_8003A9C4 = 0xE8291C04;
-
-u32 D_8003A9C8 = 0x4B244286;
-
-u32 D_8003A9CC = 0x10C00016;
-
-u32 D_8003A9D0 = 0x4BDE4206;
-
-u32 D_8003A9D4 = 0x4BDE4A4F;
-
-u32 D_8003A9D8 = 0xE8451802;
-
-u32 D_8003A9DC = 0x4BDE1086;
-
-u32 D_8003A9E0 = 0xE8521800;
-
-u32 D_8003A9E4 = 0x4BDE18CF;
-
-u32 D_8003A9E8 = 0xE8251C02;
-
-u32 D_8003A9EC = 0x4BDE3186;
-
-u32 D_8003A9F0 = 0xE8321C00;
-
-u32 D_8003A9F4 = 0x4BDE39CF;
-
-u32 D_8003A9F8 = 0xEAE80F7D;
-
-u32 D_8003A9FC = 0x4BDE5744;
-
-u32 D_8003AA00 = 0xEAE90F7C;
-
-u32 D_8003AA04 = 0x4BDE294E;
-
-u32 D_8003AA08 = 0xEAE20F7B;
-
-u32 D_8003AA0C = 0x4BDE948F;
-
-u32 D_8003AA10 = 0xEAE30F7A;
-
-u32 D_8003AA14 = 0xEAE60F7F;
-
-u32 D_8003AA18 = 0xEAE70F7E;
-
-u32 D_8003AA1C = 0xEAE50F79;
-
-u32 D_8003AA20 = 0x0800049B;
-
-u32 D_8003AA24 = 0xEAF20F78;
-
-u32 D_8003AA28 = 0xE8451802;
-
-u32 D_8003AA2C = 0xE8521800;
-
-u32 D_8003AA30 = 0xE8251C02;
-
-u32 D_8003AA34 = 0x0800049B;
-
-u32 D_8003AA38 = 0xE8321C00;
-
-u32 D_8003AA3C = 0x97390380;
-
-u32 D_8003AA40 = 0x97180380;
-
-u32 D_8003AA44 = 0x24017070;
-
-u32 D_8003AA48 = 0x8F2B0024;
-
-u32 D_8003AA4C = 0x002B0824;
-
-u32 D_8003AA50 = 0x1020FD52;
-
-u32 D_8003AA54 = 0x8F2B004C;
-
-u32 D_8003AA58 = 0x1738FFFC;
-
-u32 D_8003AA5C = 0x27390028;
-
-u32 D_8003AA60 = 0x0800047F;
-
-u32 D_8003AA64 = 0x97390380;
-
-u32 D_8003AA68 = 0x8F39001C;
-
-u32 D_8003AA6C = 0x03381022;
-
-u32 D_8003AA70 = 0x0441FD4A;
-
-u32 D_8003AA74 = 0x8C1800D8;
-
-u32 D_8003AA78 = 0x08000402;
-
-u32 D_8003AA7C = 0x936109C1;
-
-u32 D_8003AA80 = 0x08000410;
-
-u32 D_8003AA84 = 0x97390380;
-
-u32 D_8003AA88 = 0x00000000;
-
-u32 D_8003AA8C = 0x00000000;
-
-u32 D_8003AA90 = 0x00000000;
-
-u32 D_8003AA94 = 0x00000000;
-
-u32 D_8003AA98 = 0x00000000;
-
-u32 D_8003AA9C = 0x00000000;
-
-u32 D_8003AAA0 = 0x00000000;
-
-u32 D_8003AAA4 = 0x00000000;
-
-u32 D_8003AAA8 = 0x00000000;
-
-u32 D_8003AAAC = 0x00000000;
-
-u32 D_8003AAB0 = 0x00000000;
-
-u32 D_8003AAB4 = 0x00000000;
-
-u32 D_8003AAB8 = 0x00000000;
-
-u32 D_8003AABC = 0x00000000;
-
-u32 D_8003AAC0 = 0x00000000;
-
-u32 D_8003AAC4 = 0x00000000;
-
-u32 D_8003AAC8 = 0x00000000;
-
-u32 D_8003AACC = 0x00000000;
-
-u32 D_8003AAD0 = 0x00000000;
-
-u32 D_8003AAD4 = 0x00000000;
-
-u32 D_8003AAD8 = 0x00000000;
-
-u32 D_8003AADC = 0x00000000;
-
-u32 D_8003AAE0 = 0x00000000;
-
-u32 D_8003AAE4 = 0x00000000;
-
-u32 D_8003AAE8 = 0x00000000;
-
-u32 D_8003AAEC = 0x00000000;
-
-u32 D_8003AAF0 = 0x00000000;
-
-u32 D_8003AAF4 = 0x00000000;
-
-u32 D_8003AAF8 = 0x00000000;
-
-u32 D_8003AAFC = 0x00000000;
-
-u32 D_8003AB00 = 0x00000000;
-
-u32 D_8003AB04 = 0x00000000;
-
-u32 D_8003AB08 = 0x00000000;
-
-u32 D_8003AB0C = 0x00000000;
-
-u32 D_8003AB10 = 0x00000000;
-
-u32 D_8003AB14 = 0x00000000;
-
-u32 D_8003AB18 = 0x00000000;
-
-u32 D_8003AB1C = 0x00000000;
-
-u32 D_8003AB20 = 0x00000000;
-
-u32 D_8003AB24 = 0x00000000;
-
-u32 D_8003AB28 = 0x00000000;
-
-u32 D_8003AB2C = 0x00000000;
-
-u32 D_8003AB30 = 0x00000000;
-
-u32 D_8003AB34 = 0x00000000;
-
-u32 D_8003AB38 = 0x00000000;
-
-u32 D_8003AB3C = 0x00000000;
-
-u32 D_8003AB40 = 0x00000000;
-
-u32 D_8003AB44 = 0x00000000;
-
-u32 D_8003AB48 = 0x00000000;
-
-u32 D_8003AB4C = 0x00000000;
-
-u32 D_8003AB50 = 0x00000000;
-
-u32 D_8003AB54 = 0x00000000;
-
-u32 D_8003AB58 = 0x00000000;
-
-u32 D_8003AB5C = 0x00000000;
-
-u32 D_8003AB60 = 0x00000000;
-
-u32 D_8003AB64 = 0x00000000;
-
-u32 D_8003AB68 = 0x00000000;
-
-u32 D_8003AB6C = 0x00000000;
-
-u32 D_8003AB70 = 0x00000000;
-
-u32 D_8003AB74 = 0x00000000;
-
-u32 D_8003AB78 = 0x00000000;
-
-u32 D_8003AB7C = 0x00000000;
-
-u32 D_8003AB80 = 0x00000000;
-
-u32 D_8003AB84 = 0x00000000;
-
-u32 D_8003AB88 = 0x00000000;
-
-u32 D_8003AB8C = 0x00000000;
-
-u32 D_8003AB90 = 0x00000000;
-
-u32 D_8003AB94 = 0x00000000;
-
-u32 D_8003AB98 = 0x00000000;
-
-u32 D_8003AB9C = 0x00000000;
-
-u32 D_8003ABA0 = 0x00000000;
-
-u32 D_8003ABA4 = 0x00000000;
-
-u32 D_8003ABA8 = 0x00000000;
-
-u32 D_8003ABAC = 0x00000000;
-
-u32 D_8003ABB0 = 0x00000000;
-
-u32 D_8003ABB4 = 0x00000000;
-
-u32 D_8003ABB8 = 0x00000000;
-
-u32 D_8003ABBC = 0x00000000;
-
-u32 D_8003ABC0 = 0x00000000;
-
-u32 D_8003ABC4 = 0x00000000;
-
-u32 D_8003ABC8 = 0x00000000;
-
-u32 D_8003ABCC = 0x00000000;
-
-u32 D_8003ABD0 = 0x00000000;
-
-u32 D_8003ABD4 = 0x00000000;
-
-u32 D_8003ABD8 = 0x00000000;
-
-u32 D_8003ABDC = 0x00000000;
-
-u32 D_8003ABE0 = 0x00000000;
-
-u32 D_8003ABE4 = 0x00000000;
-
-u32 D_8003ABE8 = 0x00000000;
-
-u32 D_8003ABEC = 0x00000000;
-
-u32 D_8003ABF0 = 0x00000000;
-
-u32 D_8003ABF4 = 0x00000000;
-
-u32 D_8003ABF8 = 0x00000000;
-
-u32 D_8003ABFC = 0x00000000;
-
-u32 D_8003AC00 = 0x00000000;
-
-u32 D_8003AC04 = 0x00000000;
-
-u32 D_8003AC08 = 0x00000000;
-
-u32 D_8003AC0C = 0x00000000;
-
-u32 D_8003AC10 = 0x00000000;
-
-u32 D_8003AC14 = 0x00000000;
-
-u32 D_8003AC18 = 0x00000000;
-
-u32 D_8003AC1C = 0x00000000;
-
-u32 D_8003AC20 = 0x00000000;
-
-u32 D_8003AC24 = 0x00000000;
-
-u32 D_8003AC28 = 0x00000000;
-
-u32 D_8003AC2C = 0x00000000;
-
-u32 D_8003AC30 = 0x00000000;
-
-u32 D_8003AC34 = 0x00000000;
-
-u32 D_8003AC38 = 0x00000000;
-
-u32 D_8003AC3C = 0x00000000;
-
-u32 D_8003AC40 = 0x00000000;
-
-u32 D_8003AC44 = 0x00000000;
-
-u32 D_8003AC48 = 0x00000000;
-
-u32 D_8003AC4C = 0x00000000;
-
-u32 D_8003AC50 = 0x00000000;
-
-u32 D_8003AC54 = 0x00000000;
-
-u32 D_8003AC58 = 0x00000000;
-
-u32 D_8003AC5C = 0x00000000;
-
-u32 D_8003AC60 = 0x00000000;
-
-u32 D_8003AC64 = 0x00000000;
-
-u32 D_8003AC68 = 0x00000000;
-
-u32 D_8003AC6C = 0x00000000;
-
-u32 D_8003AC70 = 0x00000000;
-
-u32 D_8003AC74 = 0x00000000;
-
-u32 D_8003AC78 = 0x00000000;
-
-u32 D_8003AC7C = 0x00000000;
-
-u32 D_8003AC80 = 0x00000000;
-
-u32 D_8003AC84 = 0x00000000;
-
-u32 D_8003AC88 = 0x00000000;
-
-u32 D_8003AC8C = 0x00000000;
-
-u32 D_8003AC90 = 0x00000000;
-
-u32 D_8003AC94 = 0x00000000;
-
-u32 D_8003AC98 = 0x00000000;
-
-u32 D_8003AC9C = 0x00000000;
-
-u32 D_8003ACA0 = 0x00000000;
-
-u32 D_8003ACA4 = 0x00000000;
-
-u32 D_8003ACA8 = 0x00000000;
-
-u32 D_8003ACAC = 0x00000000;
-
-u32 D_8003ACB0 = 0x00000000;
-
-u32 D_8003ACB4 = 0x00000000;
-
-u32 D_8003ACB8 = 0x00000000;
-
-u32 D_8003ACBC = 0x00000000;
-
-u32 D_8003ACC0 = 0x00000000;
-
-u32 D_8003ACC4 = 0x00000000;
-
-u32 D_8003ACC8 = 0x00000000;
-
-u32 D_8003ACCC = 0x00000000;
-
-u32 D_8003ACD0 = 0x00000000;
-
-u32 D_8003ACD4 = 0x00000000;
-
-u32 D_8003ACD8 = 0x00000000;
-
-u32 D_8003ACDC = 0x00000000;
-
-u32 D_8003ACE0 = 0x00000000;
-
-u32 D_8003ACE4 = 0x00000000;
-
-u32 D_8003ACE8 = 0x00000000;
-
-u32 D_8003ACEC = 0x00000000;
-
-u32 D_8003ACF0 = 0x00000000;
-
-u32 D_8003ACF4 = 0x00000000;
-
-u32 D_8003ACF8 = 0x00000000;
-
-u32 D_8003ACFC = 0x00000000;
-
-u32 D_8003AD00 = 0x00000000;
-
-u32 D_8003AD04 = 0x00000000;
-
-u32 D_8003AD08 = 0x00000000;
-
-u32 D_8003AD0C = 0x00000000;
-
-u32 D_8003AD10 = 0x00000000;
-
-u32 D_8003AD14 = 0x00000000;
-
-u32 D_8003AD18 = 0x00000000;
-
-u32 D_8003AD1C = 0x00000000;
-
-u32 D_8003AD20 = 0x00000000;
-
-u32 D_8003AD24 = 0x00000000;
-
-u32 D_8003AD28 = 0x00000000;
-
-u32 D_8003AD2C = 0x00000000;
-
-u32 D_8003AD30 = 0x00000000;
-
-u32 D_8003AD34 = 0x00000000;
-
-u32 D_8003AD38 = 0x00000000;
-
-u32 D_8003AD3C = 0x00000000;
-
-u32 D_8003AD40 = 0x00000000;
-
-u32 D_8003AD44 = 0x00000000;
-
-u32 D_8003AD48 = 0x00000000;
-
-u32 D_8003AD4C = 0x00000000;
-
-u32 D_8003AD50 = 0x00000000;
-
-u32 D_8003AD54 = 0x00000000;
-
-u32 D_8003AD58 = 0x00000000;
-
-u32 D_8003AD5C = 0x00000000;
-
-u32 D_8003AD60 = 0x00000000;
-
-u32 D_8003AD64 = 0x00000000;
-
-u32 D_8003AD68 = 0x00000000;
-
-u32 D_8003AD6C = 0x00000000;
-
-u32 D_8003AD70 = 0x00000000;
-
-u32 D_8003AD74 = 0x00000000;
-
-u32 D_8003AD78 = 0x00000000;
-
-u32 D_8003AD7C = 0x00000000;
-
-u32 D_8003AD80 = 0x00000000;
-
-u32 D_8003AD84 = 0x00000000;
-
-u32 D_8003AD88 = 0x00000000;
-
-u32 D_8003AD8C = 0x00000000;
-
-u32 D_8003AD90 = 0x00000000;
-
-u32 D_8003AD94 = 0x00000000;
-
-u32 D_8003AD98 = 0x00000000;
-
-u32 D_8003AD9C = 0x00000000;
-
-u32 D_8003ADA0 = 0x00000000;
-
-u32 D_8003ADA4 = 0x00000000;
-
-u32 D_8003ADA8 = 0x00000000;
-
-u32 D_8003ADAC = 0x00000000;
-
-u32 D_8003ADB0 = 0x00000000;
-
-u32 D_8003ADB4 = 0x00000000;
-
-u32 D_8003ADB8 = 0x00000000;
-
-u32 D_8003ADBC = 0x200C1000;
-
-u32 D_8003ADC0 = 0x200B02E0;
-
-u32 D_8003ADC4 = 0x8D780000;
-
-u32 D_8003ADC8 = 0x95730004;
-
-u32 D_8003ADCC = 0x0C0007F6;
-
-u32 D_8003ADD0 = 0x95740006;
-
-u32 D_8003ADD4 = 0x359F0000;
-
-u32 D_8003ADD8 = 0x400B3000;
-
-u32 D_8003ADDC = 0x1560FFFF;
-
-u32 D_8003ADE0 = 0x400B3000;
-
-u32 D_8003ADE4 = 0x03E00008;
-
-u32 D_8003ADE8 = 0x400B2800;
-
-u32 D_8003ADEC = 0x1560FFFF;
-
-u32 D_8003ADF0 = 0x400B2800;
-
-u32 D_8003ADF4 = 0x40940000;
-
-u32 D_8003ADF8 = 0x06800003;
-
-u32 D_8003ADFC = 0x40980800;
-
-u32 D_8003AE00 = 0x03E00008;
-
-u32 D_8003AE04 = 0x40931000;
-
-u32 D_8003AE08 = 0x03E00008;
-
-u32 D_8003AE0C = 0x40931800;
-
-u32 D_8003AE10 = 0x02F65822;
-
-u32 D_8003AE14 = 0x256C0157;
-
-u32 D_8003AE18 = 0x0591009A;
-
-u32 D_8003AE1C = 0x00000000;
-
-u32 D_8003AE20 = 0x0C0007F2;
-
-u32 D_8003AE24 = 0x8C1800F0;
-
-u32 D_8003AE28 = 0x0420001A;
-
-u32 D_8003AE2C = 0x40984800;
-
-u32 D_8003AE30 = 0x1420000F;
-
-u32 D_8003AE34 = 0x035BD020;
-
-u32 D_8003AE38 = 0x8F7809C4;
-
-u32 D_8003AE3C = 0xAC1A0FF0;
-
-u32 D_8003AE40 = 0xAC180FD0;
-
-u32 D_8003AE44 = 0x24141080;
-
-u32 D_8003AE48 = 0x0C0007F6;
-
-u32 D_8003AE4C = 0x20130F47;
-
-u32 D_8003AE50 = 0x8C1800D8;
-
-u32 D_8003AE54 = 0x24140180;
-
-u32 D_8003AE58 = 0x33330FFF;
-
-u32 D_8003AE5C = 0x0314C020;
-
-u32 D_8003AE60 = 0x0C0007F6;
-
-u32 D_8003AE64 = 0x02749822;
-
-u32 D_8003AE68 = 0x080007F2;
-
-u32 D_8003AE6C = 0x201F1080;
-
-u32 D_8003AE70 = 0x8C0B0FD0;
-
-u32 D_8003AE74 = 0xAC1A0BF8;
-
-u32 D_8003AE78 = 0xAC0B0BFC;
-
-u32 D_8003AE7C = 0x200C5000;
-
-u32 D_8003AE80 = 0x8C180FF8;
-
-u32 D_8003AE84 = 0x20148000;
-
-u32 D_8003AE88 = 0x20130BFF;
-
-u32 D_8003AE8C = 0x080007F6;
-
-u32 D_8003AE90 = 0x201F1088;
-
-u32 D_8003AE94 = 0x200C4000;
-
-u32 D_8003AE98 = 0x408C2000;
-
-u32 D_8003AE9C = 0x0000000D;
-
-u32 D_8003AEA0 = 0x00000000;
-
-u32 D_8003AEA4 = 0x00000000;
-
-u32 D_8003AEA8 = 0x900100DE;
-
-u32 D_8003AEAC = 0x001913C0;
-
-u32 D_8003AEB0 = 0x0C00048D;
-
-u32 D_8003AEB4 = 0x035B1820;
-
-u32 D_8003AEB8 = 0x04400057;
-
-u32 D_8003AEBC = 0x371A0000;
-
-u32 D_8003AEC0 = 0xAC230138;
-
-u32 D_8003AEC4 = 0x20210004;
-
-u32 D_8003AEC8 = 0x0800045C;
-
-u32 D_8003AECC = 0xA00100DE;
-
-u32 D_8003AED0 = 0x200B1140;
-
-u32 D_8003AED4 = 0xAD79F0A4;
-
-u32 D_8003AED8 = 0x08000463;
-
-u32 D_8003AEDC = 0xAD78F0A8;
-
-u32 D_8003AEE0 = 0x00191402;
-
-u32 D_8003AEE4 = 0x944127FE;
-
-u32 D_8003AEE8 = 0x00390820;
-
-u32 D_8003AEEC = 0x08000463;
-
-u32 D_8003AEF0 = 0xAC380000;
-
-u32 D_8003AEF4 = 0x8C0B00F4;
-
-u32 D_8003AEF8 = 0x8C020FE0;
-
-u32 D_8003AEFC = 0x0178C022;
-
-u32 D_8003AF00 = 0x03020822;
-
-u32 D_8003AF04 = 0x04210002;
-
-u32 D_8003AF08 = 0x00000000;
-
-u32 D_8003AF0C = 0x34580000;
-
-u32 D_8003AF10 = 0x130B0048;
-
-u32 D_8003AF14 = 0xAC1800F4;
-
-u32 D_8003AF18 = 0x08000449;
-
-u32 D_8003AF1C = 0xAC0001D9;
-
-u32 D_8003AF20 = 0x943302F2;
-
-u32 D_8003AF24 = 0x0C0007F2;
-
-u32 D_8003AF28 = 0x943502F2;
-
-u32 D_8003AF2C = 0x201F118C;
-
-u32 D_8003AF30 = 0x228C0018;
-
-u32 D_8003AF34 = 0x4B00024E;
-
-u32 D_8003AF38 = 0x228B0008;
-
-u32 D_8003AF3C = 0x4B00020F;
-
-u32 D_8003AF40 = 0x22B5FFE0;
-
-u32 D_8003AF44 = 0x4B000747;
-
-u32 D_8003AF48 = 0xCAA51808;
-
-u32 D_8003AF4C = 0xCAA51C08;
-
-u32 D_8003AF50 = 0xCA832002;
-
-u32 D_8003AF54 = 0xCAA41804;
-
-u32 D_8003AF58 = 0xCAA41C04;
-
-u32 D_8003AF5C = 0xCA822000;
-
-u32 D_8003AF60 = 0x4A832F4C;
-
-u32 D_8003AF64 = 0x22940002;
-
-u32 D_8003AF68 = 0x4A83274D;
-
-u32 D_8003AF6C = 0x22B50008;
-
-u32 D_8003AF70 = 0x4A8229CE;
-
-u32 D_8003AF74 = 0x168BFFF4;
-
-u32 D_8003AF78 = 0x4A82218F;
-
-u32 D_8003AF7C = 0x168CFFED;
-
-u32 D_8003AF80 = 0x22940008;
-
-u32 D_8003AF84 = 0xEA692002;
-
-u32 D_8003AF88 = 0xEA682000;
-
-u32 D_8003AF8C = 0xEA672003;
-
-u32 D_8003AF90 = 0x03E00008;
-
-u32 D_8003AF94 = 0xEA662001;
-
-u32 D_8003AF98 = 0x332B0005;
-
-u32 D_8003AF9C = 0x15600008;
-
-u32 D_8003AFA0 = 0x33220002;
-
-u32 D_8003AFA4 = 0x8C1800F4;
-
-u32 D_8003AFA8 = 0x2014E000;
-
-u32 D_8003AFAC = 0x0C0007F6;
-
-u32 D_8003AFB0 = 0x2013003F;
-
-u32 D_8003AFB4 = 0x23180040;
-
-u32 D_8003AFB8 = 0xAC1800F4;
-
-u32 D_8003AFBC = 0x8F7809C4;
-
-u32 D_8003AFC0 = 0x01826020;
-
-u32 D_8003AFC4 = 0xAC0001D9;
-
-u32 D_8003AFC8 = 0x0C00048D;
-
-u32 D_8003AFCC = 0x332100FE;
-
-u32 D_8003AFD0 = 0x937309C1;
-
-u32 D_8003AFD4 = 0x943402F0;
-
-u32 D_8003AFD8 = 0x00191142;
-
-u32 D_8003AFDC = 0x959F0336;
-
-u32 D_8003AFE0 = 0x080007F6;
-
-u32 D_8003AFE4 = 0x0282A020;
-
-u32 D_8003AFE8 = 0x8D63EF8C;
-
-u32 D_8003AFEC = 0x3C028000;
-
-u32 D_8003AFF0 = 0x03221007;
-
-u32 D_8003AFF4 = 0x00190A02;
-
-u32 D_8003AFF8 = 0x00221006;
-
-u32 D_8003AFFC = 0x00401027;
-
-u32 D_8003B000 = 0x00621824;
-
-u32 D_8003B004 = 0x00781825;
-
-u32 D_8003B008 = 0xAD63EF8C;
-
-u32 D_8003B00C = 0x8C1900C8;
-
-u32 D_8003B010 = 0x08000488;
-
-u32 D_8003B014 = 0x8C1800CC;
-
-u32 D_8003B018 = 0x900B01DC;
-
-u32 D_8003B01C = 0x080004C1;
-
-u32 D_8003B020 = 0x900601DD;
-
-u32 D_8003B024 = 0x37FE0000;
-
-u32 D_8003B028 = 0x200B0418;
-
-u32 D_8003B02C = 0x080007ED;
-
-u32 D_8003B030 = 0x200C12F8;
-
-u32 D_8003B034 = 0x1560008D;
-
-u32 D_8003B038 = 0x20C60168;
-
-u32 D_8003B03C = 0xA01901DC;
-
-u32 D_8003B040 = 0xC80C2002;
-
-u32 D_8003B044 = 0xC8082000;
-
-u32 D_8003B048 = 0xC80D0915;
-
-u32 D_8003B04C = 0xC8090905;
-
-u32 D_8003B050 = 0x4B2C4373;
-
-u32 D_8003B054 = 0xC80E0A1A;
-
-u32 D_8003B058 = 0x4B284273;
-
-u32 D_8003B05C = 0xC80A0A0A;
-
-u32 D_8003B060 = 0x4B4C43B3;
-
-u32 D_8003B064 = 0x20140150;
-
-u32 D_8003B068 = 0x4B4842B3;
-
-u32 D_8003B06C = 0xCA873015;
-
-u32 D_8003B070 = 0x4BCC4BB3;
-
-u32 D_8003B074 = 0xC80D0A19;
-
-u32 D_8003B078 = 0x4BC84AB3;
-
-u32 D_8003B07C = 0xC8090A09;
-
-u32 D_8003B080 = 0x4B8C4B33;
-
-u32 D_8003B084 = 0xC80C0A18;
-
-u32 D_8003B088 = 0x4B884A33;
-
-u32 D_8003B08C = 0xC8080A08;
-
-u32 D_8003B090 = 0x4B276F46;
-
-u32 D_8003B094 = 0x4B274F4F;
-
-u32 D_8003B098 = 0x4B07674E;
-
-u32 D_8003B09C = 0xEA8F3016;
-
-u32 D_8003B0A0 = 0x4B07474F;
-
-u32 D_8003B0A4 = 0x8E8C00B0;
-
-u32 D_8003B0A8 = 0x4B47774E;
-
-u32 D_8003B0AC = 0x4B47574F;
-
-u32 D_8003B0B0 = 0x4B2B5ADD;
-
-u32 D_8003B0B4 = 0xAE8C00B4;
-
-u32 D_8003B0B8 = 0x4B0F7BDD;
-
-u32 D_8003B0BC = 0x1286006B;
-
-u32 D_8003B0C0 = 0x4A0B5F44;
-
-u32 D_8003B0C4 = 0x4A0B7F4D;
-
-u32 D_8003B0C8 = 0x4A0F5C0E;
-
-u32 D_8003B0CC = 0x11600002;
-
-u32 D_8003B0D0 = 0x4A0F7C4F;
-
-u32 D_8003B0D4 = 0x22940018;
-
-u32 D_8003B0D8 = 0x4B308494;
-
-u32 D_8003B0DC = 0x200B0001;
-
-u32 D_8003B0E0 = 0x4B318F50;
-
-u32 D_8003B0E4 = 0x4B509414;
-
-u32 D_8003B0E8 = 0x4B51EC50;
-
-u32 D_8003B0EC = 0x4B114776;
-
-u32 D_8003B0F0 = 0xCA873018;
-
-u32 D_8003B0F4 = 0x4B104435;
-
-u32 D_8003B0F8 = 0x4B004476;
-
-u32 D_8003B0FC = 0x4B105F44;
-
-u32 D_8003B100 = 0x4B107F4D;
-
-u32 D_8003B104 = 0x4B115ACE;
-
-u32 D_8003B108 = 0x4B117BCF;
-
-u32 D_8003B10C = 0x4BFE5AC6;
-
-u32 D_8003B110 = 0x080004D8;
-
-u32 D_8003B114 = 0x4BFE7BCF;
-
-u32 D_8003B118 = 0x4AA70190;
-
-u32 D_8003B11C = 0xC9343013;
-
-u32 D_8003B120 = 0x4AC70150;
-
-u32 D_8003B124 = 0xC9DB3801;
-
-u32 D_8003B128 = 0x4AFFF922;
-
-u32 D_8003B12C = 0xC93D3817;
-
-u32 D_8003B130 = 0x4A823D41;
-
-u32 D_8003B134 = 0xC9243814;
-
-u32 D_8003B138 = 0x4AA23549;
-
-u32 D_8003B13C = 0x112D002E;
-
-u32 D_8003B140 = 0x4AC22D49;
-
-u32 D_8003B144 = 0x4A943F01;
-
-u32 D_8003B148 = 0xC9233811;
-
-u32 D_8003B14C = 0x4AB43709;
-
-u32 D_8003B150 = 0x212BFFE8;
-
-u32 D_8003B154 = 0x4AD42F09;
-
-u32 D_8003B158 = 0x2129FFD0;
-
-u32 D_8003B15C = 0x4A1BEF67;
-
-u32 D_8003B160 = 0x48802300;
-
-u32 D_8003B164 = 0x4B0018E7;
-
-u32 D_8003B168 = 0x48802700;
-
-u32 D_8003B16C = 0x4BFFAD68;
-
-u32 D_8003B170 = 0xC9223016;
-
-u32 D_8003B174 = 0x4BFFE728;
-
-u32 D_8003B178 = 0xC9343013;
-
-u32 D_8003B17C = 0x4BFFEF40;
-
-u32 D_8003B180 = 0x4A952748;
-
-u32 D_8003B184 = 0x156DFFEA;
-
-u32 D_8003B188 = 0x4A9C1F48;
-
-u32 D_8003B18C = 0x4BBF00E7;
-
-u32 D_8003B190 = 0xC9D61206;
-
-u32 D_8003B194 = 0x4B7FCEE3;
-
-u32 D_8003B198 = 0x30AB0004;
-
-u32 D_8003B19C = 0x4A823D40;
-
-u32 D_8003B1A0 = 0x1160005F;
-
-u32 D_8003B1A4 = 0xE9DD3801;
-
-u32 D_8003B1A8 = 0x4AA23548;
-
-u32 D_8003B1AC = 0x30AC0008;
-
-u32 D_8003B1B0 = 0x4AC22D48;
-
-u32 D_8003B1B4 = 0x4BBF192C;
-
-u32 D_8003B1B8 = 0x4A943F00;
-
-u32 D_8003B1BC = 0x4AB43708;
-
-u32 D_8003B1C0 = 0x4AD42F08;
-
-u32 D_8003B1C4 = 0xC802201D;
-
-u32 D_8003B1C8 = 0x4BBF0D87;
-
-u32 D_8003B1CC = 0x4A951D88;
-
-u32 D_8003B1D0 = 0x11800053;
-
-u32 D_8003B1D4 = 0x4A9C2588;
-
-u32 D_8003B1D8 = 0x4B020D8F;
-
-u32 D_8003B1DC = 0x4A16B100;
-
-u32 D_8003B1E0 = 0x4BFFB0C0;
-
-u32 D_8003B1E4 = 0x4B42B0C8;
-
-u32 D_8003B1E8 = 0x4BBF0D87;
-
-u32 D_8003B1EC = 0x4B22B588;
-
-u32 D_8003B1F0 = 0x0800057C;
-
-u32 D_8003B1F4 = 0x4A032588;
-
-u32 D_8003B1F8 = 0x4A1BEF67;
-
-u32 D_8003B1FC = 0x4B002127;
-
-u32 D_8003B200 = 0x4BFFAD68;
-
-u32 D_8003B204 = 0x4AFFF8E1;
-
-u32 D_8003B208 = 0xC9223010;
-
-u32 D_8003B20C = 0x4BBF00E7;
-
-u32 D_8003B210 = 0xC9D61206;
-
-u32 D_8003B214 = 0x4BFFEF40;
-
-u32 D_8003B218 = 0x08000519;
-
-u32 D_8003B21C = 0x4A952748;
-
-
-
-// actual F3DEX2 start?
-
-u32 D_8003B220 = 0x4A00002C;
-
-u32 D_8003B224 = 0xC81F201B;
-
-u32 D_8003B228 = 0xC81E201C;
-
-u32 D_8003B22C = 0x20170A80;
-
-u32 D_8003B230 = 0x20160C68;
-
-u32 D_8003B234 = 0x4B1F0051;
-
-u32 D_8003B238 = 0x8C0B00F0;
-
-u32 D_8003B23C = 0x8C0C0FC4;
-
-u32 D_8003B240 = 0x20012800;
-
-u32 D_8003B244 = 0x11600006;
-
-u32 D_8003B248 = 0x40812000;
-
-u32 D_8003B24C = 0x318C0001;
-
-u32 D_8003B250 = 0x11800021;
-
-u32 D_8003B254 = 0xAC000FC4;
-
-u32 D_8003B258 = 0x08000459;
-
-u32 D_8003B25C = 0x8C1A0BF8;
-
-u32 D_8003B260 = 0x400B5800;
-
-u32 D_8003B264 = 0x316B0001;
-
-u32 D_8003B268 = 0x1560000B;
-
-u32 D_8003B26C = 0x40024800;
-
-u32 D_8003B270 = 0x8C030FE8;
-
-u32 D_8003B274 = 0x00625822;
-
-u32 D_8003B278 = 0x1D600007;
-
-u32 D_8003B27C = 0x40015000;
-
-u32 D_8003B280 = 0x8C040FEC;
-
-u32 D_8003B284 = 0x10200004;
-
-u32 D_8003B288 = 0x00245822;
-
-u32 D_8003B28C = 0x05610002;
-
-u32 D_8003B290 = 0x00000000;
-
-u32 D_8003B294 = 0x14220008;
-
-u32 D_8003B298 = 0x400B5800;
-
-u32 D_8003B29C = 0x316B0400;
-
-u32 D_8003B2A0 = 0x1560FFFD;
-
-u32 D_8003B2A4 = 0x200B0001;
-
-u32 D_8003B2A8 = 0x408B5800;
-
-u32 D_8003B2AC = 0x8C020FEC;
-
-u32 D_8003B2B0 = 0x40824000;
-
-u32 D_8003B2B4 = 0x40824800;
-
-u32 D_8003B2B8 = 0xAC0200F0;
-
-u32 D_8003B2BC = 0x00000000;
-
-u32 D_8003B2C0 = 0x00000000;
-
-u32 D_8003B2C4 = 0x00000000;
-
-u32 D_8003B2C8 = 0x8C0B00F4;
-
-u32 D_8003B2CC = 0x15600002;
-
-u32 D_8003B2D0 = 0x8C0B0FE0;
-
-u32 D_8003B2D4 = 0xAC0B00F4;
-
-u32 D_8003B2D8 = 0x8C010FD0;
-
-u32 D_8003B2DC = 0x8C0202E0;
-
-u32 D_8003B2E0 = 0x8C0302E8;
-
-u32 D_8003B2E4 = 0x00411020;
-
-u32 D_8003B2E8 = 0x00611820;
-
-u32 D_8003B2EC = 0xAC0202E0;
-
-u32 D_8003B2F0 = 0xAC0302E8;
-
-u32 D_8003B2F4 = 0x8C1900C0;
-
-u32 D_8003B2F8 = 0x0C00049C;
-
-u32 D_8003B2FC = 0x8C1800C4;
-
-u32 D_8003B300 = 0x8C1A0FF0;
-
-u32 D_8003B304 = 0x200B02E8;
-
-u32 D_8003B308 = 0x0C0007ED;
-
-u32 D_8003B30C = 0x37EC0000;
-
-u32 D_8003B310 = 0x201300A7;
-
-u32 D_8003B314 = 0x37580000;
-
-u32 D_8003B318 = 0x0C0007F6;
-
-u32 D_8003B31C = 0x241408E8;
-
-u32 D_8003B320 = 0x275A00A8;
-
-u32 D_8003B324 = 0x201BFF58;
-
-u32 D_8003B328 = 0x0C0007F2;
-
-u32 D_8003B32C = 0x40012000;
-
-u32 D_8003B330 = 0x8F790990;
-
-u32 D_8003B334 = 0x1360FFF6;
-
-u32 D_8003B338 = 0x30210080;
-
-u32 D_8003B33C = 0x00196603;
-
-u32 D_8003B340 = 0x000C5840;
-
-u32 D_8003B344 = 0x956B036E;
-
-u32 D_8003B348 = 0x14200380;
-
-u32 D_8003B34C = 0x8F780994;
-
-u32 D_8003B350 = 0x01600008;
-
-u32 D_8003B354 = 0x277B0008;
-
-u32 D_8003B358 = 0x201F118C;
-
-u32 D_8003B35C = 0x20150040;
-
-u32 D_8003B360 = 0x20140000;
-
-u32 D_8003B364 = 0x20130080;
-
-u32 D_8003B368 = 0x08000422;
-
-u32 D_8003B36C = 0xA01901D9;
-
-u32 D_8003B370 = 0x0C00048F;
-
-u32 D_8003B374 = 0x87740989;
-
-u32 D_8003B378 = 0x33330FF8;
-
-u32 D_8003B37C = 0x0014A083;
-
-u32 D_8003B380 = 0x080007F6;
-
-u32 D_8003B384 = 0x201F1188;
-
-u32 D_8003B388 = 0x8C0B01EC;
-
-u32 D_8003B38C = 0x01795824;
-
-u32 D_8003B390 = 0x01785825;
-
-u32 D_8003B394 = 0x08000463;
-
-u32 D_8003B398 = 0xAC0B01EC;
-
-u32 D_8003B39C = 0x900100DE;
-
-u32 D_8003B3A0 = 0x1020036A;
-
-u32 D_8003B3A4 = 0x2021FFFC;
-
-u32 D_8003B3A8 = 0x08000408;
-
-u32 D_8003B3AC = 0x8C3A0138;
-
-u32 D_8003B3B0 = 0xC81D1818;
-
-u32 D_8003B3B4 = 0xC81D1C1A;
-
-u32 D_8003B3B8 = 0x8C1900D8;
-
-u32 D_8003B3BC = 0x22F70010;
-
-u32 D_8003B3C0 = 0xEAFD187E;
-
-u32 D_8003B3C4 = 0xEAFD1C7F;
-
-u32 D_8003B3C8 = 0xAEF80004;
-
-u32 D_8003B3CC = 0xAEF90000;
-
-u32 D_8003B3D0 = 0x080004A7;
-
-u32 D_8003B3D4 = 0x22F70008;
-
-u32 D_8003B3D8 = 0x201F1228;
-
-u32 D_8003B3DC = 0x00185D82;
-
-u32 D_8003B3E0 = 0x316B003C;
-
-u32 D_8003B3E4 = 0x8D6B00F8;
-
-u32 D_8003B3E8 = 0x0018C200;
-
-u32 D_8003B3EC = 0x0018C202;
-
-u32 D_8003B3F0 = 0x03E00008;
-
-u32 D_8003B3F4 = 0x030BC020;
-
-u32 D_8003B3F8 = 0xAC1900C8;
-
-u32 D_8003B3FC = 0x0800048A;
-
-u32 D_8003B400 = 0xAC1800CC;
-
-u32 D_8003B404 = 0xAC1900C0;
-
-u32 D_8003B408 = 0xAC1800C4;
-
-u32 D_8003B40C = 0x201F1228;
-
-u32 D_8003B410 = 0x332B0FFF;
-
-u32 D_8003B414 = 0xA40B03E2;
-
-u32 D_8003B418 = 0x00195B02;
-
-u32 D_8003B41C = 0x316B0FFF;
-
-u32 D_8003B420 = 0xA40B03E0;
-
-u32 D_8003B424 = 0x330B0FFF;
-
-u32 D_8003B428 = 0xA40B03E6;
-
-u32 D_8003B42C = 0x00185B02;
-
-u32 D_8003B430 = 0x316B0FFF;
-
-u32 D_8003B434 = 0x03E00008;
-
-u32 D_8003B438 = 0xA40B03E4;
-
-u32 D_8003B43C = 0x201F118C;
-
-u32 D_8003B440 = 0x02F65822;
-
-u32 D_8003B444 = 0x1960034B;
-
-u32 D_8003B448 = 0x400C3000;
-
-u32 D_8003B44C = 0x8C1800F0;
-
-u32 D_8003B450 = 0x257301E8;
-
-u32 D_8003B454 = 0x1580FFFC;
-
-u32 D_8003B458 = 0x8C0C0FEC;
-
-u32 D_8003B45C = 0x40984800;
-
-u32 D_8003B460 = 0x03135820;
-
-u32 D_8003B464 = 0x018B6022;
-
-u32 D_8003B468 = 0x05810008;
-
-u32 D_8003B46C = 0x400B5800;
-
-u32 D_8003B470 = 0x316B0400;
-
-u32 D_8003B474 = 0x1560FFFD;
-
-u32 D_8003B478 = 0x8C180FE8;
-
-u32 D_8003B47C = 0x400B5000;
-
-u32 D_8003B480 = 0x1178FFFE;
-
-u32 D_8003B484 = 0x00000000;
-
-u32 D_8003B488 = 0x40984000;
-
-u32 D_8003B48C = 0x400B5000;
-
-u32 D_8003B490 = 0x01785822;
-
-u32 D_8003B494 = 0x19600002;
-
-u32 D_8003B498 = 0x01735822;
-
-u32 D_8003B49C = 0x1960FFFB;
-
-u32 D_8003B4A0 = 0x03135820;
-
-u32 D_8003B4A4 = 0xAC0B00F0;
-
-u32 D_8003B4A8 = 0x2273FFFF;
-
-u32 D_8003B4AC = 0x22D4DE18;
-
-u32 D_8003B4B0 = 0x3AD60360;
-
-u32 D_8003B4B4 = 0x080007F6;
-
-u32 D_8003B4B8 = 0x22D7FE18;
-
-u32 D_8003B4BC = 0x00000000;
-
-u32 D_8003B4C0 = 0x97340380;
-
-u32 D_8003B4C4 = 0x0C00048F;
-
-u32 D_8003B4C8 = 0x97610989;
-
-u32 D_8003B4CC = 0x0281A022;
-
-u32 D_8003B4D0 = 0x0C0007F6;
-
-u32 D_8003B4D4 = 0x2033FFFF;
-
-u32 D_8003B4D8 = 0x940501EC;
-
-u32 D_8003B4DC = 0x000108C2;
-
-u32 D_8003B4E0 = 0x03217822;
-
-u32 D_8003B4E4 = 0x95EF0380;
-
-u32 D_8003B4E8 = 0x368E0000;
-
-u32 D_8003B4EC = 0x900801D9;
-
-u32 D_8003B4F0 = 0x30A70001;
-
-u32 D_8003B4F4 = 0x15000006;
-
-u32 D_8003B4F8 = 0x000738C0;
-
-u32 D_8003B4FC = 0xA01901D9;
-
-u32 D_8003B500 = 0x20150040;
-
-u32 D_8003B504 = 0x20140000;
-
-u32 D_8003B508 = 0x0C000422;
-
-u32 D_8003B50C = 0x20130080;
-
-u32 D_8003B510 = 0xC8082008;
-
-u32 D_8003B514 = 0xC80A2009;
-
-u32 D_8003B518 = 0xC80C200A;
-
-u32 D_8003B51C = 0xC80E200B;
-
-u32 D_8003B520 = 0x4B004250;
-
-u32 D_8003B524 = 0xC8091811;
-
-u32 D_8003B528 = 0x4B0052D0;
-
-u32 D_8003B52C = 0xC80B1813;
-
-u32 D_8003B530 = 0x4B006350;
-
-u32 D_8003B534 = 0xC80D1815;
-
-u32 D_8003B538 = 0x4B0073D0;
-
-u32 D_8003B53C = 0xC80F1817;
-
-u32 D_8003B540 = 0xC8081C10;
-
-u32 D_8003B544 = 0xC80A1C12;
-
-u32 D_8003B548 = 0x0C000559;
-
-u32 D_8003B54C = 0xC80C1C14;
-
-u32 D_8003B550 = 0x0C0007F2;
-
-u32 D_8003B554 = 0xC80E1C16;
-
-u32 D_8003B558 = 0xC9D41800;
-
-u32 D_8003B55C = 0x4B356C33;
-
-u32 D_8003B560 = 0xC9D41C02;
-
-u32 D_8003B564 = 0x4B017F46;
-
-u32 D_8003B568 = 0x8DCB001C;
-
-u32 D_8003B56C = 0x4B015F4F;
-
-u32 D_8003B570 = 0xC9D61602;
-
-u32 D_8003B574 = 0x4A94674E;
-
-u32 D_8003B578 = 0x4A94474F;
-
-u32 D_8003B57C = 0x4AB46F4E;
-
-u32 D_8003B580 = 0xADCB0008;
-
-u32 D_8003B584 = 0x4AB44F4F;
-
-u32 D_8003B588 = 0x4AD475CE;
-
-u32 D_8003B58C = 0x4AD4560F;
-
-u32 D_8003B590 = 0x4B7FCEE3;
-
-u32 D_8003B594 = 0xC9D61206;
-
-u32 D_8003B598 = 0x2021FFFC;
-
-u32 D_8003B59C = 0x4B92BF44;
-
-u32 D_8003B5A0 = 0x01075822;
-
-u32 D_8003B5A4 = 0x4B92C08D;
-
-u32 D_8003B5A8 = 0xE97B07F3;
-
-u32 D_8003B5AC = 0x4B00054E;
-
-u32 D_8003B5B0 = 0xE97B03CB;
-
-u32 D_8003B5B4 = 0x4BB2B9C6;
-
-u32 D_8003B5B8 = 0xE9191C7E;
-
-u32 D_8003B5BC = 0x4BB2C18F;
-
-u32 D_8003B5C0 = 0xE9191879;
-
-u32 D_8003B5C4 = 0x4B624772;
-
-u32 D_8003B5C8 = 0xE91A0E7B;
-
-u32 D_8003B5CC = 0x4B755971;
-
-u32 D_8003B5D0 = 0xE91A0A67;
-
-u32 D_8003B5D4 = 0x4BE25932;
-
-u32 D_8003B5D8 = 0xC9C31801;
-
-u32 D_8003B5DC = 0x4BF57971;
-
-u32 D_8003B5E0 = 0x00015FC3;
-
-u32 D_8003B5E4 = 0x4B007932;
-
-u32 D_8003B5E8 = 0x316B0028;
-
-u32 D_8003B5EC = 0x4AF8C765;
-
-u32 D_8003B5F0 = 0x21EF0050;
-
-u32 D_8003B5F4 = 0x4AF7BF64;
-
-u32 D_8003B5F8 = 0x01EB4022;
-
-u32 D_8003B5FC = 0x4A05AF44;
-
-u32 D_8003B600 = 0x484A0800;
-
-u32 D_8003B604 = 0x4A05174D;
-
-u32 D_8003B608 = 0xE9171C7C;
-
-u32 D_8003B60C = 0x4A04AD4E;
-
-u32 D_8003B610 = 0xC9D41804;
-
-u32 D_8003B614 = 0x4A04108F;
-
-u32 D_8003B618 = 0xE9F71877;
-
-u32 D_8003B61C = 0x4B00C763;
-
-u32 D_8003B620 = 0xC9170F72;
-
-u32 D_8003B624 = 0x4B3F0F47;
-
-u32 D_8003B628 = 0xE9181C7B;
-
-u32 D_8003B62C = 0x4B9FAE8E;
-
-u32 D_8003B630 = 0xC9F70B5E;
-
-u32 D_8003B634 = 0x4B9F164F;
-
-u32 D_8003B638 = 0xE9F81876;
-
-u32 D_8003B63C = 0x4BFF00A7;
-
-u32 D_8003B640 = 0xC9D41C06;
-
-u32 D_8003B644 = 0x4AE6C765;
-
-u32 D_8003B648 = 0xE903107A;
-
-u32 D_8003B64C = 0x4A05D744;
-
-u32 D_8003B650 = 0xC9180F6E;
-
-u32 D_8003B654 = 0x4A05CF4D;
-
-u32 D_8003B658 = 0xE9E31270;
-
-u32 D_8003B65C = 0x4A04D14E;
-
-u32 D_8003B660 = 0xC9F80B5A;
-
-u32 D_8003B664 = 0x4A04C90F;
-
-u32 D_8003B668 = 0xA50AFFFE;
-
-u32 D_8003B66C = 0x4BFF108F;
-
-u32 D_8003B670 = 0x000A5900;
-
-u32 D_8003B674 = 0x4AE7BF64;
-
-u32 D_8003B678 = 0x484A0800;
-
-u32 D_8003B67C = 0x4AE5BF44;
-
-u32 D_8003B680 = 0xE9050F7D;
-
-u32 D_8003B684 = 0x4AE5C74D;
-
-u32 D_8003B688 = 0x21CE0020;
-
-u32 D_8003B68C = 0x4AE2BE8E;
-
-u32 D_8003B690 = 0xA50AFFFC;
-
-u32 D_8003B694 = 0x4AE2C64F;
-
-u32 D_8003B698 = 0x000A5100;
-
-u32 D_8003B69C = 0x4A12B0C5;
-
-u32 D_8003B6A0 = 0xA5EBFFD6;
-
-u32 D_8003B6A4 = 0xA5EAFFD4;
-
-u32 D_8003B6A8 = 0x4B92D744;
-
-u32 D_8003B6AC = 0xE9E50B69;
-
-u32 D_8003B6B0 = 0x4B92CE4D;
-
-u32 D_8003B6B4 = 0xE9040F7C;
-
-u32 D_8003B6B8 = 0x4B00068E;
-
-u32 D_8003B6BC = 0xE9E40B68;
-
-u32 D_8003B6C0 = 0xE903127B;
-
-u32 D_8003B6C4 = 0x4B018F47;
-
-u32 D_8003B6C8 = 0xE9E31671;
-
-u32 D_8003B6CC = 0x4B7F9F4F;
-
-u32 D_8003B6D0 = 0x4A10D68E;
-
-u32 D_8003B6D4 = 0x1C20FFA3;
-
-u32 D_8003B6D8 = 0x4A10CE4F;
-
-u32 D_8003B6DC = 0x07E0009C;
-
-u32 D_8003B6E0 = 0x4B7FCEE3;
-
-u32 D_8003B6E4 = 0xE9191C7E;
-
-u32 D_8003B6E8 = 0xE9F91879;
-
-u32 D_8003B6EC = 0xE91A0E7B;
-
-u32 D_8003B6F0 = 0x10E0FF0E;
-
-u32 D_8003B6F4 = 0xE9FA0A67;
-
-u32 D_8003B6F8 = 0xE91B07EB;
-
-u32 D_8003B6FC = 0x08000463;
-
-u32 D_8003B700 = 0xE9FB03C3;
-
-u32 D_8003B704 = 0x200D0180;
-
-u32 D_8003B708 = 0xC810181C;
-
-u32 D_8003B70C = 0xC8101C1C;
-
-u32 D_8003B710 = 0xC9BD1018;
-
-u32 D_8003B714 = 0xC811181D;
-
-u32 D_8003B718 = 0xC8111C1D;
-
-u32 D_8003B71C = 0x4B7FFCE0;
-
-u32 D_8003B720 = 0x4A100551;
-
-u32 D_8003B724 = 0xC9B2121A;
-
-u32 D_8003B728 = 0x4B1D8427;
-
-u32 D_8003B72C = 0xC9B2161A;
-
-u32 D_8003B730 = 0x4B0104E7;
-
-u32 D_8003B734 = 0xC8121437;
-
-u32 D_8003B738 = 0x4B3D8C67;
-
-u32 D_8003B73C = 0xC9B20D03;
-
-u32 D_8003B740 = 0x4B354C33;
-
-u32 D_8003B744 = 0x03E00008;
-
-u32 D_8003B748 = 0x22E80050;
-
-u32 D_8003B74C = 0x9365098D;
-
-u32 D_8003B750 = 0x937D098E;
-
-u32 D_8003B754 = 0x937E098F;
-
-u32 D_8003B758 = 0x08000574;
-
-u32 D_8003B75C = 0x201915C0;
-
-u32 D_8003B760 = 0x93650989;
-
-u32 D_8003B764 = 0x937D098A;
-
-u32 D_8003B768 = 0x937E098B;
-
-u32 D_8003B76C = 0x2019118C;
-
-u32 D_8003B770 = 0x8364098C;
-
-u32 D_8003B774 = 0x94BC0380;
-
-u32 D_8003B778 = 0x97BD0380;
-
-u32 D_8003B77C = 0x97DE0380;
-
-u32 D_8003B780 = 0x94A50380;
-
-u32 D_8003B784 = 0x20840003;
-
-u32 D_8003B788 = 0x37820000;
-
-u32 D_8003B78C = 0x0C00058C;
-
-u32 D_8003B790 = 0x37A30000;
-
-u32 D_8003B794 = 0x37A20000;
-
-u32 D_8003B798 = 0x0C00058C;
-
-u32 D_8003B79C = 0x37C30000;
-
-u32 D_8003B7A0 = 0x37C20000;
-
-u32 D_8003B7A4 = 0x37830000;
-
-u32 D_8003B7A8 = 0x0800058C;
-
-u32 D_8003B7AC = 0x373F0000;
-
-u32 D_8003B7B0 = 0x8364098B;
-
-u32 D_8003B7B4 = 0x93650989;
-
-u32 D_8003B7B8 = 0x9363098A;
-
-u32 D_8003B7BC = 0x20840003;
-
-u32 D_8003B7C0 = 0x94A20380;
-
-u32 D_8003B7C4 = 0x94630380;
-
-u32 D_8003B7C8 = 0x94A50380;
-
-u32 D_8003B7CC = 0x201F118C;
-
-u32 D_8003B7D0 = 0x8C460024;
-
-u32 D_8003B7D4 = 0x8C670024;
-
-u32 D_8003B7D8 = 0x8C0C03C4;
-
-u32 D_8003B7DC = 0x30CB7070;
-
-u32 D_8003B7E0 = 0x01675824;
-
-u32 D_8003B7E4 = 0x15600263;
-
-u32 D_8003B7E8 = 0x00C75825;
-
-u32 D_8003B7EC = 0x016C5824;
-
-u32 D_8003B7F0 = 0x1160005E;
-
-u32 D_8003B7F4 = 0xA41F03C2;
-
-u32 D_8003B7F8 = 0x24060014;
-
-u32 D_8003B7FC = 0x240F0990;
-
-u32 D_8003B800 = 0x8CCB03C8;
-
-u32 D_8003B804 = 0x8C490024;
-
-u32 D_8003B808 = 0x8C700024;
-
-u32 D_8003B80C = 0x012B4824;
-
-u32 D_8003B810 = 0x020B8024;
-
-u32 D_8003B814 = 0x11300053;
-
-u32 D_8003B818 = 0x00065840;
-
-u32 D_8003B81C = 0x11200003;
-
-u32 D_8003B820 = 0x344C0000;
-
-u32 D_8003B824 = 0x34620000;
-
-u32 D_8003B828 = 0x35830000;
-
-u32 D_8003B82C = 0xC9621830;
-
-u32 D_8003B830 = 0xC8441801;
-
-u32 D_8003B834 = 0xC8451800;
-
-u32 D_8003B838 = 0xC8661801;
-
-u32 D_8003B83C = 0xC8671800;
-
-u32 D_8003B840 = 0x4B1F10C7;
-
-u32 D_8003B844 = 0x4A022206;
-
-u32 D_8003B848 = 0x4A022A4F;
-
-u32 D_8003B84C = 0x4A03328E;
-
-u32 D_8003B850 = 0x4A033ACF;
-
-u32 D_8003B854 = 0x4A484214;
-
-u32 D_8003B858 = 0xC802201D;
-
-u32 D_8003B85C = 0x4A494A50;
-
-u32 D_8003B860 = 0x4A4A5294;
-
-u32 D_8003B864 = 0x4A4B5AD0;
-
-u32 D_8003B868 = 0x4AA84214;
-
-u32 D_8003B86C = 0x4AA94A50;
-
-u32 D_8003B870 = 0x4AAA5294;
-
-u32 D_8003B874 = 0x4AAB5AD0;
-
-u32 D_8003B878 = 0x4B6B4772;
-
-u32 D_8003B87C = 0x4B6A58F1;
-
-u32 D_8003B880 = 0x4B005B32;
-
-u32 D_8003B884 = 0x4B625F53;
-
-u32 D_8003B888 = 0x4B7D18C6;
-
-u32 D_8003B88C = 0x4B7D630F;
-
-u32 D_8003B890 = 0x4B006321;
-
-u32 D_8003B894 = 0x4B1F18E7;
-
-u32 D_8003B898 = 0x4B635744;
-
-u32 D_8003B89C = 0x4B635ACD;
-
-u32 D_8003B8A0 = 0x4B00028E;
-
-u32 D_8003B8A4 = 0x4B6B5B72;
-
-u32 D_8003B8A8 = 0x4B6A5B31;
-
-u32 D_8003B8AC = 0x4B005B72;
-
-u32 D_8003B8B0 = 0x4B9F6386;
-
-u32 D_8003B8B4 = 0x4B9F6BCF;
-
-u32 D_8003B8B8 = 0x4B3F0F47;
-
-u32 D_8003B8BC = 0x4A0A774C;
-
-u32 D_8003B8C0 = 0x4A0A7F4D;
-
-u32 D_8003B8C4 = 0x4A0B728E;
-
-u32 D_8003B8C8 = 0x4A0B7ACF;
-
-u32 D_8003B8CC = 0x4A0A6744;
-
-u32 D_8003B8D0 = 0x4A0A6F4D;
-
-u32 D_8003B8D4 = 0x4A0B630E;
-
-u32 D_8003B8D8 = 0x4A0B6B4F;
-
-u32 D_8003B8DC = 0x4A0C4744;
-
-u32 D_8003B8E0 = 0xC86E3802;
-
-u32 D_8003B8E4 = 0x4A0C4F4D;
-
-u32 D_8003B8E8 = 0xC86E1405;
-
-u32 D_8003B8EC = 0x4A0D408E;
-
-u32 D_8003B8F0 = 0xC84B3802;
-
-u32 D_8003B8F4 = 0x4A0D4A8F;
-
-u32 D_8003B8F8 = 0xC84B1405;
-
-u32 D_8003B8FC = 0x4B631744;
-
-u32 D_8003B900 = 0x4B63528D;
-
-u32 D_8003B904 = 0x4B00108E;
-
-u32 D_8003B908 = 0x4B0152A0;
-
-u32 D_8003B90C = 0x4B1F10A7;
-
-u32 D_8003B910 = 0x4B011755;
-
-u32 D_8003B914 = 0x4B0052A3;
-
-u32 D_8003B918 = 0x4B0110A7;
-
-u32 D_8003B91C = 0x4B1F10C6;
-
-u32 D_8003B920 = 0x4B623744;
-
-u32 D_8003B924 = 0x4B623F4D;
-
-u32 D_8003B928 = 0x4B63274C;
-
-u32 D_8003B92C = 0x4B632E0D;
-
-u32 D_8003B930 = 0x4B0005CE;
-
-u32 D_8003B934 = 0x4B627745;
-
-u32 D_8003B938 = 0x4B635D8D;
-
-u32 D_8003B93C = 0x20070000;
-
-u32 D_8003B940 = 0x20010002;
-
-u32 D_8003B944 = 0x35E30000;
-
-u32 D_8003B948 = 0x08000559;
-
-u32 D_8003B94C = 0x201F93F8;
-
-u32 D_8003B950 = 0xE9F91879;
-
-u32 D_8003B954 = 0xE9FA0A67;
-
-u32 D_8003B958 = 0xE9F63878;
-
-u32 D_8003B95C = 0xE9F61471;
-
-u32 D_8003B960 = 0x21EFFFD8;
-
-u32 D_8003B964 = 0x14C0FFA6;
-
-u32 D_8003B968 = 0x20C6FFFC;
-
-u32 D_8003B96C = 0x8447001A;
-
-u32 D_8003B970 = 0x4BFF00AD;
-
-u32 D_8003B974 = 0x8468001A;
-
-u32 D_8003B978 = 0x4BFF00ED;
-
-u32 D_8003B97C = 0x8C0101EC;
-
-u32 D_8003B980 = 0x900601E7;
-
-u32 D_8003B984 = 0x01075822;
-
-u32 D_8003B988 = 0x05610003;
-
-u32 D_8003B98C = 0x344B0000;
-
-u32 D_8003B990 = 0x34620000;
-
-u32 D_8003B994 = 0x35630000;
-
-u32 D_8003B998 = 0xC8441006;
-
-u32 D_8003B99C = 0xC8651006;
-
-u32 D_8003B9A0 = 0x00015A80;
-
-u32 D_8003B9A4 = 0x05610006;
-
-u32 D_8003B9A8 = 0x900901E6;
-
-u32 D_8003B9AC = 0x4A042A11;
-
-u32 D_8003B9B0 = 0xC8463002;
-
-u32 D_8003B9B4 = 0x4A042A63;
-
-u32 D_8003B9B8 = 0x0800060E;
-
-u32 D_8003B9BC = 0xC8673002;
-
-u32 D_8003B9C0 = 0xC8A63002;
-
-u32 D_8003B9C4 = 0x4A042A11;
-
-u32 D_8003B9C8 = 0xC8460313;
-
-u32 D_8003B9CC = 0x4A042A63;
-
-u32 D_8003B9D0 = 0xC8A73002;
-
-u32 D_8003B9D4 = 0xC8670313;
-
-u32 D_8003B9D8 = 0x200B03E0;
-
-u32 D_8003B9DC = 0x4A0842D3;
-
-u32 D_8003B9E0 = 0xC96C1800;
-
-u32 D_8003B9E4 = 0x4A042AA0;
-
-u32 D_8003B9E8 = 0xC8420F0F;
-
-u32 D_8003B9EC = 0x4BFE3184;
-
-u32 D_8003B9F0 = 0xC8630F0F;
-
-u32 D_8003B9F4 = 0x4BFE39C4;
-
-u32 D_8003B9F8 = 0xC8460F0E;
-
-u32 D_8003B9FC = 0x4B284B70;
-
-u32 D_8003BA00 = 0xC8670F0E;
-
-u32 D_8003BA04 = 0x4B004BB2;
-
-u32 D_8003BA08 = 0xC8461405;
-
-u32 D_8003BA0C = 0x4B084370;
-
-u32 D_8003BA10 = 0xC8671405;
-
-u32 D_8003BA14 = 0x4B0043B2;
-
-u32 D_8003BA18 = 0x4A021BD5;
-
-u32 D_8003BA1C = 0x4A063C11;
-
-u32 D_8003BA20 = 0x480A5800;
-
-u32 D_8003BA24 = 0x4B5F6F44;
-
-u32 D_8003BA28 = 0x480D5900;
-
-u32 D_8003BA2C = 0x4B5F738D;
-
-u32 D_8003BA30 = 0xC80C1C18;
-
-u32 D_8003BA34 = 0x4B00034E;
-
-u32 D_8003BA38 = 0x00260825;
-
-u32 D_8003BA3C = 0x4BBF42C5;
-
-u32 D_8003BA40 = 0x342100C8;
-
-u32 D_8003BA44 = 0x4B0001CE;
-
-u32 D_8003BA48 = 0xA2E10008;
-
-u32 D_8003BA4C = 0x4B2E7F46;
-
-u32 D_8003BA50 = 0x48849800;
-
-u32 D_8003BA54 = 0x4B2D874D;
-
-u32 D_8003BA58 = 0xEAEC1C00;
-
-u32 D_8003BA5C = 0x4B2D7C4C;
-
-u32 D_8003BA60 = 0x014D5822;
-
-u32 D_8003BA64 = 0x4B2E848F;
-
-u32 D_8003BA68 = 0x48808E00;
-
-u32 D_8003BA6C = 0x4B2D3F44;
-
-u32 D_8003BA70 = 0x48809600;
-
-u32 D_8003BA74 = 0x4B2D5F4D;
-
-u32 D_8003BA78 = 0x4B2E39CE;
-
-u32 D_8003BA7C = 0x1D600014;
-
-u32 D_8003BA80 = 0x4B2E5ACF;
-
-u32 D_8003BA84 = 0x4BBF2506;
-
-u32 D_8003BA88 = 0x4B000590;
-
-u32 D_8003BA8C = 0xA2E90009;
-
-u32 D_8003BA90 = 0x4B0005D0;
-
-u32 D_8003BA94 = 0xEAE50905;
-
-u32 D_8003BA98 = 0x4B008E10;
-
-u32 D_8003BA9C = 0xEAE50906;
-
-u32 D_8003BAA0 = 0x4B009650;
-
-u32 D_8003BAA4 = 0x0C0006B9;
-
-u32 D_8003BAA8 = 0xEAE40907;
-
-u32 D_8003BAAC = 0x4BBF9D0E;
-
-u32 D_8003BAB0 = 0xEAE7080F;
-
-u32 D_8003BAB4 = 0x4B00054F;
-
-u32 D_8003BAB8 = 0xEAEB080E;
-
-u32 D_8003BABC = 0x4B039E8E;
-
-u32 D_8003BAC0 = 0xEAE70813;
-
-u32 D_8003BAC4 = 0x4B0006CF;
-
-u32 D_8003BAC8 = 0x08000688;
-
-u32 D_8003BACC = 0xEAEB0812;
-
-u32 D_8003BAD0 = 0x4B0C52A3;
-
-u32 D_8003BAD4 = 0x480B4000;
-
-u32 D_8003BAD8 = 0x4B4C4A60;
-
-u32 D_8003BADC = 0x4BFE0710;
-
-u32 D_8003BAE0 = 0x4A0B5DD3;
-
-u32 D_8003BAE4 = 0x4B2CE744;
-
-u32 D_8003BAE8 = 0x0160582A;
-
-u32 D_8003BAEC = 0x4BBE528E;
-
-u32 D_8003BAF0 = 0x000B59C0;
-
-u32 D_8003BAF4 = 0x4B6CE744;
-
-u32 D_8003BAF8 = 0x012B4825;
-
-u32 D_8003BAFC = 0x4BBE4A4E;
-
-u32 D_8003BB00 = 0xA2E90009;
-
-u32 D_8003BB04 = 0x4B0D7F44;
-
-u32 D_8003BB08 = 0x480CB800;
-
-u32 D_8003BB0C = 0x4B0D874D;
-
-u32 D_8003BB10 = 0x22EB0001;
-
-u32 D_8003BB14 = 0x4B0E7D8E;
-
-u32 D_8003BB18 = 0xE96A0800;
-
-u32 D_8003BB1C = 0x4B0E85CF;
-
-u32 D_8003BB20 = 0xE9690802;
-
-u32 D_8003BB24 = 0x4B132390;
-
-u32 D_8003BB28 = 0x10E80015;
-
-u32 D_8003BB2C = 0x218CF8AB;
-
-u32 D_8003BB30 = 0x4B132351;
-
-u32 D_8003BB34 = 0x1D800013;
-
-u32 D_8003BB38 = 0x4B132F10;
-
-u32 D_8003BB3C = 0x4BBF6D06;
-
-u32 D_8003BB40 = 0xEAFC0905;
-
-u32 D_8003BB44 = 0x4B000610;
-
-u32 D_8003BB48 = 0xEAEE0906;
-
-u32 D_8003BB4C = 0x4B000650;
-
-u32 D_8003BB50 = 0xEAED0907;
-
-u32 D_8003BB54 = 0x0C0006B9;
-
-u32 D_8003BB58 = 0xEAE7080B;
-
-u32 D_8003BB5C = 0xEAEB080A;
-
-u32 D_8003BB60 = 0x4BBF26C5;
-
-u32 D_8003BB64 = 0xEAE7080F;
-
-u32 D_8003BB68 = 0x4B00068E;
-
-u32 D_8003BB6C = 0xEAEB080E;
-
-u32 D_8003BB70 = 0xEAE01009;
-
-u32 D_8003BB74 = 0xEAFB0808;
-
-u32 D_8003BB78 = 0x08000688;
-
-u32 D_8003BB7C = 0xEAFA0809;
-
-u32 D_8003BB80 = 0x4B132351;
-
-u32 D_8003BB84 = 0x4BBF2545;
-
-u32 D_8003BB88 = 0x39290080;
-
-u32 D_8003BB8C = 0x4B00050E;
-
-u32 D_8003BB90 = 0xA2E90009;
-
-u32 D_8003BB94 = 0x4BBF2EC5;
-
-u32 D_8003BB98 = 0xEAE01007;
-
-u32 D_8003BB9C = 0x4B00068E;
-
-u32 D_8003BBA0 = 0xEAE01009;
-
-u32 D_8003BBA4 = 0x4B000450;
-
-u32 D_8003BBA8 = 0xEAED0907;
-
-u32 D_8003BBAC = 0x4B000490;
-
-u32 D_8003BBB0 = 0xEAEE0906;
-
-u32 D_8003BBB4 = 0x4B000610;
-
-u32 D_8003BBB8 = 0xEAEE0905;
-
-u32 D_8003BBBC = 0x4B000650;
-
-u32 D_8003BBC0 = 0xEAF5080C;
-
-u32 D_8003BBC4 = 0x4B0171B3;
-
-u32 D_8003BBC8 = 0xEAF4080D;
-
-u32 D_8003BBCC = 0x4B0070B3;
-
-u32 D_8003BBD0 = 0xEAFB0810;
-
-u32 D_8003BBD4 = 0x4BDE1146;
-
-u32 D_8003BBD8 = 0xEAFA0811;
-
-u32 D_8003BBDC = 0x4BDE320F;
-
-u32 D_8003BBE0 = 0x22F70028;
-
-u32 D_8003BBE4 = 0x4BDEB246;
-
-u32 D_8003BBE8 = 0x302B0004;
-
-u32 D_8003BBEC = 0x4BDEBA8F;
-
-u32 D_8003BBF0 = 0x4BDE8B06;
-
-u32 D_8003BBF4 = 0x11600009;
-
-u32 D_8003BBF8 = 0x302B0002;
-
-u32 D_8003BBFC = 0xEAE61800;
-
-u32 D_8003BC00 = 0xEAF71801;
-
-u32 D_8003BC04 = 0xEAF61803;
-
-u32 D_8003BC08 = 0xEAF21804;
-
-u32 D_8003BC0C = 0xEAF91805;
-
-u32 D_8003BC10 = 0xEAF11806;
-
-u32 D_8003BC14 = 0xEAF81807;
-
-u32 D_8003BC18 = 0x22F70040;
-
-u32 D_8003BC1C = 0x4BDE934F;
-
-u32 D_8003BC20 = 0x1160000A;
-
-u32 D_8003BC24 = 0x302B0001;
-
-u32 D_8003BC28 = 0x22F70040;
-
-u32 D_8003BC2C = 0xEAE61C78;
-
-u32 D_8003BC30 = 0xEAE21C7A;
-
-u32 D_8003BC34 = 0xEAF71C79;
-
-u32 D_8003BC38 = 0xEAF61C7B;
-
-u32 D_8003BC3C = 0xEAF21C7C;
-
-u32 D_8003BC40 = 0xEAF11C7E;
-
-u32 D_8003BC44 = 0xEAF91C7D;
-
-u32 D_8003BC48 = 0xEAF81C7F;
-
-u32 D_8003BC4C = 0x1160FDFC;
-
-u32 D_8003BC50 = 0x941F03C2;
-
-u32 D_8003BC54 = 0xEAE80F00;
-
-u32 D_8003BC58 = 0x22F70010;
-
-u32 D_8003BC5C = 0xEAE50F79;
-
-u32 D_8003BC60 = 0x4BDEC606;
-
-u32 D_8003BC64 = 0xEAEA0F7A;
-
-u32 D_8003BC68 = 0x4BDECE4F;
-
-u32 D_8003BC6C = 0xEAE90F7B;
-
-u32 D_8003BC70 = 0xEAED0F7C;
-
-u32 D_8003BC74 = 0xEAEC0F7D;
-
-u32 D_8003BC78 = 0xEAF90F7E;
-
-u32 D_8003BC7C = 0x080004A8;
-
-u32 D_8003BC80 = 0xEAF80F7F;
-
-u32 D_8003BC84 = 0x4A140515;
-
-u32 D_8003BC88 = 0x4A000551;
-
-u32 D_8003BC8C = 0x4B011746;
-
-u32 D_8003BC90 = 0x4B01374F;
-
-u32 D_8003BC94 = 0x4B348F4C;
-
-u32 D_8003BC98 = 0x4B34974D;
-
-u32 D_8003BC9C = 0x4B35888E;
-
-u32 D_8003BCA0 = 0x4B35918F;
-
-u32 D_8003BCA4 = 0x4BBF2746;
-
-u32 D_8003BCA8 = 0x4B343F4C;
-
-u32 D_8003BCAC = 0x4B345F4D;
-
-u32 D_8003BCB0 = 0xEAE21807;
-
-u32 D_8003BCB4 = 0x4B353D0E;
-
-u32 D_8003BCB8 = 0x03E00008;
-
-u32 D_8003BCBC = 0x4B355D4F;
-
-u32 D_8003BCC0 = 0x97390380;
-
-u32 D_8003BCC4 = 0x97180380;
-
-u32 D_8003BCC8 = 0x24017070;
-
-u32 D_8003BCCC = 0x8F2B0024;
-
-u32 D_8003BCD0 = 0x002B0824;
-
-u32 D_8003BCD4 = 0x1020FD95;
-
-u32 D_8003BCD8 = 0x8F2B004C;
-
-u32 D_8003BCDC = 0x1738FFFC;
-
-u32 D_8003BCE0 = 0x27390028;
-
-u32 D_8003BCE4 = 0x0800047F;
-
-u32 D_8003BCE8 = 0x97390380;
-
-u32 D_8003BCEC = 0x8F39001C;
-
-u32 D_8003BCF0 = 0x03381022;
-
-u32 D_8003BCF4 = 0x0441FD8D;
-
-u32 D_8003BCF8 = 0x8C1800D8;
-
-u32 D_8003BCFC = 0x08000402;
-
-u32 D_8003BD00 = 0x93610989;
-
-u32 D_8003BD04 = 0x08000410;
-
-u32 D_8003BD08 = 0x97390380;
-
-u32 D_8003BD0C = 0x00000000;
-
-u32 D_8003BD10 = 0x00000000;
-
-u32 D_8003BD14 = 0x00000000;
-
-u32 D_8003BD18 = 0x00000000;
-
-u32 D_8003BD1C = 0x00000000;
-
-u32 D_8003BD20 = 0x00000000;
-
-u32 D_8003BD24 = 0x00000000;
-
-u32 D_8003BD28 = 0x00000000;
-
-u32 D_8003BD2C = 0x00000000;
-
-u32 D_8003BD30 = 0x00000000;
-
-u32 D_8003BD34 = 0x00000000;
-
-u32 D_8003BD38 = 0x00000000;
-
-u32 D_8003BD3C = 0x00000000;
-
-u32 D_8003BD40 = 0x00000000;
-
-u32 D_8003BD44 = 0x00000000;
-
-u32 D_8003BD48 = 0x00000000;
-
-u32 D_8003BD4C = 0x00000000;
-
-u32 D_8003BD50 = 0x00000000;
-
-u32 D_8003BD54 = 0x00000000;
-
-u32 D_8003BD58 = 0x00000000;
-
-u32 D_8003BD5C = 0x00000000;
-
-u32 D_8003BD60 = 0x00000000;
-
-u32 D_8003BD64 = 0x00000000;
-
-u32 D_8003BD68 = 0x00000000;
-
-u32 D_8003BD6C = 0x00000000;
-
-u32 D_8003BD70 = 0x00000000;
-
-u32 D_8003BD74 = 0x00000000;
-
-u32 D_8003BD78 = 0x00000000;
-
-u32 D_8003BD7C = 0x00000000;
-
-u32 D_8003BD80 = 0x00000000;
-
-u32 D_8003BD84 = 0x00000000;
-
-u32 D_8003BD88 = 0x00000000;
-
-u32 D_8003BD8C = 0x00000000;
-
-u32 D_8003BD90 = 0x00000000;
-
-u32 D_8003BD94 = 0x00000000;
-
-u32 D_8003BD98 = 0x00000000;
-
-u32 D_8003BD9C = 0x00000000;
-
-u32 D_8003BDA0 = 0x00000000;
-
-u32 D_8003BDA4 = 0x00000000;
-
-u32 D_8003BDA8 = 0x00000000;
-
-u32 D_8003BDAC = 0x00000000;
-
-u32 D_8003BDB0 = 0x00000000;
-
-u32 D_8003BDB4 = 0x00000000;
-
-u32 D_8003BDB8 = 0x00000000;
-
-u32 D_8003BDBC = 0x00000000;
-
-u32 D_8003BDC0 = 0x00000000;
-
-u32 D_8003BDC4 = 0x00000000;
-
-u32 D_8003BDC8 = 0x00000000;
-
-u32 D_8003BDCC = 0x00000000;
-
-u32 D_8003BDD0 = 0x00000000;
-
-u32 D_8003BDD4 = 0x00000000;
-
-u32 D_8003BDD8 = 0x00000000;
-
-u32 D_8003BDDC = 0x00000000;
-
-u32 D_8003BDE0 = 0x00000000;
-
-u32 D_8003BDE4 = 0x00000000;
-
-u32 D_8003BDE8 = 0x00000000;
-
-u32 D_8003BDEC = 0x00000000;
-
-u32 D_8003BDF0 = 0x00000000;
-
-u32 D_8003BDF4 = 0x00000000;
-
-u32 D_8003BDF8 = 0x00000000;
-
-u32 D_8003BDFC = 0x00000000;
-
-u32 D_8003BE00 = 0x00000000;
-
-u32 D_8003BE04 = 0x00000000;
-
-u32 D_8003BE08 = 0x00000000;
-
-u32 D_8003BE0C = 0x00000000;
-
-u32 D_8003BE10 = 0x00000000;
-
-u32 D_8003BE14 = 0x00000000;
-
-u32 D_8003BE18 = 0x00000000;
-
-u32 D_8003BE1C = 0x00000000;
-
-u32 D_8003BE20 = 0x00000000;
-
-u32 D_8003BE24 = 0x00000000;
-
-u32 D_8003BE28 = 0x00000000;
-
-u32 D_8003BE2C = 0x00000000;
-
-u32 D_8003BE30 = 0x00000000;
-
-u32 D_8003BE34 = 0x00000000;
-
-u32 D_8003BE38 = 0x00000000;
-
-u32 D_8003BE3C = 0x00000000;
-
-u32 D_8003BE40 = 0x00000000;
-
-u32 D_8003BE44 = 0x00000000;
-
-u32 D_8003BE48 = 0x00000000;
-
-u32 D_8003BE4C = 0x00000000;
-
-u32 D_8003BE50 = 0x00000000;
-
-u32 D_8003BE54 = 0x00000000;
-
-u32 D_8003BE58 = 0x00000000;
-
-u32 D_8003BE5C = 0x00000000;
-
-u32 D_8003BE60 = 0x00000000;
-
-u32 D_8003BE64 = 0x00000000;
-
-u32 D_8003BE68 = 0x00000000;
-
-u32 D_8003BE6C = 0x00000000;
-
-u32 D_8003BE70 = 0x00000000;
-
-u32 D_8003BE74 = 0x00000000;
-
-u32 D_8003BE78 = 0x00000000;
-
-u32 D_8003BE7C = 0x00000000;
-
-u32 D_8003BE80 = 0x00000000;
-
-u32 D_8003BE84 = 0x00000000;
-
-u32 D_8003BE88 = 0x00000000;
-
-u32 D_8003BE8C = 0x00000000;
-
-u32 D_8003BE90 = 0x00000000;
-
-u32 D_8003BE94 = 0x00000000;
-
-u32 D_8003BE98 = 0x00000000;
-
-u32 D_8003BE9C = 0x00000000;
-
-u32 D_8003BEA0 = 0x00000000;
-
-u32 D_8003BEA4 = 0x00000000;
-
-u32 D_8003BEA8 = 0x00000000;
-
-u32 D_8003BEAC = 0x00000000;
-
-u32 D_8003BEB0 = 0x00000000;
-
-u32 D_8003BEB4 = 0x00000000;
-
-u32 D_8003BEB8 = 0x00000000;
-
-u32 D_8003BEBC = 0x00000000;
-
-u32 D_8003BEC0 = 0x00000000;
-
-u32 D_8003BEC4 = 0x00000000;
-
-u32 D_8003BEC8 = 0x00000000;
-
-u32 D_8003BECC = 0x00000000;
-
-u32 D_8003BED0 = 0x00000000;
-
-u32 D_8003BED4 = 0x00000000;
-
-u32 D_8003BED8 = 0x00000000;
-
-u32 D_8003BEDC = 0x00000000;
-
-u32 D_8003BEE0 = 0x00000000;
-
-u32 D_8003BEE4 = 0x00000000;
-
-u32 D_8003BEE8 = 0x00000000;
-
-u32 D_8003BEEC = 0x00000000;
-
-u32 D_8003BEF0 = 0x00000000;
-
-u32 D_8003BEF4 = 0x00000000;
-
-u32 D_8003BEF8 = 0x00000000;
-
-u32 D_8003BEFC = 0x00000000;
-
-u32 D_8003BF00 = 0x00000000;
-
-u32 D_8003BF04 = 0x00000000;
-
-u32 D_8003BF08 = 0x00000000;
-
-u32 D_8003BF0C = 0x00000000;
-
-u32 D_8003BF10 = 0x00000000;
-
-u32 D_8003BF14 = 0x00000000;
-
-u32 D_8003BF18 = 0x00000000;
-
-u32 D_8003BF1C = 0x00000000;
-
-u32 D_8003BF20 = 0x00000000;
-
-u32 D_8003BF24 = 0x00000000;
-
-u32 D_8003BF28 = 0x00000000;
-
-u32 D_8003BF2C = 0x00000000;
-
-u32 D_8003BF30 = 0x00000000;
-
-u32 D_8003BF34 = 0x00000000;
-
-u32 D_8003BF38 = 0x00000000;
-
-u32 D_8003BF3C = 0x00000000;
-
-u32 D_8003BF40 = 0x00000000;
-
-u32 D_8003BF44 = 0x00000000;
-
-u32 D_8003BF48 = 0x00000000;
-
-u32 D_8003BF4C = 0x00000000;
-
-u32 D_8003BF50 = 0x00000000;
-
-u32 D_8003BF54 = 0x00000000;
-
-u32 D_8003BF58 = 0x00000000;
-
-u32 D_8003BF5C = 0x00000000;
-
-u32 D_8003BF60 = 0x00000000;
-
-u32 D_8003BF64 = 0x00000000;
-
-u32 D_8003BF68 = 0x00000000;
-
-u32 D_8003BF6C = 0x00000000;
-
-u32 D_8003BF70 = 0x00000000;
-
-u32 D_8003BF74 = 0x00000000;
-
-u32 D_8003BF78 = 0x00000000;
-
-u32 D_8003BF7C = 0x00000000;
-
-u32 D_8003BF80 = 0x00000000;
-
-u32 D_8003BF84 = 0x00000000;
-
-u32 D_8003BF88 = 0x00000000;
-
-u32 D_8003BF8C = 0x00000000;
-
-u32 D_8003BF90 = 0x00000000;
-
-u32 D_8003BF94 = 0x00000000;
-
-u32 D_8003BF98 = 0x00000000;
-
-u32 D_8003BF9C = 0x00000000;
-
-u32 D_8003BFA0 = 0x00000000;
-
-u32 D_8003BFA4 = 0x00000000;
-
-u32 D_8003BFA8 = 0x00000000;
-
-u32 D_8003BFAC = 0x00000000;
-
-u32 D_8003BFB0 = 0x00000000;
-
-u32 D_8003BFB4 = 0x00000000;
-
-u32 D_8003BFB8 = 0x00000000;
-
-u32 D_8003BFBC = 0x00000000;
-
-u32 D_8003BFC0 = 0x00000000;
-
-u32 D_8003BFC4 = 0x00000000;
-
-u32 D_8003BFC8 = 0x00000000;
-
-u32 D_8003BFCC = 0x00000000;
-
-u32 D_8003BFD0 = 0x00000000;
-
-u32 D_8003BFD4 = 0x00000000;
-
-u32 D_8003BFD8 = 0x00000000;
-
-u32 D_8003BFDC = 0x00000000;
-
-u32 D_8003BFE0 = 0x00000000;
-
-u32 D_8003BFE4 = 0x00000000;
-
-u32 D_8003BFE8 = 0x00000000;
-
-u32 D_8003BFEC = 0x00000000;
-
-u32 D_8003BFF0 = 0x00000000;
-
-u32 D_8003BFF4 = 0x00000000;
-
-u32 D_8003BFF8 = 0x00000000;
-
-u32 D_8003BFFC = 0x00000000;
-
-u32 D_8003C000 = 0x00000000;
-
-u32 D_8003C004 = 0x00000000;
-
-u32 D_8003C008 = 0x00000000;
-
-u32 D_8003C00C = 0x00000000;
-
-u32 D_8003C010 = 0x00000000;
-
-u32 D_8003C014 = 0x00000000;
-
-u32 D_8003C018 = 0x00000000;
-
-u32 D_8003C01C = 0x00000000;
-
-u32 D_8003C020 = 0x00000000;
-
-u32 D_8003C024 = 0x00000000;
-
-u32 D_8003C028 = 0x00000000;
-
-u32 D_8003C02C = 0x00000000;
-
-u32 D_8003C030 = 0x00000000;
-
-u32 D_8003C034 = 0x00000000;
-
-u32 D_8003C038 = 0x00000000;
-
-u32 D_8003C03C = 0x00000000;
-
-u32 D_8003C040 = 0x00000000;
-
-u32 D_8003C044 = 0x00000000;
-
-u32 D_8003C048 = 0x00000000;
-
-u32 D_8003C04C = 0x00000000;
-
-u32 D_8003C050 = 0x00000000;
-
-u32 D_8003C054 = 0x00000000;
-
-u32 D_8003C058 = 0x00000000;
-
-u32 D_8003C05C = 0x00000000;
-
-u32 D_8003C060 = 0x00000000;
-
-u32 D_8003C064 = 0x00000000;
-
-u32 D_8003C068 = 0x00000000;
-
-u32 D_8003C06C = 0x00000000;
-
-u32 D_8003C070 = 0x00000000;
-
-u32 D_8003C074 = 0x00000000;
-
-u32 D_8003C078 = 0x00000000;
-
-u32 D_8003C07C = 0x00000000;
-
-u32 D_8003C080 = 0x00000000;
-
-u32 D_8003C084 = 0x00000000;
-
-u32 D_8003C088 = 0x00000000;
-
-u32 D_8003C08C = 0x00000000;
-
-u32 D_8003C090 = 0x00000000;
-
-u32 D_8003C094 = 0x00000000;
-
-u32 D_8003C098 = 0x00000000;
-
-u32 D_8003C09C = 0x00000000;
-
-u32 D_8003C0A0 = 0x00000000;
-
-u32 D_8003C0A4 = 0x00000000;
-
-u32 D_8003C0A8 = 0x00000000;
-
-u32 D_8003C0AC = 0x00000000;
-
-u32 D_8003C0B0 = 0x00000000;
-
-u32 D_8003C0B4 = 0x00000000;
-
-u32 D_8003C0B8 = 0x00000000;
-
-u32 D_8003C0BC = 0x00000000;
-
-u32 D_8003C0C0 = 0x00000000;
-
-u32 D_8003C0C4 = 0x00000000;
-
-u32 D_8003C0C8 = 0x00000000;
-
-u32 D_8003C0CC = 0x00000000;
-
-u32 D_8003C0D0 = 0x00000000;
-
-u32 D_8003C0D4 = 0x00000000;
-
-u32 D_8003C0D8 = 0x00000000;
-
-u32 D_8003C0DC = 0x00000000;
-
-u32 D_8003C0E0 = 0x00000000;
-
-u32 D_8003C0E4 = 0x00000000;
-
-u32 D_8003C0E8 = 0x00000000;
-
-u32 D_8003C0EC = 0x00000000;
-
-u32 D_8003C0F0 = 0x00000000;
-
-u32 D_8003C0F4 = 0x00000000;
-
-u32 D_8003C0F8 = 0x00000000;
-
-u32 D_8003C0FC = 0x00000000;
-
-u32 D_8003C100 = 0x00000000;
-
-u32 D_8003C104 = 0x00000000;
-
-u32 D_8003C108 = 0x00000000;
-
-u32 D_8003C10C = 0x00000000;
-
-u32 D_8003C110 = 0x00000000;
-
-u32 D_8003C114 = 0x00000000;
-
-u32 D_8003C118 = 0x00000000;
-
-u32 D_8003C11C = 0x00000000;
-
-u32 D_8003C120 = 0x00000000;
-
-u32 D_8003C124 = 0x00000000;
-
-u32 D_8003C128 = 0x00000000;
-
-u32 D_8003C12C = 0x00000000;
-
-u32 D_8003C130 = 0x00000000;
-
-u32 D_8003C134 = 0x00000000;
-
-u32 D_8003C138 = 0x00000000;
-
-u32 D_8003C13C = 0x00000000;
-
-u32 D_8003C140 = 0x00000000;
-
-u32 D_8003C144 = 0x00000000;
-
-u32 D_8003C148 = 0x00000000;
-
-u32 D_8003C14C = 0x200C1000;
-
-u32 D_8003C150 = 0x200B02E0;
-
-u32 D_8003C154 = 0x8D780000;
-
-u32 D_8003C158 = 0x95730004;
-
-u32 D_8003C15C = 0x0C0007F6;
-
-u32 D_8003C160 = 0x95740006;
-
-u32 D_8003C164 = 0x359F0000;
-
-u32 D_8003C168 = 0x400B3000;
-
-u32 D_8003C16C = 0x1560FFFF;
-
-u32 D_8003C170 = 0x400B3000;
-
-u32 D_8003C174 = 0x03E00008;
-
-u32 D_8003C178 = 0x400B2800;
-
-u32 D_8003C17C = 0x1560FFFF;
-
-u32 D_8003C180 = 0x400B2800;
-
-u32 D_8003C184 = 0x40940000;
-
-u32 D_8003C188 = 0x06800003;
-
-u32 D_8003C18C = 0x40980800;
-
-u32 D_8003C190 = 0x03E00008;
-
-u32 D_8003C194 = 0x40931000;
-
-u32 D_8003C198 = 0x03E00008;
-
-u32 D_8003C19C = 0x40931800;
-
-u32 D_8003C1A0 = 0xC81D1818;
-
-u32 D_8003C1A4 = 0x22F70008;
-
-u32 D_8003C1A8 = 0xEAFD187F;
-
-u32 D_8003C1AC = 0x02F65822;
-
-u32 D_8003C1B0 = 0x256C01E7;
-
-u32 D_8003C1B4 = 0x059100A4;
-
-u32 D_8003C1B8 = 0x00000000;
-
-u32 D_8003C1BC = 0x0C0007F2;
-
-u32 D_8003C1C0 = 0x8C1800F0;
-
-u32 D_8003C1C4 = 0x0420001A;
-
-u32 D_8003C1C8 = 0x40984800;
-
-u32 D_8003C1CC = 0x1420000F;
-
-u32 D_8003C1D0 = 0x035BD020;
-
-u32 D_8003C1D4 = 0x8F78098C;
-
-u32 D_8003C1D8 = 0xAC1A0FF0;
-
-u32 D_8003C1DC = 0xAC180FD0;
-
-u32 D_8003C1E0 = 0x24141080;
-
-u32 D_8003C1E4 = 0x0C0007F6;
-
-u32 D_8003C1E8 = 0x20130F47;
-
-u32 D_8003C1EC = 0x8C1800D8;
-
-u32 D_8003C1F0 = 0x24140180;
-
-u32 D_8003C1F4 = 0x33330FFF;
-
-u32 D_8003C1F8 = 0x0314C020;
-
-u32 D_8003C1FC = 0x0C0007F6;
-
-u32 D_8003C200 = 0x02749822;
-
-u32 D_8003C204 = 0x080007F2;
-
-u32 D_8003C208 = 0x201F1080;
-
-u32 D_8003C20C = 0x8C0B0FD0;
-
-u32 D_8003C210 = 0xAC1A0BF8;
-
-u32 D_8003C214 = 0xAC0B0BFC;
-
-u32 D_8003C218 = 0x200C5000;
-
-u32 D_8003C21C = 0x8C180FF8;
-
-u32 D_8003C220 = 0x20148000;
-
-u32 D_8003C224 = 0x20130BFF;
-
-u32 D_8003C228 = 0x080007F6;
-
-u32 D_8003C22C = 0x201F1094;
-
-u32 D_8003C230 = 0x200C4000;
-
-u32 D_8003C234 = 0x408C2000;
-
-u32 D_8003C238 = 0x0000000D;
-
-u32 D_8003C23C = 0x00000000;
-
-u32 D_8003C240 = 0x900100DE;
-
-u32 D_8003C244 = 0x001913C0;
-
-u32 D_8003C248 = 0x0C00048F;
-
-u32 D_8003C24C = 0x035B1820;
-
-u32 D_8003C250 = 0x04400057;
-
-u32 D_8003C254 = 0x371A0000;
-
-u32 D_8003C258 = 0xAC230138;
-
-u32 D_8003C25C = 0x20210004;
-
-u32 D_8003C260 = 0x0800045C;
-
-u32 D_8003C264 = 0xA00100DE;
-
-u32 D_8003C268 = 0x200B1140;
-
-u32 D_8003C26C = 0xAD79F0A4;
-
-u32 D_8003C270 = 0x08000463;
-
-u32 D_8003C274 = 0xAD78F0A8;
-
-u32 D_8003C278 = 0x00191402;
-
-u32 D_8003C27C = 0x944127FE;
-
-u32 D_8003C280 = 0x00390820;
-
-u32 D_8003C284 = 0x08000463;
-
-u32 D_8003C288 = 0xAC380000;
-
-u32 D_8003C28C = 0x8C0B00F4;
-
-u32 D_8003C290 = 0x8C020FE0;
-
-u32 D_8003C294 = 0x0178C022;
-
-u32 D_8003C298 = 0x03020822;
-
-u32 D_8003C29C = 0x04210002;
-
-u32 D_8003C2A0 = 0x00000000;
-
-u32 D_8003C2A4 = 0x34580000;
-
-u32 D_8003C2A8 = 0x130B0048;
-
-u32 D_8003C2AC = 0xAC1800F4;
-
-u32 D_8003C2B0 = 0x08000449;
-
-u32 D_8003C2B4 = 0xAC0001D9;
-
-u32 D_8003C2B8 = 0x943302F2;
-
-u32 D_8003C2BC = 0x0C0007F2;
-
-u32 D_8003C2C0 = 0x943502F2;
-
-u32 D_8003C2C4 = 0x201F118C;
-
-u32 D_8003C2C8 = 0x228C0018;
-
-u32 D_8003C2CC = 0x4B00024E;
-
-u32 D_8003C2D0 = 0x228B0008;
-
-u32 D_8003C2D4 = 0x4B00020F;
-
-u32 D_8003C2D8 = 0x22B5FFE0;
-
-u32 D_8003C2DC = 0x4B000747;
-
-u32 D_8003C2E0 = 0xCAA51808;
-
-u32 D_8003C2E4 = 0xCAA51C08;
-
-u32 D_8003C2E8 = 0xCA832002;
-
-u32 D_8003C2EC = 0xCAA41804;
-
-u32 D_8003C2F0 = 0xCAA41C04;
-
-u32 D_8003C2F4 = 0xCA822000;
-
-u32 D_8003C2F8 = 0x4A832F4C;
-
-u32 D_8003C2FC = 0x22940002;
-
-u32 D_8003C300 = 0x4A83274D;
-
-u32 D_8003C304 = 0x22B50008;
-
-u32 D_8003C308 = 0x4A8229CE;
-
-u32 D_8003C30C = 0x168BFFF4;
-
-u32 D_8003C310 = 0x4A82218F;
-
-u32 D_8003C314 = 0x168CFFED;
-
-u32 D_8003C318 = 0x22940008;
-
-u32 D_8003C31C = 0xEA692002;
-
-u32 D_8003C320 = 0xEA682000;
-
-u32 D_8003C324 = 0xEA672003;
-
-u32 D_8003C328 = 0x03E00008;
-
-u32 D_8003C32C = 0xEA662001;
-
-u32 D_8003C330 = 0x332B0005;
-
-u32 D_8003C334 = 0x15600008;
-
-u32 D_8003C338 = 0x33220002;
-
-u32 D_8003C33C = 0x8C1800F4;
-
-u32 D_8003C340 = 0x2014E000;
-
-u32 D_8003C344 = 0x0C0007F6;
-
-u32 D_8003C348 = 0x2013003F;
-
-u32 D_8003C34C = 0x23180040;
-
-u32 D_8003C350 = 0xAC1800F4;
-
-u32 D_8003C354 = 0x8F78098C;
-
-u32 D_8003C358 = 0x01826020;
-
-u32 D_8003C35C = 0xAC0001D9;
-
-u32 D_8003C360 = 0x0C00048F;
-
-u32 D_8003C364 = 0x332100FE;
-
-u32 D_8003C368 = 0x93730989;
-
-u32 D_8003C36C = 0x943402F0;
-
-u32 D_8003C370 = 0x00191142;
-
-u32 D_8003C374 = 0x959F0336;
-
-u32 D_8003C378 = 0x080007F6;
-
-u32 D_8003C37C = 0x0282A020;
-
-u32 D_8003C380 = 0x8D63EF8C;
-
-u32 D_8003C384 = 0x3C028000;
-
-u32 D_8003C388 = 0x03221007;
-
-u32 D_8003C38C = 0x00190A02;
-
-u32 D_8003C390 = 0x00221006;
-
-u32 D_8003C394 = 0x00401027;
-
-u32 D_8003C398 = 0x00621824;
-
-u32 D_8003C39C = 0x00781825;
-
-u32 D_8003C3A0 = 0xAD63EF8C;
-
-u32 D_8003C3A4 = 0x8C1900C8;
-
-u32 D_8003C3A8 = 0x0800048A;
-
-u32 D_8003C3AC = 0x8C1800CC;
-
-u32 D_8003C3B0 = 0x4A00002C;
-
-u32 D_8003C3B4 = 0xC81F2018;
-
-u32 D_8003C3B8 = 0xC81E2019;
-
-u32 D_8003C3BC = 0xC81D201A;
-
-u32 D_8003C3C0 = 0xC81C201B;
-
-u32 D_8003C3C4 = 0x201E01C0;
-
-u32 D_8003C3C8 = 0x201704C0;
-
-u32 D_8003C3CC = 0x201607E0;
-
-u32 D_8003C3D0 = 0x4B7F0051;
-
-u32 D_8003C3D4 = 0x8C0B00F0;
-
-u32 D_8003C3D8 = 0x8C0C0FC4;
-
-u32 D_8003C3DC = 0x20012800;
-
-u32 D_8003C3E0 = 0x11600006;
-
-u32 D_8003C3E4 = 0x40812000;
-
-u32 D_8003C3E8 = 0x318C0001;
-
-u32 D_8003C3EC = 0x1180001E;
-
-u32 D_8003C3F0 = 0xAC000FC4;
-
-u32 D_8003C3F4 = 0x0800045F;
-
-u32 D_8003C3F8 = 0x8C1A0BF8;
-
-u32 D_8003C3FC = 0x400B5800;
-
-u32 D_8003C400 = 0x316B0001;
-
-u32 D_8003C404 = 0x1560000B;
-
-u32 D_8003C408 = 0x40024800;
-
-u32 D_8003C40C = 0x8C030FE8;
-
-u32 D_8003C410 = 0x00625822;
-
-u32 D_8003C414 = 0x1D600007;
-
-u32 D_8003C418 = 0x40015000;
-
-u32 D_8003C41C = 0x8C040FEC;
-
-u32 D_8003C420 = 0x10200004;
-
-u32 D_8003C424 = 0x00245822;
-
-u32 D_8003C428 = 0x05610002;
-
-u32 D_8003C42C = 0x00000000;
-
-u32 D_8003C430 = 0x14220008;
-
-u32 D_8003C434 = 0x400B5800;
-
-u32 D_8003C438 = 0x316B0400;
-
-u32 D_8003C43C = 0x1560FFFD;
-
-u32 D_8003C440 = 0x200B0001;
-
-u32 D_8003C444 = 0x408B5800;
-
-u32 D_8003C448 = 0x8C020FEC;
-
-u32 D_8003C44C = 0x40824000;
-
-u32 D_8003C450 = 0x40824800;
-
-u32 D_8003C454 = 0xAC0200F0;
-
-u32 D_8003C458 = 0x8C0B00F4;
-
-u32 D_8003C45C = 0x15600002;
-
-u32 D_8003C460 = 0x8C0B0FE0;
-
-u32 D_8003C464 = 0xAC0B00F4;
-
-u32 D_8003C468 = 0x8C010FD0;
-
-u32 D_8003C46C = 0x8C020304;
-
-u32 D_8003C470 = 0x8C03030C;
-
-u32 D_8003C474 = 0x8C040314;
-
-u32 D_8003C478 = 0x8C05031C;
-
-u32 D_8003C47C = 0x00411020;
-
-u32 D_8003C480 = 0x00611820;
-
-u32 D_8003C484 = 0xAC020304;
-
-u32 D_8003C488 = 0xAC03030C;
-
-u32 D_8003C48C = 0x00812020;
-
-u32 D_8003C490 = 0x00A12820;
-
-u32 D_8003C494 = 0xAC040314;
-
-u32 D_8003C498 = 0xAC05031C;
-
-u32 D_8003C49C = 0x8C1900C0;
-
-u32 D_8003C4A0 = 0x0C000477;
-
-u32 D_8003C4A4 = 0x8C1800C4;
-
-u32 D_8003C4A8 = 0x8C1A0FF0;
-
-u32 D_8003C4AC = 0x200B030C;
-
-u32 D_8003C4B0 = 0x0C0007ED;
-
-u32 D_8003C4B4 = 0x37EC0000;
-
-u32 D_8003C4B8 = 0x201300A7;
-
-u32 D_8003C4BC = 0x37580000;
-
-u32 D_8003C4C0 = 0x0C0007F6;
-
-u32 D_8003C4C4 = 0x24140418;
-
-u32 D_8003C4C8 = 0x275A00A8;
-
-u32 D_8003C4CC = 0x201BFF58;
-
-u32 D_8003C4D0 = 0x0C0007F2;
-
-u32 D_8003C4D4 = 0x40012000;
-
-u32 D_8003C4D8 = 0x8F7904C0;
-
-u32 D_8003C4DC = 0x1360FFF6;
-
-u32 D_8003C4E0 = 0x30210080;
-
-u32 D_8003C4E4 = 0x00196603;
-
-u32 D_8003C4E8 = 0x000C5840;
-
-u32 D_8003C4EC = 0x956B0370;
-
-u32 D_8003C4F0 = 0x1420037A;
-
-u32 D_8003C4F4 = 0x8F7804C4;
-
-u32 D_8003C4F8 = 0x01600008;
-
-u32 D_8003C4FC = 0x277B0008;
-
-u32 D_8003C500 = 0xAC1900C0;
-
-u32 D_8003C504 = 0xAC1800C4;
-
-u32 D_8003C508 = 0x201F1140;
-
-u32 D_8003C50C = 0x332B0FFF;
-
-u32 D_8003C510 = 0xA40B0208;
-
-u32 D_8003C514 = 0x00195B02;
-
-u32 D_8003C518 = 0x316B0FFF;
-
-u32 D_8003C51C = 0xA40B0204;
-
-u32 D_8003C520 = 0x330B0FFF;
-
-u32 D_8003C524 = 0xA40B020A;
-
-u32 D_8003C528 = 0x00185B02;
-
-u32 D_8003C52C = 0x316B0FFF;
-
-u32 D_8003C530 = 0x03E00008;
-
-u32 D_8003C534 = 0xA40B0206;
-
-u32 D_8003C538 = 0x0C000455;
-
-u32 D_8003C53C = 0x201403B0;
-
-u32 D_8003C540 = 0x0C0007F6;
-
-u32 D_8003C544 = 0x20130027;
-
-u32 D_8003C548 = 0xCBC21211;
-
-u32 D_8003C54C = 0x0C0007F2;
-
-u32 D_8003C550 = 0xCBC21612;
-
-u32 D_8003C554 = 0xCA882000;
-
-u32 D_8003C558 = 0x900103CB;
-
-u32 D_8003C55C = 0x4BDE42C5;
-
-u32 D_8003C560 = 0x4A024251;
-
-u32 D_8003C564 = 0x4AC80A8E;
-
-u32 D_8003C568 = 0x4AA80190;
-
-u32 D_8003C56C = 0x4B1C5AE8;
-
-u32 D_8003C570 = 0x4AC90260;
-
-u32 D_8003C574 = 0x4AEA02A3;
-
-u32 D_8003C578 = 0x4A0940D1;
-
-u32 D_8003C57C = 0x4A095AD1;
-
-u32 D_8003C580 = 0x4B7F314E;
-
-u32 D_8003C584 = 0x4A0941D0;
-
-u32 D_8003C588 = 0x4B7F51CE;
-
-u32 D_8003C58C = 0x4A0706E3;
-
-u32 D_8003C590 = 0x484B0800;
-
-u32 D_8003C594 = 0x14200002;
-
-u32 D_8003C598 = 0x4A0A5910;
-
-u32 D_8003C59C = 0x4B0B4133;
-
-u32 D_8003C5A0 = 0x316B0088;
-
-u32 D_8003C5A4 = 0x4B7D2EE3;
-
-u32 D_8003C5A8 = 0x1560FFCA;
-
-u32 D_8003C5AC = 0x4A0B2967;
-
-u32 D_8003C5B0 = 0x940203CC;
-
-u32 D_8003C5B4 = 0x800503C4;
-
-u32 D_8003C5B8 = 0x3C033510;
-
-u32 D_8003C5BC = 0x00021240;
-
-u32 D_8003C5C0 = 0x00452824;
-
-u32 D_8003C5C4 = 0x00651825;
-
-u32 D_8003C5C8 = 0xAEE30000;
-
-u32 D_8003C5CC = 0xEAFD1401;
-
-u32 D_8003C5D0 = 0xEAFD1202;
-
-u32 D_8003C5D4 = 0xAEE0000C;
-
-u32 D_8003C5D8 = 0x900503C6;
-
-u32 D_8003C5DC = 0x900603C7;
-
-u32 D_8003C5E0 = 0x900703C9;
-
-u32 D_8003C5E4 = 0x8C040264;
-
-u32 D_8003C5E8 = 0x00051880;
-
-u32 D_8003C5EC = 0x00661825;
-
-u32 D_8003C5F0 = 0x00031CC0;
-
-u32 D_8003C5F4 = 0x00621825;
-
-u32 D_8003C5F8 = 0x00075D00;
-
-u32 D_8003C5FC = 0x008B2025;
-
-u32 D_8003C600 = 0xAEE30010;
-
-u32 D_8003C604 = 0xEAFD0610;
-
-u32 D_8003C608 = 0xAEE40014;
-
-u32 D_8003C60C = 0x22F70018;
-
-u32 D_8003C610 = 0x000630C0;
-
-u32 D_8003C614 = 0x20C60274;
-
-u32 D_8003C618 = 0xC8C21800;
-
-u32 D_8003C61C = 0xCA860912;
-
-u32 D_8003C620 = 0x4B9F2945;
-
-u32 D_8003C624 = 0x4B041228;
-
-u32 D_8003C628 = 0x4B670A8E;
-
-u32 D_8003C62C = 0x4B041244;
-
-u32 D_8003C630 = 0x4B8532CE;
-
-u32 D_8003C634 = 0x10200002;
-
-u32 D_8003C638 = 0x4B7D030F;
-
-u32 D_8003C63C = 0x4A0A0A11;
-
-u32 D_8003C640 = 0x4B0A1284;
-
-u32 D_8003C644 = 0x4BDF4205;
-
-u32 D_8003C648 = 0x4B7DFA0E;
-
-u32 D_8003C64C = 0x4B004A33;
-
-u32 D_8003C650 = 0xEBCB0901;
-
-u32 D_8003C654 = 0xEBCC0900;
-
-u32 D_8003C658 = 0x8FC20000;
-
-u32 D_8003C65C = 0x8C1803C0;
-
-u32 D_8003C660 = 0x0C000455;
-
-u32 D_8003C664 = 0x00021042;
-
-u32 D_8003C668 = 0x000210C0;
-
-u32 D_8003C66C = 0x00581020;
-
-u32 D_8003C670 = 0x37030000;
-
-u32 D_8003C674 = 0x4B671890;
-
-u32 D_8003C678 = 0x4B7F088E;
-
-u32 D_8003C67C = 0x940503BA;
-
-u32 D_8003C680 = 0x48062C00;
-
-u32 D_8003C684 = 0x48074900;
-
-u32 D_8003C688 = 0x48081A00;
-
-u32 D_8003C68C = 0x48091E00;
-
-u32 D_8003C690 = 0x480A1200;
-
-u32 D_8003C694 = 0x48013F00;
-
-u32 D_8003C698 = 0x00084300;
-
-u32 D_8003C69C = 0x30A5FFFC;
-
-u32 D_8003C6A0 = 0x00063080;
-
-u32 D_8003C6A4 = 0x00A62022;
-
-u32 D_8003C6A8 = 0x1880FF8A;
-
-u32 D_8003C6AC = 0x000A5300;
-
-u32 D_8003C6B0 = 0x18E00002;
-
-u32 D_8003C6B4 = 0x3C0CE400;
-
-u32 D_8003C6B8 = 0x2084FFFC;
-
-u32 D_8003C6BC = 0x00815822;
-
-u32 D_8003C6C0 = 0x19600002;
-
-u32 D_8003C6C4 = 0x014C5025;
-
-u32 D_8003C6C8 = 0x34240000;
-
-u32 D_8003C6CC = 0x800F03C5;
-
-u32 D_8003C6D0 = 0x940D03D4;
-
-u32 D_8003C6D4 = 0x940603D2;
-
-u32 D_8003C6D8 = 0x05E10004;
-
-u32 D_8003C6DC = 0x940503D0;
-
-u32 D_8003C6E0 = 0x48055200;
-
-u32 D_8003C6E4 = 0x00052880;
-
-u32 D_8003C6E8 = 0x20A5FFFF;
-
-u32 D_8003C6EC = 0x000D7040;
-
-u32 D_8003C6F0 = 0x21CEFFFF;
-
-u32 D_8003C6F4 = 0x3C0B3D10;
-
-u32 D_8003C6F8 = 0x01CB7025;
-
-u32 D_8003C6FC = 0x000F7E00;
-
-u32 D_8003C700 = 0x34A57000;
-
-u32 D_8003C704 = 0x00052B00;
-
-u32 D_8003C708 = 0x1880002E;
-
-u32 D_8003C70C = 0x00240822;
-
-u32 D_8003C710 = 0x941D03D6;
-
-u32 D_8003C714 = 0x05E10004;
-
-u32 D_8003C718 = 0x941503CE;
-
-u32 D_8003C71C = 0x940B03D0;
-
-u32 D_8003C720 = 0x001DEC00;
-
-u32 D_8003C724 = 0x03ABE825;
-
-u32 D_8003C728 = 0x00A68025;
-
-u32 D_8003C72C = 0x00952022;
-
-u32 D_8003C730 = 0x04810010;
-
-u32 D_8003C734 = 0x48841100;
-
-u32 D_8003C738 = 0x4B223086;
-
-u32 D_8003C73C = 0x4B7D00CF;
-
-u32 D_8003C740 = 0xEBC20901;
-
-u32 D_8003C744 = 0xEBC30900;
-
-u32 D_8003C748 = 0x8FCB0000;
-
-u32 D_8003C74C = 0x02A4A820;
-
-u32 D_8003C750 = 0x05E10004;
-
-u32 D_8003C754 = 0x03ABE820;
-
-u32 D_8003C758 = 0x22ABFFFF;
-
-u32 D_8003C75C = 0x08000511;
-
-u32 D_8003C760 = 0x00AB8025;
-
-u32 D_8003C764 = 0x23ABFFFE;
-
-u32 D_8003C768 = 0x356BE000;
-
-u32 D_8003C76C = 0x000B5AC0;
-
-u32 D_8003C770 = 0x01668025;
-
-u32 D_8003C774 = 0x0135E020;
-
-u32 D_8003C778 = 0x239CFFFF;
-
-u32 D_8003C77C = 0xEAFD0700;
-
-u32 D_8003C780 = 0xAEEE0008;
-
-u32 D_8003C784 = 0xAEE2000C;
-
-u32 D_8003C788 = 0xAEEF0010;
-
-u32 D_8003C78C = 0xAEF00014;
-
-u32 D_8003C790 = 0xEAFD0418;
-
-u32 D_8003C794 = 0x015C5825;
-
-u32 D_8003C798 = 0xAEEB0020;
-
-u32 D_8003C79C = 0x01095825;
-
-u32 D_8003C7A0 = 0xAEEB0024;
-
-u32 D_8003C7A4 = 0xEAE8100A;
-
-u32 D_8003C7A8 = 0xEAFE140B;
-
-u32 D_8003C7AC = 0x0C000591;
-
-u32 D_8003C7B0 = 0x22F70030;
-
-u32 D_8003C7B4 = 0x23890001;
-
-u32 D_8003C7B8 = 0x1C80FFDC;
-
-u32 D_8003C7BC = 0x005D1020;
-
-u32 D_8003C7C0 = 0x1820FF44;
-
-u32 D_8003C7C4 = 0xEAFD0700;
-
-u32 D_8003C7C8 = 0x18E00028;
-
-u32 D_8003C7CC = 0xAEEE0008;
-
-u32 D_8003C7D0 = 0xAEE2000C;
-
-u32 D_8003C7D4 = 0x3C0B3510;
-
-u32 D_8003C7D8 = 0xAEEB0010;
-
-u32 D_8003C7DC = 0x3C0C0600;
-
-u32 D_8003C7E0 = 0xAEEC0014;
-
-u32 D_8003C7E4 = 0x3C1D3300;
-
-u32 D_8003C7E8 = 0xAEFD0018;
-
-u32 D_8003C7EC = 0x00073842;
-
-u32 D_8003C7F0 = 0x000DA842;
-
-u32 D_8003C7F4 = 0x02A7A822;
-
-u32 D_8003C7F8 = 0x00155B80;
-
-u32 D_8003C7FC = 0x216BF000;
-
-u32 D_8003C800 = 0x016C5825;
-
-u32 D_8003C804 = 0xAEEB001C;
-
-u32 D_8003C808 = 0xEAFD0420;
-
-u32 D_8003C80C = 0xAEEE0028;
-
-u32 D_8003C810 = 0xAEE3002C;
-
-u32 D_8003C814 = 0x3C0B3510;
-
-u32 D_8003C818 = 0xAEEB0030;
-
-u32 D_8003C81C = 0xA6F50032;
-
-u32 D_8003C820 = 0x3C0C0600;
-
-u32 D_8003C824 = 0xAEEC0034;
-
-u32 D_8003C828 = 0xAEFD0038;
-
-u32 D_8003C82C = 0x00075B80;
-
-u32 D_8003C830 = 0x216BF000;
-
-u32 D_8003C834 = 0x016C5825;
-
-u32 D_8003C838 = 0xAEEB003C;
-
-u32 D_8003C83C = 0x22F70058;
-
-u32 D_8003C840 = 0xEAFD0468;
-
-u32 D_8003C844 = 0x01495825;
-
-u32 D_8003C848 = 0xAEEBFFF0;
-
-u32 D_8003C84C = 0x01095825;
-
-u32 D_8003C850 = 0xAEEBFFF4;
-
-u32 D_8003C854 = 0xEAE8107E;
-
-u32 D_8003C858 = 0x0C000591;
-
-u32 D_8003C85C = 0xEAFE147F;
-
-u32 D_8003C860 = 0x21290004;
-
-u32 D_8003C864 = 0x2021FFFC;
-
-u32 D_8003C868 = 0x1820FF1A;
-
-u32 D_8003C86C = 0x000758C0;
-
-u32 D_8003C870 = 0x006B1020;
-
-u32 D_8003C874 = 0x34240000;
-
-u32 D_8003C878 = 0x080004F8;
-
-u32 D_8003C87C = 0x20010000;
-
-u32 D_8003C880 = 0x0C000455;
-
-u32 D_8003C884 = 0x201403A8;
-
-u32 D_8003C888 = 0x0C0007F6;
-
-u32 D_8003C88C = 0x333300FF;
-
-u32 D_8003C890 = 0x940B0244;
-
-u32 D_8003C894 = 0xCBC41807;
-
-u32 D_8003C898 = 0x800102AE;
-
-u32 D_8003C89C = 0xC9632000;
-
-u32 D_8003C8A0 = 0x0019CDC2;
-
-u32 D_8003C8A4 = 0x0C0007F2;
-
-u32 D_8003C8A8 = 0x87390202;
-
-u32 D_8003C8AC = 0x9283000F;
-
-u32 D_8003C8B0 = 0x8E870014;
-
-u32 D_8003C8B4 = 0x8E850010;
-
-u32 D_8003C8B8 = 0x8C6402B0;
-
-u32 D_8003C8BC = 0x00E04027;
-
-u32 D_8003C8C0 = 0x00A73024;
-
-u32 D_8003C8C4 = 0x00874824;
-
-u32 D_8003C8C8 = 0x11250027;
-
-u32 D_8003C8CC = 0x00884824;
-
-u32 D_8003C8D0 = 0x01264825;
-
-u32 D_8003C8D4 = 0xAC6902B0;
-
-u32 D_8003C8D8 = 0x200BFF81;
-
-u32 D_8003C8DC = 0x15610003;
-
-u32 D_8003C8E0 = 0xA00B02AE;
-
-u32 D_8003C8E4 = 0xEAFD0780;
-
-u32 D_8003C8E8 = 0x22F70008;
-
-u32 D_8003C8EC = 0x9684000A;
-
-u32 D_8003C8F0 = 0xEAFD0800;
-
-u32 D_8003C8F4 = 0x8E980004;
-
-u32 D_8003C8F8 = 0xA6E40002;
-
-u32 D_8003C8FC = 0x0C000455;
-
-u32 D_8003C900 = 0x96830008;
-
-u32 D_8003C904 = 0xAEF80004;
-
-u32 D_8003C908 = 0x82880001;
-
-u32 D_8003C90C = 0x92870002;
-
-u32 D_8003C910 = 0x92850003;
-
-u32 D_8003C914 = 0x20860001;
-
-u32 D_8003C918 = 0x00C83024;
-
-u32 D_8003C91C = 0x000631C0;
-
-u32 D_8003C920 = 0x00073C00;
-
-u32 D_8003C924 = 0x00661825;
-
-u32 D_8003C928 = 0x00671825;
-
-u32 D_8003C92C = 0xAEE30008;
-
-u32 D_8003C930 = 0xEAFD0608;
-
-u32 D_8003C934 = 0x04200003;
-
-u32 D_8003C938 = 0xEAFD1403;
-
-u32 D_8003C93C = 0xEAFD0710;
-
-u32 D_8003C940 = 0x22F70008;
-
-u32 D_8003C944 = 0x9681000C;
-
-u32 D_8003C948 = 0x00052E00;
-
-u32 D_8003C94C = 0xAEE50010;
-
-u32 D_8003C950 = 0x00042380;
-
-u32 D_8003C954 = 0x00812025;
-
-u32 D_8003C958 = 0xAEE40014;
-
-u32 D_8003C95C = 0xEAFD0414;
-
-u32 D_8003C960 = 0x0C000591;
-
-u32 D_8003C964 = 0x22F70018;
-
-u32 D_8003C968 = 0x03200008;
-
-u32 D_8003C96C = 0x201403C0;
-
-u32 D_8003C970 = 0x201F11A4;
-
-u32 D_8003C974 = 0x02F65822;
-
-u32 D_8003C978 = 0x19600262;
-
-u32 D_8003C97C = 0x400C3000;
-
-u32 D_8003C980 = 0x8C1800F0;
-
-u32 D_8003C984 = 0x25730320;
-
-u32 D_8003C988 = 0x1580FFFC;
-
-u32 D_8003C98C = 0x8C0C0FEC;
-
-u32 D_8003C990 = 0x40984800;
-
-u32 D_8003C994 = 0x03135820;
-
-u32 D_8003C998 = 0x018B6022;
-
-u32 D_8003C99C = 0x05810008;
-
-u32 D_8003C9A0 = 0x400B5800;
-
-u32 D_8003C9A4 = 0x316B0400;
-
-u32 D_8003C9A8 = 0x1560FFFD;
-
-u32 D_8003C9AC = 0x8C180FE8;
-
-u32 D_8003C9B0 = 0x400B5000;
-
-u32 D_8003C9B4 = 0x1178FFFE;
-
-u32 D_8003C9B8 = 0x00000000;
-
-u32 D_8003C9BC = 0x40984000;
-
-u32 D_8003C9C0 = 0x400B5000;
-
-u32 D_8003C9C4 = 0x01785822;
-
-u32 D_8003C9C8 = 0x19600002;
-
-u32 D_8003C9CC = 0x01735822;
-
-u32 D_8003C9D0 = 0x1960FFFB;
-
-u32 D_8003C9D4 = 0x03135820;
-
-u32 D_8003C9D8 = 0xAC0B00F0;
-
-u32 D_8003C9DC = 0x2273FFFF;
-
-u32 D_8003C9E0 = 0x22D4DCE0;
-
-u32 D_8003C9E4 = 0x3AD60C00;
-
-u32 D_8003C9E8 = 0x080007F6;
-
-u32 D_8003C9EC = 0x22D7FCE0;
-
-u32 D_8003C9F0 = 0x0C000455;
-
-u32 D_8003C9F4 = 0x20130017;
-
-u32 D_8003C9F8 = 0x0C0007F6;
-
-u32 D_8003C9FC = 0x201403C0;
-
-u32 D_8003CA00 = 0x940B0244;
-
-u32 D_8003CA04 = 0xCBC41807;
-
-u32 D_8003CA08 = 0x0C0007F2;
-
-u32 D_8003CA0C = 0xC9631800;
-
-u32 D_8003CA10 = 0xCA822000;
-
-u32 D_8003CA14 = 0x900201E8;
-
-u32 D_8003CA18 = 0x800301EC;
-
-u32 D_8003CA1C = 0x00431026;
-
-u32 D_8003CA20 = 0x30420180;
-
-u32 D_8003CA24 = 0x000210C2;
-
-u32 D_8003CA28 = 0x204102C0;
-
-u32 D_8003CA2C = 0x4B2411D1;
-
-u32 D_8003CA30 = 0x4B432190;
-
-u32 D_8003CA34 = 0x4B641150;
-
-u32 D_8003CA38 = 0x4B224230;
-
-u32 D_8003CA3C = 0x4B1B4132;
-
-u32 D_8003CA40 = 0x4BA26230;
-
-u32 D_8003CA44 = 0x4B1B6132;
-
-u32 D_8003CA48 = 0x4A044214;
-
-u32 D_8003CA4C = 0x4B7D2110;
-
-u32 D_8003CA50 = 0x4A883EC4;
-
-u32 D_8003CA54 = 0x4A84388D;
-
-u32 D_8003CA58 = 0x4B7DDA4E;
-
-u32 D_8003CA5C = 0x4B1E16C5;
-
-u32 D_8003CA60 = 0x4B1E488C;
-
-u32 D_8003CA64 = 0xC8292000;
-
-u32 D_8003CA68 = 0x928B0017;
-
-u32 D_8003CA6C = 0x48CB0800;
-
-u32 D_8003CA70 = 0x4A000AA7;
-
-u32 D_8003CA74 = 0x4A8A4A2C;
-
-u32 D_8003CA78 = 0x4B060EC7;
-
-u32 D_8003CA7C = 0x4AC7410F;
-
-u32 D_8003CA80 = 0x4A850EC7;
-
-u32 D_8003CA84 = 0x4AC2498F;
-
-u32 D_8003CA88 = 0x4B7D01CF;
-
-u32 D_8003CA8C = 0xEBC62000;
-
-u32 D_8003CA90 = 0xCBC61C00;
-
-u32 D_8003CA94 = 0xCBC71801;
-
-u32 D_8003CA98 = 0xCBC22002;
-
-u32 D_8003CA9C = 0xCBC50818;
-
-u32 D_8003CAA0 = 0xCBC50C19;
-
-u32 D_8003CAA4 = 0x4B632950;
-
-u32 D_8003CAA8 = 0x4B032968;
-
-u32 D_8003CAAC = 0x4A850EC7;
-
-u32 D_8003CAB0 = 0x4AA236CD;
-
-u32 D_8003CAB4 = 0x4A8236CF;
-
-u32 D_8003CAB8 = 0x4AE23ECD;
-
-u32 D_8003CABC = 0x4AC23A0F;
-
-u32 D_8003CAC0 = 0xCBC21011;
-
-u32 D_8003CAC4 = 0xCBC21412;
-
-u32 D_8003CAC8 = 0x4A8246E3;
-
-u32 D_8003CACC = 0x484B0800;
-
-u32 D_8003CAD0 = 0x316C000F;
-
-u32 D_8003CAD4 = 0x1180FE7F;
-
-u32 D_8003CAD8 = 0x316C00F0;
-
-u32 D_8003CADC = 0x1180FE7D;
-
-u32 D_8003CAE0 = 0x4AA246E0;
-
-u32 D_8003CAE4 = 0x484B0800;
-
-u32 D_8003CAE8 = 0x316C000F;
-
-u32 D_8003CAEC = 0x1180FE79;
-
-u32 D_8003CAF0 = 0x316C00F0;
-
-u32 D_8003CAF4 = 0x1180FE77;
-
-u32 D_8003CAF8 = 0x4B7D4090;
-
-u32 D_8003CAFC = 0xEBC81C00;
-
-u32 D_8003CB00 = 0x4A48416C;
-
-u32 D_8003CB04 = 0xCBC21800;
-
-u32 D_8003CB08 = 0x4A4421AC;
-
-u32 D_8003CB0C = 0x4A6210A0;
-
-u32 D_8003CB10 = 0x4A652967;
-
-u32 D_8003CB14 = 0x4A6631A7;
-
-u32 D_8003CB18 = 0x4A45422C;
-
-u32 D_8003CB1C = 0x4A46212C;
-
-u32 D_8003CB20 = 0x900102AF;
-
-u32 D_8003CB24 = 0x38210002;
-
-u32 D_8003CB28 = 0xA00102AF;
-
-u32 D_8003CB2C = 0x800C02AE;
-
-u32 D_8003CB30 = 0xEAFD0780;
-
-u32 D_8003CB34 = 0x142C0002;
-
-u32 D_8003CB38 = 0xA00102AE;
-
-u32 D_8003CB3C = 0x26F70008;
-
-u32 D_8003CB40 = 0x94070246;
-
-u32 D_8003CB44 = 0x92820014;
-
-u32 D_8003CB48 = 0x92830015;
-
-u32 D_8003CB4C = 0x96840010;
-
-u32 D_8003CB50 = 0x96850012;
-
-u32 D_8003CB54 = 0x92860016;
-
-u32 D_8003CB58 = 0x8CE70004;
-
-u32 D_8003CB5C = 0x00021540;
-
-u32 D_8003CB60 = 0x00031CC0;
-
-u32 D_8003CB64 = 0x00431025;
-
-u32 D_8003CB68 = 0x00042240;
-
-u32 D_8003CB6C = 0x00441025;
-
-u32 D_8003CB70 = 0x00451025;
-
-u32 D_8003CB74 = 0xAEE20000;
-
-u32 D_8003CB78 = 0xEAFD0600;
-
-u32 D_8003CB7C = 0x00063500;
-
-u32 D_8003CB80 = 0x00C73025;
-
-u32 D_8003CB84 = 0xAEE60004;
-
-u32 D_8003CB88 = 0xA2E10004;
-
-u32 D_8003CB8C = 0x96840004;
-
-u32 D_8003CB90 = 0xEAFD1202;
-
-u32 D_8003CB94 = 0x9685000C;
-
-u32 D_8003CB98 = 0x30847FF8;
-
-u32 D_8003CB9C = 0x00042240;
-
-u32 D_8003CBA0 = 0x000528C3;
-
-u32 D_8003CBA4 = 0x00851825;
-
-u32 D_8003CBA8 = 0x2063BFFC;
-
-u32 D_8003CBAC = 0xAEE3000C;
-
-u32 D_8003CBB0 = 0xA2E1000C;
-
-u32 D_8003CBB4 = 0xEAFD0410;
-
-u32 D_8003CBB8 = 0x26F70018;
-
-u32 D_8003CBBC = 0x4B3E01EB;
-
-u32 D_8003CBC0 = 0x200B0020;
-
-u32 D_8003CBC4 = 0x48CB0800;
-
-u32 D_8003CBC8 = 0x200C00E7;
-
-u32 D_8003CBCC = 0x4AC838A7;
-
-u32 D_8003CBD0 = 0x4A8838E7;
-
-u32 D_8003CBD4 = 0x4AA83967;
-
-u32 D_8003CBD8 = 0x4AE839A7;
-
-u32 D_8003CBDC = 0x48CC0800;
-
-u32 D_8003CBE0 = 0x4AC410A7;
-
-u32 D_8003CBE4 = 0x4A8418E7;
-
-u32 D_8003CBE8 = 0x4AA42967;
-
-u32 D_8003CBEC = 0x4AE431A7;
-
-u32 D_8003CBF0 = 0x4A022911;
-
-u32 D_8003CBF4 = 0x4A0219D1;
-
-u32 D_8003CBF8 = 0x4A031211;
-
-u32 D_8003CBFC = 0x4A032A51;
-
-u32 D_8003CC00 = 0x4BC726C7;
-
-u32 D_8003CC04 = 0x4BC446CF;
-
-u32 D_8003CC08 = 0x4B0A529D;
-
-u32 D_8003CC0C = 0x4B2B5ADD;
-
-u32 D_8003CC10 = 0x4BC44370;
-
-u32 D_8003CC14 = 0x4B7D4332;
-
-u32 D_8003CC18 = 0x4A0A56D3;
-
-u32 D_8003CC1C = 0x4BC74B70;
-
-u32 D_8003CC20 = 0x4B7D4B32;
-
-u32 D_8003CC24 = 0x4BDFDEE0;
-
-u32 D_8003CC28 = 0x4B1EF3A7;
-
-u32 D_8003CC2C = 0x4BC95370;
-
-u32 D_8003CC30 = 0x4B7D5332;
-
-u32 D_8003CC34 = 0x4B2E5EC4;
-
-u32 D_8003CC38 = 0x4B2E528D;
-
-u32 D_8003CC3C = 0x4B7DFACE;
-
-u32 D_8003CC40 = 0x4B444273;
-
-u32 D_8003CC44 = 0x4B474A73;
-
-u32 D_8003CC48 = 0x4A0C6B54;
-
-u32 D_8003CC4C = 0x4B7D6310;
-
-u32 D_8003CC50 = 0x4BC44473;
-
-u32 D_8003CC54 = 0x4BC74C73;
-
-u32 D_8003CC58 = 0x4BC95473;
-
-u32 D_8003CC5C = 0x4B5F0490;
-
-u32 D_8003CC60 = 0x4B5F6B46;
-
-u32 D_8003CC64 = 0x4B5F630F;
-
-u32 D_8003CC68 = 0x4A116EC4;
-
-u32 D_8003CC6C = 0x4A11640D;
-
-u32 D_8003CC70 = 0x4B7DFBCE;
-
-u32 D_8003CC74 = 0x4A0F03D5;
-
-u32 D_8003CC78 = 0x4A109411;
-
-u32 D_8003CC7C = 0x4A0F6EC4;
-
-u32 D_8003CC80 = 0x4A0F66CD;
-
-u32 D_8003CC84 = 0x4A106B4E;
-
-u32 D_8003CC88 = 0x4A10630F;
-
-u32 D_8003CC8C = 0x48025180;
-
-u32 D_8003CC90 = 0x30420080;
-
-u32 D_8003CC94 = 0x00411025;
-
-u32 D_8003CC98 = 0x34420A00;
-
-u32 D_8003CC9C = 0x4B4A46F2;
-
-u32 D_8003CCA0 = 0x4B4B5431;
-
-u32 D_8003CCA4 = 0x4B7D53F2;
-
-u32 D_8003CCA8 = 0x4A096EC6;
-
-u32 D_8003CCAC = 0x4A0966CF;
-
-u32 D_8003CCB0 = 0x4B2A529D;
-
-u32 D_8003CCB4 = 0x4B0B5ADD;
-
-u32 D_8003CCB8 = 0x4BC726C7;
-
-u32 D_8003CCBC = 0x4BC446CF;
-
-u32 D_8003CCC0 = 0x4B2C631D;
-
-u32 D_8003CCC4 = 0x4B0D6B5D;
-
-u32 D_8003CCC8 = 0x4B443EC7;
-
-u32 D_8003CCCC = 0x4B4826CF;
-
-u32 D_8003CCD0 = 0x4B318C5D;
-
-u32 D_8003CCD4 = 0x4B12949D;
-
-u32 D_8003CCD8 = 0x4B5066C4;
-
-u32 D_8003CCDC = 0x4B506ECD;
-
-u32 D_8003CCE0 = 0x4B4F630E;
-
-u32 D_8003CCE4 = 0x4B4F6B4F;
-
-u32 D_8003CCE8 = 0x4B508EC4;
-
-u32 D_8003CCEC = 0x4B5096CD;
-
-u32 D_8003CCF0 = 0x4B4F8C4E;
-
-u32 D_8003CCF4 = 0x4B4F948F;
-
-u32 D_8003CCF8 = 0x48031A00;
-
-u32 D_8003CCFC = 0x4B9F03D0;
-
-u32 D_8003CD00 = 0x4B9F56C4;
-
-u32 D_8003CD04 = 0x4B9F5A4D;
-
-u32 D_8003CD08 = 0x00031B80;
-
-u32 D_8003CD0C = 0x4B7D020E;
-
-u32 D_8003CD10 = 0x4BBF1128;
-
-u32 D_8003CD14 = 0x4A040111;
-
-u32 D_8003CD18 = 0x4B7D5EE0;
-
-u32 D_8003CD1C = 0x4B3D0427;
-
-u32 D_8003CD20 = 0x4B427EC6;
-
-u32 D_8003CD24 = 0x4BC4420E;
-
-u32 D_8003CD28 = 0x4BC44A4F;
-
-u32 D_8003CD2C = 0x4BC47906;
-
-u32 D_8003CD30 = 0x4B7D01CF;
-
-u32 D_8003CD34 = 0x4B1044D4;
-
-u32 D_8003CD38 = 0x4B7D4D10;
-
-u32 D_8003CD3C = 0x4B2E66C4;
-
-u32 D_8003CD40 = 0x4B2E6B4D;
-
-u32 D_8003CD44 = 0x4B7DFB0E;
-
-u32 D_8003CD48 = 0x4B2E8EC4;
-
-u32 D_8003CD4C = 0x4B2E948D;
-
-u32 D_8003CD50 = 0x4B7DFC4E;
-
-u32 D_8003CD54 = 0x4BDF6306;
-
-u32 D_8003CD58 = 0x4BDF6B4F;
-
-u32 D_8003CD5C = 0x4BDF8C46;
-
-u32 D_8003CD60 = 0x4BDF948F;
-
-u32 D_8003CD64 = 0x4B0A66C4;
-
-u32 D_8003CD68 = 0x4B0A6ECD;
-
-u32 D_8003CD6C = 0x4B0B66CE;
-
-u32 D_8003CD70 = 0x4B0B6ECF;
-
-u32 D_8003CD74 = 0x4B3D8B8E;
-
-u32 D_8003CD78 = 0x4B3D93CF;
-
-u32 D_8003CD7C = 0x4B3D16C7;
-
-u32 D_8003CD80 = 0x4BC476CC;
-
-u32 D_8003CD84 = 0x4BC47ECD;
-
-u32 D_8003CD88 = 0x4BC7740E;
-
-u32 D_8003CD8C = 0x4BC77D4F;
-
-u32 D_8003CD90 = 0xA6E20000;
-
-u32 D_8003CD94 = 0xEAE50E01;
-
-u32 D_8003CD98 = 0xEAE30E02;
-
-u32 D_8003CD9C = 0xEAE20E03;
-
-u32 D_8003CDA0 = 0xAEE30008;
-
-u32 D_8003CDA4 = 0xEAEB0A06;
-
-u32 D_8003CDA8 = 0xEAEA0A07;
-
-u32 D_8003CDAC = 0xEAE90808;
-
-u32 D_8003CDB0 = 0xEAE80809;
-
-u32 D_8003CDB4 = 0xEAEB080A;
-
-u32 D_8003CDB8 = 0xEAEA080B;
-
-u32 D_8003CDBC = 0xEAF4090C;
-
-u32 D_8003CDC0 = 0xEAF3090D;
-
-u32 D_8003CDC4 = 0xEAEB090E;
-
-u32 D_8003CDC8 = 0xEAEA090F;
-
-u32 D_8003CDCC = 0xEAF51B04;
-
-u32 D_8003CDD0 = 0xEAED1B05;
-
-u32 D_8003CDD4 = 0xEAF01B06;
-
-u32 D_8003CDD8 = 0xEAEC1B07;
-
-u32 D_8003CDDC = 0xEAEF1B08;
-
-u32 D_8003CDE0 = 0xEAF21B09;
-
-u32 D_8003CDE4 = 0xEAEE1B0A;
-
-u32 D_8003CDE8 = 0xEAF11B0B;
-
-u32 D_8003CDEC = 0x1320FEE0;
-
-u32 D_8003CDF0 = 0x26F70060;
-
-u32 D_8003CDF4 = 0x4B7D1890;
-
-u32 D_8003CDF8 = 0x20190000;
-
-u32 D_8003CDFC = 0x4B7D28D0;
-
-u32 D_8003CE00 = 0x08000630;
-
-u32 D_8003CE04 = 0x4B7D3150;
-
-u32 D_8003CE08 = 0x00000000;
-
-u32 D_8003CE0C = 0x00000000;
-
-u32 D_8003CE10 = 0x00000000;
-
-u32 D_8003CE14 = 0x00000000;
-
-u32 D_8003CE18 = 0x00000000;
-
-u32 D_8003CE1C = 0x00000000;
-
-u32 D_8003CE20 = 0x00000000;
-
-u32 D_8003CE24 = 0x00000000;
-
-u32 D_8003CE28 = 0x00000000;
-
-u32 D_8003CE2C = 0x00000000;
-
-u32 D_8003CE30 = 0x00000000;
-
-u32 D_8003CE34 = 0x00000000;
-
-u32 D_8003CE38 = 0x00000000;
-
-u32 D_8003CE3C = 0x00000000;
-
-u32 D_8003CE40 = 0x00000000;
-
-u32 D_8003CE44 = 0x00000000;
-
-u32 D_8003CE48 = 0x00000000;
-
-u32 D_8003CE4C = 0x00000000;
-
-u32 D_8003CE50 = 0x00000000;
-
-u32 D_8003CE54 = 0x00000000;
-
-u32 D_8003CE58 = 0x00000000;
-
-u32 D_8003CE5C = 0x00000000;
-
-u32 D_8003CE60 = 0x00000000;
-
-u32 D_8003CE64 = 0x00000000;
-
-u32 D_8003CE68 = 0x00000000;
-
-u32 D_8003CE6C = 0x00000000;
-
-u32 D_8003CE70 = 0x00000000;
-
-u32 D_8003CE74 = 0x00000000;
-
-u32 D_8003CE78 = 0x00000000;
-
-u32 D_8003CE7C = 0x00000000;
-
-u32 D_8003CE80 = 0x00000000;
-
-u32 D_8003CE84 = 0x00000000;
-
-u32 D_8003CE88 = 0x00000000;
-
-u32 D_8003CE8C = 0x00000000;
-
-u32 D_8003CE90 = 0x00000000;
-
-u32 D_8003CE94 = 0x00000000;
-
-u32 D_8003CE98 = 0x00000000;
-
-u32 D_8003CE9C = 0x00000000;
-
-u32 D_8003CEA0 = 0x00000000;
-
-u32 D_8003CEA4 = 0x200C1B70;
-
-u32 D_8003CEA8 = 0x080007ED;
-
-u32 D_8003CEAC = 0x200B031C;
-
-u32 D_8003CEB0 = 0x0C000455;
-
-u32 D_8003CEB4 = 0x20130017;
-
-u32 D_8003CEB8 = 0x0C0007F6;
-
-u32 D_8003CEBC = 0x201403C0;
-
-u32 D_8003CEC0 = 0x940B0244;
-
-u32 D_8003CEC4 = 0xCBC41807;
-
-u32 D_8003CEC8 = 0x0C0007F2;
-
-u32 D_8003CECC = 0xC9631800;
-
-u32 D_8003CED0 = 0x07200077;
-
-u32 D_8003CED4 = 0xCA822000;
-
-u32 D_8003CED8 = 0x4B441150;
-
-u32 D_8003CEDC = 0x4B032968;
-
-u32 D_8003CEE0 = 0x4B7D01A8;
-
-u32 D_8003CEE4 = 0x4B241211;
-
-u32 D_8003CEE8 = 0x4B0401D0;
-
-u32 D_8003CEEC = 0x4B224270;
-
-u32 D_8003CEF0 = 0x4B1B4132;
-
-u32 D_8003CEF4 = 0x4BA26270;
-
-u32 D_8003CEF8 = 0x4B1B6132;
-
-u32 D_8003CEFC = 0x4A044A54;
-
-u32 D_8003CF00 = 0x4B7D2110;
-
-u32 D_8003CF04 = 0x4B1E42C6;
-
-u32 D_8003CF08 = 0x4B7D028F;
-
-u32 D_8003CF0C = 0x4A895EC4;
-
-u32 D_8003CF10 = 0x4A8956CD;
-
-u32 D_8003CF14 = 0x4A845ACE;
-
-u32 D_8003CF18 = 0x4A84528F;
-
-u32 D_8003CF1C = 0x940B0246;
-
-u32 D_8003CF20 = 0xC9641800;
-
-u32 D_8003CF24 = 0x4AC83A50;
-
-u32 D_8003CF28 = 0x4B7F0A4E;
-
-u32 D_8003CF2C = 0x4B044A46;
-
-u32 D_8003CF30 = 0x4B241306;
-
-u32 D_8003CF34 = 0x928B0017;
-
-u32 D_8003CF38 = 0x316B0011;
-
-u32 D_8003CF3C = 0x256B0077;
-
-u32 D_8003CF40 = 0x48CB0800;
-
-u32 D_8003CF44 = 0x4A0939E7;
-
-u32 D_8003CF48 = 0x4A0C1367;
-
-u32 D_8003CF4C = 0x4B4339D0;
-
-u32 D_8003CF50 = 0x4ACB36D4;
-
-u32 D_8003CF54 = 0x4ACA2890;
-
-u32 D_8003CF58 = 0xCBC81011;
-
-u32 D_8003CF5C = 0xCBC81412;
-
-u32 D_8003CF60 = 0x4AA810A0;
-
-u32 D_8003CF64 = 0x4A882A51;
-
-u32 D_8003CF68 = 0x4A882963;
-
-u32 D_8003CF6C = 0x4A090267;
-
-u32 D_8003CF70 = 0x4A022EE3;
-
-u32 D_8003CF74 = 0x48410800;
-
-u32 D_8003CF78 = 0x30210011;
-
-u32 D_8003CF7C = 0x1420FD55;
-
-u32 D_8003CF80 = 0x4AAD4EC7;
-
-u32 D_8003CF84 = 0x4B2B5ADD;
-
-u32 D_8003CF88 = 0x4B0A529D;
-
-u32 D_8003CF8C = 0x4B5E5EC4;
-
-u32 D_8003CF90 = 0x4B5E56CD;
-
-u32 D_8003CF94 = 0x4B7F5ACE;
-
-u32 D_8003CF98 = 0x4AC709CE;
-
-u32 D_8003CF9C = 0x4B232A28;
-
-u32 D_8003CFA0 = 0x4AFE4207;
-
-u32 D_8003CFA4 = 0x4A886A05;
-
-u32 D_8003CFA8 = 0x4AA839D1;
-
-u32 D_8003CFAC = 0x48012800;
-
-u32 D_8003CFB0 = 0x48022C00;
-
-u32 D_8003CFB4 = 0x48031000;
-
-u32 D_8003CFB8 = 0x48041400;
-
-u32 D_8003CFBC = 0x30420FFF;
-
-u32 D_8003CFC0 = 0x00010B00;
-
-u32 D_8003CFC4 = 0x00220825;
-
-u32 D_8003CFC8 = 0x30840FFF;
-
-u32 D_8003CFCC = 0x00031B00;
-
-u32 D_8003CFD0 = 0x00641825;
-
-u32 D_8003CFD4 = 0x48056900;
-
-u32 D_8003CFD8 = 0x900200C9;
-
-u32 D_8003CFDC = 0x30420020;
-
-u32 D_8003CFE0 = 0x10400003;
-
-u32 D_8003CFE4 = 0x00000000;
-
-u32 D_8003CFE8 = 0x00052880;
-
-u32 D_8003CFEC = 0x2063BFFC;
-
-u32 D_8003CFF0 = 0x900202AF;
-
-u32 D_8003CFF4 = 0x38420002;
-
-u32 D_8003CFF8 = 0xA00202AF;
-
-u32 D_8003CFFC = 0x800C02AE;
-
-u32 D_8003D000 = 0xEAFD0780;
-
-u32 D_8003D004 = 0x144C0002;
-
-u32 D_8003D008 = 0xA00202AE;
-
-u32 D_8003D00C = 0x26F70008;
-
-u32 D_8003D010 = 0xEBC41200;
-
-u32 D_8003D014 = 0x8FCA0000;
-
-u32 D_8003D018 = 0x92840014;
-
-u32 D_8003D01C = 0x92860015;
-
-u32 D_8003D020 = 0x96870010;
-
-u32 D_8003D024 = 0x96880012;
-
-u32 D_8003D028 = 0x92890016;
-
-u32 D_8003D02C = 0x00042540;
-
-u32 D_8003D030 = 0x000634C0;
-
-u32 D_8003D034 = 0x00862025;
-
-u32 D_8003D038 = 0x00073A40;
-
-u32 D_8003D03C = 0x00872025;
-
-u32 D_8003D040 = 0x00882025;
-
-u32 D_8003D044 = 0xAEE40000;
-
-u32 D_8003D048 = 0xEAFD0600;
-
-u32 D_8003D04C = 0x00094D00;
-
-u32 D_8003D050 = 0x012A4825;
-
-u32 D_8003D054 = 0xAEE90004;
-
-u32 D_8003D058 = 0xA2E20004;
-
-u32 D_8003D05C = 0x96870004;
-
-u32 D_8003D060 = 0xEAFD1202;
-
-u32 D_8003D064 = 0x9688000C;
-
-u32 D_8003D068 = 0x30E77FF8;
-
-u32 D_8003D06C = 0x00073A40;
-
-u32 D_8003D070 = 0x000840C3;
-
-u32 D_8003D074 = 0x00E83025;
-
-u32 D_8003D078 = 0x20C6BFFC;
-
-u32 D_8003D07C = 0xAEE6000C;
-
-u32 D_8003D080 = 0xA2E2000C;
-
-u32 D_8003D084 = 0xAEE30010;
-
-u32 D_8003D088 = 0xEAFD0690;
-
-u32 D_8003D08C = 0xAEE10014;
-
-u32 D_8003D090 = 0xA2E20014;
-
-u32 D_8003D094 = 0xEAE7080C;
-
-u32 D_8003D098 = 0xEAE70C0D;
-
-u32 D_8003D09C = 0xA6E5001C;
-
-u32 D_8003D0A0 = 0xEAED0D0F;
-
-u32 D_8003D0A4 = 0x08000590;
-
-u32 D_8003D0A8 = 0x26F70020;
-
-u32 D_8003D0AC = 0xCA822000;
-
-u32 D_8003D0B0 = 0xCBC7091A;
-
-u32 D_8003D0B4 = 0xCBC70D1B;
-
-u32 D_8003D0B8 = 0xCBC80818;
-
-u32 D_8003D0BC = 0xCBC80C19;
-
-u32 D_8003D0C0 = 0x4B1F1245;
-
-u32 D_8003D0C4 = 0x4B7DFA8E;
-
-u32 D_8003D0C8 = 0x4B274330;
-
-u32 D_8003D0CC = 0x4B1B42F2;
-
-u32 D_8003D0D0 = 0x4BA76330;
-
-u32 D_8003D0D4 = 0x4B1B62F2;
-
-u32 D_8003D0D8 = 0x4A0B0EC6;
-
-u32 D_8003D0DC = 0x4B3D630E;
-
-u32 D_8003D0E0 = 0x4B3D5ACF;
-
-u32 D_8003D0E4 = 0x4B3F39C5;
-
-u32 D_8003D0E8 = 0x4B7DFB4E;
-
-u32 D_8003D0EC = 0x4A026EC6;
-
-u32 D_8003D0F0 = 0x4A0239CF;
-
-u32 D_8003D0F4 = 0x4A8C56C4;
-
-u32 D_8003D0F8 = 0x4A8C4ECD;
-
-u32 D_8003D0FC = 0x4A8B518E;
-
-u32 D_8003D100 = 0x4A8B494F;
-
-u32 D_8003D104 = 0x4B444210;
-
-u32 D_8003D108 = 0x4B034228;
-
-u32 D_8003D10C = 0x4B3D294E;
-
-u32 D_8003D110 = 0x4B2748B3;
-
-u32 D_8003D114 = 0x080006ED;
-
-u32 D_8003D118 = 0x4BA768B3;
-
-u32 D_8003D11C = 0x00000000;
-
-u32 D_8003D120 = 0x00000000;
-
-u32 D_8003D124 = 0x00000000;
-
-u32 D_8003D128 = 0x00000000;
-
-u32 D_8003D12C = 0x00000000;
-
-u32 D_8003D130 = 0x00000000;
-
-u32 D_8003D134 = 0x00000000;
-
-u32 D_8003D138 = 0x00000000;
-
-u32 D_8003D13C = 0x00000000;
-
-u32 D_8003D140 = 0x00000000;
-
-u32 D_8003D144 = 0x00000000;
-
-u32 D_8003D148 = 0x00000000;
-
-u32 D_8003D14C = 0x00000000;
-
-u32 D_8003D150 = 0x00000000;
-
-u32 D_8003D154 = 0x00000000;
-
-u32 D_8003D158 = 0x00000000;
-
-u32 D_8003D15C = 0x00000000;
-
-u32 D_8003D160 = 0x00000000;
-
-u32 D_8003D164 = 0x00000000;
-
-u32 D_8003D168 = 0x00000000;
-
-u32 D_8003D16C = 0x00000000;
-
-u32 D_8003D170 = 0x00000000;
-
-u32 D_8003D174 = 0x00000000;
-
-u32 D_8003D178 = 0x00000000;
-
-u32 D_8003D17C = 0x00000000;
-
-u32 D_8003D180 = 0x00000000;
-
-u32 D_8003D184 = 0x00000000;
-
-u32 D_8003D188 = 0x00000000;
-
-u32 D_8003D18C = 0x00000000;
-
-u32 D_8003D190 = 0x00000000;
-
-u32 D_8003D194 = 0x00000000;
-
-u32 D_8003D198 = 0x00000000;
-
-u32 D_8003D19C = 0x00000000;
-
-u32 D_8003D1A0 = 0x00000000;
-
-u32 D_8003D1A4 = 0x00000000;
-
-u32 D_8003D1A8 = 0x00000000;
-
-u32 D_8003D1AC = 0x00000000;
-
-u32 D_8003D1B0 = 0x00000000;
-
-u32 D_8003D1B4 = 0x00000000;
-
-u32 D_8003D1B8 = 0x00000000;
-
-u32 D_8003D1BC = 0x00000000;
-
-u32 D_8003D1C0 = 0x00000000;
-
-u32 D_8003D1C4 = 0x00000000;
-
-u32 D_8003D1C8 = 0x00000000;
-
-u32 D_8003D1CC = 0x00000000;
-
-u32 D_8003D1D0 = 0x00000000;
-
-u32 D_8003D1D4 = 0x00000000;
-
-u32 D_8003D1D8 = 0x00000000;
-
-u32 D_8003D1DC = 0x00000000;
-
-u32 D_8003D1E0 = 0x00000000;
-
-u32 D_8003D1E4 = 0x00000000;
-
-u32 D_8003D1E8 = 0x00000000;
-
-u32 D_8003D1EC = 0x00000000;
-
-u32 D_8003D1F0 = 0x00000000;
-
-u32 D_8003D1F4 = 0x00000000;
-
-u32 D_8003D1F8 = 0x00000000;
-
-u32 D_8003D1FC = 0x00000000;
-
-u32 D_8003D200 = 0x00000000;
-
-u32 D_8003D204 = 0x00000000;
-
-u32 D_8003D208 = 0x00000000;
-
-u32 D_8003D20C = 0x00000000;
-
-u32 D_8003D210 = 0x00000000;
-
-u32 D_8003D214 = 0x00000000;
-
-u32 D_8003D218 = 0x00000000;
-
-u32 D_8003D21C = 0x00000000;
-
-u32 D_8003D220 = 0x00000000;
-
-u32 D_8003D224 = 0x00000000;
-
-u32 D_8003D228 = 0x00000000;
-
-u32 D_8003D22C = 0x00000000;
-
-u32 D_8003D230 = 0x00000000;
-
-u32 D_8003D234 = 0x00000000;
-
-u32 D_8003D238 = 0x00000000;
-
-u32 D_8003D23C = 0x00000000;
-
-u32 D_8003D240 = 0x00000000;
-
-u32 D_8003D244 = 0x00000000;
-
-u32 D_8003D248 = 0x00000000;
-
-u32 D_8003D24C = 0x00000000;
-
-u32 D_8003D250 = 0x00000000;
-
-u32 D_8003D254 = 0x00000000;
-
-u32 D_8003D258 = 0x00000000;
-
-u32 D_8003D25C = 0x00000000;
-
-u32 D_8003D260 = 0x00000000;
-
-u32 D_8003D264 = 0x00000000;
-
-u32 D_8003D268 = 0x00000000;
-
-u32 D_8003D26C = 0x00000000;
-
-u32 D_8003D270 = 0x00000000;
-
-u32 D_8003D274 = 0x00000000;
-
-u32 D_8003D278 = 0x00000000;
-
-u32 D_8003D27C = 0x00000000;
-
-u32 D_8003D280 = 0x00000000;
-
-u32 D_8003D284 = 0x00000000;
-
-u32 D_8003D288 = 0x00000000;
-
-u32 D_8003D28C = 0x00000000;
-
-u32 D_8003D290 = 0x00000000;
-
-u32 D_8003D294 = 0x00000000;
-
-u32 D_8003D298 = 0x00000000;
-
-u32 D_8003D29C = 0x00000000;
-
-u32 D_8003D2A0 = 0x00000000;
-
-u32 D_8003D2A4 = 0x00000000;
-
-u32 D_8003D2A8 = 0x00000000;
-
-u32 D_8003D2AC = 0x00000000;
-
-u32 D_8003D2B0 = 0x00000000;
-
-u32 D_8003D2B4 = 0x00000000;
-
-u32 D_8003D2B8 = 0x00000000;
-
-u32 D_8003D2BC = 0x00000000;
-
-u32 D_8003D2C0 = 0x00000000;
-
-u32 D_8003D2C4 = 0x00000000;
-
-u32 D_8003D2C8 = 0x00000000;
-
-u32 D_8003D2CC = 0x00000000;
-
-u32 D_8003D2D0 = 0x00000000;
-
-u32 D_8003D2D4 = 0x00000000;
-
-u32 D_8003D2D8 = 0x00000000;
-
-u32 D_8003D2DC = 0x200C1000;
-
-u32 D_8003D2E0 = 0x200B0304;
-
-u32 D_8003D2E4 = 0x8D780000;
-
-u32 D_8003D2E8 = 0x95730004;
-
-u32 D_8003D2EC = 0x0C0007F6;
-
-u32 D_8003D2F0 = 0x95740006;
-
-u32 D_8003D2F4 = 0x359F0000;
-
-u32 D_8003D2F8 = 0x400B3000;
-
-u32 D_8003D2FC = 0x1560FFFF;
-
-u32 D_8003D300 = 0x400B3000;
-
-u32 D_8003D304 = 0x03E00008;
-
-u32 D_8003D308 = 0x400B2800;
-
-u32 D_8003D30C = 0x1560FFFF;
-
-u32 D_8003D310 = 0x400B2800;
-
-u32 D_8003D314 = 0x40940000;
-
-u32 D_8003D318 = 0x06800003;
-
-u32 D_8003D31C = 0x40980800;
-
-u32 D_8003D320 = 0x03E00008;
-
-u32 D_8003D324 = 0x40931000;
-
-u32 D_8003D328 = 0x03E00008;
-
-u32 D_8003D32C = 0x40931800;
-
-u32 D_8003D330 = 0x02F65822;
-
-u32 D_8003D334 = 0x256C031F;
-
-u32 D_8003D338 = 0x05910190;
-
-u32 D_8003D33C = 0x00000000;
-
-u32 D_8003D340 = 0x0C0007F2;
-
-u32 D_8003D344 = 0x8C1800F0;
-
-u32 D_8003D348 = 0x0420001A;
-
-u32 D_8003D34C = 0x40984800;
-
-u32 D_8003D350 = 0x1420000F;
-
-u32 D_8003D354 = 0x035BD020;
-
-u32 D_8003D358 = 0x8F7804BC;
-
-u32 D_8003D35C = 0xAC1A0FF0;
-
-u32 D_8003D360 = 0xAC180FD0;
-
-u32 D_8003D364 = 0x24141080;
-
-u32 D_8003D368 = 0x0C0007F6;
-
-u32 D_8003D36C = 0x20130F47;
-
-u32 D_8003D370 = 0x8C1800D8;
-
-u32 D_8003D374 = 0x24140180;
-
-u32 D_8003D378 = 0x33330FFF;
-
-u32 D_8003D37C = 0x0314C020;
-
-u32 D_8003D380 = 0x0C0007F6;
-
-u32 D_8003D384 = 0x02749822;
-
-u32 D_8003D388 = 0x080007F2;
-
-u32 D_8003D38C = 0x201F1080;
-
-u32 D_8003D390 = 0x8C0B0FD0;
-
-u32 D_8003D394 = 0xAC1A0BF8;
-
-u32 D_8003D398 = 0xAC0B0BFC;
-
-u32 D_8003D39C = 0x200C5000;
-
-u32 D_8003D3A0 = 0x8C180FF8;
-
-u32 D_8003D3A4 = 0x20148000;
-
-u32 D_8003D3A8 = 0x20130BFF;
-
-u32 D_8003D3AC = 0x080007F6;
-
-u32 D_8003D3B0 = 0x201F1088;
-
-u32 D_8003D3B4 = 0x200C4000;
-
-u32 D_8003D3B8 = 0x408C2000;
-
-u32 D_8003D3BC = 0x0000000D;
-
-u32 D_8003D3C0 = 0x00000000;
-
-u32 D_8003D3C4 = 0x00000000;
-
-u32 D_8003D3C8 = 0x0C000455;
-
-u32 D_8003D3CC = 0x877404B9;
-
-u32 D_8003D3D0 = 0x33330FF8;
-
-u32 D_8003D3D4 = 0x0014A083;
-
-u32 D_8003D3D8 = 0x080007F6;
-
-u32 D_8003D3DC = 0x201F11A0;
-
-u32 D_8003D3E0 = 0xAD79F0B8;
-
-u32 D_8003D3E4 = 0x08000469;
-
-u32 D_8003D3E8 = 0xAD78F0BC;
-
-u32 D_8003D3EC = 0x900100D1;
-
-u32 D_8003D3F0 = 0x8C0300D4;
-
-u32 D_8003D3F4 = 0x03002827;
-
-u32 D_8003D3F8 = 0x8C2202B0;
-
-u32 D_8003D3FC = 0xA41900D0;
-
-u32 D_8003D400 = 0x00782024;
-
-u32 D_8003D404 = 0x00583024;
-
-u32 D_8003D408 = 0x10C30058;
-
-u32 D_8003D40C = 0x00453024;
-
-u32 D_8003D410 = 0x00C43025;
-
-u32 D_8003D414 = 0xAC2602B0;
-
-u32 D_8003D418 = 0x8C1800D0;
-
-u32 D_8003D41C = 0x900100DE;
-
-u32 D_8003D420 = 0x001913C0;
-
-u32 D_8003D424 = 0x0C000455;
-
-u32 D_8003D428 = 0x035B1820;
-
-u32 D_8003D42C = 0x04400048;
-
-u32 D_8003D430 = 0x371A0000;
-
-u32 D_8003D434 = 0xAC230138;
-
-u32 D_8003D438 = 0x20210004;
-
-u32 D_8003D43C = 0x08000462;
-
-u32 D_8003D440 = 0xA00100DE;
-
-u32 D_8003D444 = 0x900100DE;
-
-u32 D_8003D448 = 0x102003CA;
-
-u32 D_8003D44C = 0x2021FFFC;
-
-u32 D_8003D450 = 0x0800041D;
-
-u32 D_8003D454 = 0x8C3A0138;
-
-u32 D_8003D458 = 0x00190C02;
-
-u32 D_8003D45C = 0x94222886;
-
-u32 D_8003D460 = 0x00591020;
-
-u32 D_8003D464 = 0x08000469;
-
-u32 D_8003D468 = 0xAC580000;
-
-u32 D_8003D46C = 0x0C000455;
-
-u32 D_8003D470 = 0x97340388;
-
-u32 D_8003D474 = 0x937304B9;
-
-u32 D_8003D478 = 0x080007F6;
-
-u32 D_8003D47C = 0x201F11A0;
-
-u32 D_8003D480 = 0x8D63F014;
-
-u32 D_8003D484 = 0x3C028000;
-
-u32 D_8003D488 = 0x03221007;
-
-u32 D_8003D48C = 0x00190A02;
-
-u32 D_8003D490 = 0x00221006;
-
-u32 D_8003D494 = 0x00401027;
-
-u32 D_8003D498 = 0x00621824;
-
-u32 D_8003D49C = 0x00781825;
-
-u32 D_8003D4A0 = 0xAD63F014;
-
-u32 D_8003D4A4 = 0x8C1900C8;
-
-u32 D_8003D4A8 = 0x08000450;
-
-u32 D_8003D4AC = 0x8C1800CC;
-
-u32 D_8003D4B0 = 0xA0180268;
-
-u32 D_8003D4B4 = 0x33010008;
-
-u32 D_8003D4B8 = 0x24210248;
-
-u32 D_8003D4BC = 0xA4010244;
-
-u32 D_8003D4C0 = 0x33210008;
-
-u32 D_8003D4C4 = 0x24210258;
-
-u32 D_8003D4C8 = 0xA4010246;
-
-u32 D_8003D4CC = 0x33010018;
-
-u32 D_8003D4D0 = 0x00010842;
-
-u32 D_8003D4D4 = 0x8C230234;
-
-u32 D_8003D4D8 = 0x33010070;
-
-u32 D_8003D4DC = 0x00010882;
-
-u32 D_8003D4E0 = 0x8C220214;
-
-u32 D_8003D4E4 = 0xAC0301FC;
-
-u32 D_8003D4E8 = 0x08000469;
-
-u32 D_8003D4EC = 0xAC0201F8;
-
-u32 D_8003D4F0 = 0xC81B1818;
-
-u32 D_8003D4F4 = 0xC81B1C1A;
-
-u32 D_8003D4F8 = 0x8C1900D8;
-
-u32 D_8003D4FC = 0x22F70010;
-
-u32 D_8003D500 = 0xEAFB187E;
-
-u32 D_8003D504 = 0xEAFB1C7F;
-
-u32 D_8003D508 = 0xAEF80004;
-
-u32 D_8003D50C = 0xAEF90000;
-
-u32 D_8003D510 = 0x08000590;
-
-u32 D_8003D514 = 0x22F70008;
-
-u32 D_8003D518 = 0x201F1140;
-
-u32 D_8003D51C = 0x00185D82;
-
-u32 D_8003D520 = 0x316B003C;
-
-u32 D_8003D524 = 0x8D6B00F8;
-
-u32 D_8003D528 = 0x0018C200;
-
-u32 D_8003D52C = 0x0018C202;
-
-u32 D_8003D530 = 0x03E00008;
-
-u32 D_8003D534 = 0x030BC020;
-
-u32 D_8003D538 = 0xAC1900C8;
-
-u32 D_8003D53C = 0x08000450;
-
-u32 D_8003D540 = 0xAC1800CC;
-
-u32 D_8003D544 = 0x00000000;
-
-u32 D_8003D548 = 0x0C000455;
-
-u32 D_8003D54C = 0x20130017;
-
-u32 D_8003D550 = 0x0C0007F6;
-
-u32 D_8003D554 = 0x201403C0;
-
-u32 D_8003D558 = 0x200C16CC;
-
-u32 D_8003D55C = 0x080007ED;
-
-u32 D_8003D560 = 0x200B0314;
-
-u32 D_8003D564 = 0x00000000;
-
-u32 D_8003D568 = 0x200C1638;
-
-u32 D_8003D56C = 0x080007ED;
-
-u32 D_8003D570 = 0x200B0314;
-
-u32 D_8003D574 = 0x0C0007F6;
-
-u32 D_8003D578 = 0x20130027;
-
-u32 D_8003D57C = 0x200103D8;
-
-u32 D_8003D580 = 0xCBC21211;
-
-u32 D_8003D584 = 0x0C0007F2;
-
-u32 D_8003D588 = 0xCBC21612;
-
-u32 D_8003D58C = 0xCA832000;
-
-u32 D_8003D590 = 0xCA84080E;
-
-u32 D_8003D594 = 0xCA840C0F;
-
-u32 D_8003D598 = 0x4B3F2145;
-
-u32 D_8003D59C = 0x4B7DF98E;
-
-u32 D_8003D5A0 = 0x4B044230;
-
-u32 D_8003D5A4 = 0x4B7D41F2;
-
-u32 D_8003D5A8 = 0x4B846230;
-
-u32 D_8003D5AC = 0x4B7D61F2;
-
-u32 D_8003D5B0 = 0x4B3D0EC6;
-
-u32 D_8003D5B4 = 0x4B1F46CC;
-
-u32 D_8003D5B8 = 0x4B1F39CD;
-
-u32 D_8003D5BC = 0x4B7DFA0E;
-
-u32 D_8003D5C0 = 0x4B9F1EC6;
-
-u32 D_8003D5C4 = 0x4B7DFA8F;
-
-u32 D_8003D5C8 = 0x900203CB;
-
-u32 D_8003D5CC = 0x4B7F0EC6;
-
-u32 D_8003D5D0 = 0x4AA346CC;
-
-u32 D_8003D5D4 = 0x30420001;
-
-u32 D_8003D5D8 = 0x4AA33B0D;
-
-u32 D_8003D5DC = 0x4B7DFACE;
-
-u32 D_8003D5E0 = 0x4B1C5AE8;
-
-u32 D_8003D5E4 = 0x4A8B1AD5;
-
-u32 D_8003D5E8 = 0x4A8C0311;
-
-u32 D_8003D5EC = 0x4B7D6323;
-
-u32 D_8003D5F0 = 0x4B7D5AE7;
-
-u32 D_8003D5F4 = 0x10400003;
-
-u32 D_8003D5F8 = 0x4AEB1B51;
-
-u32 D_8003D5FC = 0x4B6B1ED0;
-
-u32 D_8003D600 = 0x4B5B50F3;
-
-u32 D_8003D604 = 0x4A031251;
-
-u32 D_8003D608 = 0x4AE21AD1;
-
-u32 D_8003D60C = 0x4AED0ACE;
-
-u32 D_8003D610 = 0x4B7D4A63;
-
-u32 D_8003D614 = 0x4B7D5AE3;
-
-u32 D_8003D618 = 0x4AC96B51;
-
-u32 D_8003D61C = 0x4ACB6B51;
-
-u32 D_8003D620 = 0x4AC91890;
-
-u32 D_8003D624 = 0x4B3D6EE3;
-
-u32 D_8003D628 = 0x48CB0800;
-
-u32 D_8003D62C = 0x316B0088;
-
-u32 D_8003D630 = 0x1560FE7E;
-
-u32 D_8003D634 = 0x4AED1310;
-
-u32 D_8003D638 = 0x4BED78B3;
-
-u32 D_8003D63C = 0x4B9F16C5;
-
-u32 D_8003D640 = 0xE8220A00;
-
-u32 D_8003D644 = 0xE82C0A01;
-
-u32 D_8003D648 = 0x10400002;
-
-u32 D_8003D64C = 0xE83B1601;
-
-u32 D_8003D650 = 0x4B4B5273;
-
-u32 D_8003D654 = 0x4BDF1886;
-
-u32 D_8003D658 = 0x4AC92306;
-
-u32 D_8003D65C = 0x4B7DFACF;
-
-u32 D_8003D660 = 0x4BFE6306;
-
-u32 D_8003D664 = 0x4BFE5ACF;
-
-u32 D_8003D668 = 0x4A830ACF;
-
-u32 D_8003D66C = 0x48035800;
-
-u32 D_8003D670 = 0x48051100;
-
-u32 D_8003D674 = 0x48045C00;
-
-u32 D_8003D678 = 0x8C0703D0;
-
-u32 D_8003D67C = 0x48061500;
-
-u32 D_8003D680 = 0x00655822;
-
-u32 D_8003D684 = 0x05600005;
-
-u32 D_8003D688 = 0x48835800;
-
-u32 D_8003D68C = 0x00651822;
-
-u32 D_8003D690 = 0x20840020;
-
-u32 D_8003D694 = 0x080005FE;
-
-u32 D_8003D698 = 0x20E70020;
-
-u32 D_8003D69C = 0x00865822;
-
-u32 D_8003D6A0 = 0x05600004;
-
-u32 D_8003D6A4 = 0x48845C00;
-
-u32 D_8003D6A8 = 0x00862022;
-
-u32 D_8003D6AC = 0x08000605;
-
-u32 D_8003D6B0 = 0x00E63822;
-
-u32 D_8003D6B4 = 0xAC0703F0;
-
-u32 D_8003D6B8 = 0x00875822;
-
-u32 D_8003D6BC = 0x000B5940;
-
-u32 D_8003D6C0 = 0xAC0B03E4;
-
-u32 D_8003D6C4 = 0x4AED2306;
-
-u32 D_8003D6C8 = 0x4B7DFA4F;
-
-u32 D_8003D6CC = 0x4BFE6306;
-
-u32 D_8003D6D0 = 0x4BFE4A4F;
-
-u32 D_8003D6D4 = 0x900B0268;
-
-u32 D_8003D6D8 = 0x316B0008;
-
-u32 D_8003D6DC = 0x000B30C2;
-
-u32 D_8003D6E0 = 0x48044800;
-
-u32 D_8003D6E4 = 0xA00603E0;
-
-u32 D_8003D6E8 = 0x00065940;
-
-u32 D_8003D6EC = 0x2084000B;
-
-u32 D_8003D6F0 = 0x00646020;
-
-u32 D_8003D6F4 = 0x0185602A;
-
-u32 D_8003D6F8 = 0x398C0001;
-
-u32 D_8003D6FC = 0xA00C03E1;
-
-u32 D_8003D700 = 0x900B03C6;
-
-u32 D_8003D704 = 0x900C03C7;
-
-u32 D_8003D708 = 0x000B5840;
-
-u32 D_8003D70C = 0x216B02A4;
-
-u32 D_8003D710 = 0x000C6080;
-
-u32 D_8003D714 = 0x218C0294;
-
-u32 D_8003D718 = 0xC96C0A00;
-
-u32 D_8003D71C = 0xC98C1000;
-
-u32 D_8003D720 = 0x48866800;
-
-u32 D_8003D724 = 0x4B632386;
-
-u32 D_8003D728 = 0x4B7DFBCF;
-
-u32 D_8003D72C = 0x4B5C6EC6;
-
-u32 D_8003D730 = 0x4BFE76CC;
-
-u32 D_8003D734 = 0x4BFE7BCD;
-
-u32 D_8003D738 = 0x4B7DFB8E;
-
-u32 D_8003D73C = 0x4B2276D5;
-
-u32 D_8003D740 = 0x4B7D7EE0;
-
-u32 D_8003D744 = 0x4B2276E7;
-
-u32 D_8003D748 = 0x4B0C08CE;
-
-u32 D_8003D74C = 0x4B3D0EC7;
-
-u32 D_8003D750 = 0x4B2C18CD;
-
-u32 D_8003D754 = 0xE8230805;
-
-u32 D_8003D758 = 0x4B0343F0;
-
-u32 D_8003D75C = 0x4B7D43B2;
-
-u32 D_8003D760 = 0x4B7F6EC7;
-
-u32 D_8003D764 = 0x4B4C7ECC;
-
-u32 D_8003D768 = 0x4B4C740D;
-
-u32 D_8003D76C = 0x4B3C8346;
-
-u32 D_8003D770 = 0x4B7DF8CF;
-
-u32 D_8003D774 = 0x4B886EC4;
-
-u32 D_8003D778 = 0x4B881ECD;
-
-u32 D_8003D77C = 0x4B876B4E;
-
-u32 D_8003D780 = 0x4B8718CF;
-
-u32 D_8003D784 = 0xE8230808;
-
-u32 D_8003D788 = 0xE82D0809;
-
-u32 D_8003D78C = 0xC82E0807;
-
-u32 D_8003D790 = 0xC82F0806;
-
-u32 D_8003D794 = 0x4B8876C4;
-
-u32 D_8003D798 = 0x4B887ECD;
-
-u32 D_8003D79C = 0x4B87738E;
-
-u32 D_8003D7A0 = 0x4B877BCF;
-
-u32 D_8003D7A4 = 0x4B9C73A8;
-
-u32 D_8003D7A8 = 0x4B0346F2;
-
-u32 D_8003D7AC = 0x4B0D41F1;
-
-u32 D_8003D7B0 = 0x4B7D4232;
-
-u32 D_8003D7B4 = 0x4A0776C4;
-
-u32 D_8003D7B8 = 0x4A077ECD;
-
-u32 D_8003D7BC = 0x4A08744E;
-
-u32 D_8003D7C0 = 0x4A087C8F;
-
-u32 D_8003D7C4 = 0x4B5F8C46;
-
-u32 D_8003D7C8 = 0x4B5F948F;
-
-u32 D_8003D7CC = 0x4B3D0C4F;
-
-u32 D_8003D7D0 = 0x4A1169C6;
-
-u32 D_8003D7D4 = 0x4A111A0F;
-
-u32 D_8003D7D8 = 0x4A0776D5;
-
-u32 D_8003D7DC = 0x4A087EE3;
-
-u32 D_8003D7E0 = 0x4A128CA7;
-
-u32 D_8003D7E4 = 0x4B1269C6;
-
-u32 D_8003D7E8 = 0x4B121A0F;
-
-u32 D_8003D7EC = 0x4B9C3C68;
-
-u32 D_8003D7F0 = 0x4A1174D5;
-
-u32 D_8003D7F4 = 0x4A087D11;
-
-u32 D_8003D7F8 = 0x4B3F9EC4;
-
-u32 D_8003D7FC = 0x4B3FA3CD;
-
-u32 D_8003D800 = 0x4B7DFB8E;
-
-u32 D_8003D804 = 0x4B847386;
-
-u32 D_8003D808 = 0x4B847BCF;
-
-u32 D_8003D80C = 0x4B3D1EC7;
-
-u32 D_8003D810 = 0x4BBC3EE8;
-
-u32 D_8003D814 = 0x4B3D6ECE;
-
-u32 D_8003D818 = 0x4B7F99CE;
-
-u32 D_8003D81C = 0x4B7FA20F;
-
-u32 D_8003D820 = 0xE828080A;
-
-u32 D_8003D824 = 0xE827080B;
-
-u32 D_8003D828 = 0x4B1F70C5;
-
-u32 D_8003D82C = 0x4BDC18E8;
-
-u32 D_8003D830 = 0x4B3F7384;
-
-u32 D_8003D834 = 0x4B3F7BCD;
-
-u32 D_8003D838 = 0x4B7DFB8E;
-
-u32 D_8003D83C = 0x4B0E81D1;
-
-u32 D_8003D840 = 0x4B1049F3;
-
-u32 D_8003D844 = 0x080006EB;
-
-u32 D_8003D848 = 0xE827100A;
-
-u32 D_8003D84C = 0x8C0B03FC;
-
-u32 D_8003D850 = 0xEAFD0780;
-
-u32 D_8003D854 = 0x22F70010;
-
-u32 D_8003D858 = 0x01745825;
-
-u32 D_8003D85C = 0xAEEBFFF8;
-
-u32 D_8003D860 = 0xEAFD147F;
-
-u32 D_8003D864 = 0x8C0B03F8;
-
-u32 D_8003D868 = 0xAEF10004;
-
-u32 D_8003D86C = 0x22F70018;
-
-u32 D_8003D870 = 0xAEEBFFE8;
-
-u32 D_8003D874 = 0xEAFD0770;
-
-u32 D_8003D878 = 0x3C0BF400;
-
-u32 D_8003D87C = 0xAEEBFFF8;
-
-u32 D_8003D880 = 0x226B06FF;
-
-u32 D_8003D884 = 0xA6EBFFFC;
-
-u32 D_8003D888 = 0x00125880;
-
-u32 D_8003D88C = 0x216BFFFF;
-
-u32 D_8003D890 = 0x03E00008;
-
-u32 D_8003D894 = 0xA6EBFFFE;
-
-u32 D_8003D898 = 0x900B03E0;
-
-u32 D_8003D89C = 0x900C03E1;
-
-u32 D_8003D8A0 = 0x008BC820;
-
-u32 D_8003D8A4 = 0x006CC022;
-
-u32 D_8003D8A8 = 0x03195822;
-
-u32 D_8003D8AC = 0x05600007;
-
-u32 D_8003D8B0 = 0x941303E2;
-
-u32 D_8003D8B4 = 0x37320000;
-
-u32 D_8003D8B8 = 0x0C000677;
-
-u32 D_8003D8BC = 0x34B10000;
-
-u32 D_8003D8C0 = 0x00641822;
-
-u32 D_8003D8C4 = 0x08000764;
-
-u32 D_8003D8C8 = 0x00A62820;
-
-u32 D_8003D8CC = 0x4893C000;
-
-u32 D_8003D8D0 = 0x03239022;
-
-u32 D_8003D8D4 = 0x1A40000E;
-
-u32 D_8003D8D8 = 0x20740000;
-
-u32 D_8003D8DC = 0x8C1103C0;
-
-u32 D_8003D8E0 = 0x940B03F4;
-
-u32 D_8003D8E4 = 0x329F0001;
-
-u32 D_8003D8E8 = 0x13E00005;
-
-u32 D_8003D8EC = 0x022B8820;
-
-u32 D_8003D8F0 = 0x940B03F6;
-
-u32 D_8003D8F4 = 0x2294FFFF;
-
-u32 D_8003D8F8 = 0x22520001;
-
-u32 D_8003D8FC = 0x022B8822;
-
-u32 D_8003D900 = 0x4894D800;
-
-u32 D_8003D904 = 0x4B18DEC6;
-
-u32 D_8003D908 = 0x0C000671;
-
-u32 D_8003D90C = 0x4814D800;
-
-u32 D_8003D910 = 0x1180001D;
-
-u32 D_8003D914 = 0xC838090F;
-
-u32 D_8003D918 = 0x4898D100;
-
-u32 D_8003D91C = 0x4B38D6C6;
-
-u32 D_8003D920 = 0x4B7DFE4F;
-
-u32 D_8003D924 = 0xEBDB0901;
-
-u32 D_8003D928 = 0xEBD90900;
-
-u32 D_8003D92C = 0x8FCB0000;
-
-u32 D_8003D930 = 0x8C1103C0;
-
-u32 D_8003D934 = 0x33120001;
-
-u32 D_8003D938 = 0x12400005;
-
-u32 D_8003D93C = 0x00AB6020;
-
-u32 D_8003D940 = 0x940B03F6;
-
-u32 D_8003D944 = 0x4B3DD691;
-
-u32 D_8003D948 = 0x018B6022;
-
-u32 D_8003D94C = 0x022B8822;
-
-u32 D_8003D950 = 0xC8380A0E;
-
-u32 D_8003D954 = 0x22520001;
-
-u32 D_8003D958 = 0x4B58C6D1;
-
-u32 D_8003D95C = 0x4BDEDEC5;
-
-u32 D_8003D960 = 0x4B3BC651;
-
-u32 D_8003D964 = 0x4813C800;
-
-u32 D_8003D968 = 0x4B18D646;
-
-u32 D_8003D96C = 0x4B3DDE8E;
-
-u32 D_8003D970 = 0x0C000671;
-
-u32 D_8003D974 = 0x4814D100;
-
-u32 D_8003D978 = 0x35910000;
-
-u32 D_8003D97C = 0x4814C900;
-
-u32 D_8003D980 = 0x0C000671;
-
-u32 D_8003D984 = 0x4813D900;
-
-u32 D_8003D988 = 0x1F000006;
-
-u32 D_8003D98C = 0x00641822;
-
-u32 D_8003D990 = 0x8C0B03FC;
-
-u32 D_8003D994 = 0x22F70008;
-
-u32 D_8003D998 = 0xAEEBFFF8;
-
-u32 D_8003D99C = 0x080006CC;
-
-u32 D_8003D9A0 = 0xEAFD147F;
-
-u32 D_8003D9A4 = 0x34B10000;
-
-u32 D_8003D9A8 = 0x941303E2;
-
-u32 D_8003D9AC = 0x37120000;
-
-u32 D_8003D9B0 = 0x0C000671;
-
-u32 D_8003D9B4 = 0x20140000;
-
-u32 D_8003D9B8 = 0x1C600097;
-
-u32 D_8003D9BC = 0x00A62820;
-
-u32 D_8003D9C0 = 0x00035822;
-
-u32 D_8003D9C4 = 0x488BD800;
-
-u32 D_8003D9C8 = 0x940B03F4;
-
-u32 D_8003D9CC = 0x8C0503C0;
-
-u32 D_8003D9D0 = 0x4B38DEC6;
-
-u32 D_8003D9D4 = 0x4B7DFE4F;
-
-u32 D_8003D9D8 = 0x00AB2820;
-
-u32 D_8003D9DC = 0xEBDB0801;
-
-u32 D_8003D9E0 = 0x940C040C;
-
-u32 D_8003D9E4 = 0xEBD90800;
-
-u32 D_8003D9E8 = 0x8FCB0000;
-
-u32 D_8003D9EC = 0x006C1820;
-
-u32 D_8003D9F0 = 0x08000764;
-
-u32 D_8003D9F4 = 0x00AB2820;
-
-u32 D_8003D9F8 = 0x8F7804BC;
-
-u32 D_8003D9FC = 0x201F16EC;
-
-u32 D_8003DA00 = 0x08000455;
-
-u32 D_8003DA04 = 0x201403B0;
-
-u32 D_8003DA08 = 0x0C000455;
-
-u32 D_8003DA0C = 0x20130017;
-
-u32 D_8003DA10 = 0x0C0007F6;
-
-u32 D_8003DA14 = 0x201403C0;
-
-u32 D_8003DA18 = 0x200C1B8C;
-
-u32 D_8003DA1C = 0x080007ED;
-
-u32 D_8003DA20 = 0x200B0314;
-
-u32 D_8003DA24 = 0x00000000;
-
-u32 D_8003DA28 = 0x200C1638;
-
-u32 D_8003DA2C = 0x080007ED;
-
-u32 D_8003DA30 = 0x200B0314;
-
-u32 D_8003DA34 = 0xC82D080D;
-
-u32 D_8003DA38 = 0xC831080C;
-
-u32 D_8003DA3C = 0x4A1286C6;
-
-u32 D_8003DA40 = 0x4B3D76CE;
-
-u32 D_8003DA44 = 0x4B3D7ECF;
-
-u32 D_8003DA48 = 0x4B1F8ECD;
-
-u32 D_8003DA4C = 0x4B1F6A0C;
-
-u32 D_8003DA50 = 0x4BAA4351;
-
-u32 D_8003DA54 = 0x4BAA4390;
-
-u32 D_8003DA58 = 0x4B7D46E3;
-
-u32 D_8003DA5C = 0x4A0E4227;
-
-u32 D_8003DA60 = 0x4B7D6EE3;
-
-u32 D_8003DA64 = 0x4A086A27;
-
-u32 D_8003DA68 = 0x4B224AF3;
-
-u32 D_8003DA6C = 0x4B2C5A85;
-
-u32 D_8003DA70 = 0x4BDF5286;
-
-u32 D_8003DA74 = 0xE82A1007;
-
-u32 D_8003DA78 = 0xE828081B;
-
-u32 D_8003DA7C = 0x8C1803C0;
-
-u32 D_8003DA80 = 0x4B3D56C6;
-
-u32 D_8003DA84 = 0x0C000455;
-
-u32 D_8003DA88 = 0x4B2A408E;
-
-u32 D_8003DA8C = 0xAC1803C0;
-
-u32 D_8003DA90 = 0x4B7DFB4F;
-
-u32 D_8003DA94 = 0xEBC20801;
-
-u32 D_8003DA98 = 0xEBCD0800;
-
-u32 D_8003DA9C = 0x8FC50000;
-
-u32 D_8003DAA0 = 0x00B82820;
-
-u32 D_8003DAA4 = 0x4B2A3A05;
-
-u32 D_8003DAA8 = 0x4B7DF88E;
-
-u32 D_8003DAAC = 0xE8280816;
-
-u32 D_8003DAB0 = 0xE8280918;
-
-u32 D_8003DAB4 = 0xE8220817;
-
-u32 D_8003DAB8 = 0xE8220919;
-
-u32 D_8003DABC = 0x10400003;
-
-u32 D_8003DAC0 = 0x4B0C58A8;
-
-u32 D_8003DAC4 = 0x4B7F16C6;
-
-u32 D_8003DAC8 = 0x4B7F488E;
-
-u32 D_8003DACC = 0x940B03F6;
-
-u32 D_8003DAD0 = 0x940C03E2;
-
-u32 D_8003DAD4 = 0x3C02FD10;
-
-u32 D_8003DAD8 = 0x000B5842;
-
-u32 D_8003DADC = 0x216BFFFF;
-
-u32 D_8003DAE0 = 0x004B1025;
-
-u32 D_8003DAE4 = 0xAC0203F8;
-
-u32 D_8003DAE8 = 0x3C07F510;
-
-u32 D_8003DAEC = 0x000C6240;
-
-u32 D_8003DAF0 = 0x00EC3825;
-
-u32 D_8003DAF4 = 0xAC0703FC;
-
-u32 D_8003DAF8 = 0xAEE70000;
-
-u32 D_8003DAFC = 0xEAFD1401;
-
-u32 D_8003DB00 = 0xAEE70008;
-
-u32 D_8003DB04 = 0x900B03C6;
-
-u32 D_8003DB08 = 0x900C03C7;
-
-u32 D_8003DB0C = 0x000B5940;
-
-u32 D_8003DB10 = 0x000C60C0;
-
-u32 D_8003DB14 = 0x016C5825;
-
-u32 D_8003DB18 = 0xA2EB0009;
-
-u32 D_8003DB1C = 0x900B03C9;
-
-u32 D_8003DB20 = 0x8C0C0264;
-
-u32 D_8003DB24 = 0x000B5D00;
-
-u32 D_8003DB28 = 0x016C5825;
-
-u32 D_8003DB2C = 0xAEEB000C;
-
-u32 D_8003DB30 = 0xEAFD1204;
-
-u32 D_8003DB34 = 0xAEE00014;
-
-u32 D_8003DB38 = 0x22F70018;
-
-u32 D_8003DB3C = 0x940303BA;
-
-u32 D_8003DB40 = 0x940B040E;
-
-u32 D_8003DB44 = 0x84040400;
-
-u32 D_8003DB48 = 0x8C0203EC;
-
-u32 D_8003DB4C = 0x8C0F03E8;
-
-u32 D_8003DB50 = 0x940E03DE;
-
-u32 D_8003DB54 = 0x00031882;
-
-u32 D_8003DB58 = 0xA403040C;
-
-u32 D_8003DB5C = 0x006B1822;
-
-u32 D_8003DB60 = 0x8C060404;
-
-u32 D_8003DB64 = 0x940703D8;
-
-u32 D_8003DB68 = 0x940803DA;
-
-u32 D_8003DB6C = 0x940903DC;
-
-u32 D_8003DB70 = 0x00073B00;
-
-u32 D_8003DB74 = 0x00084300;
-
-u32 D_8003DB78 = 0x00026A82;
-
-u32 D_8003DB7C = 0x1DA00013;
-
-u32 D_8003DB80 = 0x00000000;
-
-u32 D_8003DB84 = 0x00641822;
-
-u32 D_8003DB88 = 0x1C600033;
-
-u32 D_8003DB8C = 0x00A62820;
-
-u32 D_8003DB90 = 0xC827080F;
-
-u32 D_8003DB94 = 0x00035822;
-
-u32 D_8003DB98 = 0x488B3900;
-
-u32 D_8003DB9C = 0x8C0503C0;
-
-u32 D_8003DBA0 = 0x4B2739C6;
-
-u32 D_8003DBA4 = 0x4B7DFA0F;
-
-u32 D_8003DBA8 = 0xEBC70801;
-
-u32 D_8003DBAC = 0xEBC80800;
-
-u32 D_8003DBB0 = 0x8FCB0000;
-
-u32 D_8003DBB4 = 0x940C03F4;
-
-u32 D_8003DBB8 = 0x941F040C;
-
-u32 D_8003DBBC = 0x00AB2820;
-
-u32 D_8003DBC0 = 0x00AC2820;
-
-u32 D_8003DBC4 = 0x08000774;
-
-u32 D_8003DBC8 = 0x007F1820;
-
-u32 D_8003DBCC = 0x01CD7022;
-
-u32 D_8003DBD0 = 0x05C1000F;
-
-u32 D_8003DBD4 = 0x304203FF;
-
-u32 D_8003DBD8 = 0x488ED800;
-
-u32 D_8003DBDC = 0x4B1B21C6;
-
-u32 D_8003DBE0 = 0x4B7DFA0F;
-
-u32 D_8003DBE4 = 0x940C0402;
-
-u32 D_8003DBE8 = 0xEBC80C00;
-
-u32 D_8003DBEC = 0xEBC70C01;
-
-u32 D_8003DBF0 = 0x8FCB0000;
-
-u32 D_8003DBF4 = 0x20840001;
-
-u32 D_8003DBF8 = 0x000B5A83;
-
-u32 D_8003DBFC = 0x008B2020;
-
-u32 D_8003DC00 = 0x008C5822;
-
-u32 D_8003DC04 = 0x05600002;
-
-u32 D_8003DC08 = 0x01AE6820;
-
-u32 D_8003DC0C = 0x21840000;
-
-u32 D_8003DC10 = 0x08000684;
-
-u32 D_8003DC14 = 0x012D5020;
-
-u32 D_8003DC18 = 0xEAFD0400;
-
-u32 D_8003DC1C = 0x000A5880;
-
-u32 D_8003DC20 = 0x01685825;
-
-u32 D_8003DC24 = 0xAEEB0008;
-
-u32 D_8003DC28 = 0xEAFD0688;
-
-u32 D_8003DC2C = 0x00095880;
-
-u32 D_8003DC30 = 0x01675825;
-
-u32 D_8003DC34 = 0xAEEB000C;
-
-u32 D_8003DC38 = 0xEAE20808;
-
-u32 D_8003DC3C = 0xEAE30809;
-
-u32 D_8003DC40 = 0xEAE4080A;
-
-u32 D_8003DC44 = 0xEAE40C0B;
-
-u32 D_8003DC48 = 0x0C000591;
-
-u32 D_8003DC4C = 0x26F70018;
-
-u32 D_8003DC50 = 0x19C0FCF6;
-
-u32 D_8003DC54 = 0x21490000;
-
-u32 D_8003DC58 = 0x004F1020;
-
-u32 D_8003DC5C = 0x84040402;
-
-u32 D_8003DC60 = 0x8C060408;
-
-u32 D_8003DC64 = 0x0800073C;
-
-u32 D_8003DC68 = 0x4A0318EC;
-
-u32 D_8003DC6C = 0x00000000;
-
-Overlay mainSegOverlay = {
-	_ovl1SegmentRomStart,
-	_ovl1DataSegmentRomEnd,
-	_ovl1SegmentStart,
-	_ovl1SegmentStart,
-	_ovl1SegmentEnd,
-	_ovl1SegmentEnd,
-	_ovl1DataSegmentEnd,
-	_ovl1DataSegmentEnd,
-	(void*)0x800F61A0,
-};
-
-u32 D_8003DC94 = 0x00000000;
-
-u32 D_8003DC98 = 0x00000000;
-
-u32 D_8003DC9C = 0x00000000;
-
-u32 D_8003DCA0 = 0x00000000;
-
-u32 D_8003DCA4 = 0x00000000;
-
-u32 D_8003DCA8 = 0x00000000;
-
-u32 D_8003DCAC = 0x80039E90;
-
-u32 D_8003DCB0 = 0x80041FF0;
-
-u32 D_8003DCB4 = 0x00000000;
-
-u32 D_8003DCB8 = 0x00000000;
-
-u32 D_8003DCBC = 0x00000000;
-
-u32 D_8003DCC0 = 0x00000000;
-
-u32 D_8003DCC4 = 0x00000000;
-
-u32 D_8003DCC8 = 0x00000000;
-
-u32 D_8003DCCC = 0x00000000;
-
-u32 D_8003DCD0 = 0x00000000;
-
-u32 D_8003DCD4 = 0x00000000;
-
-u32 D_8003DCD8 = 0x00000000;
-
-u32 D_8003DCDC = 0x00000000;
-
-u32 D_8003DCE0 = 0x00000000;
-
-u32 D_8003DCE4 = 0x00000000;
-
-u32 D_8003DCE8 = 0x00000000;
-
-u32 D_8003DCEC = 0x8003B220;
-
-u32 D_8003DCF0 = 0x80042410;
-
-u32 D_8003DCF4 = 0x00000000;
-
-u32 D_8003DCF8 = 0x00000000;
-
-u32 D_8003DCFC = 0x8003C3B0;
-
-u32 D_8003DD00 = 0x80042800;
-
-u32 D_8003DD04 = 0x00000000;
-
-u32 D_8003DD08 = 0x00000000;
-
-u32 D_8003DD0C = 0x00000000;
-
-u32 D_8003DD10 = 0x00000000;
-
-u32 D_8003DD14 = 0x00000000;
-
-u32 D_8003DD18 = 0x00000000;
-
-u32 D_8003DD1C = 0x00000000;
-
-u32 D_8003DD20 = 0x00000000;
-
-u32 D_8003DD24 = 0x00000000;
-
-u32 D_8003DD28 = 0x00000000;
-
-u32 D_8003DD2C = 0x00000000;
-
-u32 D_8003DD30 = 0x00000000;
-
-u32 D_8003DD34 = 0x00000000;
-
-Mtx D_8003DD38 = {
-	65536, 0, 1, 0,
-	0, 65536, 0, 1,
-	0, 0, 0, 0,
-	0, 0, 0, 0,
-};
-
-Gfx D_8003DD78[] = {
-	gsDPPipeSync(),
-	gsSPViewport(0x8004A530),
-	gsSPClearGeometryMode(G_ZBUFFER | G_SHADE | G_CULL_BOTH | G_FOG | G_LIGHTING | G_TEXTURE_GEN | G_TEXTURE_GEN_LINEAR | G_LOD | G_SHADING_SMOOTH),
-	gsSPClipRatio(FRUSTRATIO_1),
-	gsSPTexture(0, 0, 0, G_TX_RENDERTILE, G_OFF),
-	gsSPSetGeometryMode(G_ZBUFFER | G_SHADE | G_CULL_BACK | G_SHADING_SMOOTH),
-    gsSPMatrix(&D_8003DD38, G_MTX_NOPUSH | G_MTX_LOAD | G_MTX_PROJECTION),
-    gsSPMatrix(&D_8003DD38, G_MTX_NOPUSH | G_MTX_LOAD | G_MTX_MODELVIEW),
-    gsDPSetCycleType(G_CYC_1CYCLE),
-    gsDPPipelineMode(G_PM_NPRIMITIVE),
-    gsDPSetCombineMode(G_CC_SHADE, G_CC_SHADE),
-    gsDPSetTextureLOD(G_TL_TILE),
-    gsDPSetTextureLUT(G_TT_NONE),
-    gsDPSetTextureDetail(G_TD_CLAMP),
-    gsDPSetTexturePersp(G_TP_PERSP),
-    gsDPSetTextureFilter(G_TF_BILERP),
-    gsDPSetTextureConvert(G_TC_FILT),
-    gsDPSetCombineKey(G_CK_NONE),
-    gsDPSetAlphaCompare(G_AC_NONE),
-    gsDPSetRenderMode(G_RM_OPA_SURF, G_RM_OPA_SURF2),
-    gsDPSetColorDither(G_CD_MAGICSQ),
-    gsDPPipeSync(),
-    gsSPEndDisplayList(),
-};
-
-u32 D_8003DE48 = 0x00000000;
-
-u32 D_8003DE4C = 0x00000000;
-
-u32 D_8003DE50 = 0x00989680;
-
-u32 D_8003DE54 = 0x00000000;
-
-u32 D_8003DE58 = 0x00000000;
-
-u32 D_8003DE5C = 0x00000000;
-
-f32 D_8003DE60 = 30.0f;
-
-f32 D_8003DE64 = 4.0f / 3.0f;
-
-f32 D_8003DE68 = 100.0f;
-
-f32 D_8003DE6C = 12800.0f;
-
-f32 D_8003DE70 = 1.0f;
-
-u32 D_8003DE74 = 0x00000000;
-
-f32 D_8003DE78 = -160.0f;
-
-f32 D_8003DE7C = 160.0f;
-
-f32 D_8003DE80 = -120.0f;
-
-f32 D_8003DE84 = 120.0f;
-
-f32 D_8003DE88 = 100.0f;
-
-f32 D_8003DE8C = 12800.0f;
-
-f32 D_8003DE90 = 1.0f;
-
-u32 D_8003DE94 = 0x00000000;
-
-u32 D_8003DE98 = 0x00000000;
-
-u32 D_8003DE9C = 0x00000000;
-
-f32 D_8003DEA0 = 1500.0f;
-
-u32 D_8003DEA4 = 0x00000000;
-
-u32 D_8003DEA8 = 0x00000000;
-
-u32 D_8003DEAC = 0x00000000;
-
-u32 D_8003DEB0 = 0x00000000;
-
-f32 D_8003DEB4 = 1.0f;
-
-u32 D_8003DEB8 = 0x00000000;
-
-u32 D_8003DEBC = 0x00000000;
-
-u32 D_8003DEC0 = 0x00000000;
-
-u32 D_8003DEC4 = 0x00000000;
-
-u32 D_8003DEC8 = 0x00000000;
-
-u32 D_8003DECC = 0x00000000;
-
-u32 D_8003DED0 = 0x00000000;
-
-u32 D_8003DED4 = 0x00000000;
-
-u32 D_8003DED8 = 0x00000000;
-
-f32 D_8003DEDC = 1.0f;
-
-u32 D_8003DEE0 = 0x00000000;
-
-u32 D_8003DEE4 = 0x00000000;
-
-u32 D_8003DEE8 = 0x00000000;
-
-u32 D_8003DEEC = 0x00000000;
-
-u32 D_8003DEF0 = 0x00000000;
-
-u32 D_8003DEF4 = 0x00000000;
-
-f32 D_8003DEF8 = 1.0f;
-
-f32 D_8003DEFC = 1.0f;
-
-f32 D_8003DF00 = 1.0f;
-
-u32 D_8003DF04 = 0x00000000;
-
-u32 D_8003DF08 = 0x00000000;
-
-u32 D_8003DF0C = 0x00000000;
-
-u32 D_8003DF10 = 0x0000000A;
-
-u32 D_8003DF14 = 0x0000000A;
-
-u32 D_8003DF18 = 0x0000000A;
-
-u32 D_8003DF1C = 0x0000000A;
-
-u8 D_8003DF20[0x400] = {
-    0x2B, 0xE6, 0xDB, 0xB9, 0xB1, 0xF3, 0x53, 0x81, 0x37, 0x01, 0x7F, 0x44, 0x23, 0x3C, 0x11, 0xD0,
-    0x48, 0x80, 0xD2, 0x15, 0x6A, 0x70, 0x9B, 0x9E, 0x27, 0x51, 0xB0, 0x4B, 0x6A, 0xDB, 0x1E, 0xFF,
-    0x2D, 0xBC, 0x5B, 0x38, 0x6D, 0xCA, 0xAA, 0xF5, 0xCB, 0x42, 0x56, 0x96, 0x0F, 0x28, 0x6A, 0x59,
-    0x08, 0x08, 0xB5, 0x1B, 0xA4, 0xEC, 0xAD, 0x8B, 0x7E, 0xD9, 0x6F, 0x4E, 0x8E, 0x40, 0x92, 0xA2,
-    0x7F, 0x27, 0x88, 0x8D, 0x65, 0x5B, 0x3A, 0xF6, 0xA2, 0x41, 0xC8, 0xA6, 0x4A, 0xA6, 0xDE, 0x28,
-    0x8C, 0x48, 0xD7, 0x90, 0x9B, 0xAB, 0xAB, 0x73, 0x9A, 0x44, 0x6F, 0x73, 0xF7, 0xE4, 0x15, 0x04,
-    0xE6, 0xFF, 0xD8, 0xCD, 0xE7, 0xA9, 0x1D, 0x56, 0x4B, 0xEE, 0xD5, 0xBE, 0x32, 0x93, 0xCF, 0x7F,
-    0x39, 0x4D, 0xA8, 0x15, 0x32, 0xF7, 0x68, 0x67, 0x9B, 0x6F, 0xD6, 0x84, 0x30, 0xE3, 0xDE, 0x52,
-    0x5C, 0x09, 0x34, 0x54, 0xAB, 0x76, 0x03, 0x5A, 0xDE, 0xE0, 0xE4, 0xA0, 0x49, 0xAF, 0x0B, 0x7B,
-    0x54, 0x2C, 0x82, 0xBC, 0xF4, 0xCE, 0xA4, 0x0A, 0x99, 0x42, 0x5C, 0xD0, 0x67, 0xAC, 0x46, 0x1C,
-    0x05, 0x09, 0x4C, 0x87, 0x64, 0x61, 0x5D, 0x57, 0x4E, 0x13, 0x39, 0x67, 0x7D, 0x02, 0x26, 0x2A,
-    0x04, 0x49, 0xAE, 0xC3, 0x69, 0x7B, 0x52, 0xEC, 0x7D, 0xA7, 0x45, 0x0F, 0xB4, 0x12, 0xAE, 0x44,
-    0xA5, 0x1A, 0x22, 0xBC, 0xF3, 0xD0, 0x35, 0x3A, 0x50, 0x43, 0xEB, 0x94, 0x66, 0x9D, 0x85, 0xBB,
-    0x0C, 0x02, 0xD9, 0x0B, 0xC5, 0xE8, 0xB2, 0x1D, 0x1A, 0x20, 0x66, 0x25, 0xAE, 0xF9, 0x33, 0xEF,
-    0xD8, 0x1F, 0xA1, 0xDF, 0x68, 0xFB, 0xB0, 0xC5, 0x91, 0xA5, 0x24, 0x8D, 0xCC, 0xC1, 0x06, 0xA2,
-    0x5D, 0xFD, 0x50, 0x45, 0xE6, 0x75, 0x4A, 0x5E, 0x02, 0xB7, 0x6C, 0xFE, 0x81, 0xB8, 0x93, 0x8C,
-    0x4C, 0x29, 0x1F, 0xB1, 0x9A, 0xFA, 0x85, 0x23, 0xCE, 0x43, 0x04, 0x76, 0x5B, 0x14, 0xAF, 0x32,
-    0xDA, 0xF7, 0x48, 0xBF, 0x0C, 0x3F, 0x30, 0x6D, 0x86, 0x2C, 0xE2, 0xA3, 0x31, 0x0D, 0xFF, 0xD3,
-    0xF0, 0x76, 0x7B, 0x89, 0xED, 0xAA, 0x97, 0x0E, 0x45, 0xB3, 0x5B, 0xAB, 0x92, 0xB5, 0x79, 0xD3,
-    0xCC, 0xFA, 0x4C, 0x7E, 0xA0, 0x25, 0x1A, 0xED, 0x56, 0xF4, 0xBE, 0xCA, 0xF2, 0xD7, 0x00, 0x6A,
-    0xF3, 0x62, 0xC0, 0xBF, 0xB3, 0x56, 0x46, 0x50, 0x91, 0x69, 0x65, 0x5C, 0xC2, 0xDB, 0xE1, 0x8E,
-    0x79, 0x03, 0xD9, 0xAC, 0xED, 0x10, 0xBB, 0x5E, 0x8B, 0xC1, 0xE5, 0xD4, 0xF1, 0xD7, 0x78, 0x6E,
-    0x4D, 0x44, 0x01, 0x80, 0xDC, 0xB1, 0x0A, 0x97, 0x72, 0x1E, 0x5D, 0xB4, 0x99, 0xC0, 0xB0, 0x20,
-    0x04, 0x39, 0xF6, 0xC1, 0xF7, 0x2B, 0x3D, 0x45, 0x59, 0xEC, 0x53, 0x4F, 0x9A, 0x86, 0xE5, 0x06,
-    0x55, 0x80, 0x40, 0x9C, 0x86, 0x83, 0x90, 0x2D, 0x9B, 0xD1, 0xEF, 0xB9, 0x68, 0xDC, 0x93, 0x1D,
-    0xC2, 0x6C, 0xB6, 0xE2, 0xCF, 0x28, 0x10, 0xB6, 0x16, 0x52, 0x33, 0x60, 0xDB, 0x82, 0x41, 0xD0,
-    0xFD, 0x47, 0x1C, 0x35, 0x58, 0x94, 0x58, 0xE5, 0xBB, 0x58, 0xA2, 0x4C, 0x64, 0xCE, 0xCA, 0x94,
-    0x13, 0x47, 0x6F, 0xA7, 0xF8, 0x5E, 0xC4, 0x3B, 0x76, 0x85, 0x01, 0x61, 0x26, 0x13, 0x28, 0x23,
-    0x52, 0xFB, 0x19, 0x89, 0x35, 0xEA, 0xDD, 0x88, 0xE7, 0xBE, 0xA9, 0x2F, 0xD5, 0x82, 0x9C, 0x51,
-    0x7F, 0x51, 0x29, 0x31, 0xDA, 0xF9, 0xFA, 0x2A, 0x06, 0xC2, 0xE3, 0x14, 0xD5, 0xEC, 0x79, 0xE9,
-    0xDD, 0x75, 0x72, 0x37, 0x84, 0xFE, 0x7D, 0x72, 0x8E, 0xCD, 0xF5, 0xC3, 0xA8, 0xBA, 0xD8, 0xC6,
-    0xD4, 0x42, 0x0B, 0xEE, 0x1B, 0xE0, 0xFE, 0xF0, 0xD7, 0xA6, 0xC6, 0x22, 0x12, 0x25, 0x49, 0x81,
-    0xC0, 0x9F, 0x33, 0x9F, 0x38, 0xB7, 0x3C, 0x38, 0x42, 0xE2, 0xCB, 0xA9, 0x17, 0x03, 0x4E, 0xD3,
-    0xF5, 0xE9, 0x77, 0x0E, 0xA8, 0x08, 0x3D, 0x47, 0x86, 0x98, 0xDA, 0x20, 0xD6, 0x9F, 0x7D, 0xFC,
-    0xEE, 0xD6, 0x14, 0x5F, 0xFC, 0x5F, 0x46, 0xDF, 0xF2, 0x87, 0x39, 0x19, 0xA1, 0x62, 0x65, 0xA7,
-    0x51, 0x64, 0x20, 0x7E, 0x77, 0xFD, 0xC9, 0xA7, 0x80, 0x5F, 0xFB, 0x41, 0xA0, 0xBD, 0x9D, 0x83,
-    0xFC, 0xDF, 0xF0, 0xE3, 0x05, 0x60, 0xA6, 0xBB, 0xD9, 0xB0, 0x59, 0x43, 0xFF, 0xD1, 0x30, 0x7C,
-    0xB2, 0xAF, 0x8D, 0x8C, 0xC7, 0x53, 0xF1, 0x47, 0x92, 0x38, 0xA3, 0x57, 0x9C, 0x0A, 0x72, 0x95,
-    0x4A, 0x7C, 0xE4, 0xBF, 0x84, 0x49, 0x15, 0x2E, 0x2D, 0x0C, 0x68, 0x7C, 0x3B, 0x84, 0xEF, 0x05,
-    0x1C, 0x78, 0x78, 0x2E, 0x6B, 0x11, 0x1D, 0x27, 0x3D, 0x87, 0x61, 0x1B, 0x3C, 0x93, 0x5A, 0x99,
-    0xE8, 0x89, 0x95, 0xD2, 0xE1, 0x3B, 0x0C, 0x99, 0x11, 0x9F, 0x67, 0x3C, 0xC7, 0x98, 0xF2, 0x8B,
-    0xCD, 0x21, 0x66, 0x19, 0x73, 0x96, 0xEB, 0xAA, 0x89, 0xCB, 0x4B, 0x8F, 0x33, 0xA5, 0x74, 0x4B,
-    0x2F, 0xA1, 0x2A, 0x55, 0x48, 0x5A, 0xAD, 0x13, 0x0A, 0x43, 0xB4, 0xFA, 0xD5, 0x3F, 0xAC, 0x79,
-    0x85, 0xD6, 0x9E, 0x9D, 0x60, 0x73, 0x74, 0xB2, 0x98, 0xE1, 0xBA, 0xB2, 0x6B, 0x29, 0xEB, 0x71,
-    0x7B, 0xF2, 0x40, 0x37, 0x12, 0x90, 0x1F, 0x9D, 0x62, 0x1E, 0xAE, 0x6C, 0x3A, 0x8B, 0xCB, 0x1E,
-    0x71, 0xBD, 0x2C, 0x9C, 0xBE, 0x8E, 0x10, 0x98, 0x70, 0xD4, 0x11, 0x4F, 0x07, 0x2F, 0x8F, 0x09,
-    0x57, 0xE8, 0x0D, 0x3E, 0xC2, 0xDC, 0xA1, 0xF5, 0x6C, 0x46, 0x24, 0x26, 0x16, 0x8A, 0xEA, 0xED,
-    0x1F, 0x00, 0x74, 0x00, 0x18, 0x2F, 0xF6, 0xCC, 0xC6, 0x02, 0x78, 0xC7, 0xB8, 0x2C, 0x2A, 0xE3,
-    0x91, 0xD4, 0xFC, 0xF4, 0xCF, 0x7A, 0x3D, 0xC3, 0x14, 0xCC, 0x3F, 0xA0, 0xDE, 0x22, 0x0D, 0x1A,
-    0x6D, 0xBD, 0xD1, 0x27, 0xBA, 0x6D, 0x08, 0x4D, 0x50, 0x07, 0xE7, 0x6B, 0x71, 0x36, 0x8C, 0x35,
-    0x18, 0x2B, 0xFD, 0xA8, 0x74, 0x7A, 0xB9, 0x0E, 0x3B, 0x5D, 0x17, 0xBF, 0xB3, 0x3A, 0x00, 0x83,
-    0xEF, 0x26, 0xF6, 0x2D, 0x31, 0xF9, 0x9E, 0xC1, 0xC9, 0x58, 0xE7, 0xC7, 0x41, 0x71, 0x36, 0x07,
-    0xB7, 0x63, 0x29, 0xFB, 0x06, 0x63, 0x05, 0x0B, 0xF0, 0x88, 0x1B, 0x6B, 0x4F, 0xB8, 0xD8, 0x75,
-    0x34, 0x70, 0x21, 0x92, 0xE2, 0x95, 0xC6, 0x70, 0x63, 0xDA, 0xEA, 0xB4, 0x77, 0xBC, 0xCF, 0x3E,
-    0xC9, 0xE0, 0xC4, 0x65, 0xE0, 0xC5, 0x40, 0xEA, 0xC8, 0xF1, 0xF8, 0xD3, 0x34, 0x32, 0x54, 0x61,
-    0x5F, 0xAD, 0x6E, 0x34, 0x17, 0xA3, 0x2E, 0xB3, 0xA3, 0xC3, 0x5C, 0xD2, 0x54, 0x16, 0x5A, 0x55,
-    0xAC, 0x94, 0xAD, 0x53, 0xAA, 0x96, 0xCA, 0xEE, 0x64, 0x30, 0xB7, 0x87, 0x82, 0x36, 0xF1, 0xDD,
-    0x31, 0x69, 0xCD, 0xA9, 0x91, 0xA5, 0x18, 0xC5, 0xF9, 0xE4, 0x8F, 0x97, 0x90, 0xE6, 0xCE, 0x23,
-    0xB8, 0x59, 0x0D, 0xB1, 0x95, 0x4E, 0x2E, 0xB5, 0x8A, 0x57, 0x7C, 0xC9, 0x77, 0x88, 0xC4, 0x21,
-    0x7A, 0xFE, 0x75, 0x83, 0x81, 0x63, 0x10, 0x3E, 0xB6, 0x0F, 0x7E, 0x36, 0x5E, 0x62, 0x6E, 0xD2,
-    0x55, 0x19, 0x4F, 0x16, 0xC4, 0xBA, 0xA4, 0xAF, 0xE9, 0xB9, 0x0E, 0x25, 0x9A, 0xF8, 0x09, 0x0F,
-    0x03, 0x12, 0xBD, 0x7A, 0xB6, 0xE9, 0x3E, 0x37, 0x24, 0x8A, 0xF8, 0x97, 0xF3, 0xE8, 0x4D, 0xE5,
-    0x4A, 0x60, 0x3F, 0xA4, 0x2B, 0x24, 0xC0, 0xDC, 0x17, 0x66, 0x69, 0x21, 0xB5, 0x8F, 0x8D, 0x6E,
-    0x96, 0x07, 0xF4, 0xD1, 0x1C, 0xC8, 0xE1, 0x18, 0xDD, 0xEB, 0x22, 0x01, 0x9E, 0xDF, 0xC8, 0x8A,
-};
-
-u32 D_8003E320 = 0x00000000;
-
-u32 D_8003E324 = 0x00000001;
-
-u32 D_8003E328 = 0x00000000;
-
-u32 D_8003E32C = 0x00000000;
-
-u32 D_8003E330 = 0x00000032;
-
-u32 D_8003E334 = 0x00640096;
-
-u32 D_8003E338 = 0x00C900FB;
-
-u32 D_8003E33C = 0x012D015F;
-
-u32 D_8003E340 = 0x019201C4;
-
-u32 D_8003E344 = 0x01F60228;
-
-u32 D_8003E348 = 0x025B028D;
-
-u32 D_8003E34C = 0x02BF02F1;
-
-u32 D_8003E350 = 0x03240356;
-
-u32 D_8003E354 = 0x038803BA;
-
-u32 D_8003E358 = 0x03ED041F;
-
-u32 D_8003E35C = 0x04510483;
-
-u32 D_8003E360 = 0x04B604E8;
-
-u32 D_8003E364 = 0x051A054C;
-
-u32 D_8003E368 = 0x057F05B1;
-
-u32 D_8003E36C = 0x05E30615;
-
-u32 D_8003E370 = 0x0647067A;
-
-u32 D_8003E374 = 0x06AC06DE;
-
-u32 D_8003E378 = 0x07100742;
-
-u32 D_8003E37C = 0x077507A7;
-
-u32 D_8003E380 = 0x07D9080B;
-
-u32 D_8003E384 = 0x083D086F;
-
-u32 D_8003E388 = 0x08A208D4;
-
-u32 D_8003E38C = 0x09060938;
-
-u32 D_8003E390 = 0x096A099C;
-
-u32 D_8003E394 = 0x09CE0A00;
-
-u32 D_8003E398 = 0x0A330A65;
-
-u32 D_8003E39C = 0x0A970AC9;
-
-u32 D_8003E3A0 = 0x0AFB0B2D;
-
-u32 D_8003E3A4 = 0x0B5F0B91;
-
-u32 D_8003E3A8 = 0x0BC30BF5;
-
-u32 D_8003E3AC = 0x0C270C59;
-
-u32 D_8003E3B0 = 0x0C8B0CBD;
-
-u32 D_8003E3B4 = 0x0CEF0D21;
-
-u32 D_8003E3B8 = 0x0D530D85;
-
-u32 D_8003E3BC = 0x0DB70DE9;
-
-u32 D_8003E3C0 = 0x0E1B0E4D;
-
-u32 D_8003E3C4 = 0x0E7F0EB1;
-
-u32 D_8003E3C8 = 0x0EE30F15;
-
-u32 D_8003E3CC = 0x0F470F79;
-
-u32 D_8003E3D0 = 0x0FAB0FDD;
-
-u32 D_8003E3D4 = 0x100E1040;
-
-u32 D_8003E3D8 = 0x107210A4;
-
-u32 D_8003E3DC = 0x10D61108;
-
-u32 D_8003E3E0 = 0x1139116B;
-
-u32 D_8003E3E4 = 0x119D11CF;
-
-u32 D_8003E3E8 = 0x12011232;
-
-u32 D_8003E3EC = 0x12641296;
-
-u32 D_8003E3F0 = 0x12C812F9;
-
-u32 D_8003E3F4 = 0x132B135D;
-
-u32 D_8003E3F8 = 0x138E13C0;
-
-u32 D_8003E3FC = 0x13F21423;
-
-u32 D_8003E400 = 0x14551487;
-
-u32 D_8003E404 = 0x14B814EA;
-
-u32 D_8003E408 = 0x151B154D;
-
-u32 D_8003E40C = 0x157F15B0;
-
-u32 D_8003E410 = 0x15E21613;
-
-u32 D_8003E414 = 0x16451676;
-
-u32 D_8003E418 = 0x16A816D9;
-
-u32 D_8003E41C = 0x170A173C;
-
-u32 D_8003E420 = 0x176D179F;
-
-u32 D_8003E424 = 0x17D01802;
-
-u32 D_8003E428 = 0x18331864;
-
-u32 D_8003E42C = 0x189618C7;
-
-u32 D_8003E430 = 0x18F8192A;
-
-u32 D_8003E434 = 0x195B198C;
-
-u32 D_8003E438 = 0x19BD19EF;
-
-u32 D_8003E43C = 0x1A201A51;
-
-u32 D_8003E440 = 0x1A821AB3;
-
-u32 D_8003E444 = 0x1AE41B16;
-
-u32 D_8003E448 = 0x1B471B78;
-
-u32 D_8003E44C = 0x1BA91BDA;
-
-u32 D_8003E450 = 0x1C0B1C3C;
-
-u32 D_8003E454 = 0x1C6D1C9E;
-
-u32 D_8003E458 = 0x1CCF1D00;
-
-u32 D_8003E45C = 0x1D311D62;
-
-u32 D_8003E460 = 0x1D931DC4;
-
-u32 D_8003E464 = 0x1DF51E25;
-
-u32 D_8003E468 = 0x1E561E87;
-
-u32 D_8003E46C = 0x1EB81EE9;
-
-u32 D_8003E470 = 0x1F191F4A;
-
-u32 D_8003E474 = 0x1F7B1FAC;
-
-u32 D_8003E478 = 0x1FDC200D;
-
-u32 D_8003E47C = 0x203E206E;
-
-u32 D_8003E480 = 0x209F20D0;
-
-u32 D_8003E484 = 0x21002131;
-
-u32 D_8003E488 = 0x21612192;
-
-u32 D_8003E48C = 0x21C221F3;
-
-u32 D_8003E490 = 0x22232254;
-
-u32 D_8003E494 = 0x228422B4;
-
-u32 D_8003E498 = 0x22E52315;
-
-u32 D_8003E49C = 0x23452376;
-
-u32 D_8003E4A0 = 0x23A623D6;
-
-u32 D_8003E4A4 = 0x24072437;
-
-u32 D_8003E4A8 = 0x24672497;
-
-u32 D_8003E4AC = 0x24C724F7;
-
-u32 D_8003E4B0 = 0x25282558;
-
-u32 D_8003E4B4 = 0x258825B8;
-
-u32 D_8003E4B8 = 0x25E82618;
-
-u32 D_8003E4BC = 0x26482678;
-
-u32 D_8003E4C0 = 0x26A826D8;
-
-u32 D_8003E4C4 = 0x27072737;
-
-u32 D_8003E4C8 = 0x27672797;
-
-u32 D_8003E4CC = 0x27C727F6;
-
-u32 D_8003E4D0 = 0x28262856;
-
-u32 D_8003E4D4 = 0x288628B5;
-
-u32 D_8003E4D8 = 0x28E52915;
-
-u32 D_8003E4DC = 0x29442974;
-
-u32 D_8003E4E0 = 0x29A329D3;
-
-u32 D_8003E4E4 = 0x2A022A32;
-
-u32 D_8003E4E8 = 0x2A612A91;
-
-u32 D_8003E4EC = 0x2AC02AEF;
-
-u32 D_8003E4F0 = 0x2B1F2B4E;
-
-u32 D_8003E4F4 = 0x2B7D2BAD;
-
-u32 D_8003E4F8 = 0x2BDC2C0B;
-
-u32 D_8003E4FC = 0x2C3A2C69;
-
-u32 D_8003E500 = 0x2C982CC8;
-
-u32 D_8003E504 = 0x2CF72D26;
-
-u32 D_8003E508 = 0x2D552D84;
-
-u32 D_8003E50C = 0x2DB32DE2;
-
-u32 D_8003E510 = 0x2E112E3F;
-
-u32 D_8003E514 = 0x2E6E2E9D;
-
-u32 D_8003E518 = 0x2ECC2EFB;
-
-u32 D_8003E51C = 0x2F292F58;
-
-u32 D_8003E520 = 0x2F872FB5;
-
-u32 D_8003E524 = 0x2FE43013;
-
-u32 D_8003E528 = 0x30413070;
-
-u32 D_8003E52C = 0x309E30CD;
-
-u32 D_8003E530 = 0x30FB312A;
-
-u32 D_8003E534 = 0x31583186;
-
-u32 D_8003E538 = 0x31B531E3;
-
-u32 D_8003E53C = 0x32113240;
-
-u32 D_8003E540 = 0x326E329C;
-
-u32 D_8003E544 = 0x32CA32F8;
-
-u32 D_8003E548 = 0x33263354;
-
-u32 D_8003E54C = 0x338233B0;
-
-u32 D_8003E550 = 0x33DE340C;
-
-u32 D_8003E554 = 0x343A3468;
-
-u32 D_8003E558 = 0x349634C4;
-
-u32 D_8003E55C = 0x34F2351F;
-
-u32 D_8003E560 = 0x354D357B;
-
-u32 D_8003E564 = 0x35A835D6;
-
-u32 D_8003E568 = 0x36043631;
-
-u32 D_8003E56C = 0x365F368C;
-
-u32 D_8003E570 = 0x36BA36E7;
-
-u32 D_8003E574 = 0x37143742;
-
-u32 D_8003E578 = 0x376F379C;
-
-u32 D_8003E57C = 0x37CA37F7;
-
-u32 D_8003E580 = 0x38243851;
-
-u32 D_8003E584 = 0x387E38AB;
-
-u32 D_8003E588 = 0x38D83906;
-
-u32 D_8003E58C = 0x3932395F;
-
-u32 D_8003E590 = 0x398C39B9;
-
-u32 D_8003E594 = 0x39E63A13;
-
-u32 D_8003E598 = 0x3A403A6C;
-
-u32 D_8003E59C = 0x3A993AC6;
-
-u32 D_8003E5A0 = 0x3AF23B1F;
-
-u32 D_8003E5A4 = 0x3B4C3B78;
-
-u32 D_8003E5A8 = 0x3BA53BD1;
-
-u32 D_8003E5AC = 0x3BFD3C2A;
-
-u32 D_8003E5B0 = 0x3C563C83;
-
-u32 D_8003E5B4 = 0x3CAF3CDB;
-
-u32 D_8003E5B8 = 0x3D073D33;
-
-u32 D_8003E5BC = 0x3D603D8C;
-
-u32 D_8003E5C0 = 0x3DB83DE4;
-
-u32 D_8003E5C4 = 0x3E103E3C;
-
-u32 D_8003E5C8 = 0x3E683E93;
-
-u32 D_8003E5CC = 0x3EBF3EEB;
-
-u32 D_8003E5D0 = 0x3F173F43;
-
-u32 D_8003E5D4 = 0x3F6E3F9A;
-
-u32 D_8003E5D8 = 0x3FC53FF1;
-
-u32 D_8003E5DC = 0x401D4048;
-
-u32 D_8003E5E0 = 0x4073409F;
-
-u32 D_8003E5E4 = 0x40CA40F6;
-
-u32 D_8003E5E8 = 0x4121414C;
-
-u32 D_8003E5EC = 0x417741A2;
-
-u32 D_8003E5F0 = 0x41CE41F9;
-
-u32 D_8003E5F4 = 0x4224424F;
-
-u32 D_8003E5F8 = 0x427A42A5;
-
-u32 D_8003E5FC = 0x42D042FA;
-
-u32 D_8003E600 = 0x43254350;
-
-u32 D_8003E604 = 0x437B43A5;
-
-u32 D_8003E608 = 0x43D043FB;
-
-u32 D_8003E60C = 0x44254450;
-
-u32 D_8003E610 = 0x447A44A5;
-
-u32 D_8003E614 = 0x44CF44FA;
-
-u32 D_8003E618 = 0x4524454E;
-
-u32 D_8003E61C = 0x457845A3;
-
-u32 D_8003E620 = 0x45CD45F7;
-
-u32 D_8003E624 = 0x4621464B;
-
-u32 D_8003E628 = 0x4675469F;
-
-u32 D_8003E62C = 0x46C946F3;
-
-u32 D_8003E630 = 0x471C4746;
-
-u32 D_8003E634 = 0x4770479A;
-
-u32 D_8003E638 = 0x47C347ED;
-
-u32 D_8003E63C = 0x48164840;
-
-u32 D_8003E640 = 0x48694893;
-
-u32 D_8003E644 = 0x48BC48E6;
-
-u32 D_8003E648 = 0x490F4938;
-
-u32 D_8003E64C = 0x4961498A;
-
-u32 D_8003E650 = 0x49B449DD;
-
-u32 D_8003E654 = 0x4A064A2F;
-
-u32 D_8003E658 = 0x4A584A81;
-
-u32 D_8003E65C = 0x4AA94AD2;
-
-u32 D_8003E660 = 0x4AFB4B24;
-
-u32 D_8003E664 = 0x4B4C4B75;
-
-u32 D_8003E668 = 0x4B9E4BC6;
-
-u32 D_8003E66C = 0x4BEF4C17;
-
-u32 D_8003E670 = 0x4C3F4C68;
-
-u32 D_8003E674 = 0x4C904CB8;
-
-u32 D_8003E678 = 0x4CE14D09;
-
-u32 D_8003E67C = 0x4D314D59;
-
-u32 D_8003E680 = 0x4D814DA9;
-
-u32 D_8003E684 = 0x4DD14DF9;
-
-u32 D_8003E688 = 0x4E214E48;
-
-u32 D_8003E68C = 0x4E704E98;
-
-u32 D_8003E690 = 0x4EBF4EE7;
-
-u32 D_8003E694 = 0x4F0F4F36;
-
-u32 D_8003E698 = 0x4F5E4F85;
-
-u32 D_8003E69C = 0x4FAC4FD4;
-
-u32 D_8003E6A0 = 0x4FFB5022;
-
-u32 D_8003E6A4 = 0x50495070;
-
-u32 D_8003E6A8 = 0x509750BF;
-
-u32 D_8003E6AC = 0x50E5510C;
-
-u32 D_8003E6B0 = 0x5133515A;
-
-u32 D_8003E6B4 = 0x518151A8;
-
-u32 D_8003E6B8 = 0x51CE51F5;
-
-u32 D_8003E6BC = 0x521C5242;
-
-u32 D_8003E6C0 = 0x5269528F;
-
-u32 D_8003E6C4 = 0x52B552DC;
-
-u32 D_8003E6C8 = 0x53025328;
-
-u32 D_8003E6CC = 0x534E5375;
-
-u32 D_8003E6D0 = 0x539B53C1;
-
-u32 D_8003E6D4 = 0x53E7540D;
-
-u32 D_8003E6D8 = 0x54335458;
-
-u32 D_8003E6DC = 0x547E54A4;
-
-u32 D_8003E6E0 = 0x54CA54EF;
-
-u32 D_8003E6E4 = 0x5515553A;
-
-u32 D_8003E6E8 = 0x55605585;
-
-u32 D_8003E6EC = 0x55AB55D0;
-
-u32 D_8003E6F0 = 0x55F5561A;
-
-u32 D_8003E6F4 = 0x56405665;
-
-u32 D_8003E6F8 = 0x568A56AF;
-
-u32 D_8003E6FC = 0x56D456F9;
-
-u32 D_8003E700 = 0x571D5742;
-
-u32 D_8003E704 = 0x5767578C;
-
-u32 D_8003E708 = 0x57B057D5;
-
-u32 D_8003E70C = 0x57F9581E;
-
-u32 D_8003E710 = 0x58425867;
-
-u32 D_8003E714 = 0x588B58AF;
-
-u32 D_8003E718 = 0x58D458F8;
-
-u32 D_8003E71C = 0x591C5940;
-
-u32 D_8003E720 = 0x59645988;
-
-u32 D_8003E724 = 0x59AC59D0;
-
-u32 D_8003E728 = 0x59F35A17;
-
-u32 D_8003E72C = 0x5A3B5A5E;
-
-u32 D_8003E730 = 0x5A825AA5;
-
-u32 D_8003E734 = 0x5AC95AEC;
-
-u32 D_8003E738 = 0x5B105B33;
-
-u32 D_8003E73C = 0x5B565B79;
-
-u32 D_8003E740 = 0x5B9D5BC0;
-
-u32 D_8003E744 = 0x5BE35C06;
-
-u32 D_8003E748 = 0x5C295C4B;
-
-u32 D_8003E74C = 0x5C6E5C91;
-
-u32 D_8003E750 = 0x5CB45CD6;
-
-u32 D_8003E754 = 0x5CF95D1B;
-
-u32 D_8003E758 = 0x5D3E5D60;
-
-u32 D_8003E75C = 0x5D835DA5;
-
-u32 D_8003E760 = 0x5DC75DE9;
-
-u32 D_8003E764 = 0x5E0B5E2D;
-
-u32 D_8003E768 = 0x5E505E71;
-
-u32 D_8003E76C = 0x5E935EB5;
-
-u32 D_8003E770 = 0x5ED75EF9;
-
-u32 D_8003E774 = 0x5F1A5F3C;
-
-u32 D_8003E778 = 0x5F5E5F7F;
-
-u32 D_8003E77C = 0x5FA05FC2;
-
-u32 D_8003E780 = 0x5FE36004;
-
-u32 D_8003E784 = 0x60266047;
-
-u32 D_8003E788 = 0x60686089;
-
-u32 D_8003E78C = 0x60AA60CB;
-
-u32 D_8003E790 = 0x60EC610D;
-
-u32 D_8003E794 = 0x612D614E;
-
-u32 D_8003E798 = 0x616F618F;
-
-u32 D_8003E79C = 0x61B061D0;
-
-u32 D_8003E7A0 = 0x61F16211;
-
-u32 D_8003E7A4 = 0x62316251;
-
-u32 D_8003E7A8 = 0x62716292;
-
-u32 D_8003E7AC = 0x62B262D2;
-
-u32 D_8003E7B0 = 0x62F26311;
-
-u32 D_8003E7B4 = 0x63316351;
-
-u32 D_8003E7B8 = 0x63716390;
-
-u32 D_8003E7BC = 0x63B063CF;
-
-u32 D_8003E7C0 = 0x63EF640E;
-
-u32 D_8003E7C4 = 0x642D644D;
-
-u32 D_8003E7C8 = 0x646C648B;
-
-u32 D_8003E7CC = 0x64AA64C9;
-
-u32 D_8003E7D0 = 0x64E86507;
-
-u32 D_8003E7D4 = 0x65266545;
-
-u32 D_8003E7D8 = 0x65636582;
-
-u32 D_8003E7DC = 0x65A065BF;
-
-u32 D_8003E7E0 = 0x65DD65FC;
-
-u32 D_8003E7E4 = 0x661A6639;
-
-u32 D_8003E7E8 = 0x66576675;
-
-u32 D_8003E7EC = 0x669366B1;
-
-u32 D_8003E7F0 = 0x66CF66ED;
-
-u32 D_8003E7F4 = 0x670B6729;
-
-u32 D_8003E7F8 = 0x67466764;
-
-u32 D_8003E7FC = 0x6782679F;
-
-u32 D_8003E800 = 0x67BD67DA;
-
-u32 D_8003E804 = 0x67F76815;
-
-u32 D_8003E808 = 0x6832684F;
-
-u32 D_8003E80C = 0x686C6889;
-
-u32 D_8003E810 = 0x68A668C3;
-
-u32 D_8003E814 = 0x68E068FD;
-
-u32 D_8003E818 = 0x69196936;
-
-u32 D_8003E81C = 0x6953696F;
-
-u32 D_8003E820 = 0x698C69A8;
-
-u32 D_8003E824 = 0x69C469E1;
-
-u32 D_8003E828 = 0x69FD6A19;
-
-u32 D_8003E82C = 0x6A356A51;
-
-u32 D_8003E830 = 0x6A6D6A89;
-
-u32 D_8003E834 = 0x6AA56AC1;
-
-u32 D_8003E838 = 0x6ADC6AF8;
-
-u32 D_8003E83C = 0x6B136B2F;
-
-u32 D_8003E840 = 0x6B4A6B66;
-
-u32 D_8003E844 = 0x6B816B9C;
-
-u32 D_8003E848 = 0x6BB86BD3;
-
-u32 D_8003E84C = 0x6BEE6C09;
-
-u32 D_8003E850 = 0x6C246C3F;
-
-u32 D_8003E854 = 0x6C596C74;
-
-u32 D_8003E858 = 0x6C8F6CA9;
-
-u32 D_8003E85C = 0x6CC46CDE;
-
-u32 D_8003E860 = 0x6CF96D13;
-
-u32 D_8003E864 = 0x6D2D6D48;
-
-u32 D_8003E868 = 0x6D626D7C;
-
-u32 D_8003E86C = 0x6D966DB0;
-
-u32 D_8003E870 = 0x6DCA6DE3;
-
-u32 D_8003E874 = 0x6DFD6E17;
-
-u32 D_8003E878 = 0x6E306E4A;
-
-u32 D_8003E87C = 0x6E636E7D;
-
-u32 D_8003E880 = 0x6E966EAF;
-
-u32 D_8003E884 = 0x6EC96EE2;
-
-u32 D_8003E888 = 0x6EFB6F14;
-
-u32 D_8003E88C = 0x6F2D6F46;
-
-u32 D_8003E890 = 0x6F5F6F77;
-
-u32 D_8003E894 = 0x6F906FA9;
-
-u32 D_8003E898 = 0x6FC16FDA;
-
-u32 D_8003E89C = 0x6FF2700A;
-
-u32 D_8003E8A0 = 0x7023703B;
-
-u32 D_8003E8A4 = 0x7053706B;
-
-u32 D_8003E8A8 = 0x7083709B;
-
-u32 D_8003E8AC = 0x70B370CB;
-
-u32 D_8003E8B0 = 0x70E270FA;
-
-u32 D_8003E8B4 = 0x71127129;
-
-u32 D_8003E8B8 = 0x71417158;
-
-u32 D_8003E8BC = 0x716F7186;
-
-u32 D_8003E8C0 = 0x719E71B5;
-
-u32 D_8003E8C4 = 0x71CC71E3;
-
-u32 D_8003E8C8 = 0x71FA7211;
-
-u32 D_8003E8CC = 0x7227723E;
-
-u32 D_8003E8D0 = 0x7255726B;
-
-u32 D_8003E8D4 = 0x72827298;
-
-u32 D_8003E8D8 = 0x72AF72C5;
-
-u32 D_8003E8DC = 0x72DB72F1;
-
-u32 D_8003E8E0 = 0x7307731D;
-
-u32 D_8003E8E4 = 0x73337349;
-
-u32 D_8003E8E8 = 0x735F7375;
-
-u32 D_8003E8EC = 0x738A73A0;
-
-u32 D_8003E8F0 = 0x73B573CB;
-
-u32 D_8003E8F4 = 0x73E073F6;
-
-u32 D_8003E8F8 = 0x740B7420;
-
-u32 D_8003E8FC = 0x7435744A;
-
-u32 D_8003E900 = 0x745F7474;
-
-u32 D_8003E904 = 0x7489749E;
-
-u32 D_8003E908 = 0x74B274C7;
-
-u32 D_8003E90C = 0x74DB74F0;
-
-u32 D_8003E910 = 0x75047519;
-
-u32 D_8003E914 = 0x752D7541;
-
-u32 D_8003E918 = 0x75557569;
-
-u32 D_8003E91C = 0x757D7591;
-
-u32 D_8003E920 = 0x75A575B9;
-
-u32 D_8003E924 = 0x75CC75E0;
-
-u32 D_8003E928 = 0x75F47607;
-
-u32 D_8003E92C = 0x761B762E;
-
-u32 D_8003E930 = 0x76417654;
-
-u32 D_8003E934 = 0x7668767B;
-
-u32 D_8003E938 = 0x768E76A0;
-
-u32 D_8003E93C = 0x76B376C6;
-
-u32 D_8003E940 = 0x76D976EB;
-
-u32 D_8003E944 = 0x76FE7710;
-
-u32 D_8003E948 = 0x77237735;
-
-u32 D_8003E94C = 0x7747775A;
-
-u32 D_8003E950 = 0x776C777E;
-
-u32 D_8003E954 = 0x779077A2;
-
-u32 D_8003E958 = 0x77B477C5;
-
-u32 D_8003E95C = 0x77D777E9;
-
-u32 D_8003E960 = 0x77FA780C;
-
-u32 D_8003E964 = 0x781D782E;
-
-u32 D_8003E968 = 0x78407851;
-
-u32 D_8003E96C = 0x78627873;
-
-u32 D_8003E970 = 0x78847895;
-
-u32 D_8003E974 = 0x78A678B6;
-
-u32 D_8003E978 = 0x78C778D8;
-
-u32 D_8003E97C = 0x78E878F9;
-
-u32 D_8003E980 = 0x79097919;
-
-u32 D_8003E984 = 0x792A793A;
-
-u32 D_8003E988 = 0x794A795A;
-
-u32 D_8003E98C = 0x796A797A;
-
-u32 D_8003E990 = 0x798A7999;
-
-u32 D_8003E994 = 0x79A979B9;
-
-u32 D_8003E998 = 0x79C879D8;
-
-u32 D_8003E99C = 0x79E779F6;
-
-u32 D_8003E9A0 = 0x7A057A15;
-
-u32 D_8003E9A4 = 0x7A247A33;
-
-u32 D_8003E9A8 = 0x7A427A50;
-
-u32 D_8003E9AC = 0x7A5F7A6E;
-
-u32 D_8003E9B0 = 0x7A7D7A8B;
-
-u32 D_8003E9B4 = 0x7A9A7AA8;
-
-u32 D_8003E9B8 = 0x7AB67AC5;
-
-u32 D_8003E9BC = 0x7AD37AE1;
-
-u32 D_8003E9C0 = 0x7AEF7AFD;
-
-u32 D_8003E9C4 = 0x7B0B7B19;
-
-u32 D_8003E9C8 = 0x7B267B34;
-
-u32 D_8003E9CC = 0x7B427B4F;
-
-u32 D_8003E9D0 = 0x7B5D7B6A;
-
-u32 D_8003E9D4 = 0x7B777B84;
-
-u32 D_8003E9D8 = 0x7B927B9F;
-
-u32 D_8003E9DC = 0x7BAC7BB9;
-
-u32 D_8003E9E0 = 0x7BC57BD2;
-
-u32 D_8003E9E4 = 0x7BDF7BEB;
-
-u32 D_8003E9E8 = 0x7BF87C05;
-
-u32 D_8003E9EC = 0x7C117C1D;
-
-u32 D_8003E9F0 = 0x7C297C36;
-
-u32 D_8003E9F4 = 0x7C427C4E;
-
-u32 D_8003E9F8 = 0x7C5A7C66;
-
-u32 D_8003E9FC = 0x7C717C7D;
-
-u32 D_8003EA00 = 0x7C897C94;
-
-u32 D_8003EA04 = 0x7CA07CAB;
-
-u32 D_8003EA08 = 0x7CB77CC2;
-
-u32 D_8003EA0C = 0x7CCD7CD8;
-
-u32 D_8003EA10 = 0x7CE37CEE;
-
-u32 D_8003EA14 = 0x7CF97D04;
-
-u32 D_8003EA18 = 0x7D0F7D19;
-
-u32 D_8003EA1C = 0x7D247D2F;
-
-u32 D_8003EA20 = 0x7D397D43;
-
-u32 D_8003EA24 = 0x7D4E7D58;
-
-u32 D_8003EA28 = 0x7D627D6C;
-
-u32 D_8003EA2C = 0x7D767D80;
-
-u32 D_8003EA30 = 0x7D8A7D94;
-
-u32 D_8003EA34 = 0x7D9D7DA7;
-
-u32 D_8003EA38 = 0x7DB07DBA;
-
-u32 D_8003EA3C = 0x7DC37DCD;
-
-u32 D_8003EA40 = 0x7DD67DDF;
-
-u32 D_8003EA44 = 0x7DE87DF1;
-
-u32 D_8003EA48 = 0x7DFA7E03;
-
-u32 D_8003EA4C = 0x7E0C7E14;
-
-u32 D_8003EA50 = 0x7E1D7E26;
-
-u32 D_8003EA54 = 0x7E2E7E37;
-
-u32 D_8003EA58 = 0x7E3F7E47;
-
-u32 D_8003EA5C = 0x7E4F7E57;
-
-u32 D_8003EA60 = 0x7E5F7E67;
-
-u32 D_8003EA64 = 0x7E6F7E77;
-
-u32 D_8003EA68 = 0x7E7F7E86;
-
-u32 D_8003EA6C = 0x7E8E7E95;
-
-u32 D_8003EA70 = 0x7E9D7EA4;
-
-u32 D_8003EA74 = 0x7EAB7EB3;
-
-u32 D_8003EA78 = 0x7EBA7EC1;
-
-u32 D_8003EA7C = 0x7EC87ECF;
-
-u32 D_8003EA80 = 0x7ED57EDC;
-
-u32 D_8003EA84 = 0x7EE37EE9;
-
-u32 D_8003EA88 = 0x7EF07EF6;
-
-u32 D_8003EA8C = 0x7EFD7F03;
-
-u32 D_8003EA90 = 0x7F097F0F;
-
-u32 D_8003EA94 = 0x7F157F1B;
-
-u32 D_8003EA98 = 0x7F217F27;
-
-u32 D_8003EA9C = 0x7F2D7F32;
-
-u32 D_8003EAA0 = 0x7F387F3D;
-
-u32 D_8003EAA4 = 0x7F437F48;
-
-u32 D_8003EAA8 = 0x7F4D7F53;
-
-u32 D_8003EAAC = 0x7F587F5D;
-
-u32 D_8003EAB0 = 0x7F627F67;
-
-u32 D_8003EAB4 = 0x7F6B7F70;
-
-u32 D_8003EAB8 = 0x7F757F79;
-
-u32 D_8003EABC = 0x7F7E7F82;
-
-u32 D_8003EAC0 = 0x7F877F8B;
-
-u32 D_8003EAC4 = 0x7F8F7F93;
-
-u32 D_8003EAC8 = 0x7F977F9B;
-
-u32 D_8003EACC = 0x7F9F7FA3;
-
-u32 D_8003EAD0 = 0x7FA77FAA;
-
-u32 D_8003EAD4 = 0x7FAE7FB1;
-
-u32 D_8003EAD8 = 0x7FB57FB8;
-
-u32 D_8003EADC = 0x7FBC7FBF;
-
-u32 D_8003EAE0 = 0x7FC27FC5;
-
-u32 D_8003EAE4 = 0x7FC87FCB;
-
-u32 D_8003EAE8 = 0x7FCE7FD0;
-
-u32 D_8003EAEC = 0x7FD37FD6;
-
-u32 D_8003EAF0 = 0x7FD87FDA;
-
-u32 D_8003EAF4 = 0x7FDD7FDF;
-
-u32 D_8003EAF8 = 0x7FE17FE3;
-
-u32 D_8003EAFC = 0x7FE57FE7;
-
-u32 D_8003EB00 = 0x7FE97FEB;
-
-u32 D_8003EB04 = 0x7FED7FEE;
-
-u32 D_8003EB08 = 0x7FF07FF2;
-
-u32 D_8003EB0C = 0x7FF37FF4;
-
-u32 D_8003EB10 = 0x7FF67FF7;
-
-u32 D_8003EB14 = 0x7FF87FF9;
-
-u32 D_8003EB18 = 0x7FFA7FFB;
-
-u32 D_8003EB1C = 0x7FFC7FFC;
-
-u32 D_8003EB20 = 0x7FFD7FFE;
-
-u32 D_8003EB24 = 0x7FFE7FFF;
-
-u32 D_8003EB28 = 0x7FFF7FFF;
-
-u32 D_8003EB2C = 0x7FFF8000;
-
-u32 D_8003EB30 = 0x80007FFF;
-
-u32 D_8003EB34 = 0x7FFF7FFF;
-
-u32 D_8003EB38 = 0x7FFF7FFE;
-
-u32 D_8003EB3C = 0x7FFE7FFD;
-
-u32 D_8003EB40 = 0x7FFC7FFC;
-
-u32 D_8003EB44 = 0x7FFB7FFA;
-
-u32 D_8003EB48 = 0x7FF97FF8;
-
-u32 D_8003EB4C = 0x7FF77FF6;
-
-u32 D_8003EB50 = 0x7FF47FF3;
-
-u32 D_8003EB54 = 0x7FF27FF0;
-
-u32 D_8003EB58 = 0x7FEE7FED;
-
-u32 D_8003EB5C = 0x7FEB7FE9;
-
-u32 D_8003EB60 = 0x7FE77FE5;
-
-u32 D_8003EB64 = 0x7FE37FE1;
-
-u32 D_8003EB68 = 0x7FDF7FDD;
-
-u32 D_8003EB6C = 0x7FDA7FD8;
-
-u32 D_8003EB70 = 0x7FD67FD3;
-
-u32 D_8003EB74 = 0x7FD07FCE;
-
-u32 D_8003EB78 = 0x7FCB7FC8;
-
-u32 D_8003EB7C = 0x7FC57FC2;
-
-u32 D_8003EB80 = 0x7FBF7FBC;
-
-u32 D_8003EB84 = 0x7FB87FB5;
-
-u32 D_8003EB88 = 0x7FB17FAE;
-
-u32 D_8003EB8C = 0x7FAA7FA7;
-
-u32 D_8003EB90 = 0x7FA37F9F;
-
-u32 D_8003EB94 = 0x7F9B7F97;
-
-u32 D_8003EB98 = 0x7F937F8F;
-
-u32 D_8003EB9C = 0x7F8B7F87;
-
-u32 D_8003EBA0 = 0x7F827F7E;
-
-u32 D_8003EBA4 = 0x7F797F75;
-
-u32 D_8003EBA8 = 0x7F707F6B;
-
-u32 D_8003EBAC = 0x7F677F62;
-
-u32 D_8003EBB0 = 0x7F5D7F58;
-
-u32 D_8003EBB4 = 0x7F537F4D;
-
-u32 D_8003EBB8 = 0x7F487F43;
-
-u32 D_8003EBBC = 0x7F3D7F38;
-
-u32 D_8003EBC0 = 0x7F327F2D;
-
-u32 D_8003EBC4 = 0x7F277F21;
-
-u32 D_8003EBC8 = 0x7F1B7F15;
-
-u32 D_8003EBCC = 0x7F0F7F09;
-
-u32 D_8003EBD0 = 0x7F037EFD;
-
-u32 D_8003EBD4 = 0x7EF67EF0;
-
-u32 D_8003EBD8 = 0x7EE97EE3;
-
-u32 D_8003EBDC = 0x7EDC7ED5;
-
-u32 D_8003EBE0 = 0x7ECF7EC8;
-
-u32 D_8003EBE4 = 0x7EC17EBA;
-
-u32 D_8003EBE8 = 0x7EB37EAB;
-
-u32 D_8003EBEC = 0x7EA47E9D;
-
-u32 D_8003EBF0 = 0x7E957E8E;
-
-u32 D_8003EBF4 = 0x7E867E7F;
-
-u32 D_8003EBF8 = 0x7E777E6F;
-
-u32 D_8003EBFC = 0x7E677E5F;
-
-u32 D_8003EC00 = 0x7E577E4F;
-
-u32 D_8003EC04 = 0x7E477E3F;
-
-u32 D_8003EC08 = 0x7E377E2E;
-
-u32 D_8003EC0C = 0x7E267E1D;
-
-u32 D_8003EC10 = 0x7E147E0C;
-
-u32 D_8003EC14 = 0x7E037DFA;
-
-u32 D_8003EC18 = 0x7DF17DE8;
-
-u32 D_8003EC1C = 0x7DDF7DD6;
-
-u32 D_8003EC20 = 0x7DCD7DC3;
-
-u32 D_8003EC24 = 0x7DBA7DB0;
-
-u32 D_8003EC28 = 0x7DA77D9D;
-
-u32 D_8003EC2C = 0x7D947D8A;
-
-u32 D_8003EC30 = 0x7D807D76;
-
-u32 D_8003EC34 = 0x7D6C7D62;
-
-u32 D_8003EC38 = 0x7D587D4E;
-
-u32 D_8003EC3C = 0x7D437D39;
-
-u32 D_8003EC40 = 0x7D2F7D24;
-
-u32 D_8003EC44 = 0x7D197D0F;
-
-u32 D_8003EC48 = 0x7D047CF9;
-
-u32 D_8003EC4C = 0x7CEE7CE3;
-
-u32 D_8003EC50 = 0x7CD87CCD;
-
-u32 D_8003EC54 = 0x7CC27CB7;
-
-u32 D_8003EC58 = 0x7CAB7CA0;
-
-u32 D_8003EC5C = 0x7C947C89;
-
-u32 D_8003EC60 = 0x7C7D7C71;
-
-u32 D_8003EC64 = 0x7C667C5A;
-
-u32 D_8003EC68 = 0x7C4E7C42;
-
-u32 D_8003EC6C = 0x7C367C29;
-
-u32 D_8003EC70 = 0x7C1D7C11;
-
-u32 D_8003EC74 = 0x7C057BF8;
-
-u32 D_8003EC78 = 0x7BEB7BDF;
-
-u32 D_8003EC7C = 0x7BD27BC5;
-
-u32 D_8003EC80 = 0x7BB97BAC;
-
-u32 D_8003EC84 = 0x7B9F7B92;
-
-u32 D_8003EC88 = 0x7B847B77;
-
-u32 D_8003EC8C = 0x7B6A7B5D;
-
-u32 D_8003EC90 = 0x7B4F7B42;
-
-u32 D_8003EC94 = 0x7B347B26;
-
-u32 D_8003EC98 = 0x7B197B0B;
-
-u32 D_8003EC9C = 0x7AFD7AEF;
-
-u32 D_8003ECA0 = 0x7AE17AD3;
-
-u32 D_8003ECA4 = 0x7AC57AB6;
-
-u32 D_8003ECA8 = 0x7AA87A9A;
-
-u32 D_8003ECAC = 0x7A8B7A7D;
-
-u32 D_8003ECB0 = 0x7A6E7A5F;
-
-u32 D_8003ECB4 = 0x7A507A42;
-
-u32 D_8003ECB8 = 0x7A337A24;
-
-u32 D_8003ECBC = 0x7A157A05;
-
-u32 D_8003ECC0 = 0x79F679E7;
-
-u32 D_8003ECC4 = 0x79D879C8;
-
-u32 D_8003ECC8 = 0x79B979A9;
-
-u32 D_8003ECCC = 0x7999798A;
-
-u32 D_8003ECD0 = 0x797A796A;
-
-u32 D_8003ECD4 = 0x795A794A;
-
-u32 D_8003ECD8 = 0x793A792A;
-
-u32 D_8003ECDC = 0x79197909;
-
-u32 D_8003ECE0 = 0x78F978E8;
-
-u32 D_8003ECE4 = 0x78D878C7;
-
-u32 D_8003ECE8 = 0x78B678A6;
-
-u32 D_8003ECEC = 0x78957884;
-
-u32 D_8003ECF0 = 0x78737862;
-
-u32 D_8003ECF4 = 0x78517840;
-
-u32 D_8003ECF8 = 0x782E781D;
-
-u32 D_8003ECFC = 0x780C77FA;
-
-u32 D_8003ED00 = 0x77E977D7;
-
-u32 D_8003ED04 = 0x77C577B4;
-
-u32 D_8003ED08 = 0x77A27790;
-
-u32 D_8003ED0C = 0x777E776C;
-
-u32 D_8003ED10 = 0x775A7747;
-
-u32 D_8003ED14 = 0x77357723;
-
-u32 D_8003ED18 = 0x771076FE;
-
-u32 D_8003ED1C = 0x76EB76D9;
-
-u32 D_8003ED20 = 0x76C676B3;
-
-u32 D_8003ED24 = 0x76A0768E;
-
-u32 D_8003ED28 = 0x767B7668;
-
-u32 D_8003ED2C = 0x76547641;
-
-u32 D_8003ED30 = 0x762E761B;
-
-u32 D_8003ED34 = 0x760775F4;
-
-u32 D_8003ED38 = 0x75E075CC;
-
-u32 D_8003ED3C = 0x75B975A5;
-
-u32 D_8003ED40 = 0x7591757D;
-
-u32 D_8003ED44 = 0x75697555;
-
-u32 D_8003ED48 = 0x7541752D;
-
-u32 D_8003ED4C = 0x75197504;
-
-u32 D_8003ED50 = 0x74F074DB;
-
-u32 D_8003ED54 = 0x74C774B2;
-
-u32 D_8003ED58 = 0x749E7489;
-
-u32 D_8003ED5C = 0x7474745F;
-
-u32 D_8003ED60 = 0x744A7435;
-
-u32 D_8003ED64 = 0x7420740B;
-
-u32 D_8003ED68 = 0x73F673E0;
-
-u32 D_8003ED6C = 0x73CB73B5;
-
-u32 D_8003ED70 = 0x73A0738A;
-
-u32 D_8003ED74 = 0x7375735F;
-
-u32 D_8003ED78 = 0x73497333;
-
-u32 D_8003ED7C = 0x731D7307;
-
-u32 D_8003ED80 = 0x72F172DB;
-
-u32 D_8003ED84 = 0x72C572AF;
-
-u32 D_8003ED88 = 0x72987282;
-
-u32 D_8003ED8C = 0x726B7255;
-
-u32 D_8003ED90 = 0x723E7227;
-
-u32 D_8003ED94 = 0x721171FA;
-
-u32 D_8003ED98 = 0x71E371CC;
-
-u32 D_8003ED9C = 0x71B5719E;
-
-u32 D_8003EDA0 = 0x7186716F;
-
-u32 D_8003EDA4 = 0x71587141;
-
-u32 D_8003EDA8 = 0x71297112;
-
-u32 D_8003EDAC = 0x70FA70E2;
-
-u32 D_8003EDB0 = 0x70CB70B3;
-
-u32 D_8003EDB4 = 0x709B7083;
-
-u32 D_8003EDB8 = 0x706B7053;
-
-u32 D_8003EDBC = 0x703B7023;
-
-u32 D_8003EDC0 = 0x700A6FF2;
-
-u32 D_8003EDC4 = 0x6FDA6FC1;
-
-u32 D_8003EDC8 = 0x6FA96F90;
-
-u32 D_8003EDCC = 0x6F776F5F;
-
-u32 D_8003EDD0 = 0x6F466F2D;
-
-u32 D_8003EDD4 = 0x6F146EFB;
-
-u32 D_8003EDD8 = 0x6EE26EC9;
-
-u32 D_8003EDDC = 0x6EAF6E96;
-
-u32 D_8003EDE0 = 0x6E7D6E63;
-
-u32 D_8003EDE4 = 0x6E4A6E30;
-
-u32 D_8003EDE8 = 0x6E176DFD;
-
-u32 D_8003EDEC = 0x6DE36DCA;
-
-u32 D_8003EDF0 = 0x6DB06D96;
-
-u32 D_8003EDF4 = 0x6D7C6D62;
-
-u32 D_8003EDF8 = 0x6D486D2D;
-
-u32 D_8003EDFC = 0x6D136CF9;
-
-u32 D_8003EE00 = 0x6CDE6CC4;
-
-u32 D_8003EE04 = 0x6CA96C8F;
-
-u32 D_8003EE08 = 0x6C746C59;
-
-u32 D_8003EE0C = 0x6C3F6C24;
-
-u32 D_8003EE10 = 0x6C096BEE;
-
-u32 D_8003EE14 = 0x6BD36BB8;
-
-u32 D_8003EE18 = 0x6B9C6B81;
-
-u32 D_8003EE1C = 0x6B666B4A;
-
-u32 D_8003EE20 = 0x6B2F6B13;
-
-u32 D_8003EE24 = 0x6AF86ADC;
-
-u32 D_8003EE28 = 0x6AC16AA5;
-
-u32 D_8003EE2C = 0x6A896A6D;
-
-u32 D_8003EE30 = 0x6A516A35;
-
-u32 D_8003EE34 = 0x6A1969FD;
-
-u32 D_8003EE38 = 0x69E169C4;
-
-u32 D_8003EE3C = 0x69A8698C;
-
-u32 D_8003EE40 = 0x696F6953;
-
-u32 D_8003EE44 = 0x69366919;
-
-u32 D_8003EE48 = 0x68FD68E0;
-
-u32 D_8003EE4C = 0x68C368A6;
-
-u32 D_8003EE50 = 0x6889686C;
-
-u32 D_8003EE54 = 0x684F6832;
-
-u32 D_8003EE58 = 0x681567F7;
-
-u32 D_8003EE5C = 0x67DA67BD;
-
-u32 D_8003EE60 = 0x679F6782;
-
-u32 D_8003EE64 = 0x67646746;
-
-u32 D_8003EE68 = 0x6729670B;
-
-u32 D_8003EE6C = 0x66ED66CF;
-
-u32 D_8003EE70 = 0x66B16693;
-
-u32 D_8003EE74 = 0x66756657;
-
-u32 D_8003EE78 = 0x6639661A;
-
-u32 D_8003EE7C = 0x65FC65DD;
-
-u32 D_8003EE80 = 0x65BF65A0;
-
-u32 D_8003EE84 = 0x65826563;
-
-u32 D_8003EE88 = 0x65456526;
-
-u32 D_8003EE8C = 0x650764E8;
-
-u32 D_8003EE90 = 0x64C964AA;
-
-u32 D_8003EE94 = 0x648B646C;
-
-u32 D_8003EE98 = 0x644D642D;
-
-u32 D_8003EE9C = 0x640E63EF;
-
-u32 D_8003EEA0 = 0x63CF63B0;
-
-u32 D_8003EEA4 = 0x63906371;
-
-u32 D_8003EEA8 = 0x63516331;
-
-u32 D_8003EEAC = 0x631162F2;
-
-u32 D_8003EEB0 = 0x62D262B2;
-
-u32 D_8003EEB4 = 0x62926271;
-
-u32 D_8003EEB8 = 0x62516231;
-
-u32 D_8003EEBC = 0x621161F1;
-
-u32 D_8003EEC0 = 0x61D061B0;
-
-u32 D_8003EEC4 = 0x618F616F;
-
-u32 D_8003EEC8 = 0x614E612D;
-
-u32 D_8003EECC = 0x610D60EC;
-
-u32 D_8003EED0 = 0x60CB60AA;
-
-u32 D_8003EED4 = 0x60896068;
-
-u32 D_8003EED8 = 0x60476026;
-
-u32 D_8003EEDC = 0x60045FE3;
-
-u32 D_8003EEE0 = 0x5FC25FA0;
-
-u32 D_8003EEE4 = 0x5F7F5F5E;
-
-u32 D_8003EEE8 = 0x5F3C5F1A;
-
-u32 D_8003EEEC = 0x5EF95ED7;
-
-u32 D_8003EEF0 = 0x5EB55E93;
-
-u32 D_8003EEF4 = 0x5E715E50;
-
-u32 D_8003EEF8 = 0x5E2D5E0B;
-
-u32 D_8003EEFC = 0x5DE95DC7;
-
-u32 D_8003EF00 = 0x5DA55D83;
-
-u32 D_8003EF04 = 0x5D605D3E;
-
-u32 D_8003EF08 = 0x5D1B5CF9;
-
-u32 D_8003EF0C = 0x5CD65CB4;
-
-u32 D_8003EF10 = 0x5C915C6E;
-
-u32 D_8003EF14 = 0x5C4B5C29;
-
-u32 D_8003EF18 = 0x5C065BE3;
-
-u32 D_8003EF1C = 0x5BC05B9D;
-
-u32 D_8003EF20 = 0x5B795B56;
-
-u32 D_8003EF24 = 0x5B335B10;
-
-u32 D_8003EF28 = 0x5AEC5AC9;
-
-u32 D_8003EF2C = 0x5AA55A82;
-
-u32 D_8003EF30 = 0x5A5E5A3B;
-
-u32 D_8003EF34 = 0x5A1759F3;
-
-u32 D_8003EF38 = 0x59D059AC;
-
-u32 D_8003EF3C = 0x59885964;
-
-u32 D_8003EF40 = 0x5940591C;
-
-u32 D_8003EF44 = 0x58F858D4;
-
-u32 D_8003EF48 = 0x58AF588B;
-
-u32 D_8003EF4C = 0x58675842;
-
-u32 D_8003EF50 = 0x581E57F9;
-
-u32 D_8003EF54 = 0x57D557B0;
-
-u32 D_8003EF58 = 0x578C5767;
-
-u32 D_8003EF5C = 0x5742571D;
-
-u32 D_8003EF60 = 0x56F956D4;
-
-u32 D_8003EF64 = 0x56AF568A;
-
-u32 D_8003EF68 = 0x56655640;
-
-u32 D_8003EF6C = 0x561A55F5;
-
-u32 D_8003EF70 = 0x55D055AB;
-
-u32 D_8003EF74 = 0x55855560;
-
-u32 D_8003EF78 = 0x553A5515;
-
-u32 D_8003EF7C = 0x54EF54CA;
-
-u32 D_8003EF80 = 0x54A4547E;
-
-u32 D_8003EF84 = 0x54585433;
-
-u32 D_8003EF88 = 0x540D53E7;
-
-u32 D_8003EF8C = 0x53C1539B;
-
-u32 D_8003EF90 = 0x5375534E;
-
-u32 D_8003EF94 = 0x53285302;
-
-u32 D_8003EF98 = 0x52DC52B5;
-
-u32 D_8003EF9C = 0x528F5269;
-
-u32 D_8003EFA0 = 0x5242521C;
-
-u32 D_8003EFA4 = 0x51F551CE;
-
-u32 D_8003EFA8 = 0x51A85181;
-
-u32 D_8003EFAC = 0x515A5133;
-
-u32 D_8003EFB0 = 0x510C50E5;
-
-u32 D_8003EFB4 = 0x50BF5097;
-
-u32 D_8003EFB8 = 0x50705049;
-
-u32 D_8003EFBC = 0x50224FFB;
-
-u32 D_8003EFC0 = 0x4FD44FAC;
-
-u32 D_8003EFC4 = 0x4F854F5E;
-
-u32 D_8003EFC8 = 0x4F364F0F;
-
-u32 D_8003EFCC = 0x4EE74EBF;
-
-u32 D_8003EFD0 = 0x4E984E70;
-
-u32 D_8003EFD4 = 0x4E484E21;
-
-u32 D_8003EFD8 = 0x4DF94DD1;
-
-u32 D_8003EFDC = 0x4DA94D81;
-
-u32 D_8003EFE0 = 0x4D594D31;
-
-u32 D_8003EFE4 = 0x4D094CE1;
-
-u32 D_8003EFE8 = 0x4CB84C90;
-
-u32 D_8003EFEC = 0x4C684C3F;
-
-u32 D_8003EFF0 = 0x4C174BEF;
-
-u32 D_8003EFF4 = 0x4BC64B9E;
-
-u32 D_8003EFF8 = 0x4B754B4C;
-
-u32 D_8003EFFC = 0x4B244AFB;
-
-u32 D_8003F000 = 0x4AD24AA9;
-
-u32 D_8003F004 = 0x4A814A58;
-
-u32 D_8003F008 = 0x4A2F4A06;
-
-u32 D_8003F00C = 0x49DD49B4;
-
-u32 D_8003F010 = 0x498A4961;
-
-u32 D_8003F014 = 0x4938490F;
-
-u32 D_8003F018 = 0x48E648BC;
-
-u32 D_8003F01C = 0x48934869;
-
-u32 D_8003F020 = 0x48404816;
-
-u32 D_8003F024 = 0x47ED47C3;
-
-u32 D_8003F028 = 0x479A4770;
-
-u32 D_8003F02C = 0x4746471C;
-
-u32 D_8003F030 = 0x46F346C9;
-
-u32 D_8003F034 = 0x469F4675;
-
-u32 D_8003F038 = 0x464B4621;
-
-u32 D_8003F03C = 0x45F745CD;
-
-u32 D_8003F040 = 0x45A34578;
-
-u32 D_8003F044 = 0x454E4524;
-
-u32 D_8003F048 = 0x44FA44CF;
-
-u32 D_8003F04C = 0x44A5447A;
-
-u32 D_8003F050 = 0x44504425;
-
-u32 D_8003F054 = 0x43FB43D0;
-
-u32 D_8003F058 = 0x43A5437B;
-
-u32 D_8003F05C = 0x43504325;
-
-u32 D_8003F060 = 0x42FA42D0;
-
-u32 D_8003F064 = 0x42A5427A;
-
-u32 D_8003F068 = 0x424F4224;
-
-u32 D_8003F06C = 0x41F941CE;
-
-u32 D_8003F070 = 0x41A24177;
-
-u32 D_8003F074 = 0x414C4121;
-
-u32 D_8003F078 = 0x40F640CA;
-
-u32 D_8003F07C = 0x409F4073;
-
-u32 D_8003F080 = 0x4048401D;
-
-u32 D_8003F084 = 0x3FF13FC5;
-
-u32 D_8003F088 = 0x3F9A3F6E;
-
-u32 D_8003F08C = 0x3F433F17;
-
-u32 D_8003F090 = 0x3EEB3EBF;
-
-u32 D_8003F094 = 0x3E933E68;
-
-u32 D_8003F098 = 0x3E3C3E10;
-
-u32 D_8003F09C = 0x3DE43DB8;
-
-u32 D_8003F0A0 = 0x3D8C3D60;
-
-u32 D_8003F0A4 = 0x3D333D07;
-
-u32 D_8003F0A8 = 0x3CDB3CAF;
-
-u32 D_8003F0AC = 0x3C833C56;
-
-u32 D_8003F0B0 = 0x3C2A3BFD;
-
-u32 D_8003F0B4 = 0x3BD13BA5;
-
-u32 D_8003F0B8 = 0x3B783B4C;
-
-u32 D_8003F0BC = 0x3B1F3AF2;
-
-u32 D_8003F0C0 = 0x3AC63A99;
-
-u32 D_8003F0C4 = 0x3A6C3A40;
-
-u32 D_8003F0C8 = 0x3A1339E6;
-
-u32 D_8003F0CC = 0x39B9398C;
-
-u32 D_8003F0D0 = 0x395F3932;
-
-u32 D_8003F0D4 = 0x390638D8;
-
-u32 D_8003F0D8 = 0x38AB387E;
-
-u32 D_8003F0DC = 0x38513824;
-
-u32 D_8003F0E0 = 0x37F737CA;
-
-u32 D_8003F0E4 = 0x379C376F;
-
-u32 D_8003F0E8 = 0x37423714;
-
-u32 D_8003F0EC = 0x36E736BA;
-
-u32 D_8003F0F0 = 0x368C365F;
-
-u32 D_8003F0F4 = 0x36313604;
-
-u32 D_8003F0F8 = 0x35D635A8;
-
-u32 D_8003F0FC = 0x357B354D;
-
-u32 D_8003F100 = 0x351F34F2;
-
-u32 D_8003F104 = 0x34C43496;
-
-u32 D_8003F108 = 0x3468343A;
-
-u32 D_8003F10C = 0x340C33DE;
-
-u32 D_8003F110 = 0x33B03382;
-
-u32 D_8003F114 = 0x33543326;
-
-u32 D_8003F118 = 0x32F832CA;
-
-u32 D_8003F11C = 0x329C326E;
-
-u32 D_8003F120 = 0x32403211;
-
-u32 D_8003F124 = 0x31E331B5;
-
-u32 D_8003F128 = 0x31863158;
-
-u32 D_8003F12C = 0x312A30FB;
-
-u32 D_8003F130 = 0x30CD309E;
-
-u32 D_8003F134 = 0x30703041;
-
-u32 D_8003F138 = 0x30132FE4;
-
-u32 D_8003F13C = 0x2FB52F87;
-
-u32 D_8003F140 = 0x2F582F29;
-
-u32 D_8003F144 = 0x2EFB2ECC;
-
-u32 D_8003F148 = 0x2E9D2E6E;
-
-u32 D_8003F14C = 0x2E3F2E11;
-
-u32 D_8003F150 = 0x2DE22DB3;
-
-u32 D_8003F154 = 0x2D842D55;
-
-u32 D_8003F158 = 0x2D262CF7;
-
-u32 D_8003F15C = 0x2CC82C98;
-
-u32 D_8003F160 = 0x2C692C3A;
-
-u32 D_8003F164 = 0x2C0B2BDC;
-
-u32 D_8003F168 = 0x2BAD2B7D;
-
-u32 D_8003F16C = 0x2B4E2B1F;
-
-u32 D_8003F170 = 0x2AEF2AC0;
-
-u32 D_8003F174 = 0x2A912A61;
-
-u32 D_8003F178 = 0x2A322A02;
-
-u32 D_8003F17C = 0x29D329A3;
-
-u32 D_8003F180 = 0x29742944;
-
-u32 D_8003F184 = 0x291528E5;
-
-u32 D_8003F188 = 0x28B52886;
-
-u32 D_8003F18C = 0x28562826;
-
-u32 D_8003F190 = 0x27F627C7;
-
-u32 D_8003F194 = 0x27972767;
-
-u32 D_8003F198 = 0x27372707;
-
-u32 D_8003F19C = 0x26D826A8;
-
-u32 D_8003F1A0 = 0x26782648;
-
-u32 D_8003F1A4 = 0x261825E8;
-
-u32 D_8003F1A8 = 0x25B82588;
-
-u32 D_8003F1AC = 0x25582528;
-
-u32 D_8003F1B0 = 0x24F724C7;
-
-u32 D_8003F1B4 = 0x24972467;
-
-u32 D_8003F1B8 = 0x24372407;
-
-u32 D_8003F1BC = 0x23D623A6;
-
-u32 D_8003F1C0 = 0x23762345;
-
-u32 D_8003F1C4 = 0x231522E5;
-
-u32 D_8003F1C8 = 0x22B42284;
-
-u32 D_8003F1CC = 0x22542223;
-
-u32 D_8003F1D0 = 0x21F321C2;
-
-u32 D_8003F1D4 = 0x21922161;
-
-u32 D_8003F1D8 = 0x21312100;
-
-u32 D_8003F1DC = 0x20D0209F;
-
-u32 D_8003F1E0 = 0x206E203E;
-
-u32 D_8003F1E4 = 0x200D1FDC;
-
-u32 D_8003F1E8 = 0x1FAC1F7B;
-
-u32 D_8003F1EC = 0x1F4A1F19;
-
-u32 D_8003F1F0 = 0x1EE91EB8;
-
-u32 D_8003F1F4 = 0x1E871E56;
-
-u32 D_8003F1F8 = 0x1E251DF5;
-
-u32 D_8003F1FC = 0x1DC41D93;
-
-u32 D_8003F200 = 0x1D621D31;
-
-u32 D_8003F204 = 0x1D001CCF;
-
-u32 D_8003F208 = 0x1C9E1C6D;
-
-u32 D_8003F20C = 0x1C3C1C0B;
-
-u32 D_8003F210 = 0x1BDA1BA9;
-
-u32 D_8003F214 = 0x1B781B47;
-
-u32 D_8003F218 = 0x1B161AE4;
-
-u32 D_8003F21C = 0x1AB31A82;
-
-u32 D_8003F220 = 0x1A511A20;
-
-u32 D_8003F224 = 0x19EF19BD;
-
-u32 D_8003F228 = 0x198C195B;
-
-u32 D_8003F22C = 0x192A18F8;
-
-u32 D_8003F230 = 0x18C71896;
-
-u32 D_8003F234 = 0x18641833;
-
-u32 D_8003F238 = 0x180217D0;
-
-u32 D_8003F23C = 0x179F176D;
-
-u32 D_8003F240 = 0x173C170A;
-
-u32 D_8003F244 = 0x16D916A8;
-
-u32 D_8003F248 = 0x16761645;
-
-u32 D_8003F24C = 0x161315E2;
-
-u32 D_8003F250 = 0x15B0157F;
-
-u32 D_8003F254 = 0x154D151B;
-
-u32 D_8003F258 = 0x14EA14B8;
-
-u32 D_8003F25C = 0x14871455;
-
-u32 D_8003F260 = 0x142313F2;
-
-u32 D_8003F264 = 0x13C0138E;
-
-u32 D_8003F268 = 0x135D132B;
-
-u32 D_8003F26C = 0x12F912C8;
-
-u32 D_8003F270 = 0x12961264;
-
-u32 D_8003F274 = 0x12321201;
-
-u32 D_8003F278 = 0x11CF119D;
-
-u32 D_8003F27C = 0x116B1139;
-
-u32 D_8003F280 = 0x110810D6;
-
-u32 D_8003F284 = 0x10A41072;
-
-u32 D_8003F288 = 0x1040100E;
-
-u32 D_8003F28C = 0x0FDD0FAB;
-
-u32 D_8003F290 = 0x0F790F47;
-
-u32 D_8003F294 = 0x0F150EE3;
-
-u32 D_8003F298 = 0x0EB10E7F;
-
-u32 D_8003F29C = 0x0E4D0E1B;
-
-u32 D_8003F2A0 = 0x0DE90DB7;
-
-u32 D_8003F2A4 = 0x0D850D53;
-
-u32 D_8003F2A8 = 0x0D210CEF;
-
-u32 D_8003F2AC = 0x0CBD0C8B;
-
-u32 D_8003F2B0 = 0x0C590C27;
-
-u32 D_8003F2B4 = 0x0BF50BC3;
-
-u32 D_8003F2B8 = 0x0B910B5F;
-
-u32 D_8003F2BC = 0x0B2D0AFB;
-
-u32 D_8003F2C0 = 0x0AC90A97;
-
-u32 D_8003F2C4 = 0x0A650A33;
-
-u32 D_8003F2C8 = 0x0A0009CE;
-
-u32 D_8003F2CC = 0x099C096A;
-
-u32 D_8003F2D0 = 0x09380906;
-
-u32 D_8003F2D4 = 0x08D408A2;
-
-u32 D_8003F2D8 = 0x086F083D;
-
-u32 D_8003F2DC = 0x080B07D9;
-
-u32 D_8003F2E0 = 0x07A70775;
-
-u32 D_8003F2E4 = 0x07420710;
-
-u32 D_8003F2E8 = 0x06DE06AC;
-
-u32 D_8003F2EC = 0x067A0647;
-
-u32 D_8003F2F0 = 0x061505E3;
-
-u32 D_8003F2F4 = 0x05B1057F;
-
-u32 D_8003F2F8 = 0x054C051A;
-
-u32 D_8003F2FC = 0x04E804B6;
-
-u32 D_8003F300 = 0x04830451;
-
-u32 D_8003F304 = 0x041F03ED;
-
-u32 D_8003F308 = 0x03BA0388;
-
-u32 D_8003F30C = 0x03560324;
-
-u32 D_8003F310 = 0x02F102BF;
-
-u32 D_8003F314 = 0x028D025B;
-
-u32 D_8003F318 = 0x022801F6;
-
-u32 D_8003F31C = 0x01C40192;
-
-u32 D_8003F320 = 0x015F012D;
-
-u32 D_8003F324 = 0x00FB00C9;
-
-u32 D_8003F328 = 0x00960064;
-
-u32 D_8003F32C = 0x00320000;
-
-u32 D_8003F330 = 0x00000000;
-
-u32 D_8003F334 = 0x00000000;
-
-
-
-u32 D_8003F338 = 0x00000001;
-
-u32 D_8003F33C = 0x00000000;
-
-u32 D_8003F340 = 0x00000000;
-
-u32 D_8003F344 = 0x00000001;
-
-u32 D_8003F348 = 0x00000000;
-
-u32 D_8003F34C = 0x00000000;
-
-
-u32 D_8003F350 = 0x8004ADC0;
-u32 D_8003F354 = 0x0004B000;
-u32 D_8003F358 = 0x7D001018;
-u32 D_8003F35C = 0x40400818;
-u32 D_8003F360 = 0x00003214;
-
-u32 *D_8003F364 = _data_3E1400SegmentRomStart;
-u32 *D_8003F368 = _data_3E6BC0SegmentRomStart;
-u32 *D_8003F36C = _data_3E6BC0SegmentRomStart;
-u32 *D_8003F370 = _data_2A8CB0SegmentRomStart;
-u32 *D_8003F374 = _data_2B1510SegmentRomStart;
-u32 *D_8003F378 = _data_2B1510SegmentRomStart;
-u32 *D_8003F37C = _data_250320SegmentRomStart;
-
-u32 D_8003F380 = 0x00100808;
-u32 D_8003F384 = 0x00000000;
-u32 D_8003F388 = 0x00000000;
-u32 D_8003F38C = 0x00000000;
-
-u32 D_8003F390 = 0x00000000;
-u32 D_8003F394 = 0x00000000;
-u32 D_8003F398 = 0x00000000;
-u32 D_8003F39C = 0x00000000;
-
-u32 *D_8003F3A0 = _data_49F590SegmentRomStart;
-u32 *D_8003F3A4 = _data_49F590SegmentRomEnd;
-
-u32 *D_8003F3A8 = _data_4A0340SegmentRomStart;
-u32 *D_8003F3AC = _data_4A0340SegmentRomEnd;
-
-u32 *D_8003F3B0 = _data_4A3B60SegmentRomStart;
-u32 *D_8003F3B4 = _data_4A3B60SegmentRomEnd;
-
-u32 D_8003F3B8 = 0x00000000;
-u32 D_8003F3BC = 0x00000000;
-
-u32 D_8003F3C0 = 0x00000000;
-
-u32 D_8003F3C4 = 0x00000000;
-
-u32 D_8003F3C8 = 0x00000003;
-
-u32 D_8003F3CC = 0x00000000;
-
-u32 D_8003F3D0 = 0x00000000;
-
-u32 D_8003F3D4 = 0x00000000;
-
-u32 D_8003F3D8 = 0x00000000;
-
-u32 D_8003F3DC = 0x00000003;
-
-u32 D_8003F3E0 = 0x00000003;
-
-u32 D_8003F3E4 = 0x00000000;
-
-u32 D_8003F3E8 = 0x00000003;
-
-u32 D_8003F3EC = 0x00000003;
-
-u32 D_8003F3F0 = 0x00000000;
-
-u32 D_8003F3F4 = 0x00000003;
-
-u32 D_8003F3F8 = 0x00000003;
-
-u32 D_8003F3FC = 0x00000003;
-
-u32 D_8003F400 = 0x00000000;
-
-u32 D_8003F404 = 0x00000003;
-
-u32 D_8003F408 = 0x00000000;
-
-u32 D_8003F40C = 0x00000006;
-
-u32 D_8003F410 = 0x00000003;
-
-u32 D_8003F414 = 0x00000003;
-
-u32 D_8003F418 = 0x00000003;
-
-u32 D_8003F41C = 0x00000006;
-
-u32 D_8003F420 = 0x00000000;
-
-u32 D_8003F424 = 0x00000006;
-
-u32 D_8003F428 = 0x00000003;
-
-u32 D_8003F42C = 0x00000006;
-
-u32 D_8003F430 = 0x00000005;
-
-u32 D_8003F434 = 0x00000006;
-
-u32 D_8003F438 = 0x00000005;
-
-u32 D_8003F43C = 0x00000006;
-
-u32 D_8003F440 = 0x00000001;
-
-u32 D_8003F444 = 0x00000001;
-
-u32 D_8003F448 = 0x00000001;
-
-u32 D_8003F44C = 0x00000000;
-
-u32 D_8003F450 = 0x00000001;
-
-u32 D_8003F454 = 0x00000001;
-
-u32 D_8003F458 = 0x00000001;
-
-u32 D_8003F45C = 0x00000000;
-
-u32 D_8003F460 = 0x00000000;
-
-u32 D_8003F464 = 0x00000000;
-
-u32 D_8003F468 = 0x00000001;
-
-u32 D_8003F46C = 0x00000000;
-
-u32 D_8003F470 = 0x00000000;
-
-u32 D_8003F474 = 0x00000001;
-
-u32 D_8003F478 = 0x00000000;
-
-u32 D_8003F47C = 0x00000000;
-
-u32 D_8003F480 = 0x00000001;
-
-u32 D_8003F484 = 0x00000000;
-
-u32 D_8003F488 = 0x00000001;
-
-u32 D_8003F48C = 0x00000001;
-
-u32 D_8003F490 = 0x00000001;
-
-u32 D_8003F494 = 0x00000000;
-
-u32 D_8003F498 = 0x00000001;
-
-u32 D_8003F49C = 0x00000000;
-
-u32 D_8003F4A0 = 0x00000001;
-
-u32 D_8003F4A4 = 0x00000000;
-
-u32 D_8003F4A8 = 0x00000001;
-
-u32 D_8003F4AC = 0x00000001;
-
-u32 D_8003F4B0 = 0x00000000;
-
-u32 D_8003F4B4 = 0x00000001;
-
-u32 D_8003F4B8 = 0x00000001;
-
-u32 D_8003F4BC = 0x00000000;
-
-u32 D_8003F4C0 = 0x00000000;
-
-u32 D_8003F4C4 = 0x00000001;
-
-u32 D_8003F4C8 = 0x00000001;
-
-u32 D_8003F4CC = 0x00000001;
-
-u32 D_8003F4D0 = 0x00000000;
-
-u32 D_8003F4D4 = 0x00000001;
-
-u32 D_8003F4D8 = 0x00000000;
-
-u32 D_8003F4DC = 0x00000000;
-
-u32 D_8003F4E0 = 0x00000001;
-
-u32 D_8003F4E4 = 0x00000001;
-
-u32 D_8003F4E8 = 0x00000000;
-
-u32 D_8003F4EC = 0x00000001;
-
-u32 D_8003F4F0 = 0x00000000;
-
-u32 D_8003F4F4 = 0x00000001;
-
-u32 D_8003F4F8 = 0x00000001;
-
-u32 D_8003F4FC = 0x00000000;
-
-u32 D_8003F500 = 0x00000001;
-
-u32 D_8003F504 = 0x00000001;
-
-u32 D_8003F508 = 0x00000000;
-
-u32 D_8003F50C = 0x00000001;
-
-u32 D_8003F510 = 0x00000001;
-
-u32 D_8003F514 = 0x00000001;
-
-u32 D_8003F518 = 0x00000001;
-
-u32 D_8003F51C = 0x00000000;
-
-u32 D_8003F520 = 0x00000001;
-
-u32 D_8003F524 = 0x00000000;
-
-u32 D_8003F528 = 0x00000001;
-
-u32 D_8003F52C = 0x00000000;
-
-u32 D_8003F530 = 0x00000000;
-
-u32 D_8003F534 = 0x00000001;
-
-u32 D_8003F538 = 0x00000000;
-
-u32 D_8003F53C = 0x00000000;
-
-u32 D_8003F540 = 0x00000001;
-
-u32 D_8003F544 = 0x00000001;
-
-u32 D_8003F548 = 0x00000001;
-
-u32 D_8003F54C = 0x00000001;
-
-u32 D_8003F550 = 0x00000001;
-
-u32 D_8003F554 = 0x00000001;
-
-u32 D_8003F558 = 0x00000001;
-
-u32 D_8003F55C = 0x00000000;
-
-u32 D_8003F560 = 0x00000001;
-
-u32 D_8003F564 = 0x00000001;
-
-u32 D_8003F568 = 0x00000001;
-
-u32 D_8003F56C = 0x00000001;
-
-u32 D_8003F570 = 0x00000000;
-
-u32 D_8003F574 = 0x00000001;
-
-u32 D_8003F578 = 0x00000001;
-
-u32 D_8003F57C = 0x00000000;
-
-u32 D_8003F580 = 0x00000001;
-
-u32 D_8003F584 = 0x00000001;
-
-u32 D_8003F588 = 0x00000001;
-
-u32 D_8003F58C = 0x00000001;
-
-u32 D_8003F590 = 0x00000001;
-
-u32 D_8003F594 = 0x00000001;
-
-u32 D_8003F598 = 0x00000000;
-
-u32 D_8003F59C = 0x00000000;
-
-u32 D_8003F5A0 = 0x00000000;
-
-u32 D_8003F5A4 = 0x00000001;
-
-u32 D_8003F5A8 = 0x00000000;
-
-u32 D_8003F5AC = 0x00000001;
-
-u32 D_8003F5B0 = 0x00000001;
-
-u32 D_8003F5B4 = 0x00000001;
-
-u32 D_8003F5B8 = 0x00000001;
-
-u32 D_8003F5BC = 0x00000000;
-
-u32 D_8003F5C0 = 0x00000000;
-
-u32 D_8003F5C4 = 0x00000000;
-
-u32 D_8003F5C8 = 0x00000000;
-
-u32 D_8003F5CC = 0x00000001;
-
-u32 D_8003F5D0 = 0x00000001;
-
-u32 D_8003F5D4 = 0x00000000;
-
-u32 D_8003F5D8 = 0x00000001;
-
-u32 D_8003F5DC = 0x00000000;
-
-u32 D_8003F5E0 = 0x00000000;
-
-u32 D_8003F5E4 = 0x00000000;
-
-u32 D_8003F5E8 = 0x00000001;
-
-u32 D_8003F5EC = 0x00000001;
-
-u32 D_8003F5F0 = 0x00000001;
-
-u32 D_8003F5F4 = 0x00000001;
-
-u32 D_8003F5F8 = 0x00000001;
-
-u32 D_8003F5FC = 0x00000000;
-
-u32 D_8003F600 = 0x00000001;
-
-u32 D_8003F604 = 0x00000001;
-
-u32 D_8003F608 = 0x00000000;
-
-u32 D_8003F60C = 0x00000001;
-
-u32 D_8003F610 = 0x00000001;
-
-u32 D_8003F614 = 0x00000000;
-
-u32 D_8003F618 = 0x00000001;
-
-u32 D_8003F61C = 0x00000000;
-
-u32 D_8003F620 = 0x00000001;
-
-u32 D_8003F624 = 0x00000001;
-
-u32 D_8003F628 = 0x00000000;
-
-u32 D_8003F62C = 0x00000001;
-
-u32 D_8003F630 = 0x00000001;
-
-u32 D_8003F634 = 0x00000000;
-
-u32 D_8003F638 = 0x00000000;
-
-u32 D_8003F63C = 0x00000000;
-
-u32 D_8003F640 = 0x00000000;
-
-u32 D_8003F644 = 0x00000000;
-
-u32 D_8003F648 = 0x00000000;
-
-u32 D_8003F64C = 0x00000001;
-
-u32 D_8003F650 = 0x00000000;
-
-u32 D_8003F654 = 0x00000000;
-
-u32 D_8003F658 = 0x00000000;
-
-u32 D_8003F65C = 0x00000000;
-
-f32 D_8003F660 = 1.0f;
-
-f32 D_8003F664 = 10.0f;
-
-f32 D_8003F668 = 100.0f;
-
-f32 D_8003F66C = 1000.0f;
-
-f32 D_8003F670 = 10000.0f;
-
-f32 D_8003F674 = 100000.0f;
-
-f32 D_8003F678 = 1000000.0f;
-
-f32 D_8003F67C = 10000000.0f;
-
-f32 D_8003F680 = 100000000.0f;
-
-f32 D_8003F684 = 1000000000.0f;
-
-u32 D_8003F688 = 0x00000000;
-
-u32 D_8003F68C = 0xFFFFFFFF;
-
-u32 D_8003F690 = 0xFFFFFFFF;
-
-u32 D_8003F694 = 0xFFFFFFFF;
-
-u32 D_8003F698 = 0xFFFFFFFF;
-
-u32 D_8003F69C = 0xFFFFFFFF;
-
-u32 D_8003F6A0 = 0xFFFFFFFF;
-
-u32 D_8003F6A4 = 0xFFFFFFFF;
-
-u32 D_8003F6A8 = 0xFFFFFFFF;
-
-u32 D_8003F6AC = 0xFF29FFFF;
-
-u32 D_8003F6B0 = 0xFF2BFFFF;
-
-u32 D_8003F6B4 = 0x2526FF2A;
-
-u32 D_8003F6B8 = 0xFF272CFF;
-
-u32 D_8003F6BC = 0x00010203;
-
-u32 D_8003F6C0 = 0x04050607;
-
-u32 D_8003F6C4 = 0x080924FF;
-
-u32 D_8003F6C8 = 0xFFFFFF28;
-
-u32 D_8003F6CC = 0xFF0A0B0C;
-
-u32 D_8003F6D0 = 0x0D0E0F10;
-
-u32 D_8003F6D4 = 0x11121314;
-
-u32 D_8003F6D8 = 0x15161718;
-
-u32 D_8003F6DC = 0x191A1B1C;
-
-u32 D_8003F6E0 = 0x1D1E1F20;
-
-u32 D_8003F6E4 = 0x212223FF;
-
-u32 D_8003F6E8 = 0xFFFFFFFF;
-
-u32 D_8003F6EC = 0xFF0A0B0C;
-
-u32 D_8003F6F0 = 0x0D0E0F10;
-
-u32 D_8003F6F4 = 0x11121314;
-
-u32 D_8003F6F8 = 0x15161718;
-
-u32 D_8003F6FC = 0x191A1B1C;
-
-u32 D_8003F700 = 0x1D1E1F20;
-
-u32 D_8003F704 = 0x212223FF;
-
-u32 D_8003F708 = 0xFFFFFFFF;
-
-u32 D_8003F70C = 0x70871C30;
-
-u32 D_8003F710 = 0x8988A250;
-
-u32 D_8003F714 = 0x88808290;
-
-u32 D_8003F718 = 0x88831C90;
-
-u32 D_8003F71C = 0x888402F8;
-
-u32 D_8003F720 = 0x88882210;
-
-u32 D_8003F724 = 0x71CF9C10;
-
-u32 D_8003F728 = 0xF9CF9C70;
-
-u32 D_8003F72C = 0x8228A288;
-
-u32 D_8003F730 = 0xF200A288;
-
-u32 D_8003F734 = 0x0BC11C78;
-
-u32 D_8003F738 = 0x0A222208;
-
-u32 D_8003F73C = 0x8A222288;
-
-u32 D_8003F740 = 0x71C21C70;
-
-u32 D_8003F744 = 0x23C738F8;
-
-u32 D_8003F748 = 0x5228A480;
-
-u32 D_8003F74C = 0x8A282280;
-
-u32 D_8003F750 = 0x8BC822F0;
-
-u32 D_8003F754 = 0xFA282280;
-
-u32 D_8003F758 = 0x8A28A480;
-
-u32 D_8003F75C = 0x8BC738F8;
-
-u32 D_8003F760 = 0xF9C89C08;
-
-u32 D_8003F764 = 0x82288808;
-
-u32 D_8003F768 = 0x82088808;
-
-u32 D_8003F76C = 0xF2EF8808;
-
-u32 D_8003F770 = 0x82288888;
-
-u32 D_8003F774 = 0x82288888;
-
-u32 D_8003F778 = 0x81C89C70;
-
-u32 D_8003F77C = 0x8A08A270;
-
-u32 D_8003F780 = 0x920DA288;
-
-u32 D_8003F784 = 0xA20AB288;
-
-u32 D_8003F788 = 0xC20AAA88;
-
-u32 D_8003F78C = 0xA208A688;
-
-u32 D_8003F790 = 0x9208A288;
-
-u32 D_8003F794 = 0x8BE8A270;
-
-u32 D_8003F798 = 0xF1CF1CF8;
-
-u32 D_8003F79C = 0x8A28A220;
-
-u32 D_8003F7A0 = 0x8A28A020;
-
-u32 D_8003F7A4 = 0xF22F1C20;
-
-u32 D_8003F7A8 = 0x82AA0220;
-
-u32 D_8003F7AC = 0x82492220;
-
-u32 D_8003F7B0 = 0x81A89C20;
-
-u32 D_8003F7B4 = 0x8A28A288;
-
-u32 D_8003F7B8 = 0x8A28A288;
-
-u32 D_8003F7BC = 0x8A289488;
-
-u32 D_8003F7C0 = 0x8A2A8850;
-
-u32 D_8003F7C4 = 0x894A9420;
-
-u32 D_8003F7C8 = 0x894AA220;
-
-u32 D_8003F7CC = 0x70852220;
-
-u32 D_8003F7D0 = 0xF8011000;
-
-u32 D_8003F7D4 = 0x08020800;
-
-u32 D_8003F7D8 = 0x10840400;
-
-u32 D_8003F7DC = 0x20040470;
-
-u32 D_8003F7E0 = 0x40840400;
-
-u32 D_8003F7E4 = 0x80020800;
-
-u32 D_8003F7E8 = 0xF8011000;
-
-u32 D_8003F7EC = 0x70800000;
-
-u32 D_8003F7F0 = 0x88822200;
-
-u32 D_8003F7F4 = 0x08820400;
-
-u32 D_8003F7F8 = 0x108F8800;
-
-u32 D_8003F7FC = 0x20821000;
-
-u32 D_8003F800 = 0x00022200;
-
-u32 D_8003F804 = 0x20800020;
-
-u32 D_8003F808 = 0x00000000;
-
-u32 D_8003F80C = 0x80040D10;
-
-u32 D_8003F810 = 0x80040D1C;
-
-u32 D_8003F814 = 0x80040D30;
-
-u32 D_8003F818 = 0x80040D48;
-
-u32 D_8003F81C = 0x80040D60;
-
-u32 D_8003F820 = 0x80040D78;
-
-u32 D_8003F824 = 0x80040D90;
-
-u32 D_8003F828 = 0x80040DA4;
-
-u32 D_8003F82C = 0x80040DB8;
-
-u32 D_8003F830 = 0x80040DD0;
-
-u32 D_8003F834 = 0x80040DE8;
-
-u32 D_8003F838 = 0x80040E00;
-
-u32 D_8003F83C = 0x80040E18;
-
-u32 D_8003F840 = 0x80040E2C;
-
-u32 D_8003F844 = 0x80040E3C;
-
-u32 D_8003F848 = 0x80040E58;
-
-u32 D_8003F84C = 0x80040E74;
-
-u32 D_8003F850 = 0x80040E8C;
-
-char *D_8003F854[] = {
-	0x80040EA8,
-	0x80040EC0,
-	0x80040ED4,
-	0x80040EE8,
-	0x80040EF4,
-	0x80040F00,
-};
-
-u32 D_8003F86C = 0x0000001E;
-
-u32 D_8003F870 = 0x00000019;
-
-u32 D_8003F874 = 0x00000000;
-
-u32 D_8003F878 = 0x00000000;
-
-u32 D_8003F87C = 0x00000000;
-
-u32 D_8003F880 = 0x00000003;
-
-u32 D_8003F884 = 0x00000FA0;
-
-u32 D_8003F888 = 0x00000000;
-
-u32 D_8003F88C = 0x00000870;
-
-u32 D_8003F890 = 0x00002666;
-
-u32 D_8003F894 = 0xFFFFD99A;
-
-u32 D_8003F898 = 0x00000000;
-
-u32 D_8003F89C = 0x00000000;
-
-u32 D_8003F8A0 = 0x00000000;
-
-u32 D_8003F8A4 = 0x00000000;
-
-u32 D_8003F8A8 = 0x000002F8;
-
-u32 D_8003F8AC = 0x000005F0;
-
-u32 D_8003F8B0 = 0x00000CCC;
-
-u32 D_8003F8B4 = 0xFFFFF334;
-
-u32 D_8003F8B8 = 0x00003FFF;
-
-u32 D_8003F8BC = 0x00000000;
-
-u32 D_8003F8C0 = 0x00000000;
-
-u32 D_8003F8C4 = 0x00000000;
-
-u32 D_8003F8C8 = 0x00000000;
-
-u32 D_8003F8CC = 0x00000960;
-
-u32 D_8003F8D0 = 0x00001388;
-
-u32 D_8003F8D4 = 0x00000000;
-
-u32 D_8003F8D8 = 0x00000000;
-
-u32 D_8003F8DC = 0x00000000;
-
-u32 D_8003F8E0 = 0x00000000;
-
-u32 D_8003F8E4 = 0x00005000;
-
-u32 D_8003F8E8 = 0x00000004;
-
-u32 D_8003F8EC = 0x00000FA0;
-
-u32 D_8003F8F0 = 0x00000000;
-
-u32 D_8003F8F4 = 0x00000A50;
-
-u32 D_8003F8F8 = 0x00002666;
-
-u32 D_8003F8FC = 0xFFFFD99A;
-
-u32 D_8003F900 = 0x00000000;
-
-u32 D_8003F904 = 0x00000000;
-
-u32 D_8003F908 = 0x00000000;
-
-u32 D_8003F90C = 0x00000000;
-
-u32 D_8003F910 = 0x00000370;
-
-u32 D_8003F914 = 0x00000870;
-
-u32 D_8003F918 = 0x00000CCC;
-
-u32 D_8003F91C = 0xFFFFF334;
-
-u32 D_8003F920 = 0x00003FFF;
-
-u32 D_8003F924 = 0x00000000;
-
-u32 D_8003F928 = 0x00000000;
-
-u32 D_8003F92C = 0x00000000;
-
-u32 D_8003F930 = 0x00000A50;
-
-u32 D_8003F934 = 0x00000E38;
-
-u32 D_8003F938 = 0x00000CCC;
-
-u32 D_8003F93C = 0xFFFFF334;
-
-u32 D_8003F940 = 0x00003FFF;
-
-u32 D_8003F944 = 0x00000000;
-
-u32 D_8003F948 = 0x00000000;
-
-u32 D_8003F94C = 0x00000000;
-
-u32 D_8003F950 = 0x00000000;
-
-u32 D_8003F954 = 0x00000EB0;
-
-u32 D_8003F958 = 0x00001F40;
-
-u32 D_8003F95C = 0x00000000;
-
-u32 D_8003F960 = 0x00000000;
-
-u32 D_8003F964 = 0x00000000;
-
-u32 D_8003F968 = 0x00000000;
-
-u32 D_8003F96C = 0x00005000;
-
-u32 D_8003F970 = 0x00000001;
-
-u32 D_8003F974 = 0x00001F40;
-
-u32 D_8003F978 = 0x00000000;
-
-u32 D_8003F97C = 0x00001BF8;
-
-u32 D_8003F980 = 0x00002EE0;
-
-u32 D_8003F984 = 0x00000000;
-
-u32 D_8003F988 = 0x00007FFF;
-
-u32 D_8003F98C = 0x00000000;
-
-u32 D_8003F990 = 0x00000000;
-
-u32 D_8003F994 = 0x00000000;
-
-u32 D_8003F998 = 0x00000001;
-
-u32 D_8003F99C = 0x00000320;
-
-u32 D_8003F9A0 = 0x00000000;
-
-u32 D_8003F9A4 = 0x000000C8;
-
-u32 D_8003F9A8 = 0x00004000;
-
-u32 D_8003F9AC = 0x00000000;
-
-u32 D_8003F9B0 = 0x00007FFF;
-
-u32 D_8003F9B4 = 0x00001DB0;
-
-u32 D_8003F9B8 = 0x000002BC;
-
-u32 D_8003F9BC = 0x00000000;
-
-u32 D_8003F9C0 = 0x00000001;
-
-u32 D_8003F9C4 = 0x00000320;
-
-u32 D_8003F9C8 = 0x00000000;
-
-u32 D_8003F9CC = 0x000000C8;
-
-u32 D_8003F9D0 = 0x00000000;
-
-u32 D_8003F9D4 = 0x00005FFF;
-
-u32 D_8003F9D8 = 0x00007FFF;
-
-u32 D_8003F9DC = 0x0000017C;
-
-u32 D_8003F9E0 = 0x000001F4;
-
-u32 D_8003F9E4 = 0x00000000;
-
-u32 D_8003F9E8 = 0x00000000;
-
-u32 D_8003F9EC = 0x00000000;
-
-u32 D_8003F9F0 = 0x00000000;
-
-u32 D_8003F9F4 = 0x00000000;
-
-u32 D_8003F9F8 = 0x00000000;
-
-u32 D_8003F9FC = 0x00000000;
-
-u32 D_8003FA00 = 0x00000000;
-
-u32 D_8003FA04 = 0x00000000;
-
-u32 D_8003FA08 = 0x00000000;
-
-u32 D_8003FA0C = 0x00000000;
-
-u32 D_8003FA10 = 0x7FFF7FFC;
-
-u32 D_8003FA14 = 0x7FF57FE8;
-
-u32 D_8003FA18 = 0x7FD77FC0;
-
-u32 D_8003FA1C = 0x7FA57F84;
-
-u32 D_8003FA20 = 0x7F5F7F34;
-
-u32 D_8003FA24 = 0x7F057ED0;
-
-u32 D_8003FA28 = 0x7E977E58;
-
-u32 D_8003FA2C = 0x7E157DCD;
-
-u32 D_8003FA30 = 0x7D7F7D2D;
-
-u32 D_8003FA34 = 0x7CD67C7A;
-
-u32 D_8003FA38 = 0x7C1A7BB4;
-
-u32 D_8003FA3C = 0x7B497ADA;
-
-u32 D_8003FA40 = 0x7A6679ED;
-
-u32 D_8003FA44 = 0x796F78ED;
-
-u32 D_8003FA48 = 0x786677DA;
-
-u32 D_8003FA4C = 0x774976B4;
-
-u32 D_8003FA50 = 0x761A757B;
-
-u32 D_8003FA54 = 0x74D87430;
-
-u32 D_8003FA58 = 0x738472D3;
-
-u32 D_8003FA5C = 0x721E7164;
-
-u32 D_8003FA60 = 0x70A66FE3;
-
-u32 D_8003FA64 = 0x6F1C6E51;
-
-u32 D_8003FA68 = 0x6D816CAD;
-
-u32 D_8003FA6C = 0x6BD56AF9;
-
-u32 D_8003FA70 = 0x6A186933;
-
-u32 D_8003FA74 = 0x684A675D;
-
-u32 D_8003FA78 = 0x666C6577;
-
-u32 D_8003FA7C = 0x647E6381;
-
-u32 D_8003FA80 = 0x6280617C;
-
-u32 D_8003FA84 = 0x60735F67;
-
-u32 D_8003FA88 = 0x5E575D43;
-
-u32 D_8003FA8C = 0x5C2C5B11;
-
-u32 D_8003FA90 = 0x59F258D0;
-
-u32 D_8003FA94 = 0x57AA5681;
-
-u32 D_8003FA98 = 0x55555425;
-
-u32 D_8003FA9C = 0x52F251BC;
-
-u32 D_8003FAA0 = 0x50824F46;
-
-u32 D_8003FAA4 = 0x4E064CC3;
-
-u32 D_8003FAA8 = 0x4B7D4A35;
-
-u32 D_8003FAAC = 0x48E9479B;
-
-u32 D_8003FAB0 = 0x464944F5;
-
-u32 D_8003FAB4 = 0x439E4245;
-
-u32 D_8003FAB8 = 0x40E93F8A;
-
-u32 D_8003FABC = 0x3E293CC6;
-
-u32 D_8003FAC0 = 0x3B6039F8;
-
-u32 D_8003FAC4 = 0x388D3721;
-
-u32 D_8003FAC8 = 0x35B23441;
-
-u32 D_8003FACC = 0x32CE3159;
-
-u32 D_8003FAD0 = 0x2FE22E69;
-
-u32 D_8003FAD4 = 0x2CEF2B72;
-
-u32 D_8003FAD8 = 0x29F42875;
-
-u32 D_8003FADC = 0x26F32570;
-
-u32 D_8003FAE0 = 0x23EC2266;
-
-u32 D_8003FAE4 = 0x20DF1F57;
-
-u32 D_8003FAE8 = 0x1DCE1C43;
-
-u32 D_8003FAEC = 0x1AB7192A;
-
-u32 D_8003FAF0 = 0x179C160E;
-
-u32 D_8003FAF4 = 0x147E12ED;
-
-u32 D_8003FAF8 = 0x115C0FCA;
-
-u32 D_8003FAFC = 0x0E380CA5;
-
-u32 D_8003FB00 = 0x0B11097D;
-
-u32 D_8003FB04 = 0x07E90654;
-
-u32 D_8003FB08 = 0x04C0032A;
-
-u32 D_8003FB0C = 0x01950000;
-
-u32 D_8003FB10 = 0x00000000;
-
-u32 D_8003FB14 = 0x00000000;
-
-f32 D_8003FB18 = 1.0f;
-
-u32 D_8003FB1C = 0xFFFF0000;
-
-u32 D_8003FB20 = 0x00000001;
-
-u32 D_8003FB24 = 0x00000001;
-
-u32 D_8003FB28 = 0x00000000;
-
-u32 D_8003FB2C = 0x00000000;
-
-u32 __osHwIntTable = 0x00000000;
-
-u32 D_8003FB34 = 0x00000000;
-
-u32 D_8003FB38 = 0x00000000;
-
-u32 D_8003FB3C = 0x00000000;
-
-u32 D_8003FB40 = 0x00000000;
-
-u32 D_8003FB44 = 0x00000000;
-
-u32 D_8003FB48 = 0x00000000;
-
-u32 D_8003FB4C = 0x00000000;
-
-u32 __osThreadTail = 0x00000000;
-
-u32 D_8003FB54 = 0xFFFFFFFF;
-
-u32 __osRunQueue = 0x8003FB50;
-
-u32 __osActiveQueue = 0x8003FB50;
-
-u32 __osRunningThread = 0x00000000;
-
-u32 __osFaultedThread = 0x00000000;
-
-u32 D_8003FB68 = 0x00000000;
-
-u32 D_8003FB6C = 0x00000000;
-
-u32 D_8003FB70 = 0x00000000;
-
-u32 D_8003FB74 = 0x00000000;
-
-u32 D_8003FB78 = 0x00000000;
-
-u32 D_8003FB7C = 0x00000000;
-
-u32 D_8003FB80 = 0x00000000;
-
-u32 D_8003FB84 = 0x00000000;
-
-u32 D_8003FB88 = 0x00000000;
-
-u32 D_8003FB8C = 0x00000000;
-
-u64 osClockRate = 0x0000000003B9ACA0;
-
-u32 osViClock = 0x02E6D354;
-
-u32 __osShutdown = 0x00000000;
-
-u32 __OSGlobalIntMask = 0x003FFF01;
-
-u32 D_8003FBA4 = 0x00000000;
-
-u32 D_8003FBA8 = 0x00000000;
-
-u32 D_8003FBAC = 0x00000000;
-
-u32 D_8003FBB0 = 0xFFFFFFFF;
-
-u32 D_8003FBB4 = 0xFA000000;
-
-u32 D_8003FBB8 = 0x00000000;
-
-u32 D_8003FBBC = 0x00000000;
-
-OSDevMgr __osPiDevMgr = {0};
-u32 __osPiTable = 0x00000000;
-
-u32 __osCurrentHandle = 0x80098E00;
-
-u32 D_8003FBE4 = 0x80098E78;
-
-u32 D_8003FBE8 = 0x00000000;
-
-u32 D_8003FBEC = 0x00000000;
-
-u32 D_8003FBF0 = 0x00000000;
-
-u32 D_8003FBF4 = 0x00000000;
-
-u32 D_8003FBF8 = 0x00000000;
-
-u32 D_8003FBFC = 0x00000000;
-
-u32 D_8003FC00 = 0x20202020;
-
-u32 D_8003FC04 = 0x20202020;
-
-u32 D_8003FC08 = 0x20202020;
-
-u32 D_8003FC0C = 0x20202020;
-
-u32 D_8003FC10 = 0x20202020;
-
-u32 D_8003FC14 = 0x20202020;
-
-u32 D_8003FC18 = 0x20202020;
-
-u32 D_8003FC1C = 0x20202020;
-
-u32 D_8003FC20 = 0x00000000;
-
-u32 D_8003FC24 = 0x30303030;
-
-u32 D_8003FC28 = 0x30303030;
-
-u32 D_8003FC2C = 0x30303030;
-
-u32 D_8003FC30 = 0x30303030;
-
-u32 D_8003FC34 = 0x30303030;
-
-u32 D_8003FC38 = 0x30303030;
-
-u32 D_8003FC3C = 0x30303030;
-
-u32 D_8003FC40 = 0x30303030;
-
-u32 D_8003FC44 = 0x00000000;
-
-u32 D_8003FC48 = 0x00000000;
-
-u32 D_8003FC4C = 0x00000000;
-
