+++ src/ovl0/ovl0_2_5.c
@@ -1,3 +1,5 @@
+#define	OS_K0_TO_PHYSICAL(x)	(u32)(((char *)(x)-0x80000000))
+
 #include <ultra64.h>
 #include <macros.h>
 #include <PR/os_message.h>
@@ -16,21 +18,12 @@ void func_80005350(void *arg0) {
     D_8004A48C = func_80000A44;
 }
 
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void *func_80005378(void *arg0) {
-    void *temp_v0;
+extern u32* D_80049308;
 
-    D_80049308 = (s32) (*arg0 + 4);
-    temp_v0 = *arg0;
-    *arg0 = (void *) (temp_v0 + 8);
-    temp_v0->unk4 = 0;
-    temp_v0->unk0 = 0xDB06003C;
-    return temp_v0;
+void func_80005378(Gfx **arg0) {
+    D_80049308 = &(*arg0)->words.w1;
+    gSPSegment((*arg0)++, 0x0F, 0x00000000);
 }
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80005378.s")
-#endif
 
 extern s32 D_8004A440;
 
@@ -43,13 +36,6 @@ void func_800053B4(u16 arg0, u16 arg1) {
     D_8004A444 = arg0;
     D_8004A446 = arg1;
 }
- 
-struct Unk_Ovl0_2_5 { // used a lot in this file
-    u32 unk0;
-    u32 unk4;
-    u32 unk8;
-    u32 unkC;
-};
 
 extern struct Unk_Ovl0_2_5 D_8004A408;
 
@@ -558,7 +544,6 @@ void func_80005A98(struct Unk80005A98 *arg0, s32 arg1, u32 arg2, s32 arg3, s32 a
 GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80005A98.s")
 #endif
 
-
 #ifdef MIPS_TO_C
 u32 func_80005C64(void) {
     u16 phi_v1;
@@ -590,7 +575,6 @@ u32 func_80005C64(void) {
 #else
 GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80005C64.s")
 #endif
-
 
 #ifdef MIPS_TO_C
 void func_80005CC0(s32 arg0, s32 arg1) {
@@ -633,66 +617,42 @@ GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80005CC0.s")
 #endif
 
 
-// comment out the last block in kirby.039DF0.2.c if you're testing matching
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_80005DE4(Gfx *arg0, u32 arg1) {
-    u32 temp_a1;
-    u32 temp_a1_2;
-    u32 temp_a1_3;
-    u32 temp_a1_4;
-    u32 temp_a1_5;
-    u32 temp_a1_6;
-    u32 temp_a1_7;
-    s32 phi_v0;
+extern Gfx* D_8004A44C;
 
-    phi_v0 = 0;
-    if (arg1 < 0x10) {
-        phi_v0 = 0;
-        switch(arg1) {
-            // goto **(&jtbl_800401C4 + (arg1 * 4));
-            case 0:
-                temp_a1 = arg0->words.w0;
-                arg0->words.w0 = temp_a1 + 8;
-                temp_a1->unk0 = 0xE1000000;
-                temp_a1->unk4 = &D_00041FF0;
-                temp_a1_2 = arg0->words.w0;
-                arg0->words.w0 = temp_a1_2 + 8;
-                temp_a1_2->unk0 = 0xDD0007FF;
-                temp_a1_2->unk4 = &D_00039E90;
-                phi_v0 = 0;
-                break; // goes to case 15
-            default:
-                temp_a1_3 = arg0->words.w0;
-                arg0->words.w0 = temp_a1_3 + 8;
-                temp_a1_3->unk0 = 0xE1000000;
-                temp_a1_3->unk4 = &D_00042410;
-                temp_a1_4 = arg0->words.w0;
-                arg0->words.w0 = temp_a1_4 + 8;
-                temp_a1_4->unk0 = 0xDD0007FF;
-                temp_a1_4->unk4 = &D_0003B220;
-                phi_v0 = 0;
-                break;
-            case 9:
-            case 10:
-                temp_a1_5 = arg0->words.w0;
-                arg0->words.w0 = temp_a1_5 + 8;
-                temp_a1_5->unk0 = 0xE1000000;
-                temp_a1_5->unk4 = &D_00042800;
-                temp_a1_6 = arg0->words.w0;
-                arg0->words.w0 = temp_a1_6 + 8;
-                temp_a1_6->unk0 = 0xDD0007FF;
-                temp_a1_6->unk4 = &D_0003C3B0;
-                phi_v0 = 1;
-            case 11:
-            case 12:
-            case 13:
-            case 14:
-            case 15:
-        }
+extern long long int gspF3DEX2_fifoDataStart[];
+extern long long int gspF3DEX2_fifoTextStart[];
+
+extern long long int gspL3DEX2_fifoDataStart[];
+extern long long int gspL3DEX2_fifoTextStart[];
+
+extern long long int gspS2DEX2_fifoDataStart[];
+extern long long int gspS2DEX2_fifoTextStart[];
+
+
+// comment out the last block in kirby.039DF0.2.c if you're testing matching
+#if 0
+// Matches, but can't be used until the rodata for this file is sorted out
+void func_80005DE4(Gfx **arg0, u32 arg1) {
+    s32 phi_v0 = 0;
+    switch (arg1) { 
+        case 0:
+            // F3DEX2
+            gSPLoadUcodeL((*arg0)++, gspF3DEX2_fifo);
+            break;
+        case 1: case 2: case 3: case 4: case 5: case 6: case 7: case 8:
+            // L3DEX2
+            gSPLoadUcodeL((*arg0)++, gspL3DEX2_fifo);
+            break;
+        case 9: case 10:
+            // S2DEX2
+            gSPLoadUcodeL((*arg0)++, gspS2DEX2_fifo);
+            phi_v0 = 1;
+            break;
+        case 11: case 12: case 13: case 14: case 15: default:
+            break;
     }
     if (phi_v0 == 0) {
-        gSPDisplayList(*arg0++, D_8004A44C);
+        gSPDisplayList((*arg0)++, D_8004A44C);
     }
 }
 
@@ -1547,77 +1507,53 @@ void func_80007380(void *arg0) {
 GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007380.s")
 #endif
 
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_800075DC(s16 arg0, s16 arg1) {
-    D_800492D8 = arg0;
-    *(void *)0x800492DA = arg1;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_800075DC.s")
-#endif
+extern s16 D_800492D8;
+extern s16 D_800492DA;
 
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void *func_800075F8(void) {
+void func_800075DC(u16 arg0, u16 arg1) {
+    D_800492D8 = arg0;
+    D_800492DA = arg1;
+}
+
+extern s32 D_800492DC;
+
+void func_800075F8(void) {
     if (D_800492DC != 2) {
         D_800492DC = 1;
     }
-    return &D_800492DC;
 }
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_800075F8.s")
-#endif
 
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-? func_80007620(void) {
+u32 func_80007620(void) {
     if (D_800492DC == 2) {
         return 1;
     }
     return 0;
 }
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007620.s")
-#endif
 
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
+extern OSMesgQueue D_800492E8;
+
 void func_80007648(void) {
     osSendMesg(&D_800492E8, 0, 0);
 }
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007648.s")
-#endif
 
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
+extern s32 D_8004A460;
+
 void func_80007674(s32 arg0) {
     if ((arg0 == 1) || (arg0 == 2)) {
         D_8004A460 = arg0;
-    } else {
-
     }
 }
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007674.s")
-#endif
 
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-? func_80007694(s32 arg0) {
+extern s32 D_8004A454[];
+
+u32 func_80007694(s32 arg0) {
     if ((arg0 == 1) || (arg0 == 2)) {
-        if (((arg0 * 4) + 0x80050000)->unk-5BAC == 0) {
+        if (D_8004A454[arg0] == 0) {
             return 1;
         }
-    } else {
-
     }
     return 0;
 }
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007694.s")
-#endif
 
 #ifdef MIPS_TO_C
 //generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
@@ -1684,284 +1620,3 @@ loop_3:
 #else
 GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_800076D0.s")
 #endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_80007824(void *arg0) {
-    arg0->unkC = (s32) arg0->unk4;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007824.s")
-#endif
-
-extern const char D_80040210[];
-u32 func_8000783C(struct Unk_Ovl0_2_5 *arg0, u32 arg1, u32 arg2) {
-    u32 phi_v1;
-    u32 temp_v0;
-
-    if (arg2 != 0) {
-        temp_v0 = arg2 - 1;
-        phi_v1 = (arg0->unkC + temp_v0) & ~temp_v0;
-    } else {
-        phi_v1 = arg0->unkC;
-    }
-    temp_v0 = phi_v1 + arg1;
-    arg0->unkC = temp_v0;
-    if (arg0->unk8 < temp_v0) {
-        fatal_printf("ml : alloc overflow #%d\n", arg0->unk0);
-        while (1);
-    }
-    return phi_v1;
-}
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_800078A4(void *arg0, ?32 arg1, s32 arg2, s32 arg3) {
-    arg0->unk0 = arg1;
-    arg0->unkC = arg2;
-    arg0->unk4 = arg2;
-    arg0->unk8 = (s32) (arg2 + arg3);
-    bzero(arg2, arg3);
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_800078A4.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-u32 func_800078F0(u32 arg0) {
-    s32 temp_v0;
-
-    if (D_8004A504 == 3) {
-        return arg0;
-    }
-    temp_v0 = ((((arg0 >> 0x10) & 0xF800) | ((arg0 >> 0xD) & 0x7C0)) | ((arg0 >> 0xA) & 0x3E)) | ((arg0 >> 7) & 1);
-    return (temp_v0 << 0x10) | temp_v0;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_800078F0.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_80007944(s32 arg0, ?32 arg1, ?32 arg2) {
-    ?32 sp44;
-    ?32 sp40;
-    s32 sp3C;
-    ?32 sp1C;
-    ?32 sp18;
-
-    D_8004A518.unk0 = arg0;
-    sp18 = 5;
-    sp1C = 0x64;
-    D_8004A518.unk4 = arg1;
-    D_8004A518.unk8 = arg2;
-    sp3C = arg0;
-    sp40 = arg1;
-    sp44 = arg2;
-    func_80000980(&sp18);
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007944.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void *func_80007998(s32 arg0) {
-    D_8004A510 = (s32) (D_8004A510 | arg0);
-    if ((arg0 & 0x20) != 0) {
-        D_8004A504 = 3;
-    }
-    if ((arg0 & 0x10) != 0) {
-        D_8004A504 = 2;
-    }
-    D_8004A514 = 1;
-    return &D_8004A510;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007998.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_800079E4(s32 arg0) {
-    D_8004A508 = arg0;
-    D_8004A514 = 1;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_800079E4.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_800079FC(s32 arg0) {
-    D_8004A50C = arg0;
-    D_8004A514 = 1;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_800079FC.s")
-#endif
-
-extern s16 D_8004A524, D_8004A526, D_8004A528, D_8004A52A;
-extern u32 D_8004A514;
-void func_80007A14(s16 arg0, s16 arg1, s16 arg2, s16 arg3) {
-    D_8004A524 = arg0;
-    D_8004A526 = arg1;
-    D_8004A528 = arg2;
-    D_8004A52A = arg3;
-    D_8004A514 = 1;
-}
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void *func_80007A74(void *arg0) {
-    arg0->unk24 = (s32) D_8004A508;
-    arg0->unk28 = (s32) D_8004A50C;
-    arg0->unk2C = (s32) D_8004A510;
-    arg0->unk30 = (s16) D_8004A524;
-    arg0->unk32 = (s16) D_8004A526;
-    arg0->unk34 = (s16) D_8004A528;
-    arg0->unk36 = (s16) D_8004A52A;
-    D_8004A510 = 0;
-    *(void *)0x8004A514 = 0;
-    return &D_8004A510;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007A74.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_80007ADC(void *arg0) {
-    void *temp_a1;
-
-    temp_a1 = arg0;
-    if (D_8004A514 != 0) {
-        arg0->unk0 = 4;
-        arg0->unk4 = 0x32;
-        arg0->unk14 = 0;
-        arg0->unk20 = 0;
-        arg0 = temp_a1;
-        func_80007A74(temp_a1);
-        osSendMesg(&D_80048CB8, arg0, 0);
-    }
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007ADC.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_80007B38(s32 arg0, s32 arg1, ? arg2) {
-    ?32 sp1C;
-    ?32 sp18;
-
-    D_8004A510 = 0;
-    D_8004A504 = 2;
-    func_80007998(arg2);
-    func_800079E4(arg0);
-    func_800079FC(arg1);
-    sp18 = 4;
-    sp1C = 0x64;
-    func_80007A74(&sp18);
-    func_80000980(&sp18);
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007B38.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_80007BA4(void *arg0) {
-    s32 temp_a0;
-    s32 temp_a1;
-    s32 temp_a2;
-    void *temp_a3;
-
-    temp_a3 = arg0;
-    temp_a1 = temp_a3->unk4;
-    temp_a2 = temp_a3->unk8;
-    temp_a0 = arg0->unk0;
-    arg0 = temp_a3;
-    func_80007944(temp_a0, temp_a1, temp_a2, temp_a3);
-    D_8004A500 = (s32) arg0->unkC;
-    func_80007B38(arg0->unk10, arg0->unk14, arg0->unk18, arg0);
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007BA4.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-f32 func_80007BF4(void *arg0, f32 arg1, f32 arg2, f32 arg3, f32 arg4) {
-    f32 temp_f0;
-    f32 temp_f2;
-
-    temp_f0 = (arg1 + arg3) * 0.5f;
-    arg0->unk0 = (s16) (s32) ((arg3 - temp_f0) * 4.0f);
-    arg0->unkC = (u16)0x1FF;
-    temp_f2 = (arg2 + arg4) * 0.5f;
-    arg0->unk4 = (s16) arg0->unkC;
-    arg0->unk2 = (s16) (s32) ((arg4 - temp_f2) * 4.0f);
-    arg0->unk8 = (s16) (s32) (temp_f0 * 4.0f);
-    arg0->unkA = (s16) (s32) (temp_f2 * 4.0f);
-    return temp_f0;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007BF4.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-s32 func_80007C9C(void *arg0) {
-    s32 temp_v0;
-    s32 temp_v0_2;
-
-    temp_v0 = (s32) (D_8004A508 * 0x20000) >> 0x10;
-    arg0->unk8 = (s16) temp_v0;
-    arg0->unk0 = (s16) temp_v0;
-    arg0->unkC = (u16)0x1FF;
-    temp_v0_2 = (s32) (D_8004A50C * 0x20000) >> 0x10;
-    arg0->unkA = (s16) temp_v0_2;
-    arg0->unk2 = (s16) temp_v0_2;
-    arg0->unk4 = (s16) arg0->unkC;
-    return temp_v0_2;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007C9C.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-void func_80007CE8(s32 arg0) {
-    *(void *)0x8003DD30 = arg0;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007CE8.s")
-#endif
-
-#ifdef MIPS_TO_C
-//generated by mips_to_c commit e0e006e8858ba357d1dcb4dc64f038b7df278aa6
-f32 func_80007CF4(Gfx *dlist) {
-    gSPSegment(*dlist++, 0x00, 0x00000000);
-    func_80005378(dlist);
-    gDPSetDepthImage(*dlist++, D_8004A500);
-    func_80007C9C(&D_8004A530);
-    gSPDisplayList(*dlist++, D_8003DD78);
-
-    temp_v1 = sp34;
-    sp34 = sp34 + 8;
-    temp_v1->unk0 = (s32) (((((s32) (((f32) ((s32) D_8004A508 / 0x140) * 10.0f) * 4.0f) & 0xFFF) << 0xC) | 0xED000000) | ((s32) ((10.0f * (f32) ((s32) D_8004A50C / 0xF0)) * 4.0f) & 0xFFF));
-    temp_v1->unk4 = (s32) ((((s32) (((f32) D_8004A508 - (10.0f * (f32) ((s32) D_8004A508 / 0x140))) * 4.0f) & 0xFFF) << 0xC) | ((s32) (((f32) D_8004A50C - (10.0f * (f32) ((s32) D_8004A50C / 0xF0))) * 4.0f) & 0xFFF));
-    phi_return = 10.0f;
-    if (D_8003DD30 != NULL) {
-        phi_return = D_8003DD30(dlist, D_8004A50C, D_8003DD30, &D_8004A508);
-    }
-    *dlist = sp34;
-    return phi_return;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/ovl0/ovl0_2_5/func_80007CF4.s")
-#endif
-
-
