+++ undefined_syms.txt
@@ -332,6 +332,7 @@ D_800492DC = 0x800492DC;
 D_800492E8 = 0x800492E8;
 D_800492C0 = 0x800492C0;
 D_800492D8 = 0x800492D8;
+D_800492DA = 0x800492DA;
 D_8004A42C = 0x8004A42C;
 func_80006854 = 0x80006854;
 func_80000510 = 0x80000510;
